Module boa_engine::builtins

Boa’s ECMAScript built-in object implementations, e.g. Object, String, Math, Array, etc.

Modules

array

Boa’s implementation of ECMAScript’s global Array object.

array_buffer

Boa’s implementation of ECMAScript’s global ArrayBuffer object.

async_function

Boa’s implementation of ECMAScript’s global AsyncFunction object.

async_generator

Boa’s implementation of ECMAScript’s global AsyncGenerator object.

async_generator_function

Boa’s implementation of ECMAScript’s AsyncGeneratorFunction object.

bigint

Boa’s implementation of ECMAScript’s global BigInt object.

boolean

Boa’s implementation of ECMAScript’s global Boolean object.

dataview

Boa’s implementation of ECMAScript’s global DataView object.

date

Boa’s implementation of ECMAScript’s Date object.

error

Boa’s implementation of ECMAScript’s global Error object.

escape

Boa’s implementation of ECMAScript’s string escaping functions.

eval

Boa’s implementation of ECMAScript’s global eval function.

function

Boa’s implementation of ECMAScript’s global Function object and Native Functions.

generator

Boa’s implementation of ECMAScript’s global Generator object.

generator_function

Boa’s implementation of ECMAScript’s global GeneratorFunction object.

intl

Boa’s implementation of ECMAScript’s global Intl object.

iterable

Boa’s implementation of ECMAScript’s IteratorRecord and iterator prototype objects.

json

Boa’s implementation of ECMAScript’s global JSON object.

map

Boa’s implementation of ECMAScript’s global Map object.

math

Boa’s implementation of ECMAScript’s global Math object.

number

Boa’s implementation of ECMAScript’s global Number object.

object

Boa’s implementation of ECMAScript’s global Object object.

options 🔒

Utilities to parse, validate and get options in builtins.

promise

Boa’s implementation of ECMAScript’s global Promise object.

proxy

Boa’s implementation of ECMAScript’s global Proxy object.

reflect

Boa’s implementation of ECMAScript’s global Reflect object.

regexp

Boa’s implementation of ECMAScript’s global RegExp object.

set

Boa’s implementation of ECMAScript’s global Set object.

string

Boa’s implementation of ECMAScript’s global String object.

symbol

Boa’s implementation of ECMAScript’s global Symbol object.

typed_array

Boa’s implementation of ECMAScript’s global TypedArray objects.

uri

Boa’s implementation of ECMAScript’s URI Handling Functions.

weak

Boa’s implementation of ECMAScript’s WeakRef object.

weak_map

Boa’s implementation of ECMAScript’s WeakMap builtin object.

weak_set

Boa’s implementation of ECMAScript’s WeakSet builtin object.

Structs

BuiltInBuilder 🔒

Builder for creating built-in objects, like Array.

BuiltInCallable 🔒

BuiltInConstructorWithPrototype 🔒

Callable 🔒

Marker for a callable object.

Constructor 🔒

Marker for a constructor function.

ConstructorNoProto 🔒

Marker for a constructor function without a custom prototype for its instances.

OrdinaryFunction 🔒

Marker for an ordinary function.

OrdinaryObject 🔒

Marker for an ordinary object.

Enums

BuiltInObjectInitializer 🔒

Traits

ApplyToObject 🔒

Applies the pending builder data to the object.

BuiltInConstructor 🔒

A built-in object that is also a constructor.

BuiltInObject 🔒

A built-in object.

IntrinsicObject 🔒

A Well-Known Intrinsic Object.

IsConstructor 🔒

Indicates if the marker is a constructor.

Functions

global_binding 🔒

set_default_global_bindings 🔒

Abstract operation SetDefaultGlobalBindings ( realmRec ).