Struct boa_engine::builtins::array::Array

pub(crate) struct Array;

JavaScript Array built-in implementation.

Implementations

impl Array

pub(crate) fn array_create( length: u64, prototype: Option<JsObject>, context: &mut Context<'_> ) -> JsResult<JsObject>

Utility for constructing Array objects.

More information:

ECMAScript reference

pub(crate) fn create_array_from_list<I>( elements: I, context: &mut Context<'_> ) -> JsObjectwhere I: IntoIterator<Item = JsValue>,

Utility for constructing Array objects from an iterator of JsValues.

More information:

ECMAScript reference

fn is_concat_spreadable(o: &JsValue, context: &mut Context<'_>) -> JsResult<bool>

Utility function for concatenating array objects.

Returns a Boolean valued property that if true indicates that an object should be flattened to its array elements by Array.prototype.concat.

fn get_species( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

get Array [ @@species ]

The Array [ @@species ] accessor property returns the Array constructor.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn array_species_create( original_array: &JsObject, length: u64, context: &mut Context<'_> ) -> JsResult<JsObject>

Utility function used to specify the creation of a new Array object using a constructor function that is derived from original_array.

see: https://tc39.es/ecma262/#sec-arrayspeciescreate

pub(crate) fn from( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.from(arrayLike)

The Array.from() static method creates a new, shallow-copied Array instance from an array-like or iterable object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn is_array( _: &JsValue, args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Array.isArray( arg )

The isArray function takes one argument arg, and returns the Boolean value true if the argument is an object whose class internal property is “Array”; otherwise it returns false.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn of( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.of(...items)

The Array.of method creates a new Array instance from a variable number of arguments, regardless of the number or type of arguments.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn at( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

‘Array.prototype.at(index)’

The at() method takes an integer value and returns the item at that index, allowing for positive and negative integers. Negative integers count back from the last item in the array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn concat( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.concat(...arguments)

When the concat method is called with zero or more arguments, it returns an array containing the array elements of the object followed by the array elements of each argument in order.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn push( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.push( ...items )

The arguments are appended to the end of the array, in the order in which they appear. The new length of the array is returned as the result of the call.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn pop( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.pop()

The last element of the array is removed from the array and returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn for_each( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.forEach( callbackFn [ , thisArg ] )

This method executes the provided callback function for each element in the array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn join( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.join( separator )

The elements of the array are converted to Strings, and these Strings are then concatenated, separated by occurrences of the separator. If no separator is provided, a single comma is used as the separator.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_string( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.toString( separator )

The toString function is intentionally generic; it does not require that its this value be an Array object. Therefore it can be transferred to other kinds of objects for use as a method.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn reverse( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.reverse()

The elements of the array are rearranged so as to reverse their order. The object is returned as the result of the call.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn shift( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.shift()

The first element of the array is removed from the array and returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn unshift( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.unshift( ...items )

The arguments are prepended to the start of the array, such that their order within the array is the same as the order in which they appear in the argument list.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn every( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.every( callback, [ thisArg ] )

The every method executes the provided callback function once for each element present in the array until it finds the one where callback returns a falsy value. It returns false if it finds such element, otherwise it returns true.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn map( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.map( callback, [ thisArg ] )

For each element in the array the callback function is called, and a new array is constructed from the return values of these calls.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn index_of( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.indexOf( searchElement[, fromIndex ] )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn last_index_of( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.lastIndexOf( searchElement[, fromIndex ] )

lastIndexOf compares searchElement to the elements of the array in descending order using the Strict Equality Comparison algorithm, and if found at one or more indices, returns the largest such index; otherwise, -1 is returned.

The optional second argument fromIndex defaults to the array’s length minus one (i.e. the whole array is searched). If it is greater than or equal to the length of the array, the whole array will be searched. If it is negative, it is used as the offset from the end of the array to compute fromIndex. If the computed index is less than 0, -1 is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn find( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.find( callback, [thisArg] )

The find method executes the callback function once for each index of the array until the callback returns a truthy value. If so, find immediately returns the value of that element. Otherwise, find returns undefined.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn find_index( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.findIndex( predicate [ , thisArg ] )

This method executes the provided predicate function for each element of the array. If the predicate function returns true for an element, this method returns the index of the element. If all elements return false, the value -1 is returned.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn find_last( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.findLast( predicate, [thisArg] )

findLast calls predicate once for each element of the array, in descending order, until it finds one where predicate returns true. If such an element is found, findLast immediately returns that element value. Otherwise, findLast returns undefined.

More information:

ECMAScript proposal

pub(crate) fn find_last_index( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.findLastIndex( predicate [ , thisArg ] )

findLastIndex calls predicate once for each element of the array, in descending order, until it finds one where predicate returns true. If such an element is found, findLastIndex immediately returns the index of that element value. Otherwise, findLastIndex returns -1.

More information:

ECMAScript proposal

pub(crate) fn flat( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.flat( [depth] )

This method creates a new array with all sub-array elements concatenated into it recursively up to the specified depth.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn flat_map( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.flatMap( callback, [ thisArg ] )

This method returns a new array formed by applying a given callback function to each element of the array, and then flattening the result by one level. It is identical to a map() followed by a flat() of depth 1, but slightly more efficient than calling those two methods separately.

More information:

ECMAScript reference

MDN documentation

fn flatten_into_array( target: &JsObject, source: &JsObject, source_len: u64, start: u64, depth: u64, mapper_function: Option<&JsObject>, this_arg: &JsValue, context: &mut Context<'_> ) -> JsResult<u64>

Abstract method FlattenIntoArray.

More information:

ECMAScript reference

pub(crate) fn fill( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.fill( value[, start[, end]] )

The method fills (modifies) all the elements of an array from start index (default 0) to an end index (default array length) with a static value. It returns the modified array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn includes_value( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.includes( valueToFind [, fromIndex] )

Determines whether an array includes a certain value among its entries, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn slice( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.slice( [begin[, end]] )

The slice method takes two arguments, start and end, and returns an array containing the elements of the array from element start up to, but not including, element end (or through the end of the array if end is undefined). If start is negative, it is treated as length + start where length is the length of the array. If end is negative, it is treated as length + end where length is the length of the array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_locale_string( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.toLocaleString ( [ locales [ , options ] ] ).

Returns a string representing the elements of the array. The elements are converted to strings using their toLocaleString methods and these strings are separated by a locale-specific string (such as a comma “,”).

More information:

MDN documentation

pub(crate) fn splice( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.splice ( start, [deleteCount[, ...items]] )

Splices an array by following The deleteCount elements of the array starting at integer index start are replaced by the elements of items. An Array object containing the deleted elements (if any) is returned.

pub(crate) fn filter( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.filter( callback, [ thisArg ] )

For each element in the array the callback function is called, and a new array is constructed for every value whose callback returned a truthy value.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn some( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.some ( callbackfn [ , thisArg ] )

The some method tests whether at least one element in the array passes the test implemented by the provided callback function. It returns a Boolean value, true if the callback function returns a truthy value for at least one element in the array. Otherwise, false.

Caution: Calling this method on an empty array returns false for any condition!

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sort( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.sort ( comparefn )

The sort method sorts the elements of an array in place and returns the sorted array. The default sort order is ascending, built upon converting the elements into strings, then comparing their sequences of UTF-16 code units values.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn reduce( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.reduce( callbackFn [ , initialValue ] )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn reduce_right( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.reduceRight( callbackFn [ , initialValue ] )

The reduceRight method traverses right to left starting from the last defined value in the array, accumulating a value using a given callback function. It returns the accumulated value.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn copy_within( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.copyWithin ( target, start [ , end ] )

The copyWithin() method shallow copies part of an array to another location in the same array and returns it without modifying its length.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn values( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.values( )

The values method returns an iterable that iterates over the values in the array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn keys( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.keys( )

The keys method returns an iterable that iterates over the indexes in the array.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn entries( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Array.prototype.entries( )

The entries method returns an iterable that iterates over the key-value pairs in the array.

More information:

ECMAScript reference

MDN documentation

pub(super) fn get_relative_start( context: &mut Context<'_>, arg: Option<&JsValue>, len: u64 ) -> JsResult<u64>

Represents the algorithm to calculate relativeStart (or k) in array functions.

pub(super) fn get_relative_end( context: &mut Context<'_>, arg: Option<&JsValue>, len: u64 ) -> JsResult<u64>

Represents the algorithm to calculate relativeEnd (or final) in array functions.

pub(crate) fn unscopables_object() -> JsObject

Array.prototype [ @@unscopables ]

The initial value of the ‘unscopables’ data property is an ordinary object with the following boolean properties set to true: ‘at’, ‘copyWithin’, ‘entries’, ‘fill’, ‘find’, ‘findIndex’, ‘flat’, ‘flatMap’, ‘includes’, ‘keys’, ‘values’

More information:

ECMAScript reference

MDN documentation