Module boa_engine::builtins::array_buffer

Boa’s implementation of ECMAScript’s global ArrayBuffer and SharedArrayBuffer objects

More information:

ECMAScript reference

MDN documentation

Modules

shared 🔒

utils 🔒

Structs

ArrayBuffer

The internal representation of an ArrayBuffer object.

SharedArrayBuffer

The internal representation of a SharedArrayBuffer object.

SliceRange 🔒

Utility struct to return the result of the get_slice_range function.

Enums

BufferRef 🔒

BufferRefMut 🔒

Functions

create_byte_data_block 🔒

CreateByteDataBlock ( size ) abstract operation.

get_slice_range 🔒

Gets the slice copy range from the original length, the relative start and the end.