Function boa_engine::builtins::array_buffer::utils::memcpy

pub(crate) unsafe fn memcpy(
    src: SliceRef<'_>,
    dest: SliceRefMut<'_>,
    count: usize
)

Copies count bytes from the buffer src into the buffer dest, using the atomic ordering order if any of the buffers are atomic.

Safety

Both src and dest must have at least count bytes to read and write, respectively.

The region of memory referenced by src must not overlap with the region of memory referenced by dest. This is guaranteed if either of them are slices (you cannot borrow and mutably borrow a slice at the same time), but cannot be guaranteed for atomic slices.