Struct boa_engine::builtins::boolean::Boolean

pub(crate) struct Boolean;

Boolean implementation.

Implementations

impl Boolean

pub(crate) const LENGTH: usize

The amount of arguments this function object takes.

pub(crate) fn constructor(
new_target: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

[[Construct]] Create a new boolean object

[[Call]] Creates a new boolean primitive

fn this_boolean_value(value: &JsValue, context: &mut Context) -> JsResult<bool>

An Utility function used to get the internal [[BooleanData]].

More information:

ECMAScript reference

pub(crate) fn to_string(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

The toString() method returns a string representing the specified Boolean object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn value_of(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

The valueOf() method returns the primitive value of a Boolean object.

More information:

ECMAScript reference

MDN documentation