Struct boa_engine::builtins::dataview::DataView

pub struct DataView {
    pub(crate) viewed_array_buffer: JsObject,
    pub(crate) byte_length: u64,
    pub(crate) byte_offset: u64,
}

The internal representation of a DataView object.

Fields
viewed_array_buffer: JsObjectbyte_length: u64byte_offset: u64
Implementations

impl DataView

pub(crate) fn get_buffer( this: &JsValue, _args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.1 get DataView.prototype.buffer

The buffer accessor property represents the ArrayBuffer or SharedArrayBuffer referenced by the DataView at construction time.

More information:

ECMAScript reference

MDN

pub(crate) fn get_byte_length( this: &JsValue, _args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.1 get DataView.prototype.byteLength

The byteLength accessor property represents the length (in bytes) of the dataview.

More information:

ECMAScript reference

MDN

pub(crate) fn get_byte_offset( this: &JsValue, _args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.1 get DataView.prototype.byteOffset

The byteOffset accessor property represents the offset (in bytes) of this view from the start of its ArrayBuffer or SharedArrayBuffer.

More information:

ECMAScript reference

MDN

fn get_view_value( view: &JsValue, request_index: &JsValue, is_little_endian: &JsValue, t: TypedArrayKind, context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.1.1 GetViewValue ( view, requestIndex, isLittleEndian, type )

The abstract operation GetViewValue takes arguments view, requestIndex, isLittleEndian, and type. It is used by functions on DataView instances to retrieve values from the view’s buffer.

More information:

ECMAScript reference

pub(crate) fn get_big_int64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.5 DataView.prototype.getBigInt64 ( byteOffset [ , littleEndian ] )

The getBigInt64() method gets a signed 64-bit integer (long long) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_big_uint64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.6 DataView.prototype.getBigUint64 ( byteOffset [ , littleEndian ] )

The getBigUint64() method gets an unsigned 64-bit integer (unsigned long long) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_float32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.7 DataView.prototype.getBigUint64 ( byteOffset [ , littleEndian ] )

The getFloat32() method gets a signed 32-bit float (float) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_float64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.8 DataView.prototype.getFloat64 ( byteOffset [ , littleEndian ] )

The getFloat64() method gets a signed 64-bit float (double) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_int8( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.9 DataView.prototype.getInt8 ( byteOffset [ , littleEndian ] )

The getInt8() method gets a signed 8-bit integer (byte) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_int16( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.10 DataView.prototype.getInt16 ( byteOffset [ , littleEndian ] )

The getInt16() method gets a signed 16-bit integer (short) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_int32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.11 DataView.prototype.getInt32 ( byteOffset [ , littleEndian ] )

The getInt32() method gets a signed 32-bit integer (long) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_uint8( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.12 DataView.prototype.getUint8 ( byteOffset [ , littleEndian ] )

The getUint8() method gets an unsigned 8-bit integer (unsigned byte) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_uint16( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.13 DataView.prototype.getUint16 ( byteOffset [ , littleEndian ] )

The getUint16() method gets an unsigned 16-bit integer (unsigned short) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn get_uint32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.14 DataView.prototype.getUint32 ( byteOffset [ , littleEndian ] )

The getUint32() method gets an unsigned 32-bit integer (unsigned long) at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

fn set_view_value( view: &JsValue, request_index: &JsValue, is_little_endian: &JsValue, t: TypedArrayKind, value: &JsValue, context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.1.1 SetViewValue ( view, requestIndex, isLittleEndian, type )

The abstract operation SetViewValue takes arguments view, requestIndex, isLittleEndian, type, and value. It is used by functions on DataView instances to store values into the view’s buffer.

More information:

ECMAScript reference

pub(crate) fn set_big_int64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.15 DataView.prototype.setBigInt64 ( byteOffset, value [ , littleEndian ] )

The setBigInt64() method stores a signed 64-bit integer (long long) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_big_uint64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.16 DataView.prototype.setBigUint64 ( byteOffset, value [ , littleEndian ] )

The setBigUint64() method stores an unsigned 64-bit integer (unsigned long long) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_float32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.17 DataView.prototype.setFloat32 ( byteOffset, value [ , littleEndian ] )

The setFloat32() method stores a signed 32-bit float (float) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_float64( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.18 DataView.prototype.setFloat64 ( byteOffset, value [ , littleEndian ] )

The setFloat64() method stores a signed 64-bit float (double) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_int8( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.19 DataView.prototype.setInt8 ( byteOffset, value [ , littleEndian ] )

The setInt8() method stores a signed 8-bit integer (byte) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_int16( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.20 DataView.prototype.setInt16 ( byteOffset, value [ , littleEndian ] )

The setInt16() method stores a signed 16-bit integer (short) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_int32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.21 DataView.prototype.setInt32 ( byteOffset, value [ , littleEndian ] )

The setInt32() method stores a signed 32-bit integer (long) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_uint8( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.22 DataView.prototype.setUint8 ( byteOffset, value [ , littleEndian ] )

The setUint8() method stores an unsigned 8-bit integer (byte) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_uint16( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.23 DataView.prototype.setUint16 ( byteOffset, value [ , littleEndian ] )

The setUint16() method stores an unsigned 16-bit integer (unsigned short) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN

pub(crate) fn set_uint32( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

25.3.4.24 DataView.prototype.setUint32 ( byteOffset, value [ , littleEndian ] )

The setUint32() method stores an unsigned 32-bit integer (unsigned long) value at the specified byte offset from the start of the DataView.

More information:

ECMAScript reference

MDN