Macro boa_engine::builtins::date::get_mut_date

macro_rules! get_mut_date {
    (let $var:ident = $val:expr) => { ... };
}

Gets a mutable reference to the inner Date object of val and stores it on var, or returns a TypeError if val is not a Date object.