Struct boa_engine::builtins::date::Date

pub struct Date(Option<i64>);

The internal representation of a Date object.

Tuple Fields
0: Option<i64>
Implementations

impl Date

pub(crate) const fn new(dt: Option<i64>) -> Self

Creates a new Date.

pub(crate) fn utc_now(hooks: &dyn HostHooks) -> Self

Creates a new Date from the current UTC time of the host.

fn as_value(&self) -> JsValue

Converts the Date into a JsValue, mapping None to NaN and Some(datetime) to JsValue::from(datetime.timestamp_millis()).

impl Date

fn construct_date( values: &[JsValue], context: &mut Context<'_> ) -> JsResult<Option<NaiveDateTime>>

Gets the timestamp from a list of component values.

pub(crate) fn now( _: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.now()

The static Date.now() method returns the number of milliseconds elapsed since January 1, 1970 00:00:00 UTC.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn parse( _: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.parse()

The Date.parse() method parses a string representation of a date, and returns the number of milliseconds since January 1, 1970, 00:00:00 UTC or NaN if the string is unrecognized or, in some cases, contains illegal date values.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn utc( _: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.UTC()

The Date.UTC() method accepts parameters similar to the Date constructor, but treats them as UTC.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_date<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getDate ( ) and Date.prototype.getUTCDate ( ).

The getDate() method returns the day of the month for the specified date.

pub(crate) fn get_day<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getDay ( ) and Date.prototype.getUTCDay ( ).

The getDay() method returns the day of the week for the specified date, where 0 represents Sunday.

pub(crate) fn get_year( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getYear().

The getYear() method returns the year in the specified date according to local time. Because getYear() does not return full years (“year 2000 problem”), it is no longer used and has been replaced by the getFullYear() method.

More information:

MDN documentation

pub(crate) fn get_full_year<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getFullYear ( ) and Date.prototype.getUTCFullYear ( ).

The getFullYear() method returns the year of the specified date.

pub(crate) fn get_hours<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getHours ( ) and Date.prototype.getUTCHours ( ).

The getHours() method returns the hour for the specified date.

pub(crate) fn get_milliseconds<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getMilliseconds ( ) and Date.prototype.getUTCMilliseconds ( ).

The getMilliseconds() method returns the milliseconds in the specified date.

pub(crate) fn get_minutes<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getMinutes ( ) and Date.prototype.getUTCMinutes ( ).

The getMinutes() method returns the minutes in the specified date.

pub(crate) fn get_month<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getMonth ( ) and Date.prototype.getUTCMonth ( ).

The getMonth() method returns the month in the specified date, as a zero-based value (where zero indicates the first month of the year).

pub(crate) fn get_seconds<const LOCAL: bool>( this: &JsValue, _args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getSeconds ( ) and Date.prototype.getUTCSeconds ( ).

The getSeconds() method returns the seconds in the specified date.

pub(crate) fn get_time( this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getTime().

The getTime() method returns the number of milliseconds since the Unix Epoch.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_timezone_offset( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.getTimeZoneOffset().

The getTimezoneOffset() method returns the time zone difference, in minutes, from current locale (host system settings) to UTC.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn set_date<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setDate ( date ) and Date.prototype.setUTCDate ( date ).

The setDate() method sets the day of the Date object relative to the beginning of the currently set month.

pub(crate) fn set_full_year<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setFullYear ( year [ , month [ , date ] ] ) and Date.prototype.setUTCFullYear ( year [ , month [ , date ] ] ).

The setFullYear() method sets the full year for a specified date and returns the new timestamp.

pub(crate) fn set_hours<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setHours ( hour [ , min [ , sec [ , ms ] ] ] ) and Date.prototype.setUTCHours ( hour [ , min [ , sec [ , ms ] ] ] ).

The setHours() method sets the hours for a specified date, and returns the number of milliseconds since January 1, 1970 00:00:00 UTC until the time represented by the updated Date instance.

pub(crate) fn set_milliseconds<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

[Date.prototype.setMilliseconds ( ms )local and Date.prototype.setUTCMilliseconds ( ms ).

The setMilliseconds() method sets the milliseconds for a specified date according to local time.

pub(crate) fn set_minutes<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setMinutes ( min [ , sec [ , ms ] ] ) and Date.prototype.setUTCMinutes ( min [ , sec [ , ms ] ] ).

The setMinutes() method sets the minutes for a specified date.

pub(crate) fn set_month<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setMonth ( month [ , date ] ) and Date.prototype.setUTCMonth ( month [ , date ] ).

The setMonth() method sets the month for a specified date according to the currently set year.

pub(crate) fn set_seconds<const LOCAL: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

[Date.prototype.setSeconds ( sec [ , ms ] )local and Date.prototype.setUTCSeconds ( sec [ , ms ] ).

The setSeconds() method sets the seconds for a specified date.

pub(crate) fn set_year( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setYear().

The setYear() method sets the year for a specified date according to local time.

Note

The Self::set_full_year method is preferred for nearly all purposes, because it avoids the “year 2000 problem.”

More information:

MDN documentation

pub(crate) fn set_time( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.setTime().

The setTime() method sets the Date object to the time represented by a number of milliseconds since January 1, 1970, 00:00:00 UTC.

More information:

MDN documentation

pub(crate) fn to_date_string( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toDateString().

The toDateString() method returns the date portion of a Date object in English.

More information:

MDN documentation

pub(crate) fn to_iso_string( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toISOString().

The toISOString() method returns a string in simplified extended ISO format (ISO 8601).

More information:

MDN documentation

pub(crate) fn to_json( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toJSON().

The toJSON() method returns a string representation of the Date object.

More information:

MDN documentation

pub(crate) fn to_locale_date_string( _this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toLocaleDateString().

The toLocaleDateString() method returns the date portion of the given Date instance according to language-specific conventions.

More information:

MDN documentation

pub(crate) fn to_locale_string( _this: &JsValue, _: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toLocaleString().

The toLocaleString() method returns a string representing the specified Date object.

More information:

MDN documentation

pub(crate) fn to_locale_time_string( _this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toLocaleTimeString().

The toLocaleTimeString() method returns the time portion of a Date object in human readable form in American English.

More information:

MDN documentation

pub(crate) fn to_string( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toString().

The toString() method returns a string representing the specified Date object.

More information:

MDN documentation

pub(crate) fn to_time_string( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toTimeString().

The toTimeString() method returns the time portion of a Date object in human readable form in American English.

More information:

MDN documentation

pub(crate) fn to_utc_string( this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.toUTCString().

The toUTCString() method returns a string representing the specified Date object.

More information:

MDN documentation

pub(crate) fn value_of( this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype.valueOf().

The valueOf() method returns the primitive value of a Date object.

More information:

MDN documentation

pub(crate) fn to_primitive( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Date.prototype [ @@toPrimitive ] ( hint ).

The [@@toPrimitive]() method converts a Date object to a primitive value.

More information:

MDN documentation