Module boa_engine::builtins::date::utils

Structs

DateParser ğŸ”’

Parses a date string according to the Date Time String Format.

Constants

HOURS_PER_DAY ğŸ”’

MINUTES_PER_HOUR ğŸ”’

MS_PER_DAY ğŸ”’

MS_PER_HOUR ğŸ”’

MS_PER_MINUTE ğŸ”’

MS_PER_SECOND ğŸ”’

SECONDS_PER_MINUTE ğŸ”’

Functions

date_from_time ğŸ”’

Abstract operation DateFromTime ( t )

date_string ğŸ”’

Abstract operation DateString ( tv )

day ğŸ”’

Abstract operation Day ( t )

day_from_year ğŸ”’

Abstract operation DayFromYear ( y )

day_within_year ğŸ”’

Abstract operation DayWithinYear ( t )

days_in_year ğŸ”’

Abstract operation DaysInYear ( y )

hour_from_time ğŸ”’

Abstract operation HourFromTime ( t )

in_leap_year ğŸ”’

Abstract operation InLeapYear ( t )

local_time ğŸ”’

Abstract operation LocalTime ( t )

local_timezone_offset_seconds ğŸ”’

make_date ğŸ”’

Abstract operation MakeDate ( day, time )

make_day ğŸ”’

Abstract operation MakeDay ( year, month, date )

make_full_year ğŸ”’

Abstract operation MakeFullYear ( year )

make_time ğŸ”’

Abstract operation MakeTime ( hour, min, sec, ms )

min_from_time ğŸ”’

Abstract operation MinFromTime ( t )

month_from_time ğŸ”’

Abstract operation MonthFromTime ( t )

ms_from_time ğŸ”’

Abstract operation msFromTime ( t )

pad_five ğŸ”’

pad_four ğŸ”’

pad_six ğŸ”’

pad_three ğŸ”’

pad_two ğŸ”’

parse_date ğŸ”’

Parse a date string according to the steps specified in Date.parse.

sec_from_time ğŸ”’

Abstract operation SecFromTime ( t )

time_clip ğŸ”’

Abstract operation TimeClip ( time )

time_from_year ğŸ”’

Abstract operation TimeFromYear ( y )

time_string ğŸ”’

Abstract operation TimeString ( tv )

time_within_day ğŸ”’

Abstract operation TimeWithinDay ( t )

time_zone_string ğŸ”’

Abstract operation TimeZoneString ( tv )

to_date_string_t ğŸ”’

Abstract operation ToDateString ( tv )

utc_t ğŸ”’

Abstract operation UTC ( t )

week_day ğŸ”’

Abstract operation WeekDay ( t )

year_from_time ğŸ”’

Abstract operation YearFromTime ( t )