Module boa_engine::builtins::function

Boa’s implementation of ECMAScript’s global Function object and Native Functions.

Objects wrap Functions and expose them via call/construct slots.

The Function object is used for matching text with a pattern.

More information:

ECMAScript reference

MDN documentation

Modules

arguments 🔒

Structs

BoundFunction

Binds a Function Object when bind is called.

BuiltInFunctionObject

The internal representation of a Function object.

Enums

ClassFieldDefinition

Record containing the field definition of classes.

ConstructorKind

Represents the [[ConstructorKind]] internal slot of function objects.

Function

Boa representation of a Function Object.

ThisMode

Represents the [[ThisMode]] internal slot of function objects.

Functions

make_builtin_fn 🔒

Creates a new member function of a Object or prototype.

set_function_name 🔒

Abstract operation SetFunctionName