Enum boa_engine::builtins::function::Function

pub enum Function {
    Native {
        function: NativeFunctionSignature,
        constructor: Option<ConstructorKind>,
    },
    Closure {
        function: Box<dyn ClosureFunctionSignature>,
        constructor: Option<ConstructorKind>,
        captures: Captures,
    },
    Ordinary {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        constructor_kind: ConstructorKind,
        home_object: Option<JsObject>,
        fields: Vec<ClassFieldDefinition>,
        private_methods: Vec<(Sym, PrivateElement)>,
    },
    Async {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
        promise_capability: PromiseCapability,
    },
    Generator {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
    },
    AsyncGenerator {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
    },
}

Boa representation of a Function Object.

FunctionBody is specific to this interpreter, it will either be Rust code or JavaScript code (AST Node).

https://tc39.es/ecma262/#sec-ecmascript-function-objects

Variants

Native

Fields

function: NativeFunctionSignature

constructor: Option<ConstructorKind>

Closure

Fields

function: Box<dyn ClosureFunctionSignature>

constructor: Option<ConstructorKind>

captures: Captures

Ordinary

Fields

code: Gc<CodeBlock>

environments: DeclarativeEnvironmentStack

constructor_kind: ConstructorKind

The [[ConstructorKind]] internal slot.

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

fields: Vec<ClassFieldDefinition>

The [[Fields]] internal slot.

private_methods: Vec<(Sym, PrivateElement)>

The [[PrivateMethods]] internal slot.

Async

Fields

code: Gc<CodeBlock>

environments: DeclarativeEnvironmentStack

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

promise_capability: PromiseCapability

Generator

Fields

code: Gc<CodeBlock>

environments: DeclarativeEnvironmentStack

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

AsyncGenerator

Fields

code: Gc<CodeBlock>

environments: DeclarativeEnvironmentStack

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

Implementations

impl Function

pub fn is_constructor(&self) -> bool

Returns true if the function object is a constructor.

pub(crate) fn is_derived_constructor(&self) -> bool

Returns true if the function object is a derived constructor.

pub(crate) fn get_home_object(&self) -> Option<&JsObject>

Returns a reference to the function [[HomeObject]] slot if present.

pub(crate) fn set_home_object(&mut self, object: JsObject)

Sets the [[HomeObject]] slot if present.

pub(crate) fn get_fields(&self) -> &[ClassFieldDefinition]

Returns the values of the [[Fields]] internal slot.

pub(crate) fn push_field(&mut self, key: PropertyKey, value: JsFunction)

Pushes a value to the [[Fields]] internal slot if present.

pub(crate) fn push_field_private(&mut self, key: Sym, value: JsFunction)

Pushes a private value to the [[Fields]] internal slot if present.

pub(crate) fn get_private_methods(&self) -> &[(Sym, PrivateElement)]

Returns the values of the [[PrivateMethods]] internal slot.

pub(crate) fn push_private_method(&mut self, name: Sym, method: PrivateElement)

Pushes a private method to the [[PrivateMethods]] internal slot if present.

pub(crate) fn get_promise_capability(&self) -> Option<&PromiseCapability>

Returns the promise capability if the function is an async function.