Enum boa_engine::builtins::function::FunctionKind

pub(crate) enum FunctionKind {
    Native {
        function: NativeFunction,
        constructor: Option<ConstructorKind>,
    },
    Ordinary {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        constructor_kind: ConstructorKind,
        home_object: Option<JsObject>,
        fields: ThinVec<ClassFieldDefinition>,
        private_methods: ThinVec<(PrivateName, PrivateElement)>,
        class_object: Option<JsObject>,
    },
    Async {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
        class_object: Option<JsObject>,
    },
    Generator {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
        class_object: Option<JsObject>,
    },
    AsyncGenerator {
        code: Gc<CodeBlock>,
        environments: DeclarativeEnvironmentStack,
        home_object: Option<JsObject>,
        class_object: Option<JsObject>,
    },
}

Variants

Native

Fields

function: NativeFunction

The rust function.

constructor: Option<ConstructorKind>

The kind of the function constructor if it is a constructor.

A rust function.

Ordinary

Fields

code: Gc<CodeBlock>

The code block containing the compiled function.

environments: DeclarativeEnvironmentStack

The [[Environment]] internal slot.

constructor_kind: ConstructorKind

The [[ConstructorKind]] internal slot.

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

fields: ThinVec<ClassFieldDefinition>

The [[Fields]] internal slot.

private_methods: ThinVec<(PrivateName, PrivateElement)>

The [[PrivateMethods]] internal slot.

class_object: Option<JsObject>

The class object that this function is associated with.

A bytecode function.

Async

Fields

code: Gc<CodeBlock>

The code block containing the compiled function.

environments: DeclarativeEnvironmentStack

The [[Environment]] internal slot.

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

class_object: Option<JsObject>

The class object that this function is associated with.

A bytecode async function.

Generator

Fields

code: Gc<CodeBlock>

The code block containing the compiled function.

environments: DeclarativeEnvironmentStack

The [[Environment]] internal slot.

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

class_object: Option<JsObject>

The class object that this function is associated with.

A bytecode generator function.

AsyncGenerator

Fields

code: Gc<CodeBlock>

The code block containing the compiled function.

environments: DeclarativeEnvironmentStack

The [[Environment]] internal slot.

home_object: Option<JsObject>

The [[HomeObject]] internal slot.

class_object: Option<JsObject>

The class object that this function is associated with.

A bytecode async generator function.