Enum boa_engine::builtins::function::ThisMode

pub enum ThisMode {
    Lexical,
    Strict,
    Global,
}

Represents the [[ThisMode]] internal slot of function objects.

More information:

ECMAScript specification

Variants

Lexical

The this value refers to the this value of a lexically enclosing function.

Strict

The this value is used exactly as provided by an invocation of the function.

Global

The this value of undefined or null is interpreted as a reference to the global object, and any other this value is first passed to ToObject.

Implementations

impl ThisMode

pub const fn is_lexical(&self) -> bool

Returns true if the this mode is Lexical.

pub const fn is_strict(&self) -> bool

Returns true if the this mode is Strict.

pub const fn is_global(&self) -> bool

Returns true if the this mode is Global.