Struct boa_engine::builtins::function::Function

pub struct Function {
    kind: FunctionKind,
    realm: Realm,
}

Boa representation of a Function Object.

FunctionBody is specific to this interpreter, it will either be Rust code or JavaScript code (AST Node).

https://tc39.es/ecma262/#sec-ecmascript-function-objects

Fields
kind: FunctionKindrealm: Realm
Implementations

impl Function

pub fn is_constructor(&self) -> bool

Returns true if the function object is a constructor.

pub fn codeblock(&self) -> Option<&CodeBlock>

Returns the codeblock of the function, or None if the function is a NativeFunction.

pub(crate) fn new(kind: FunctionKind, realm: Realm) -> Self

Creates a new Function.

pub(crate) const fn is_derived_constructor(&self) -> bool

Returns true if the function object is a derived constructor.

pub(crate) fn class_field_initializer_name(&self) -> Option<Sym>

Returns the [[ClassFieldInitializerName]] internal slot of the function.

pub(crate) const fn get_home_object(&self) -> Option<&JsObject>

Returns a reference to the function [[HomeObject]] slot if present.

pub(crate) fn set_home_object(&mut self, object: JsObject)

Sets the [[HomeObject]] slot if present.

pub(crate) fn get_fields(&self) -> &[ClassFieldDefinition]

Returns the values of the [[Fields]] internal slot.

pub(crate) fn push_field(&mut self, key: PropertyKey, value: JsFunction)

Pushes a value to the [[Fields]] internal slot if present.

pub(crate) fn push_field_private(&mut self, key: PrivateName, value: JsFunction)

Pushes a private value to the [[Fields]] internal slot if present.

pub(crate) fn get_private_methods(&self) -> &[(PrivateName, PrivateElement)]

Returns the values of the [[PrivateMethods]] internal slot.

pub(crate) fn push_private_method( &mut self, name: PrivateName, method: PrivateElement )

Pushes a private method to the [[PrivateMethods]] internal slot if present.

pub(crate) fn set_class_object(&mut self, object: JsObject)

Sets the class object.

pub const fn realm(&self) -> &Realm

Gets the Realm from where this function originates.

pub(crate) const fn kind(&self) -> &FunctionKind

Gets a reference to the FunctionKind of the Function.

pub(crate) fn kind_mut(&mut self) -> &mut FunctionKind

Gets a mutable reference to the FunctionKind of the Function.