Module boa_engine::builtins::generator

Boa’s implementation of ECMAScript’s global Generator object.

A Generator is an instance of a generator function and conforms to both the Iterator and Iterable interfaces.

More information:

ECMAScript reference

MDN documentation

Structs

Generator

The internal representation of a Generator object.

GeneratorContext 🔒

Holds all information that a generator needs to continue it’s execution.

Enums

GeneratorState 🔒

Indicates the state of a generator.