Struct boa_engine::builtins::generator::Generator

pub struct Generator {
    pub(crate) state: GeneratorState,
}

The internal representation of a Generator object.

Fields
state: GeneratorState

The [[GeneratorState]] internal slot.

Implementations

impl Generator

const NAME: &str = "Generator"

pub(crate) fn next( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Generator.prototype.next ( value )

The next() method returns an object with two properties done and value. You can also provide a parameter to the next method to send a value to the generator.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn return( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Generator.prototype.return ( value )

The return() method returns the given value and finishes the generator.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn throw( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Generator.prototype.throw ( exception )

The throw() method resumes the execution of a generator by throwing an error into it and returns an object with two properties done and value.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn generator_resume( gen: &JsValue, value: JsValue, context: &mut Context<'_> ) -> JsResult<JsValue>

27.5.3.3 GeneratorResume ( generator, value, generatorBrand )

More information:

ECMAScript reference

pub(crate) fn generator_resume_abrupt( gen: &JsValue, abrupt_completion: JsResult<JsValue>, context: &mut Context<'_> ) -> JsResult<JsValue>

27.5.3.4 GeneratorResumeAbrupt ( generator, abruptCompletion, generatorBrand )

More information:

ECMAScript reference