Function boa_engine::builtins::intl::get_number_option

pub(crate) fn get_number_option(
    options: &JsObject, 
    property: &str, 
    minimum: f64, 
    maximum: f64, 
    fallback: Option<f64>, 
    context: &mut Context
) -> JsResult<Option<f64>>

Abstract operation GetNumberOption ( options, property, minimum, maximum, fallback )

Extracts the value of the property named property from the provided options object, converts it to a Number value, checks whether it is in the allowed range, and fills in a fallback value if necessary.

More information:

ECMAScript reference