Function boa_engine::builtins::intl::unicode_extension_components

fn unicode_extension_components(extension: &str) -> UniExtRecord

Abstract operation UnicodeExtensionComponents ( extension )

Returns the attributes and keywords from extension, which must be a String value whose contents are a Unicode locale extension sequence.

More information:

ECMAScript reference