Module boa_engine::builtins::intl::locale::utils

Functions

best_available_locale 🔒

Abstract operation BestAvailableLocale ( availableLocales, locale )

best_fit_matcher 🔒

Abstract operation BestFitMatcher ( availableLocales, requestedLocales )

best_fit_supported_locales 🔒

Abstract operation BestFitSupportedLocales ( availableLocales, requestedLocales )

best_locale_for_provider 🔒

Returns the locale resolved by the provider after using the ICU4X fallback algorithm with candidate (if the provider supports this), or None if the locale is not supported.

canonicalize_locale_list 🔒

Abstract operation CanonicalizeLocaleList ( locales )

default_locale 🔒

Abstract operation DefaultLocale ( )

lookup_matcher 🔒

Abstract operation LookupMatcher ( availableLocales, requestedLocales )

lookup_supported_locales 🔒

Abstract operation LookupSupportedLocales ( availableLocales, requestedLocales )

resolve_locale 🔒

Abstract operation ResolveLocale ( availableLocales, requestedLocales, options, relevantExtensionKeys, localeData )

supported_locales 🔒

Abstract operation SupportedLocales ( availableLocales, requestedLocales, options )

validate_extension 🔒

Validates that the unicode extension key with value is a valid extension value for the language.