Function boa_engine::builtins::intl::locale::utils::best_available_locale

pub(crate) fn best_available_locale<M: KeyedDataMarker>(
    candidate: LanguageIdentifier,
    provider: &(impl DataProvider<M> + ?Sized)
) -> Option<LanguageIdentifier>

Abstract operation BestAvailableLocale ( availableLocales, locale )

Compares the provided argument locale, which must be a String value with a structurally valid and canonicalized Unicode BCP 47 locale identifier, against the locales in availableLocales and returns either the longest non-empty prefix of locale that is an element of availableLocales, or undefined if there is no such element.

We only work with language identifiers, which have the same semantics but are a bit easier to manipulate.

More information:

ECMAScript reference