Function boa_engine::builtins::intl::locale::utils::canonicalize_locale_list

pub(crate) fn canonicalize_locale_list(
    locales: &JsValue,
    context: &mut Context<'_>
) -> JsResult<Vec<Locale>>

Abstract operation CanonicalizeLocaleList ( locales )

Converts an array of JsValues containing structurally valid Unicode BCP 47 locale identifiers into their canonical form.

For efficiency, this returns [Locale]s instead of Strings, since Locale allows us to modify individual parts of the locale without scanning the whole string again.

More information:

ECMAScript reference