Module boa_engine::builtins::intl::options

Structs

IntlOptions 🔒

IntlOptions aggregates the locale_matcher selector and any other object property needed for Intl object constructors.

ParseLocaleMatcherError 🔒

Enums

LocaleMatcher 🔒

Traits

OptionType 🔒

A type used as an option parameter inside the Intl spec.

OptionTypeParsable 🔒

Functions

coerce_options_to_object 🔒

Abstract operation CoerceOptionsToObject ( options )

default_number_option 🔒

Abstract operation DefaultNumberOption ( value, minimum, maximum, fallback )

get_number_option 🔒

Abstract operation GetNumberOption ( options, property, minimum, maximum, fallback )

get_option 🔒

Abstract operation GetOption ( options, property, type, values, fallback )

get_options_object 🔒

Abstract operation GetOptionsObject ( options )