Function boa_engine::builtins::intl::options::default_number_option

pub(super) fn default_number_option(
    value: &JsValue,
    minimum: f64,
    maximum: f64,
    fallback: Option<f64>,
    context: &mut Context<'_>
) -> JsResult<Option<f64>>

Abstract operation DefaultNumberOption ( value, minimum, maximum, fallback )

Converts value to a Number value, checks whether it is in the allowed range, and fills in a fallback value if necessary.