Function boa_engine::builtins::intl::options::get_options_object

pub(super) fn get_options_object(options: &JsValue) -> JsResult<JsObject>

Abstract operation GetOptionsObject ( options )

Returns a JsObject suitable for use with get_option, either options itself or a default empty JsObject. It throws a TypeError if options is not undefined and not a JsObject.