Trait boa_engine::builtins::intl::options::OptionType

pub(super) trait OptionType: Sized {
    fn from_value(value: JsValue, context: &mut Context<'_>) -> JsResult<Self>;
}

A type used as an option parameter inside the Intl spec.

Required Methods

fn from_value(value: JsValue, context: &mut Context<'_>) -> JsResult<Self>

Parses a JsValue into an instance of Self.

Roughly equivalent to the algorithm steps of 9.12.13.3-7, but allows for parsing steps instead of returning a pure string, number or boolean.