Module boa_engine::builtins::number

This module implements the global Number object.

The Number JavaScript object is a wrapper object allowing you to work with numerical values. A Number object is created using the Number() constructor. A primitive type object number is created using the Number() function.

The JavaScript Number type is double-precision 64-bit binary format IEEE 754 value. In more recent implementations, JavaScript also supports integers with arbitrary precision using the BigInt type.

More information:

ECMAScript reference

MDN documentation

Modules

conversions

Structs

Number

Number implementation.

Constants

BUF_SIZE

Functions

f64_to_exponential

Helper function that formats a float as a ES6-style exponential number string.

f64_to_exponential_with_precision

Helper function that formats a float as a ES6-style exponential number string with a given precision.