Trait boa_engine::builtins::options::OptionType

pub(crate) trait OptionType: Sized {
    // Required method
    fn from_value(value: JsValue, context: &mut Context) -> JsResult<Self>;
}

A type used as an option parameter for get_option.

Required Methods

fn from_value(value: JsValue, context: &mut Context) -> JsResult<Self>

Parses a JsValue into an instance of Self.

Roughly equivalent to the algorithm steps of 9.12.13.3-7, but allows for parsing steps instead of returning a pure string, number or boolean.

Object Safety

This trait is not object safe.