Module boa_engine::builtins::proxy

Boa’s implementation of ECMAScript’s global Proxy object.

The Proxy object enables you to create a proxy for another object, which can intercept and redefine fundamental operations for that object.

More information:

ECMAScript reference

MDN documentation

Structs

Proxy

Javascript Proxy object.

Functions

proxy_exotic_call 🔒

10.5.12 [[Call]] ( thisArgument, argumentsList )

proxy_exotic_construct 🔒

[[Construct]] ( argumentsList, newTarget )

proxy_exotic_define_own_property 🔒

10.5.6 [[DefineOwnProperty]] ( P, Desc )

proxy_exotic_delete 🔒

10.5.10 [[Delete]] ( P )

proxy_exotic_get 🔒

10.5.8 [[Get]] ( P, Receiver )

proxy_exotic_get_own_property 🔒

10.5.5 [[GetOwnProperty]] ( P )

proxy_exotic_get_prototype_of 🔒

10.5.1 [[GetPrototypeOf]] ( )

proxy_exotic_has_property 🔒

10.5.7 [[HasProperty]] ( P )

proxy_exotic_is_extensible 🔒

10.5.3 [[IsExtensible]] ( )

proxy_exotic_own_property_keys 🔒

10.5.11 [[OwnPropertyKeys]] ( )

proxy_exotic_prevent_extensions 🔒

10.5.4 [[PreventExtensions]] ( )

proxy_exotic_set 🔒

10.5.9 [[Set]] ( P, V, Receiver )

proxy_exotic_set_prototype_of 🔒

10.5.2 [[SetPrototypeOf]] ( V )