Struct boa_engine::builtins::set::ordered_set::OrderedSet

pub struct OrderedSet<S = RandomState> {
    inner: IndexSet<MapKey, S>,
    lock: u32,
    empty_count: usize,
}

A type wrapping indexmap::IndexSet

Fields
inner: IndexSet<MapKey, S>lock: u32empty_count: usize
Implementations

impl OrderedSet

pub fn new() -> Self

Creates a new empty OrderedSet.

pub fn with_capacity(capacity: usize) -> Self

Creates a new empty OrderedSet with the specified capacity.

pub fn full_len(&self) -> usize

Return the number of elements in the set, including empty elements.

Computes in O(1) time.

pub fn len(&self) -> usize

Return the number of elements in the set.

Computes in O(1) time.

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

Computes in O(1) time.

pub fn add(&mut self, value: JsValue) -> bool

Insert a value pair in the set.

If an equivalent value already exists in the set: ???

If no equivalent value existed in the set: the new value is inserted, last in order, and false

Computes in O(1) time (amortized average).

pub fn delete(&mut self, value: &JsValue) -> bool

Delete the value from the set and return true if successful

Return false if value is not in set.

Computes in O(n) time (average).

pub fn clear(&mut self)

Removes all elements in the set, while preserving its capacity.

pub fn contains(&self, value: &JsValue) -> bool

Checks if a given value is present in the set

Return true if value is present in set, false otherwise.

Computes in O(n) time (average).

pub fn get_index(&self, index: usize) -> Option<&JsValue>

Get a key-value pair by index Valid indices are 0 <= index < self.len() Computes in O(1) time.

pub fn iter(&self) -> impl Iterator<Item = &JsValue>

Return an iterator over the values of the set, in their order

pub(crate) fn lock(&mut self, set: JsObject) -> SetLock

Increases the lock counter and returns a lock object that will decrement the counter when dropped.

This allows objects to be removed from the set during iteration without affecting the indexes until the iteration has completed.

fn unlock(&mut self)

Decreases the lock counter and, if 0, removes all empty entries.