Struct boa_engine::builtins::set::Set

pub(crate) struct Set;

Implementations

impl Set

pub(crate) fn set_create( prototype: Option<JsObject>, context: &mut Context<'_> ) -> JsObject

Utility for constructing Set objects.

pub(crate) fn create_set_from_list<I>( elements: I, context: &mut Context<'_> ) -> JsObjectwhere I: IntoIterator<Item = JsValue>,

Utility for constructing Set objects from an iterator of JsValue’s.

fn get_species( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

get Set [ @@species ]

The Set[Symbol.species] accessor property returns the Set constructor.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn add( this: &JsValue, args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.add( value )

This method adds an entry with value into the set. Returns the set object

More information:

ECMAScript reference

MDN documentation

pub(crate) fn clear( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.clear( )

This method removes all entries from the set.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn delete( this: &JsValue, args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.delete( value )

This method removes the entry for the given value if it exists. Returns true if there was an element, false otherwise.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn entries( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.entries( )

This method returns an iterator over the entries of the set

More information:

ECMAScript reference

MDN documentation

pub(crate) fn for_each( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.forEach( callbackFn [ , thisArg ] )

This method executes the provided callback function for each value in the set

More information:

ECMAScript reference

MDN documentation

pub(crate) fn has( this: &JsValue, args: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

Map.prototype.has( key )

This method checks if the map contains an entry with the given key.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn values( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Set.prototype.values( )

This method returns an iterator over the values of the set

More information:

ECMAScript reference

MDN documentation

fn size_getter( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

pub(crate) fn get_size(set: &JsValue) -> JsResult<usize>

Helper function to get the size of the Set object.

pub(crate) fn get_size_full(set: &JsValue) -> JsResult<usize>

Helper function to get the full size of the Set object.