Struct boa_engine::builtins::string::String

pub(crate) struct String;

JavaScript String implementation.

Implementations

impl String

pub(crate) const MAX_STRING_LENGTH: usize = 4_294_967_295usize

JavaScript strings must be between 0 and less than positive Infinity and cannot be a negative number. The range of allowed values can be described like this: [0, +∞).

The resulting string can also not be larger than the maximum string size, which can differ in JavaScript engines. In Boa it is 2^32 - 1

fn string_create( value: JsString, prototype: JsObject, context: &mut Context<'_> ) -> JsObject

Abstract function StringCreate( value, prototype ).

Call this function if you want to create a String exotic object.

More information:

ECMAScript reference

fn this_string_value(this: &JsValue) -> JsResult<JsString>

Abstract operation thisStringValue( value )

More information:

ECMAScript reference

pub(crate) fn from_code_point( _: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.fromCodePoint(num1[, ...[, numN]])

The static String.fromCodePoint() method returns a string created by using the specified sequence of code points.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn raw( _: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.raw( template, ...substitutions )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn from_char_code( _: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.fromCharCode(...codeUnits)

Construct a String from one or more code points (as numbers). More information:

ECMAScript reference

pub(crate) fn to_string( this: &JsValue, _: &[JsValue], _: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.toString ( )

More information:

ECMAScript reference

pub(crate) fn char_at( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.charAt( index )

The String object’s charAt() method returns a new string consisting of the single UTF-16 code unit located at the specified offset into the string.

Characters in a string are indexed from left to right. The index of the first character is 0, and the index of the last character—in a string called stringName—is stringName.length - 1. If the index you supply is out of this range, JavaScript returns an empty string.

If no index is provided to charAt(), the default is 0.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn at( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.at ( index )

This String object’s at() method returns a String consisting of the single UTF-16 code unit located at the specified position. Returns undefined if the given index cannot be found.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn code_point_at( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.codePointAt( index )

The codePointAt() method returns an integer between 0 to 1114111 (0x10FFFF) representing the UTF-16 code unit at the given index.

If no UTF-16 surrogate pair begins at the index, the code point at the index is returned.

codePointAt() returns undefined if the given index is less than 0, or if it is equal to or greater than the length of the string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn char_code_at( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.charCodeAt( index )

The charCodeAt() method returns an integer between 0 and 65535 representing the UTF-16 code unit at the given index.

Unicode code points range from 0 to 1114111 (0x10FFFF). The first 128 Unicode code points are a direct match of the ASCII character encoding.

charCodeAt() returns NaN if the given index is less than 0, or if it is equal to or greater than the length of the string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn concat( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.concat( str1[, ...strN] )

The concat() method concatenates the string arguments to the calling string and returns a new string.

Changes to the original string or the returned string don’t affect the other.

If the arguments are not of the type string, they are converted to string values before concatenating.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn repeat( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.repeat( count )

The repeat() method constructs and returns a new string which contains the specified number of copies of the string on which it was called, concatenated together.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn slice( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.slice( beginIndex [, endIndex] )

The slice() method extracts a section of a string and returns it as a new string, without modifying the original string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn starts_with( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.startWith( searchString[, position] )

The startsWith() method determines whether a string begins with the characters of a specified string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn ends_with( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.endsWith( searchString[, length] )

The endsWith() method determines whether a string ends with the characters of a specified string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn includes( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.includes( searchString[, position] )

The includes() method determines whether one string may be found within another string, returning true or false as appropriate.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn replace( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.replace( regexp|substr, newSubstr|function )

The replace() method returns a new string with some or all matches of a pattern replaced by a replacement.

The pattern can be a string or a RegExp, and the replacement can be a string or a function to be called for each match. If pattern is a string, only the first occurrence will be replaced.

The original string is left unchanged.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn replace_all( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

22.1.3.18 String.prototype.replaceAll ( searchValue, replaceValue )

The replaceAll() method returns a new string with all matches of a pattern replaced by a replacement.

The pattern can be a string or a RegExp, and the replacement can be a string or a function to be called for each match.

The original string is left unchanged.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn index_of( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.indexOf( searchValue[, fromIndex] )

The indexOf() method returns the index within the calling String object of the first occurrence of the specified value, starting the search at fromIndex.

Returns -1 if the value is not found.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn last_index_of( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.lastIndexOf( searchValue[, fromIndex] )

The lastIndexOf() method returns the index within the calling String object of the last occurrence of the specified value, searching backwards from fromIndex.

Returns -1 if the value is not found.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn locale_compare( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.localeCompare ( that [ , locales [ , options ] ] )

Returns a number indicating whether a reference string comes before, or after, or is the same as the given string in sort order.

More information:

MDN documentation

pub(crate) fn match( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.match( regexp )

The match() method retrieves the result of matching a string against a regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn string_pad( object: &JsValue, max_length: &JsValue, fill_string: &JsValue, placement: Placement, context: &mut Context<'_> ) -> JsResult<JsValue>

Abstract operation StringPad ( O, maxLength, fillString, placement ).

Performs the actual string padding for padStart/End.

More information:

ECMAScript reference

pub(crate) fn pad_end( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.padEnd( targetLength[, padString] )

The padEnd() method pads the current string with a given string (repeated, if needed) so that the resulting string reaches a given length.

The padding is applied from the end of the current string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn pad_start( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.padStart( targetLength [, padString] )

The padStart() method pads the current string with another string (multiple times, if needed) until the resulting string reaches the given length.

The padding is applied from the start of the current string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trim( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.trim()

The trim() method removes whitespace from both ends of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trim_start( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.trimStart()

The trimStart() method removes whitespace from the beginning of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn trim_end( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.trimEnd()

The trimEnd() method removes whitespace from the end of a string.

Whitespace in this context is all the whitespace characters (space, tab, no-break space, etc.) and all the line terminator characters (LF, CR, etc.).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_case<const UPPER: bool>( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.toUpperCase() and String.prototype.toLowerCase()

The case methods return the calling string value converted to uppercase or lowercase.

The value will be converted to a string if it isn’t one.

More information:

MDN documentation

pub(crate) fn to_locale_case<const UPPER: bool>( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.toLocaleLowerCase ( [ locales ] ) and String.prototype.toLocaleUpperCase ( [ locales ] )

pub(crate) fn substring( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.substring( indexStart[, indexEnd] )

The substring() method returns the part of the string between the start and end indexes, or to the end of the string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn split( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.split ( separator, limit )

The split() method divides a String into an ordered list of substrings, puts these substrings into an array, and returns the array. The division is done by searching for a pattern; where the pattern is provided as the first parameter in the method’s call.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn value_of( this: &JsValue, _args: &[JsValue], _context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.valueOf()

The valueOf() method returns the primitive value of a String object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match_all( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.matchAll( regexp )

The matchAll() method returns an iterator of all results matching a string against a regular expression, including capturing groups.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn normalize( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.normalize( [ form ] )

The normalize() method normalizes a string into a form specified in the Unicode® Standard Annex #15

More information:

ECMAScript reference

MDN documentation

pub(crate) fn search( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.search( regexp )

The search() method executes a search for a match between a regular expression and this String object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn iterator( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

impl String

pub(crate) fn substr( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.substr( start[, length] )

The substr() method returns a portion of the string, starting at the specified index and extending for a given number of characters afterward.

More information:

ECMAScript reference

MDN documentation

https://tc39.es/ecma262/#sec-string.prototype.substr

pub(crate) fn create_html( string: &JsValue, tag: &[u16], attribute_and_value: Option<(&[u16], &JsValue)>, context: &mut Context<'_> ) -> JsResult<JsValue>

CreateHTML(string, tag, attribute, value)

More information:

ECMAScript reference

pub(crate) fn anchor( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.anchor( name )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn big( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.big( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn blink( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.blink( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn bold( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.bold( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn fixed( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.fixed( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn fontcolor( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.fontcolor( color )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn fontsize( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.fontsize( size )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn italics( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.italics( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn link( this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.link( url )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn small( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.small( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn strike( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.strike( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sub( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.sub( )

More information:

ECMAScript reference

MDN documentation

pub(crate) fn sup( this: &JsValue, _: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

String.prototype.sup( )

More information:

ECMAScript reference

MDN documentation