Struct boa_engine::builtins::BuiltInConstructorWithPrototype

struct BuiltInConstructorWithPrototype<'ctx> {

    realm: &'ctx Realm,
    function: NativeFunctionPointer,
    name: JsString,
    length: usize,
    object_property_table: PropertyTableInner,
    object_storage: Vec<JsValue>,
    object: JsObject,
    prototype_property_table: PropertyTableInner,
    prototype_storage: Vec<JsValue>,
    prototype: JsObject,
    __proto__: JsPrototype,
    inherits: Option<JsObject>,
    attributes: Attribute,

}

Fields
realm: &'ctx Realmfunction: NativeFunctionPointername: JsStringlength: usizeobject_property_table: PropertyTableInnerobject_storage: Vec<JsValue>object: JsObjectprototype_property_table: PropertyTableInnerprototype_storage: Vec<JsValue>prototype: JsObject__proto__: JsPrototypeinherits: Option<JsObject>attributes: Attribute
Implementations

impl BuiltInConstructorWithPrototype<'_>

const fn length(self, length: usize) -> Self

Specify how many arguments the constructor function takes.

Default is 0.

fn name<N: Into<JsString>>(self, name: N) -> Self

Specify the name of the constructor function.

Default is ""

fn static_method<B>( self, function: NativeFunctionPointer, binding: B, length: usize ) -> Selfwhere B: Into<FunctionBinding>,

Adds a new static method to the builtin object.

fn static_property<K, V>(self, key: K, value: V, attribute: Attribute) -> Selfwhere K: Into<PropertyKey>, V: Into<JsValue>,

Adds a new static data property to the builtin object.

fn static_accessor<K>( self, key: K, get: Option<JsFunction>, set: Option<JsFunction>, attribute: Attribute ) -> Selfwhere K: Into<PropertyKey>,

Adds a new static accessor property to the builtin object.

fn prototype(self, prototype: JsObject) -> Self

Specify the [[Prototype]] internal field of the builtin object.

Default is Function.prototype for constructors and Object.prototype for statics.

fn method<B>( self, function: NativeFunctionPointer, binding: B, length: usize ) -> Selfwhere B: Into<FunctionBinding>,

Adds a new method to the constructor’s prototype.

fn property<K, V>(self, key: K, value: V, attribute: Attribute) -> Selfwhere K: Into<PropertyKey>, V: Into<JsValue>,

Adds a new data property to the constructor’s prototype.

fn accessor<K>( self, key: K, get: Option<JsFunction>, set: Option<JsFunction>, attribute: Attribute ) -> Selfwhere K: Into<PropertyKey>,

Adds new accessor property to the constructor’s prototype.

fn inherits(self, prototype: JsPrototype) -> Self

Specifies the parent prototype which objects created by this constructor inherit from.

Default is Object.prototype.

const fn constructor_attributes(self, attributes: Attribute) -> Self

Specifies the property attributes of the prototype’s “constructor” property.

fn build(self)

fn build_without_prototype(self)