Module boa_engine::builtins::temporal

The ECMAScript Temporal stage 3 built-in implementation.

More information:

Modules

calendar 🔒

An implementation of the Temporal proposal’s Calendar builtin.

date_equations 🔒

This file represents all equations listed under section 13.4 of the Temporal Specification

duration 🔒

fields 🔒

A Rust native implementation of the fields object used in Temporal.

instant 🔒

Boa’s implementation of ECMAScript’s Temporal.Instant builtin object.

now 🔒

Boa’s implementation of Temporal.Now ECMAScript Builtin object.

options 🔒

Temporal Option types.

plain_date 🔒

Boa’s implementation of the ECMAScript Temporal.PlainDate builtin object.

plain_date_time 🔒

Boa’s implementation of the ECMAScript Temporal.PlainDateTime builtin object.

plain_month_day 🔒

Boa’s implementation of the ECMAScript Temporal.PlainMonthDay builtin object.

plain_time 🔒

Boa’s implementation of the ECMAScript Temporal.PlainTime builtin object.

plain_year_month 🔒

Boa’s implementation of the Temporal.PlainYearMonth builtin object.

time_zone 🔒

zoned_date_time 🔒

Structs

Calendar

The Temporal.Calendar object.

Duration

The Temporal.Duration object.

Instant

The Temporal.Instant object.

Now

JavaScript Temporal.Now object.

PlainDate

The Temporal.PlainDate object.

PlainDateTime

The Temporal.PlainDateTime object.

PlainMonthDay

The Temporal.PlainMonthDay object.

PlainTime

The Temporal.PlainTime object.

PlainYearMonth

The Temporal.PlainYearMonth object.

Temporal 🔒

The Temporal builtin object.

TimeZone

The Temporal.TimeZone object.

ZonedDateTime

The Temporal.ZonedDateTime object.

Enums

DateTimeValues 🔒

Constants

MIS_PER_DAY 🔒

Microseconds per day constant: 8.64e+10

MS_PER_DAY 🔒

Milliseconds per day constant: 8.64e+7

NS_PER_DAY 🔒

Nanoseconds per day constant: 8.64e+13

Functions

apply_unsigned_rounding_mode 🔒

13.27 ApplyUnsignedRoundingMode ( x, r1, r2, unsignedRoundingMode )

copy_data_properties 🔒

14.6 CopyDataProperties ( target, source, excludedKeys [ , excludedValues ] )

epoch_days_to_epoch_ms 🔒

13.2 ISODateToEpochDays ( year, month, date )

get_diff_settings 🔒

13.47 GetDifferenceSettings ( operation, options, unitGroup, disallowedUnits, fallbackSmallestUnit, smallestLargestDefaultUnit )

iterator_to_list_of_types 🔒

Abstract Operation 13.1 IteratorToListOfType

ns_max_instant 🔒

ns_min_instant 🔒

round_number_to_increment 🔒

13.28 RoundNumberToIncrement ( x, increment, roundingMode )

round_to_increment_as_if_positive 🔒

13.29 RoundNumberToIncrementAsIfPositive ( x, increment, roundingMode )

snapshot_own_properties 🔒

14.7 SnapshotOwnProperties ( source, proto [ , excludedKeys [ , excludedValues ] ] )

to_integer_if_integral 🔒

Abstract operation 13.45 ToIntegerIfIntegral( argument )

to_integer_with_truncation 🔒

13.44 ToIntegerWithTruncation ( argument )

to_positive_integer_with_trunc 🔒

13.43 ToPositiveIntegerWithTruncation ( argument )

to_relative_temporal_object 🔒

13.21 ToRelativeTemporalObject ( options )

to_zero_padded_decimal_string 🔒

Abstract operation ToZeroPaddedDecimalString ( n, minLength )

validate_temporal_rounding_increment 🔒

13.7 ToTemporalOverflow (options) 13.10 ToTemporalRoundingMode ( normalizedOptions, fallback ) 13.17 ValidateTemporalRoundingIncrement ( increment, dividend, inclusive )