Struct boa_engine::builtins::temporal::calendar::object::JsCustomCalendar

pub(crate) struct JsCustomCalendar {
    calendar: JsObject,
}

A user-defined, custom calendar that is only known at runtime and executed at runtime.

A user-defined calendar implements all of the CalendarProtocolMethods and therefore satisfies the requirements to be used as a calendar.

Fields
calendar: JsObject
Implementations

impl JsCustomCalendar

pub(crate) fn new(calendar: &JsObject) -> Self