Trait boa_engine::builtins::temporal::calendar::BuiltinCalendar

pub(crate) trait BuiltinCalendar {

    // Required methods
    fn date_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn year_month_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn month_day_from_fields(
        &self,
        fields: &mut TemporalFields,
        overflow: ArithmeticOverflow,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn date_add(
        &self,
        date: &PlainDate,
        duration: &DurationRecord,
        overflow: ArithmeticOverflow,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn date_until(
        &self,
        one: &PlainDate,
        two: &PlainDate,
        largest_unit: TemporalUnit,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn era(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn era_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn month(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn month_code(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn day(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn day_of_week(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn day_of_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn week_of_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn year_of_week(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn days_in_week(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn days_in_month(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn days_in_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn months_in_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn in_leap_year(
        &self,
        date_like: &IsoDateRecord,
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
    fn resolve_fields(
        &self,
        fields: &mut TemporalFields,
        type: &str
    ) -> JsResult<()>;
    fn field_descriptors(&self, type: &[String]) -> Vec<(String, bool)>;
    fn field_keys_to_ignore(
        &self,
        additional_keys: Vec<PropertyKey>
    ) -> Vec<PropertyKey>;
    fn debug_name(&self) -> &str;

}

A trait for implementing a Builtin Calendar’s Calendar Protocol in Rust.

Required Methods

fn date_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut Context<'_> ) -> JsResult<JsValue>

Creates a Temporal.PlainDate object from provided fields.

fn year_month_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut Context<'_> ) -> JsResult<JsValue>

Creates a Temporal.PlainYearMonth object from the provided fields.

fn month_day_from_fields( &self, fields: &mut TemporalFields, overflow: ArithmeticOverflow, context: &mut Context<'_> ) -> JsResult<JsValue>

Creates a Temporal.PlainMonthDay object from the provided fields.

fn date_add( &self, date: &PlainDate, duration: &DurationRecord, overflow: ArithmeticOverflow, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns a Temporal.PlainDate based off an added date.

fn date_until( &self, one: &PlainDate, two: &PlainDate, largest_unit: TemporalUnit, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns a Temporal.Duration representing the duration between two dates.

fn era( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the era for a given temporaldatelike.

fn era_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the era year for a given temporaldatelike

fn year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the year for a given temporaldatelike

fn month( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the month for a given temporaldatelike

fn month_code( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the monthCode for a given temporaldatelike

fn day( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the day for a given temporaldatelike

fn day_of_week( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns a value representing the day of the week for a date.

fn day_of_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns a value representing the day of the year for a given calendar.

fn week_of_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns a value representing the week of the year for a given calendar.

fn year_of_week( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the year of a given week.

fn days_in_week( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the days in a week for a given calendar.

fn days_in_month( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the days in a month for a given calendar.

fn days_in_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the days in a year for a given calendar.

fn months_in_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns the months in a year for a given calendar.

fn in_leap_year( &self, date_like: &IsoDateRecord, context: &mut Context<'_> ) -> JsResult<JsValue>

Returns whether a value is within a leap year according to the designated calendar.

fn resolve_fields( &self, fields: &mut TemporalFields, type: &str ) -> JsResult<()>

Resolve the TemporalFields for the implemented Calendar

fn field_descriptors(&self, type: &[String]) -> Vec<(String, bool)>

Return this calendar’s a fieldName and whether it is required depending on type (date, day-month).

fn field_keys_to_ignore( &self, additional_keys: Vec<PropertyKey> ) -> Vec<PropertyKey>

Return the fields to ignore for this Calendar based on provided keys.

fn debug_name(&self) -> &str

Debug name