Struct boa_engine::builtins::temporal::duration::record::DurationRecord

pub(crate) struct DurationRecord {
    date: DateDuration,
    time: TimeDuration,
}

The DurationRecord is a native Rust implementation of the Duration builtin object internal fields and is primarily defined by Abtract Operation 7.5.1-5.

Fields
date: DateDurationtime: TimeDuration
Implementations

impl DurationRecord

pub(crate) const fn new(date: DateDuration, time: TimeDuration) -> Self

pub(crate) const fn partial() -> Self

pub(crate) fn from_date_duration(date: DateDuration) -> Self

pub(crate) const fn from_day_and_time(day: f64, time: TimeDuration) -> Self

pub(crate) fn from_partial_js_object( duration_like: &JsValue, context: &mut Context<'_> ) -> JsResult<Self>

Equivalent to 7.5.13 ToTemporalPartialDurationRecord ( temporalDurationLike )

Takes an unknown JsObject and attempts to create a partial duration

impl DurationRecord

fn balance_hours(&mut self)

Balance/bubble the current unit from one step down.

fn balance_minutes(&mut self)

Balance/bubble the current unit from one step down.

fn balance_seconds(&mut self)

Balance/bubble the current unit from one step down.

fn balance_milliseconds(&mut self)

Balance/bubble the current unit from one step down.

fn balance_microseconds(&mut self)

Balance/bubble the current unit from one step down.

impl DurationRecord

pub(crate) const fn date(&self) -> DateDuration

Return this DurationRecord’s DateDuration

pub(crate) const fn time(&self) -> TimeDuration

Return this DurationRecord’s TimeDuration

pub(crate) fn set_time_duration(&mut self, time: TimeDuration)

Set this DurationRecord’s TimeDuration.

pub(crate) fn set_years(&mut self, y: f64)

Set the value for years.

pub(crate) const fn years(&self) -> f64

Return the value for years.

pub(crate) fn set_months(&mut self, mo: f64)

Set the value for months.

pub(crate) const fn months(&self) -> f64

Return the value for months.

pub(crate) fn set_weeks(&mut self, w: f64)

Set the value for weeks.

pub(crate) const fn weeks(&self) -> f64

Return the value for weeks.

pub(crate) fn set_days(&mut self, d: f64)

Set the value for days.

pub(crate) const fn days(&self) -> f64

Return the value for days.

pub(crate) fn set_hours(&mut self, h: f64)

Set the value for hours.

pub(crate) const fn hours(&self) -> f64

Return the value for hours.

pub(crate) fn set_minutes(&mut self, m: f64)

Set the value for minutes.

pub(crate) const fn minutes(&self) -> f64

Return the value for minutes.

pub(crate) fn set_seconds(&mut self, s: f64)

Set the value for seconds.

pub(crate) const fn seconds(&self) -> f64

Return the value for seconds.

pub(crate) fn set_milliseconds(&mut self, ms: f64)

Set the value for milliseconds.

pub(crate) const fn milliseconds(&self) -> f64

Return the value for milliseconds.

pub(crate) fn set_microseconds(&mut self, mis: f64)

Set the value for microseconds.

pub(crate) const fn microseconds(&self) -> f64

Return the value for microseconds.

pub(crate) fn set_nanoseconds(&mut self, ns: f64)

Set the value for nanoseconds.

pub(crate) const fn nanoseconds(&self) -> f64

Return the value for nanoseconds.

impl DurationRecord

pub(crate) fn abs(&self) -> Self

impl DurationRecord

pub(crate) fn as_object(&self, context: &mut Context<'_>) -> JsResult<JsObject>

Creates a Duration object from the current DurationRecord.

fn time_values(&self) -> Vec<f64>

Returns the duration time values as a vec

fn is_overfowed(&self) -> bool

Determines if the DurationRecord has overflowed.

pub(crate) fn is_positive_overflow(&self) -> bool

pub(crate) fn is_negative_overflow(&self) -> bool

pub(crate) fn duration_sign(&self) -> i32

7.5.10 DurationSign ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds )

Determines the sign for the current self.

pub(crate) fn is_valid_duration(&self) -> bool

7.5.11 IsValidDuration ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds )

Checks if the current DurationRecord is a valid self.

pub(crate) fn default_temporal_largest_unit(&self) -> TemporalUnit

7.5.12 DefaultTemporalLargestUnit ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds )

fn total_duration_nanoseconds(&self, offset_shift: f64) -> f64

7.5.17 TotalDurationNanoseconds ( days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, offsetShift )

pub(crate) fn balance_time_duration( &mut self, largest_unit: TemporalUnit, relative_to: Option<&JsValue> ) -> JsResult<()>

Abstract Operation 7.5.18 BalanceTimeDuration ( days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, largestUnit [ , relativeTo ] )

pub(crate) fn balance_possibly_infinite_duration( &mut self, largest_unit: TemporalUnit, relative_to: Option<&JsValue> ) -> JsResult<()>

Abstract Operation 7.5.19 BalancePossiblyInfiniteDuration ( days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, largestUnit [ , relativeTo ] )

pub(crate) fn unbalance_duration_relative( &mut self, largest_unit: TemporalUnit, relative_to: &JsValue, context: &mut Context<'_> ) -> JsResult<()>

7.5.20 UnbalanceDurationRelative ( years, months, weeks, days, largestUnit, relativeTo )

pub(crate) fn balance_date_duration_relative( &mut self, largest_unit: TemporalUnit, relative_to: &JsValue, context: &mut Context<'_> ) -> JsResult<()>

BalanceDateDurationRelative

pub(crate) fn round_duration( &mut self, increment: f64, unit: TemporalUnit, rounding_mode: RoundingMode, relative_to: Option<&JsValue>, context: &mut Context<'_> ) -> JsResult<f64>

Abstract Operation 7.5.26 RoundDuration ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, increment, unit, roundingMode [ , relativeTo ] )

pub(crate) fn adjust_rounded_duration_days( &mut self, increment: f64, unit: TemporalUnit, rounding_mode: RoundingMode, relative_to: Option<&JsValue>, context: &mut Context<'_> ) -> JsResult<()>

7.5.27 AdjustRoundedDurationDays ( years, months, weeks, days, hours, minutes, seconds, milliseconds, microseconds, nanoseconds, increment, unit, roundingMode, relativeTo )