Trait boa_engine::builtins::IntrinsicObject

pub(crate) trait IntrinsicObject {
    // Required methods
    fn init(realm: &Realm);
    fn get(intrinsics: &Intrinsics) -> JsObject;
}

A Well-Known Intrinsic Object.

Well-known intrinsics are built-in objects that are explicitly referenced by the algorithms of the specification and which usually have realm-specific identities.

Required Methods

fn init(realm: &Realm)

Initializes the intrinsic object.

This is where the methods, properties, static methods and the constructor of a built-in must be initialized to be accessible from ECMAScript.

fn get(intrinsics: &Intrinsics) -> JsObject

Gets the intrinsic object.