Trait boa_engine::builtins::typed_array::element::atomic::Atomic

pub(crate) trait Atomic {
    type Plain;

    // Required methods
    fn load(&self, order: Ordering) -> Self::Plain;
    fn store(&self, val: Self::Plain, order: Ordering);
    fn add(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn bit_and(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn compare_exchange(
        &self,
        expected: Self::Plain,
        replacement: Self::Plain,
        order: Ordering
    ) -> Self::Plain;
    fn swap(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn bit_or(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn sub(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn bit_xor(&self, val: Self::Plain, order: Ordering) -> Self::Plain;
    fn is_lock_free() -> bool;
}

An atomic type that supports atomic operations.

Required Associated Types

type Plain

The “plain” type of the atomic e.g. AtomicU8::Plain == u8

Required Methods

fn load(&self, order: Ordering) -> Self::Plain

Loads the value of this atomic.

fn store(&self, val: Self::Plain, order: Ordering)

Stores value on this atomic.

fn add(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Computes the + operation between self and value, storing the result on self and returning the old value. This operation wraps on overflow.

fn bit_and(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Computes the & operation between self and value, storing the result on self and returning the old value.

fn compare_exchange( &self, expected: Self::Plain, replacement: Self::Plain, order: Ordering ) -> Self::Plain

Compares the current value of self with expected, storing replacement if they’re equal and returning its old value in all cases.

fn swap(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Swaps self with value, returning the old value of self.

fn bit_or(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Computes the | operation between self and value, storing the result on self and returning the old value.

fn sub(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Computes the - operation between self and value, storing the result on self and returning the old value. This operation wraps on overflow.

fn bit_xor(&self, val: Self::Plain, order: Ordering) -> Self::Plain

Computes the ^ operation between self and value, storing the result on self and returning the old value.

fn is_lock_free() -> bool

Checks if this atomic does not use any locks to support atomic operations.