Module boa_engine::builtins::uri

Boa’s implementation of ECMAScript’s URI Handling Functions.

Uniform Resource Identifiers, or URIs, are Strings that identify resources (e.g. web pages or files) and transport protocols by which to access them (e.g. HTTP or FTP) on the Internet. The ECMAScript language itself does not provide any support for using URIs except for functions that encode and decode URIs as described in 19.2.6.2, 19.2.6.3, 19.2.6.4 and 19.2.6.5

More information:

ECMAScript reference

Modules

consts 🔒

URI handling function constants

Structs

DecodeUri 🔒

URI Handling Functions

DecodeUriComponent 🔒

EncodeUri 🔒

EncodeUriComponent 🔒

UriFunctions

Intrinsics for the URI Handling Functions.

Functions

decode 🔒

The Decode ( string, reservedSet ) abstract operation.

decode_hex_byte 🔒

Decodes a byte from two unicode code units.

decode_uri 🔒

Builtin JavaScript decodeURI ( encodedURI ) function.

decode_uri_component 🔒

Builtin JavaScript decodeURIComponent ( encodedURIComponent ) function.

encode 🔒

The Encode ( string, unescapedSet ) abstract operation

encode_uri 🔒

Builtin JavaScript encodeURI ( uri ) function.

encode_uri_component 🔒

Builtin JavaScript encodeURIComponent ( uriComponent ) function.

mark 🔒

mark 🔒

mark 🔒

mark 🔒