Module boa_engine::console

Boa’s implementation of JavaScript’s console Web API object.

The console object can be accessed from any global object.

The specifics of how it works varies from browser to browser, but there is a de facto set of features that are typically provided.

More information:

MDN documentation

WHATWG console specification

Structs

Console 🔒

This is the internal console object state.

Enums

LogMessage 🔒

This represents the different types of log messages.

Functions

formatter

This represents the console formatter.

logger 🔒

Helper function for logging messages.