Module boa_engine::context

The ECMAScript context.

Modules

intrinsics

Data structures that contain intrinsic objects and constructors.

Structs

Context

ECMAScript context. It is the primary way to interact with the runtime.

ContextBuilder

Builder for the Context type.

DefaultHooks

Default implementation of HostHooks, which doesn’t carry any state.

PhantomData

Zero-sized type used to mark things that “act like” they own a T.

Enums

MaybeShared

A Cow-like pointer where the Owned variant is an Rc.

Traits

HostHooks

Host Hooks customizable by the host code or engine.