Struct boa_engine::context::icu::IntlProvider

pub(crate) struct IntlProvider {
    inner_provider: ErasedProvider,
    locale_canonicalizer: LocaleCanonicalizer,
    locale_expander: LocaleExpander,
    string_normalizers: StringNormalizers,
    case_mapper: CaseMapper,
}

Custom [DataProvider] for Intl that caches some utilities.

Fields
inner_provider: ErasedProviderlocale_canonicalizer: LocaleCanonicalizerlocale_expander: LocaleExpanderstring_normalizers: StringNormalizerscase_mapper: CaseMapper
Implementations

impl IntlProvider

pub(crate) fn try_new_with_buffer_provider( provider: impl BufferProvider + 'static ) -> Result<IntlProvider, IcuError>

Creates a new IntlProvider from a [BufferProvider].

Errors

Returns an error if any of the tools required cannot be constructed.

pub(crate) fn try_new_with_any_provider( provider: impl AnyProvider + 'static ) -> Result<IntlProvider, IcuError>

Creates a new IntlProvider from an [AnyProvider].

Errors

Returns an error if any of the tools required cannot be constructed.

pub(crate) const fn locale_canonicalizer(&self) -> &LocaleCanonicalizer

Gets the [LocaleCanonicalizer] tool.

pub(crate) const fn locale_expander(&self) -> &LocaleExpander

Gets the [LocaleExpander] tool.

pub(crate) const fn string_normalizers(&self) -> &StringNormalizers

Gets the StringNormalizers tools.

pub(crate) const fn case_mapper(&self) -> &CaseMapper

Gets the [CaseMapper] tool.