Struct boa_engine::context::ContextBuilder

pub struct ContextBuilder<'icu, 'hooks, 'queue> {
    interner: Option<Interner>,
    host_hooks: Option<&'hooks dyn HostHooks>,
    job_queue: Option<&'queue dyn JobQueue>,
    icu: Option<Icu<'icu>>,
    instructions_remaining: usize,
}

Builder for the Context type.

This builder allows custom initialization of the Interner within the context. Additionally, if the intl feature is enabled, ContextBuilder becomes the only way to create a new Context, since now it requires a valid data provider for the Intl functionality. The required data in a valid provider is specified in BoaProvider

Fields
interner: Option<Interner>host_hooks: Option<&'hooks dyn HostHooks>job_queue: Option<&'queue dyn JobQueue>icu: Option<Icu<'icu>>instructions_remaining: usize
Implementations

impl<'icu, 'hooks, 'queue> ContextBuilder<'icu, 'hooks, 'queue>

pub fn new() -> Self

Creates a new ContextBuilder with a default empty Interner and a default BoaProvider if the intl feature is enabled.

pub fn interner(self, interner: Interner) -> Self

Initializes the context Interner to the provided interner.

This is useful when you want to initialize an Interner with a collection of words before parsing.

pub fn icu_provider(
self,
provider: BoaProvider<'_>
) -> Result<ContextBuilder<'_, 'hooks, 'queue>, LocaleTransformError>

Provides an icu data provider to the Context.

This function is only available if the intl feature is enabled.

Errors

This returns Err if the provided provider doesn’t have the required locale information to construct both a LocaleCanonicalizer and a LocaleExpander. Note that this doesn’t mean that the provider will successfully construct all Intl services; that check is made until the creation of an instance of a service.

pub fn host_hooks(
self,
host_hooks: &dyn HostHooks
) -> ContextBuilder<'icu, '_, 'queue>

Initializes the HostHooks for the context.

pub fn job_queue(
self,
job_queue: &dyn JobQueue
) -> ContextBuilder<'icu, 'hooks, '_>

Initializes the JobQueue for the context.

pub const fn instructions_remaining(self, instructions_remaining: usize) -> Self

Specifies the number of instructions remaining to the Context.

This function is only available if the fuzz feature is enabled.

pub fn build<'host>(self) -> Context<'host>where
'icu: 'host,
'hooks: 'host,
'queue: 'host,

Builds a new Context with the provided parameters, and defaults all missing parameters to their default values.