Struct boa_engine::environments::compile::CompileTimeEnvironment

pub(crate) struct CompileTimeEnvironment {
    outer: Option<Rc<Self>>,
    environment_index: u32,
    bindings: RefCell<FxHashMap<Identifier, CompileTimeBinding>>,
    function_scope: bool,
}

A compile time environment maps bound identifiers to their binding positions.

A compile time environment also indicates, if it is a function environment.

Fields
outer: Option<Rc<Self>>environment_index: u32bindings: RefCell<FxHashMap<Identifier, CompileTimeBinding>>function_scope: bool
Implementations

impl CompileTimeEnvironment

pub(crate) fn new_global() -> Self

Creates a new global compile time environment.

pub(crate) fn new(parent: Rc<Self>, function_scope: bool) -> Self

Creates a new compile time environment.

pub(crate) fn has_lex_binding(&self, name: Identifier) -> bool

Check if environment has a lexical binding with the given name.

pub(crate) fn has_binding(&self, name: Identifier) -> bool

Check if the environment has a binding with the given name.

pub(crate) fn get_identifier_reference( &self, name: Identifier ) -> IdentifierReference

Get the binding locator for a binding with the given name. Fall back to the global environment if the binding is not found.

pub(crate) fn num_bindings(&self) -> u32

Returns the number of bindings in this environment.

pub(crate) fn environment_index(&self) -> u32

Returns the index of this environment.

pub(crate) const fn is_function(&self) -> bool

Check if the environment is a function environment.

pub(crate) const fn is_global(&self) -> bool

Check if the environment is a global environment.

pub(crate) fn get_binding(&self, name: Identifier) -> Option<BindingLocator>

Get the locator for a binding name.

pub(crate) fn create_mutable_binding( &self, name: Identifier, function_scope: bool ) -> BindingLocator

Create a mutable binding.

pub(crate) fn create_immutable_binding( &self, name: Identifier, strict: bool ) -> BindingLocator

Crate an immutable binding.

pub(crate) fn set_mutable_binding( &self, name: Identifier ) -> Result<BindingLocator, BindingLocatorError>

Return the binding locator for a mutable binding.

pub(crate) fn set_mutable_binding_var( &self, name: Identifier ) -> Result<BindingLocator, BindingLocatorError>

Return the binding locator for a set operation on an existing var binding.

pub(crate) fn outer(&self) -> Option<Rc<Self>>

Gets the outer environment of this environment.