Struct boa_engine::environments::runtime::BindingLocator

pub(crate) struct BindingLocator {
    name: Identifier,
    environment_index: u32,
    binding_index: u32,
    global: bool,
    mutate_immutable: bool,
    silent: bool,
}

A binding locator contains all information about a binding that is needed to resolve it at runtime.

Binding locators get created at bytecode compile time and are accessible at runtime via the crate::vm::CodeBlock.

Fields
name: Identifierenvironment_index: u32binding_index: u32global: boolmutate_immutable: boolsilent: bool
Implementations

impl BindingLocator

pub(crate) const fn declarative( name: Identifier, environment_index: u32, binding_index: u32 ) -> Self

Creates a new declarative binding locator that has knows indices.

pub(super) const fn global(name: Identifier) -> Self

Creates a binding locator that indicates that the binding is on the global object.

pub(super) const fn mutate_immutable(name: Identifier) -> Self

Creates a binding locator that indicates that it was attempted to mutate an immutable binding. At runtime this should always produce a type error.

pub(super) const fn silent(name: Identifier) -> Self

Creates a binding locator that indicates that any action is silently ignored.

pub(crate) const fn name(&self) -> Identifier

Returns the name of the binding.

pub(crate) const fn is_global(&self) -> bool

Returns if the binding is located on the global object.

pub(crate) const fn environment_index(&self) -> u32

Returns the environment index of the binding.

pub(crate) const fn binding_index(&self) -> u32

Returns the binding index of the binding.

pub(crate) const fn is_silent(&self) -> bool

Returns if the binding is a silent operation.

pub(crate) fn throw_mutate_immutable( &self, context: &mut Context<'_> ) -> Result<(), JsNativeError>

Helper method to throws an error if the binding access is illegal.