Struct boa_engine::environments::runtime::DeclarativeEnvironment

pub(crate) struct DeclarativeEnvironment {
    bindings: Cell<Vec<Option<JsValue>>>,
    compile: Gc<Cell<CompileTimeEnvironment>>,
    poisoned: Cell<bool>,
    slots: Option<EnvironmentSlots>,
}

A declarative environment holds binding values at runtime.

Bindings are stored in a fixed size list of optional values. If a binding is not initialized, the value is None.

Optionally, an environment can hold a this value. The this value is present only if the environment is a function environment.

Code evaluation at runtime (e.g. the eval built-in function) can add bindings to existing, compiled function environments. This makes it impossible to determine the location of all bindings at compile time. To dynamically check for added bindings at runtime, a reference to the corresponding compile time environment is needed.

Checking all environments for potential added bindings at runtime on every get/set would offset the performance improvement of determining binding locations at compile time. To minimize this, each environment holds a poisoned flag. If bindings where added at runtime, the current environment and all inner environments are marked as poisoned. All poisoned environments have to be checked for added bindings.

Fields
bindings: Cell<Vec<Option<JsValue>>>compile: Gc<Cell<CompileTimeEnvironment>>poisoned: Cell<bool>slots: Option<EnvironmentSlots>
Implementations

impl DeclarativeEnvironment

pub(crate) fn slots(&self) -> Option<&EnvironmentSlots>

Returns the internal slot data of the current environment.

pub(crate) fn get(&self, index: usize) -> JsValue

Get the binding value from the environment by it’s index.

Panics

Panics if the binding value is out of range or not initialized.

pub(crate) fn set(&self, index: usize, value: JsValue)

Set the binding value at the specified index.

Panics

Panics if the binding value is out of range or not initialized.