Struct boa_engine::environments::runtime::DeclarativeEnvironmentStack

pub struct DeclarativeEnvironmentStack {
    stack: Vec<Gc<DeclarativeEnvironment>>,
}

A declarative environment stack holds all declarative environments at runtime.

Environments themselves are garbage collected, because they must be preserved for function calls.

Fields
stack: Vec<Gc<DeclarativeEnvironment>>
Implementations

impl DeclarativeEnvironmentStack

pub(crate) fn new(
global_compile_environment: Gc<GcCell<CompileTimeEnvironment>>
) -> Self

Create a new environment stack with the most outer declarative environment.

pub(crate) fn extend_outer_function_environment(&mut self)

Extends the length of the next outer function environment to the number of compiled bindings.

This is only useful when compiled bindings are added after the initial compilation (eval).

pub(crate) fn has_lex_binding_until_function_environment(
&self,
names: &FxHashSet<Identifier>
) -> Option<Identifier>

Check if any of the provided binding names are defined as lexical bindings.

Start at the current environment. Stop at the next outer function environment.

pub(crate) fn pop_to_global(&mut self) -> Vec<Gc<DeclarativeEnvironment>>

Pop all current environments except the global environment.

pub(crate) fn len(&self) -> usize

Get the number of current environments.

pub(crate) fn truncate(&mut self, len: usize)

Truncate current environments to the given number.

pub(crate) fn extend(&mut self, other: Vec<Gc<DeclarativeEnvironment>>)

Extend the current environment stack with the given environments.

pub(crate) fn set_global_binding_number(&mut self, binding_number: usize)

Set the number of bindings on the global environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn get_this_environment(&self) -> &EnvironmentSlots

GetThisEnvironment

Returns the environment that currently provides a this biding.

More information:

ECMAScript specification

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_declarative(
&mut self,
num_bindings: usize,
compile_environment: Gc<GcCell<CompileTimeEnvironment>>
) -> usize

Push a declarative environment on the environments stack and return it’s index.

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_function(
&mut self,
num_bindings: usize,
compile_environment: Gc<GcCell<CompileTimeEnvironment>>,
this: Option<JsValue>,
function_object: JsObject,
new_target: Option<JsObject>,
lexical: bool
)

Push a function environment on the environments stack.

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_function_inherit(
&mut self,
num_bindings: usize,
compile_environment: Gc<GcCell<CompileTimeEnvironment>>
)

Push a function environment that inherits it’s internal slots from the outer environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn pop(&mut self) -> Gc<DeclarativeEnvironment>

Pop environment from the environments stack.

pub(crate) fn current_function_slots(&self) -> &EnvironmentSlots

Get the most outer function environment slots.

Panics

Panics if no environment exists on the stack.

pub(crate) fn current(&mut self) -> Gc<DeclarativeEnvironment>

Get the most outer environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn current_compile_environment(
&self
) -> Gc<GcCell<CompileTimeEnvironment>>

Get the compile environment for the current runtime environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn poison_current(&mut self)

Mark that there may be added bindings in the current environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn poison_all(&mut self)

Mark that there may be added binding in all environments.

pub(crate) fn get_value_optional(
&self,
environment_index: usize,
binding_index: usize,
name: Identifier
) -> Option<JsValue>

Get the value of a binding.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn get_value_if_global_poisoned(
&self,
name: Identifier
) -> Option<JsValue>

Get the value of a binding by it’s name.

This only considers function environments that are poisoned. All other bindings are accessed via indices.

pub(crate) fn put_value(
&mut self,
environment_index: usize,
binding_index: usize,
value: JsValue
)

Set the value of a binding.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn put_value_if_initialized(
&mut self,
environment_index: usize,
binding_index: usize,
name: Identifier,
value: JsValue
) -> bool

Set the value of a binding if it is initialized. Return true if the value has been set.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn put_value_if_uninitialized(
&mut self,
environment_index: usize,
binding_index: usize,
value: JsValue
)

Set the value of a binding if it is uninitialized.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn put_value_if_global_poisoned(
&mut self,
name: Identifier,
value: &JsValue
) -> bool

Set the value of a binding by it’s name.

This only considers function environments that are poisoned. All other bindings are set via indices.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn is_only_global_property(&mut self, name: Identifier) -> bool

Checks if the name only exists as a global property.

A binding could be marked as global, and at the same time, exist in a deeper environment context; if the global context is poisoned, an eval call could have added a binding that is not global with the same name as the global binding. This double checks that the binding name is truly a global property.

Panics

Panics if the environment or binding index are out of range.