Struct boa_engine::environments::runtime::EnvironmentStack

pub(crate) struct EnvironmentStack {
    stack: Vec<Environment>,
    private_stack: Vec<Gc<PrivateEnvironment>>,
}

The environment stack holds all environments at runtime.

Environments themselves are garbage collected, because they must be preserved for function calls.

Fields
stack: Vec<Environment>private_stack: Vec<Gc<PrivateEnvironment>>
Implementations

impl EnvironmentStack

pub(crate) fn new(global: Gc<DeclarativeEnvironment>) -> Self

Create a new environment stack.

pub(crate) fn replace_global(&mut self, global: Gc<DeclarativeEnvironment>)

Replaces the current global with a new global environment.

pub(crate) fn extend_outer_function_environment(&mut self)

Extends the length of the next outer function environment to the number of compiled bindings.

This is only useful when compiled bindings are added after the initial compilation (eval).

pub(crate) fn has_lex_binding_until_function_environment( &self, names: &FxHashSet<Identifier> ) -> Option<Identifier>

Check if any of the provided binding names are defined as lexical bindings.

Start at the current environment. Stop at the next outer function environment.

pub(crate) fn is_next_outer_function_environment_global(&self) -> bool

Check if the next outer function environment is the global environment.

pub(crate) fn pop_to_global(&mut self) -> Vec<Environment>

Pop all current environments except the global environment.

pub(crate) fn len(&self) -> usize

Get the number of current environments.

pub(crate) fn truncate(&mut self, len: usize)

Truncate current environments to the given number.

pub(crate) fn extend(&mut self, other: Vec<Environment>)

Extend the current environment stack with the given environments.

pub(crate) fn get_this_environment(&self) -> &DeclarativeEnvironmentKind

GetThisEnvironment

Returns the environment that currently provides a this biding.

More information:

ECMAScript specification

Panics

Panics if no environment exists on the stack.

pub(crate) fn get_this_binding(&self) -> JsResult<JsValue>

GetThisBinding

Returns the current this binding of the environment.

More information:

ECMAScript specification

pub(crate) fn push_object(&mut self, object: JsObject) -> usize

Push a new object environment on the environments stack and return it’s index.

pub(crate) fn push_lexical( &mut self, compile_environment: Gc<GcRefCell<CompileTimeEnvironment>> ) -> u32

Push a lexical environment on the environments stack and return it’s index.

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_function( &mut self, compile_environment: Gc<GcRefCell<CompileTimeEnvironment>>, function_slots: FunctionSlots )

Push a function environment on the environments stack.

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_function_inherit( &mut self, compile_environment: Gc<GcRefCell<CompileTimeEnvironment>> )

Push a function environment that inherits it’s internal slots from the outer function environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn push_module( &mut self, compile_environment: Gc<GcRefCell<CompileTimeEnvironment>> )

Push a module environment on the environments stack.

Panics

Panics if no environment exists on the stack.

pub(crate) fn pop(&mut self) -> Environment

Pop environment from the environments stack.

pub(crate) fn current(&self) -> Environment

Get the most outer environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn current_compile_environment( &self ) -> Gc<GcRefCell<CompileTimeEnvironment>>

Get the compile environment for the current runtime environment.

Panics

Panics if no environment exists on the stack.

pub(crate) fn poison_until_last_function(&mut self)

Mark that there may be added bindings from the current environment to the next function environment.

pub(crate) fn put_lexical_value( &mut self, environment_index: u32, binding_index: u32, value: JsValue )

Set the value of a lexical binding.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn put_value_if_uninitialized( &mut self, environment_index: u32, binding_index: u32, value: JsValue )

Set the value of a binding if it is uninitialized.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn push_private(&mut self, environment: Gc<PrivateEnvironment>)

Push a private environment to the private environment stack.

pub(crate) fn pop_private(&mut self)

Pop a private environment from the private environment stack.

pub(crate) fn resolve_private_identifier( &self, identifier: JsString ) -> Option<PrivateName>

ResolvePrivateIdentifier ( privEnv, identifier )

More information:

ECMAScript specification

pub(crate) fn private_name_descriptions(&self) -> Vec<&JsString>

Return all private name descriptions in all private environments.