Enum boa_engine::error::TryNativeError

pub enum TryNativeError {
    InvalidPropertyType(&'static str),
    InvalidMessageEncoding,
    InvalidConstructor,
    InaccessibleProperty {
        property: &'static str,
        source: JsError,
    },
    InvalidErrorsIndex {
        index: u64,
        source: JsError,
    },
    NotAnErrorObject(JsValue),
    InaccessibleRealm {
        source: JsError,
    },
}

The error type returned by the JsError::try_native method.

Variants

InvalidPropertyType(&'static str)

A property of the error object has an invalid type.

InvalidMessageEncoding

The message of the error object could not be decoded.

InvalidConstructor

The constructor property of the error object was invalid.

InaccessibleProperty

Fields

property: &'static str

The name of the property that could not be accessed.

source: JsError

The source error.

A property of the error object is not accessible.

InvalidErrorsIndex

Fields

index: u64

The index of the error that could not be accessed.

source: JsError

The source error.

An inner error of an aggregate error is not accessible.

NotAnErrorObject(JsValue)

The error value is not an error object.

InaccessibleRealm

Fields

source: JsError

The source error.

The original realm of the error object was inaccessible.