Module boa_engine::native_function

Boa’s wrappers for native Rust functions to be compatible with ECMAScript calls.

NativeFunction is the main type of this module, providing APIs to create native callables from native Rust functions and closures.

Structs

Closure 🔒

NativeFunction

A callable Rust function that can be invoked by the engine.

Enums

Inner 🔒

Traits

TraceableClosure 🔒

Type Definitions

NativeFunctionPointer

The required signature for all native built-in function pointers.