Module boa_engine::object::builtins

Contains all the Rust representations of JavaScript objects.

Modules

jsarray 🔒

This module implements a wrapper for the Array Builtin JavaScript Object

jsarraybuffer 🔒

This module implements a wrapper for the ArrayBuffer Builtin JavaScript Object

jsdataview 🔒

This module implements a wrapper for the DataView Builtin JavaScript Object

jsdate 🔒

jsfunction 🔒

This module implements a wrapper for the Function Builtin JavaScript Object

jsgenerator 🔒

This module implements a wrapper for the Generator Builtin JavaScript Object

jsmap 🔒

This module implements a wrapper for the Map Builtin JavaScript Object

jsmap_iterator 🔒

This module implements a wrapper for the MapIterator Builtin JavaScript Object

jsproxy 🔒

This module implements a wrapper for the Proxy Builtin JavaScript Object

jsregexp 🔒

This module implements a wrapper for the RegExp Builtin JavaScript Object

jsset 🔒

This module implements a wrapper for the Set Builtin JavaScript Object

jsset_iterator 🔒

This module implements a wrapper for the SetIterator Builtin JavaScript Object

jstypedarray 🔒

This module implements a wrapper for the TypedArray Builtin JavaScript Object

Structs

JsArray

JsArray provides a wrapper for Boa’s implementation of the JavaScript Array object.

JsArrayBuffer

JsArrayBuffer provides a wrapper for Boa’s implementation of the JavaScript ArrayBuffer object

JsDataView

JsDataView Provides a wrapper for Boa’s implementation of the JavaScript DataView object

JsDate

JsDate is a wrapper for JavaScript JsDate builtin object

JsFloat32Array

JavaScript Float32Array rust object.

JsFloat64Array

JavaScript Float64Array rust object.

JsFunction

JavaScript Function rust object.

JsGenerator

JsGenerator provides a wrapper for Boa’s implementation of the JavaScript Generator builtin object

JsInt8Array

JavaScript Int8Array rust object.

JsInt16Array

JavaScript Int16Array rust object.

JsInt32Array

JavaScript Int32Array rust object.

JsMap

JsMap provides a wrapper for Boa’s implementation of the JavaScript Map object.

JsMapIterator

JavaScript MapIterator rust object

JsProxy

JavaScript Proxy rust object.

JsRegExp

JsRegExp provides a wrapper for Boa’s implementation of the JavaScript RegExp builtin object

JsRevocableProxy

JavaScript Proxy rust object that can be disabled.

JsSet

JsSet provides a wrapper for Boa’s implementation of the JavaScript Set object.

JsSetIterator

JsSetIterator provides a wrapper for Boa’s implementation of the JavaScript SetIterator object

JsTypedArray

JavaScript TypedArray rust object.

JsUint8Array

JavaScript Uint8Array rust object.

JsUint16Array

JavaScript Uint16Array rust object.

JsUint32Array

JavaScript Uint32Array rust object.