Struct boa_engine::object::builtins::jsproxy::JsProxy

pub struct JsProxy {
    inner: JsObject,
}

JavaScript Proxy rust object.

This is a wrapper type for the Proxy API that allows customizing essential behaviour for an object, like property accesses or the delete operator.

The only way to construct this type is to use the JsProxyBuilder type; also accessible from JsProxy::builder.

Fields
inner: JsObject
Implementations

impl JsProxy

pub fn builder(target: JsObject) -> JsProxyBuilder