Struct boa_engine::object::builtins::jsset::JsSet

pub struct JsSet {
    inner: JsObject,
}

JsSet provides a wrapper for Boa’s implementation of the ECMAScript Set object.

Fields
inner: JsObject
Implementations

impl JsSet

pub fn new(context: &mut Context<'_>) -> Self

Create a new empty set.

Doesn’t matches JavaScript new Set() as it doesn’t takes an iterator similar to Rust initialization.

pub fn size(&self) -> JsResult<usize>

Returns the size of the Set as an integer.

Same as JavaScript’s set.size.

pub fn add<T>(&self, value: T, context: &mut Context<'_>) -> JsResult<JsValue>where T: Into<JsValue>,

Appends value to the Set object. Returns the Set object with added value.

Same as JavaScript’s set.add(value).

pub fn add_items( &self, items: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Adds slice as a single element. Returns the Set object with added slice.

Same as JavaScript’s set.add(["one", "two", "three"])

pub fn clear(&self, context: &mut Context<'_>) -> JsResult<JsValue>

Removes all elements from the Set object. Returns Undefined.

Same as JavaScript’s set.clear().

pub fn delete<T>(&self, value: T, context: &mut Context<'_>) -> JsResult<bool>where T: Into<JsValue>,

Removes the element associated to the value. Returns a boolean asserting whether an element was successfully removed or not.

Same as JavaScript’s set.delete(value).

pub fn has<T>(&self, value: T, context: &mut Context<'_>) -> JsResult<bool>where T: Into<JsValue>,

Returns a boolean asserting whether an element is present with the given value in the Set object or not.

Same as JavaScript’s set.has(value).

pub fn values(&self, context: &mut Context<'_>) -> JsResult<JsSetIterator>

Returns a new iterator object that yields the values for each element in the Set object in insertion order.

Same as JavaScript’s set.values().

pub fn keys(&self, context: &mut Context<'_>) -> JsResult<JsSetIterator>

Alias for Set.prototype.values() Returns a new iterator object that yields the values for each element in the Set object in insertion order.

Same as JavaScript’s set.keys().

pub fn for_each( &self, callback: JsFunction, this_arg: JsValue, context: &mut Context<'_> ) -> JsResult<JsValue>

Calls callbackFn once for each value present in the Set object, in insertion order. Returns Undefined.

Same as JavaScript’s set.forEach(values).

pub fn from_object(object: JsObject) -> JsResult<Self>

Utility: Creates JsSet from JsObject, if not a Set throw TypeError.

pub fn from_iter<I>(elements: I, context: &mut Context<'_>) -> Selfwhere I: IntoIterator<Item = JsValue>,

Utility: Creates a JsSet from a <IntoIterator<Item = JsValue> convertible object.