Struct boa_engine::object::builtins::JsArrayBuffer

pub struct JsArrayBuffer { /* private fields */ }

JsArrayBuffer provides a wrapper for Boa’s implementation of the ECMAScript ArrayBuffer object

Implementations

impl JsArrayBuffer

pub fn new(byte_length: usize, context: &mut Context<'_>) -> JsResult<Self>

Create a new array buffer with byte length.

// Creates a blank array buffer of n bytes
let array_buffer = JsArrayBuffer::new(4, context)?;

assert_eq!(array_buffer.take()?, vec![0_u8; 4]);

pub fn from_byte_block( byte_block: Vec<u8>, context: &mut Context<'_> ) -> JsResult<Self>

Create a new array buffer from byte block.

This uses the passed byte block as the internal storage, it does not clone it!

The byte_length will be set to byte_block.len().

// Create a buffer from a chunk of data
let data_block: Vec<u8> = (0..5).collect();
let array_buffer = JsArrayBuffer::from_byte_block(data_block, context)?;

assert_eq!(array_buffer.take()?, (0..5).collect::<Vec<u8>>());

pub fn from_object(object: JsObject) -> JsResult<Self>

Create a JsArrayBuffer from a JsObject, if the object is not an array buffer throw a TypeError.

This does not clone the fields of the array buffer, it only does a shallow clone of the object.

pub fn byte_length(&self, context: &mut Context<'_>) -> usize

Returns the byte length of the array buffer.

// Create a buffer from a chunk of data
let data_block: Vec<u8> = (0..5).collect();
let array_buffer = JsArrayBuffer::from_byte_block(data_block, context)?;

// Take the inner buffer
let buffer_length = array_buffer.byte_length(context);

assert_eq!(buffer_length, 5);

pub fn take(&self) -> JsResult<Vec<u8>>

Take the inner ArrayBuffer’s array_buffer_data field and replace it with None

Note: This causes the pre-existing JsArrayBuffer to become detached.

// Create a buffer from a chunk of data
let data_block: Vec<u8> = (0..5).collect();
let array_buffer = JsArrayBuffer::from_byte_block(data_block, context)?;

// Take the inner buffer
let internal_buffer = array_buffer.take()?;

assert_eq!(internal_buffer, (0..5).collect::<Vec<u8>>());

// Anymore interaction with the buffer will return an error
let detached_err = array_buffer.take();
assert!(detached_err.is_err());