Module boa_engine::object::internal_methods

This module defines the object internal methods.

More information:

ECMAScript reference

Modules

immutable_prototype 🔒

string 🔒

Structs

InternalMethodContext 🔒

A lightweight wrapper around Context used in InternalObjectMethods.

InternalObjectMethods

The internal representation of the internal methods of a JsObject.

Enums

CallValue 🔒

The return value of an internal method ([[Call]] or [[Construct]]).

Statics

ORDINARY_INTERNAL_METHODS 🔒

Definitions of the internal object methods for ordinary objects.

Functions

get_prototype_from_constructor 🔒

Abstract operation GetPrototypeFromConstructor

is_compatible_property_descriptor 🔒

Abstract operation IsCompatiblePropertyDescriptor

non_existant_call 🔒

non_existant_construct 🔒

ordinary_define_own_property 🔒

Abstract operation OrdinaryDefineOwnProperty.

ordinary_delete 🔒

Abstract operation OrdinaryDelete.

ordinary_get 🔒

Abstract operation OrdinaryGet.

ordinary_get_own_property 🔒

Abstract operation OrdinaryGetOwnProperty.

ordinary_get_prototype_of 🔒

Abstract operation OrdinaryGetPrototypeOf.

ordinary_has_property 🔒

Abstract operation OrdinaryHasProperty.

ordinary_is_extensible 🔒

Abstract operation OrdinaryIsExtensible.

ordinary_own_property_keys 🔒

Abstract operation OrdinaryOwnPropertyKeys.

ordinary_prevent_extensions 🔒

Abstract operation OrdinaryPreventExtensions.

ordinary_set 🔒

Abstract operation OrdinarySet.

ordinary_set_prototype_of 🔒

Abstract operation OrdinarySetPrototypeOf.

validate_and_apply_property_descriptor 🔒

Abstract operation ValidateAndApplyPropertyDescriptor