Module boa_engine::object::internal_methods::array

Statics

ARRAY_EXOTIC_INTERNAL_METHODS 🔒

Definitions of the internal object methods for array exotic objects.

Functions

array_exotic_define_own_property 🔒

Define an own property for an array exotic object.

array_set_length 🔒

Abstract operation ArraySetLength ( A, Desc )