Function boa_engine::object::jsobject::coerce_gc

fn coerce_gc<T: NativeObject>(
    ptr: Gc<VTableObject<T>>
) -> Gc<VTableObject<dyn NativeObject>>

Upcasts the reference to an object from a specific type T to an erased type dyn NativeObject.