Struct boa_engine::object::jsobject::RecursionLimiter

pub struct RecursionLimiter {
    top_level: bool,
    ptr: usize,
    pub visited: bool,
    pub live: bool,
}

Prevents infinite recursion during Debug::fmt, JSON.stringify, and other conversions. This uses a thread local, so is not safe to use where the object graph will be traversed by multiple threads!

Fields
top_level: bool

If this was the first JsObject in the tree.

ptr: usize

The ptr being kept in the HashSet, so we can delete it when we drop.

visited: bool

If this JsObject has been visited before in the graph, but not in the current branch.

live: bool

If this JsObject has been visited in the current branch of the graph.

Implementations

impl RecursionLimiter

pub fn new(o: &JsObject) -> Self

Determines if the specified JsObject has been visited, and returns a struct that will free it when dropped.

This is done by maintaining a thread-local hashset containing the pointers of JsObject values that have been visited. The first JsObject visited will clear the hashset, while any others will check if they are contained by the hashset.