Module boa_engine::object::property_map

Structs

Iter

An iterator over the property entries of an Object

Keys

An iterator over the keys (PropertyKey) of an Object.

OrderedHashMap 🔒

Wrapper around indexmap::IndexMap for usage in PropertyMap.

PropertyMap

A PropertyMap contains all the properties of an object.

StringProperties

An iterator over the String property entries of an Object

StringPropertyKeys

An iterator over the string keys (RcString) of an Object.

StringPropertyValues

An iterator over the string values (Property) of an Object.

SymbolProperties

An iterator over the Symbol property entries of an Object

SymbolPropertyKeys

An iterator over the keys (RcSymbol) of an Object.

SymbolPropertyValues

An iterator over the Symbol values (Property) of an Object.

Values

An iterator over the values (Property) of an Object.

Enums

IndexProperties

An iterator over the indexed property entries of an Object.

IndexPropertyKeys

An iterator over the index keys (u32) of an Object.

IndexPropertyValues

An iterator over the index values (Property) of an Object.

IndexedProperties 🔒

This represents all the indexed properties.

Type Definitions

GlobalPropertyMap 🔒

Type alias to make it easier to work with the string properties on the global object.