Enum boa_engine::object::property_map::IndexedProperties

enum IndexedProperties {
    Dense(Vec<JsValue>),
    Sparse(FxHashMap<u32, PropertyDescriptor>),
}

This represents all the indexed properties.

The index properties can be stored in two storage methods:

Dense Storage

Sparse Storage

By default it is dense storage.

Variants

Dense(Vec<JsValue>)

Dense storage holds a contiguous array of properties where the index in the array is the key of the property. These are known to be data descriptors with a value field, writable field set to true, configurable field set to true, enumerable field set to true.

Since we know the properties of the property descriptors (and they are all the same) we can omit it and just store only the value field and construct the data property descriptor on demand.

This storage method is used by default.

Sparse(FxHashMap<u32, PropertyDescriptor>)

Sparse storage this storage is used as a backup if the element keys are not continuous or the property descriptors are not data descriptors with with a value field, writable field set to true, configurable field set to true, enumerable field set to true.

This method uses more space, since we also have to store the property descriptors, not just the value. It is also slower because we need to to a hash lookup.

Implementations

impl IndexedProperties

fn get(&self, key: u32) -> Option<PropertyDescriptor>

Get a property descriptor if it exists.

fn convert_dense_to_sparse(
vec: &mut Vec<JsValue>
) -> FxHashMap<u32, PropertyDescriptor>

Helper function for converting from a dense storage type to sparse storage type.

fn insert(
&mut self,
key: u32,
property: PropertyDescriptor
) -> Option<PropertyDescriptor>

Inserts a property descriptor with the specified key.

fn remove(&mut self, key: u32) -> Option<PropertyDescriptor>

Inserts a property descriptor with the specified key.

fn contains_key(&self, key: u32) -> bool

Check if we contain the key to a property descriptor.

fn iter(&self) -> IndexProperties<'_>

fn keys(&self) -> IndexPropertyKeys<'_>

fn values(&self) -> IndexPropertyValues<'_>