Struct boa_engine::object::shape::SharedShape

pub struct SharedShape {
    inner: Gc<Inner>,
}

Represents a shared object shape.

Fields
inner: Gc<Inner>
Implementations

impl SharedShape

fn property_table(&self) -> &PropertyTable

fn property_count(&self) -> u32

Return the property count that this shape owns in the PropertyTable.

fn property_index(&self) -> u32

Return the index to the property in the the PropertyTable.

pub fn transition_count(&self) -> u16

Getter for the transition count field.

pub fn previous(&self) -> Option<&Self>

Getter for the previous field.

pub fn prototype(&self) -> JsPrototype

Get the prototype of the shape.

pub(crate) fn property(&self) -> (PropertyKey, Slot)

Get the property this SharedShape refers to.

fn flags(&self) -> ShapeFlags

Get the flags of the shape.

fn forward_transitions(&self) -> &ForwardTransition

Getter for the ForwardTransition field.

pub fn has_prototype(&self, prototype: &JsObject) -> bool

Check if the shape has the given prototype.

fn new(inner: Inner) -> Self

Create a new SharedShape.

pub(crate) fn root() -> Self

Create a root SharedShape.

pub(crate) fn change_prototype_transition(&self, prototype: JsPrototype) -> Self

Create a SharedShape change prototype transition.

pub(crate) fn insert_property_transition(&self, key: TransitionKey) -> Self

Create a SharedShape insert property transition.

pub(crate) fn change_attributes_transition( &self, key: TransitionKey ) -> ChangeTransition<Self>

Create a SharedShape change prototype transition, returning ChangeTransition.

fn rollback_before( &self, key: &PropertyKey ) -> (Self, Option<JsPrototype>, IndexMap<PropertyKey, SlotAttributes>)

Rollback to shape before the insertion of the PropertyKey that is provided.

This returns the shape before the insertion, if it sees a prototype transition it will return the lastest one, ignoring any others, None otherwise. It also will return the property transitions ordered from latest to oldest that it sees.

NOTE: In the transitions it does not include the property that we are rolling back.

NOTE: The prototype transitions if it sees a property insert and then later an attribute change it will condense into one property insert transition with the new attribute in the change attribute transition, in the same place that the property was inserted initially.

pub(crate) fn remove_property_transition(&self, key: &PropertyKey) -> Self

Remove a property from SharedShape, returning the new SharedShape.

pub(crate) fn lookup(&self, key: &PropertyKey) -> Option<Slot>

Do a property lookup, returns None if property not found.

pub(crate) fn keys(&self) -> Vec<PropertyKey>

Gets all keys first strings then symbols in creation order.

pub(crate) fn to_unique(&self) -> UniqueShape

Returns a new UniqueShape with the properties of the SharedShape.

pub(crate) fn to_addr_usize(&self) -> usize

Return location in memory of the SharedShape.