Struct boa_engine::object::JsObject

pub struct JsObject { /* private fields */ }

Garbage collected Object.

Implementations

impl JsObject

pub fn with_object_proto(intrinsics: &Intrinsics) -> Self

Creates a new ordinary object with its prototype set to the Object prototype.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate(%Object.prototype%).

pub fn with_null_proto() -> Self

Creates a new ordinary object, with its prototype set to null.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate(null).

pub fn from_proto_and_data<O: Into<Option<Self>>>( prototype: O, data: ObjectData ) -> Self

Creates a new object with the provided prototype and object data.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate, with the difference that the additionalInternalSlotsList parameter is automatically set by the ObjectData provided.

pub fn borrow(&self) -> Ref<'_, Object>

Immutably borrows the Object.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the object is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'_, Object, Object>

Mutably borrows the Object.

The borrow lasts until the returned RefMut exits scope. The object cannot be borrowed while this borrow is active.

Panics

Panics if the object is currently borrowed.

pub fn try_borrow(&self) -> StdResult<Ref<'_, Object>, BorrowError>

Immutably borrows the Object, returning an error if the value is currently mutably borrowed.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

This is the non-panicking variant of borrow.

pub fn try_borrow_mut( &self ) -> StdResult<RefMut<'_, Object, Object>, BorrowMutError>

Mutably borrows the object, returning an error if the value is currently borrowed.

The borrow lasts until the returned GcCellRefMut exits scope. The object be borrowed while this borrow is active.

This is the non-panicking variant of borrow_mut.

pub fn equals(lhs: &Self, rhs: &Self) -> bool

Checks if the garbage collected memory is the same.

pub fn is<T>(&self) -> bool
where T: NativeObject,

Return true if it is a native object and the native type is T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_ref<T>(&self) -> Option<Ref<'_, T>>
where T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_mut<T>(&self) -> Option<RefMut<'_, Object, T>>
where T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently borrowed.

pub fn prototype(&self) -> JsPrototype

Get the prototype of the object.

Panics

Panics if the object is currently mutably borrowed.

pub fn set_prototype(&self, prototype: JsPrototype) -> bool

Set the prototype of the object.

Panics

Panics if the object is currently mutably borrowed

pub fn is_array(&self) -> bool

Checks if it’s an Array object.

pub fn is_data_view(&self) -> bool

Checks if it’s a DataView object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_array_iterator(&self) -> bool

Checks if it is an ArrayIterator object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_array_buffer(&self) -> bool

Checks if it’s an ArrayBuffer object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_map(&self) -> bool

Checks if it is a Map object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_map_iterator(&self) -> bool

Checks if it’s a MapIterator object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_set(&self) -> bool

Checks if it is a Set object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_set_iterator(&self) -> bool

Checks if it is a SetIterator object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_string(&self) -> bool

Checks if it’s a String object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_function(&self) -> bool

Checks if it’s a Function object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_generator(&self) -> bool

Checks if it’s a Generator object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_symbol(&self) -> bool

Checks if it’s a Symbol object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_error(&self) -> bool

Checks if it’s an Error object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_boolean(&self) -> bool

Checks if it’s a Boolean object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_number(&self) -> bool

Checks if it’s a Number object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_bigint(&self) -> bool

Checks if it’s a BigInt object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_date(&self) -> bool

Checks if it’s a Date object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_regexp(&self) -> bool

Checks if it’s a RegExp object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_array(&self) -> bool

Checks if it’s a TypedArray object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint8_array(&self) -> bool

Checks if it’s a Uint8Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int8_array(&self) -> bool

Checks if it’s a Int8Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint16_array(&self) -> bool

Checks if it’s a Uint16Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int16_array(&self) -> bool

Checks if it’s a Int16Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint32_array(&self) -> bool

Checks if it’s a Uint32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int32_array(&self) -> bool

Checks if it’s a Int32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_float32_array(&self) -> bool

Checks if it’s a Float32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_float64_array(&self) -> bool

Checks if it’s a Float64Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_promise(&self) -> bool

Checks if it’s a Promise object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_ordinary(&self) -> bool

Checks if it’s an ordinary object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_proxy(&self) -> bool

Checks if it’s a proxy object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

Panics

Panics if the object is currently mutably borrowed.

pub fn to_property_descriptor( &self, context: &mut Context<'_> ) -> JsResult<PropertyDescriptor>

The abstract operation ToPropertyDescriptor.

More information:

ECMAScript reference

pub fn copy_data_properties<K>( &self, source: &JsValue, excluded_keys: Vec<K>, context: &mut Context<'_> ) -> JsResult<()>
where K: Into<PropertyKey>,

7.3.25 CopyDataProperties ( target, source, excludedItems )

More information:

ECMAScript

pub fn insert_property<K, P>(&self, key: K, property: P) -> bool
where K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name, than true is returned with that field, otherwise false is returned.

pub fn is_callable(&self) -> bool

It determines if Object is a callable function with a [[Call]] internal method.

More information:

ECMAScript reference

pub fn is_constructor(&self) -> bool

It determines if Object is a function object with a [[Construct]] internal method.

More information:

ECMAScript reference

pub fn is_global(&self) -> bool

Returns true if the JsObject is the global for a Realm

impl JsObject

pub fn is_extensible(&self, context: &mut Context<'_>) -> JsResult<bool>

Check if object is extensible.

More information:

ECMAScript reference

pub fn get<K>(&self, key: K, context: &mut Context<'_>) -> JsResult<JsValue>
where K: Into<PropertyKey>,

Get property from object or throw.

More information:

ECMAScript reference

pub fn set<K, V>( &self, key: K, value: V, throw: bool, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>, V: Into<JsValue>,

set property of object or throw if bool flag is passed.

More information:

ECMAScript reference

pub fn create_data_property<K, V>( &self, key: K, value: V, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>, V: Into<JsValue>,

Create data property

More information:

ECMAScript reference

pub fn create_data_property_or_throw<K, V>( &self, key: K, value: V, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>, V: Into<JsValue>,

Create data property or throw

More information:

ECMAScript reference

pub fn define_property_or_throw<K, P>( &self, key: K, desc: P, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Define property or throw.

More information:

ECMAScript reference

pub fn delete_property_or_throw<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>,

Defines the property or throws a TypeError if the operation fails.

More information:

ECMAScript reference

pub fn has_property<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>,

Check if object has property.

More information:

ECMAScript reference

pub fn has_own_property<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>
where K: Into<PropertyKey>,

Check if object has an own property.

More information:

ECMAScript reference

pub fn call( &self, this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Call ( F, V [ , argumentsList ] )

Panics

Panics if the object is currently mutably borrowed.

More information:

ECMAScript reference

pub fn construct( &self, args: &[JsValue], new_target: Option<&Self>, context: &mut Context<'_> ) -> JsResult<Self>

Construct ( F [ , argumentsList [ , newTarget ] ] )

Construct an instance of this object with the specified arguments.

Panics

Panics if the object is currently mutably borrowed.

More information:

ECMAScript reference

pub fn set_integrity_level( &self, level: IntegrityLevel, context: &mut Context<'_> ) -> JsResult<bool>

Make the object sealed or frozen.

More information:

ECMAScript reference

pub fn test_integrity_level( &self, level: IntegrityLevel, context: &mut Context<'_> ) -> JsResult<bool>

Check if the object is sealed or frozen.

More information:

ECMAScript reference