Struct boa_engine::object::JsRevocableProxy

pub struct JsRevocableProxy {
    proxy: JsProxy,
    revoker: JsFunction,
}

JavaScript Proxy rust object that can be disabled.

Safe interface for the Proxy.revocable method that creates a proxy that can be disabled using the JsRevocableProxy::revoke method. The internal proxy is accessible using the Deref operator.

The only way to construct this type is to use the JsProxyBuilder type; also accessible from JsProxy::builder; with the JsProxyBuilder::build_revocable method.

Fields
proxy: JsProxyrevoker: JsFunction
Implementations

impl JsRevocableProxy

pub fn revoke(self, context: &mut Context) -> JsResult<()>

Disables the traps of the internal proxy object, essentially making it unusable and throwing TypeErrors for all the traps.