Struct boa_engine::object::Object

pub struct Object<T: ?Sized> {
    pub(crate) properties: PropertyMap,
    pub(crate) extensible: bool,
    private_elements: ThinVec<(PrivateName, PrivateElement)>,
    pub(crate) data: T,
}

The internal representation of a JavaScript object.

Fields
properties: PropertyMap

The collection of properties contained in the object

extensible: bool

Whether it can have new properties added to it.

private_elements: ThinVec<(PrivateName, PrivateElement)>

The [[PrivateElements]] internal slot.

data: T

The inner object data

Implementations

impl<T: ?Sized> Object<T>

pub const fn shape(&self) -> &Shape

Returns the shape of the object.

pub const fn data(&self) -> &T

Returns the data of the object.

pub fn prototype(&self) -> JsPrototype

Gets the prototype instance of this object.

pub fn set_prototype<O: Into<JsPrototype>>(&mut self, prototype: O) -> bool

Sets the prototype instance of the object.

More information

pub const fn properties(&self) -> &PropertyMap

Returns the properties of the object.

pub(crate) fn properties_mut(&mut self) -> &mut PropertyMap

pub(crate) fn insert<K, P>(&mut self, key: K, property: P) -> boolwhere K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name, then true is returned otherwise, false is returned.

pub(crate) fn remove(&mut self, key: &PropertyKey) -> bool

Helper function for property removal without checking if it’s configurable.

Returns true if the property was removed, false otherwise.

pub(crate) fn append_private_element( &mut self, name: PrivateName, element: PrivateElement )

Append a private element to an object.

impl Object<dyn NativeObject>

pub fn is<T: NativeObject>(&self) -> bool

Return true if it is a native object and the native type is T.

pub fn downcast_ref<T: NativeObject>(&self) -> Option<&T>

Downcast a reference to the object, if the object is type native object type T.

pub fn downcast_mut<T: NativeObject>(&mut self) -> Option<&mut T>

Downcast a mutable reference to the object, if the object is type native object type T.

pub(crate) fn is_arguments(&self) -> bool

Checks if this object is an Arguments object.

pub fn is_typed_uint8_array(&self) -> bool

Checks if it a Uint8Array object.

pub fn is_typed_int8_array(&self) -> bool

Checks if it a Int8Array object.

pub fn is_typed_uint16_array(&self) -> bool

Checks if it a Uint16Array object.

pub fn is_typed_int16_array(&self) -> bool

Checks if it a Int16Array object.

pub fn is_typed_uint32_array(&self) -> bool

Checks if it a Uint32Array object.

pub fn is_typed_int32_array(&self) -> bool

Checks if it a Int32Array object.

pub fn is_typed_float32_array(&self) -> bool

Checks if it a Float32Array object.

pub fn is_typed_float64_array(&self) -> bool

Checks if it a Float64Array object.