Struct boa_engine::prelude::Source

pub struct Source<'path, R> {
    pub(crate) reader: R,
    pub(crate) path: Option<&'path Path>,
}

A source of ECMAScript code.

Sources can be created from plain strs, file Paths or more generally, any Read instance.

Fields
reader: Rpath: Option<&'path Path>
Implementations

impl<'bytes> Source<'static, &'bytes [u8]>

pub fn from_bytes<T>(source: &'bytes T) -> Source<'static, &'bytes [u8]>where T: AsRef<[u8]> + ?Sized,

Creates a new Source from any type equivalent to a slice of bytes e.g. &str, Vec<u8>, Box<[u8]> or a plain slice &[u8].

Examples

let code = r#"var array = [5, 4, 3, 2, 1];"#;
let source = Source::from_bytes(code);

impl<'path> Source<'path, BufReader<File>>

pub fn from_filepath( source: &'path Path ) -> Result<Source<'path, BufReader<File>>, Error>

Creates a new Source from a Path to a file.

Errors

See File::open.

Examples

let path = Path::new("script.js");
let source = Source::from_filepath(path)?;

impl<'path, R> Source<'path, R>where R: Read,

pub const fn from_reader( reader: R, path: Option<&'path Path> ) -> Source<'path, R>

Creates a new Source from a Read instance and an optional Path.

Examples

let strictler = r#""use strict""#;

let path = Path::new("no_strict.js");
let file = File::open(path)?;
let strict = strictler.as_bytes().chain(file);

let source = Source::from_reader(strict, Some(path));