Module boa_engine::property

Boa’s implementation of ECMAScript’s Property Descriptor.

The Property Descriptor type is used to explain the manipulation and reification of Object property attributes. Values of the Property Descriptor type are Records. Each field’s name is an attribute name and its value is a corresponding attribute value as specified in 6.1.7.1. In addition, any field may be present or absent. The schema name used within this specification to tag literal descriptions of Property Descriptor records is PropertyDescriptor.

More information:

MDN documentation

ECMAScript reference

Structs

Attribute

This struct constains the property flags as described in the ECMAScript specification.

PropertyDescriptor

This represents an ECMAScript Property AKA The Property Descriptor.

PropertyDescriptorBuilder

A builder for PropertyDescriptor.

Enums

DescriptorKind

DescriptorKind represents the different kinds of property descriptors.

PropertyKey

This abstracts away the need for IsPropertyKey by transforming the PropertyKey values into an enum with both valid types: String and Symbol