Struct boa_engine::property::Attribute

pub struct Attribute(<Attribute as PublicFlags>::Internal);

This struct constains the property flags as described in the ECMAScript specification.

It contains the following flags:

[[Writable]] (WRITABLE) - If false, attempts by ECMAScript code to change the property’s [[Value]] attribute using [[Set]] will not succeed.

[[Enumerable]] (ENUMERABLE) - If the property will be enumerated by a for-in enumeration.

[[Configurable]] (CONFIGURABLE) - If false, attempts to delete the property, change the property to be an accessor property, or change its attributes (other than [[Value]], or changing [[Writable]] to false) will fail.

Tuple Fields
0: <Attribute as PublicFlags>::Internal
Implementations

impl Attribute

pub const WRITABLE: Self = _

The Writable attribute decides whether the value associated with the property can be changed or not, from its initial value.

pub const ENUMERABLE: Self = _

If the property can be enumerated by a for-in loop.

pub const CONFIGURABLE: Self = _

If the property descriptor can be changed later.

pub const READONLY: Self = _

The property is not writable.

pub const NON_ENUMERABLE: Self = _

The property can not be enumerated in a for-in loop.

pub const PERMANENT: Self = _

The property descriptor cannot be changed.

impl Attribute

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u8

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u8) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u8) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u8) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesn’t correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove won’t truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference won’t truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl Attribute

pub const fn iter(&self) -> Iter<Attribute>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<Attribute>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl Attribute

pub fn clear(&mut self)

Clear all flags.

pub fn set_writable(&mut self, value: bool)

Sets the writable flag.

pub const fn writable(self) -> bool

Gets the writable flag.

pub fn set_enumerable(&mut self, value: bool)

Sets the enumerable flag.

pub const fn enumerable(self) -> bool

Gets the enumerable flag.

pub fn set_configurable(&mut self, value: bool)

Sets the configurable flag.

pub const fn configurable(self) -> bool

Gets the configurable flag.