Struct boa_engine::string::common::StaticJsStrings

pub(crate) struct StaticJsStrings;

List of commonly used strings in Javascript code.

Any strings defined here are used as a static JsString instead of allocating on the heap.

Implementations

impl StaticJsStrings

const fn find_index(candidate: &[u16]) -> usize

pub(crate) fn get_string(string: &[u16]) -> Option<JsString>

Gets the JsString corresponding to string, or None if the string doesn’t exist inside the static array.

pub(crate) fn get(index: usize) -> Option<&'static [u16]>

Gets the &[u16] slice corresponding to the provided index, or None if the index provided exceeds the size of the static array.

pub(crate) const EMPTY_STRING: JsString = _

Gets the static JsString for "".

pub(crate) const SYMBOL_ASYNC_ITERATOR: JsString = _

Gets the static JsString for "Symbol.asyncIterator".

pub(crate) const SYMBOL_HAS_INSTANCE: JsString = _

Gets the static JsString for "Symbol.hasInstance".

pub(crate) const SYMBOL_IS_CONCAT_SPREADABLE: JsString = _

Gets the static JsString for "Symbol.isConcatSpreadable".

pub(crate) const SYMBOL_ITERATOR: JsString = _

Gets the static JsString for "Symbol.iterator".

pub(crate) const SYMBOL_MATCH: JsString = _

Gets the static JsString for "Symbol.match".

pub(crate) const SYMBOL_MATCH_ALL: JsString = _

Gets the static JsString for "Symbol.matchAll".

pub(crate) const SYMBOL_REPLACE: JsString = _

Gets the static JsString for "Symbol.replace".

pub(crate) const SYMBOL_SEARCH: JsString = _

Gets the static JsString for "Symbol.search".

pub(crate) const SYMBOL_SPECIES: JsString = _

Gets the static JsString for "Symbol.species".

pub(crate) const SYMBOL_SPLIT: JsString = _

Gets the static JsString for "Symbol.split".

pub(crate) const SYMBOL_TO_PRIMITIVE: JsString = _

Gets the static JsString for "Symbol.toPrimitive".

pub(crate) const SYMBOL_TO_STRING_TAG: JsString = _

Gets the static JsString for "Symbol.toStringTag".

pub(crate) const SYMBOL_UNSCOPABLES: JsString = _

Gets the static JsString for "Symbol.unscopables".

pub(crate) const FN_SYMBOL_ASYNC_ITERATOR: JsString = _

Gets the static JsString for "[Symbol.asyncIterator]".

pub(crate) const FN_SYMBOL_HAS_INSTANCE: JsString = _

Gets the static JsString for "[Symbol.hasInstance]".

pub(crate) const FN_SYMBOL_IS_CONCAT_SPREADABLE: JsString = _

Gets the static JsString for "[Symbol.isConcatSpreadable]".

pub(crate) const FN_SYMBOL_ITERATOR: JsString = _

Gets the static JsString for "[Symbol.iterator]".

pub(crate) const FN_SYMBOL_MATCH: JsString = _

Gets the static JsString for "[Symbol.match]".

pub(crate) const FN_SYMBOL_MATCH_ALL: JsString = _

Gets the static JsString for "[Symbol.matchAll]".

pub(crate) const FN_SYMBOL_REPLACE: JsString = _

Gets the static JsString for "[Symbol.replace]".

pub(crate) const FN_SYMBOL_SEARCH: JsString = _

Gets the static JsString for "[Symbol.search]".

pub(crate) const FN_SYMBOL_SPECIES: JsString = _

Gets the static JsString for "[Symbol.species]".

pub(crate) const FN_SYMBOL_SPLIT: JsString = _

Gets the static JsString for "[Symbol.split]".

pub(crate) const FN_SYMBOL_TO_PRIMITIVE: JsString = _

Gets the static JsString for "[Symbol.toPrimitive]".

pub(crate) const FN_SYMBOL_TO_STRING_TAG: JsString = _

Gets the static JsString for "[Symbol.toStringTag]".

pub(crate) const FN_SYMBOL_UNSCOPABLES: JsString = _

Gets the static JsString for "[Symbol.unscopables]".

pub(crate) const ARRAY: JsString = _

Gets the static JsString for "Array".

pub(crate) const ARRAY_BUFFER: JsString = _

Gets the static JsString for "ArrayBuffer".

pub(crate) const ASYNC_FUNCTION: JsString = _

Gets the static JsString for "AsyncFunction".

pub(crate) const ASYNC_GENERATOR: JsString = _

Gets the static JsString for "AsyncGenerator".

pub(crate) const ASYNC_GENERATOR_FUNCTION: JsString = _

Gets the static JsString for "AsyncGeneratorFunction".

pub(crate) const BIG_INT: JsString = _

Gets the static JsString for "BigInt".

pub(crate) const BOOLEAN: JsString = _

Gets the static JsString for "Boolean".

pub(crate) const DATA_VIEW: JsString = _

Gets the static JsString for "DataView".

pub(crate) const DATE: JsString = _

Gets the static JsString for "Date".

pub(crate) const ERROR: JsString = _

Gets the static JsString for "Error".

pub(crate) const AGGREGATE_ERROR: JsString = _

Gets the static JsString for "AggregateError".

pub(crate) const EVAL_ERROR: JsString = _

Gets the static JsString for "EvalError".

pub(crate) const RANGE_ERROR: JsString = _

Gets the static JsString for "RangeError".

pub(crate) const REFERENCE_ERROR: JsString = _

Gets the static JsString for "ReferenceError".

pub(crate) const SYNTAX_ERROR: JsString = _

Gets the static JsString for "SyntaxError".

pub(crate) const TYPE_ERROR: JsString = _

Gets the static JsString for "TypeError".

pub(crate) const URI_ERROR: JsString = _

Gets the static JsString for "URIError".

pub(crate) const ESCAPE: JsString = _

Gets the static JsString for "escape".

pub(crate) const UNESCAPE: JsString = _

Gets the static JsString for "unescape".

pub(crate) const EVAL: JsString = _

Gets the static JsString for "eval".

pub(crate) const FUNCTION: JsString = _

Gets the static JsString for "Function".

pub(crate) const GENERATOR: JsString = _

Gets the static JsString for "Generator".

pub(crate) const GENERATOR_FUNCTION: JsString = _

Gets the static JsString for "GeneratorFunction".

pub(crate) const INTL: JsString = _

Gets the static JsString for "Intl".

pub(crate) const COLLATOR: JsString = _

Gets the static JsString for "Collator".

pub(crate) const LIST_FORMAT: JsString = _

Gets the static JsString for "ListFormat".

pub(crate) const LOCALE: JsString = _

Gets the static JsString for "Locale".

pub(crate) const PLURAL_RULES: JsString = _

Gets the static JsString for "PluralRules".

pub(crate) const SEGMENTER: JsString = _

Gets the static JsString for "Segmenter".

pub(crate) const DATE_TIME_FORMAT: JsString = _

Gets the static JsString for "DateTimeFormat".

pub(crate) const JSON: JsString = _

Gets the static JsString for "JSON".

pub(crate) const MAP: JsString = _

Gets the static JsString for "Map".

pub(crate) const MATH: JsString = _

Gets the static JsString for "Math".

pub(crate) const NUMBER: JsString = _

Gets the static JsString for "Number".

pub(crate) const IS_FINITE: JsString = _

Gets the static JsString for "isFinite".

pub(crate) const IS_NAN: JsString = _

Gets the static JsString for "isNaN".

pub(crate) const PARSE_INT: JsString = _

Gets the static JsString for "parseInt".

pub(crate) const PARSE_FLOAT: JsString = _

Gets the static JsString for "parseFloat".

pub(crate) const OBJECT: JsString = _

Gets the static JsString for "Object".

pub(crate) const PROMISE: JsString = _

Gets the static JsString for "Promise".

pub(crate) const PROXY: JsString = _

Gets the static JsString for "Proxy".

pub(crate) const REFLECT: JsString = _

Gets the static JsString for "Reflect".

pub(crate) const REG_EXP: JsString = _

Gets the static JsString for "RegExp".

pub(crate) const SET: JsString = _

Gets the static JsString for "Set".

pub(crate) const STRING: JsString = _

Gets the static JsString for "String".

pub(crate) const SYMBOL: JsString = _

Gets the static JsString for "Symbol".

pub(crate) const TYPED_ARRAY: JsString = _

Gets the static JsString for "TypedArray".

pub(crate) const INT8_ARRAY: JsString = _

Gets the static JsString for "Int8Array".

pub(crate) const UINT8_ARRAY: JsString = _

Gets the static JsString for "Uint8Array".

pub(crate) const UINT8_CLAMPED_ARRAY: JsString = _

Gets the static JsString for "Uint8ClampedArray".

pub(crate) const INT16_ARRAY: JsString = _

Gets the static JsString for "Int16Array".

pub(crate) const UINT16_ARRAY: JsString = _

Gets the static JsString for "Uint16Array".

pub(crate) const INT32_ARRAY: JsString = _

Gets the static JsString for "Int32Array".

pub(crate) const UINT32_ARRAY: JsString = _

Gets the static JsString for "Uint32Array".

pub(crate) const BIG_INT64_ARRAY: JsString = _

Gets the static JsString for "BigInt64Array".

pub(crate) const BIG_UINT64_ARRAY: JsString = _

Gets the static JsString for "BigUint64Array".

pub(crate) const FLOAT32_ARRAY: JsString = _

Gets the static JsString for "Float32Array".

pub(crate) const FLOAT64_ARRAY: JsString = _

Gets the static JsString for "Float64Array".

pub(crate) const ENCODE_URI: JsString = _

Gets the static JsString for "encodeURI".

pub(crate) const ENCODE_URI_COMPONENT: JsString = _

Gets the static JsString for "encodeURIComponent".

pub(crate) const DECODE_URI: JsString = _

Gets the static JsString for "decodeURI".

pub(crate) const DECODE_URI_COMPONENT: JsString = _

Gets the static JsString for "decodeURIComponent".

pub(crate) const WEAK_REF: JsString = _

Gets the static JsString for "WeakRef".

pub(crate) const WEAK_MAP: JsString = _

Gets the static JsString for "WeakMap".

pub(crate) const WEAK_SET: JsString = _

Gets the static JsString for "WeakSet".