Struct boa_engine::JsBigInt

pub struct JsBigInt {
    inner: Rc<RawBigInt>,
}

JavaScript bigint primitive rust type.

Fields
inner: Rc<RawBigInt>
Implementations

impl JsBigInt

pub fn new<T: Into<Self>>(value: T) -> Self

Create a new JsBigInt.

pub fn zero() -> Self

Create a JsBigInt with value 0.

pub fn is_zero(&self) -> bool

Check if is zero.

pub fn one() -> Self

Create a JsBigInt with value 1.

pub fn is_one(&self) -> bool

Check if is one.

pub fn to_string_radix(&self, radix: u32) -> String

Convert bigint to string with radix.

pub fn to_f64(&self) -> f64

Converts the BigInt to a f64 type.

Returns f64::INFINITY if the BigInt is too big.

pub fn from_string_radix(buf: &str, radix: u32) -> Option<Self>

Converts a string to a BigInt with the specified radix.

pub fn from_string(string: &str) -> Option<Self>

This function takes a string and converts it to BigInt type.

More information:

ECMAScript reference

pub fn same_value_zero(x: &Self, y: &Self) -> bool

Checks for SameValueZero equality.

More information:

ECMAScript reference

pub fn same_value(x: &Self, y: &Self) -> bool

Checks for SameValue equality.

More information:

ECMAScript reference

pub fn equal(x: &Self, y: &Self) -> bool

Checks for mathematical equality.

The abstract operation BigInt::equal takes arguments x (a BigInt) and y (a BigInt). It returns true if x and y have the same mathematical integer value and false otherwise.

More information:

ECMAScript reference

pub fn pow(x: &Self, y: &Self) -> JsResult<Self>

Returns x to the power y.

pub fn shift_right(x: &Self, y: &Self) -> JsResult<Self>

Performs the >> operation.

pub fn shift_left(x: &Self, y: &Self) -> JsResult<Self>

Performs the << operation.

pub fn mod_floor(x: &Self, y: &Self) -> Self

Floored integer modulo.

Examples

assert_eq!((8).mod_floor(&3), 2);
assert_eq!((8).mod_floor(&-3), -1);

pub fn add(x: &Self, y: &Self) -> Self

Performs the + operation.

pub fn sub(x: &Self, y: &Self) -> Self

Performs the - operation.

pub fn mul(x: &Self, y: &Self) -> Self

Performs the * operation.

pub fn div(x: &Self, y: &Self) -> Self

Performs the / operation.

pub fn rem(x: &Self, y: &Self) -> Self

Performs the % operation.

pub fn bitand(x: &Self, y: &Self) -> Self

Performs the & operation.

pub fn bitor(x: &Self, y: &Self) -> Self

Performs the | operation.

pub fn bitxor(x: &Self, y: &Self) -> Self

Performs the ^ operation.

pub fn neg(x: &Self) -> Self

Performs the unary - operation.

pub fn not(x: &Self) -> Self

Performs the unary ! operation.

pub(crate) fn as_inner(&self) -> &RawBigInt