Struct boa_engine::JsString

pub struct JsString {
    ptr: TaggedJsString,
}

A UTF-16–encoded, reference counted, immutable string.

This is pretty similar to a Rc<[u16]>, but without the length metadata associated with the Rc fat pointer. Instead, the length of every string is stored on the heap, along with its reference counter and its data.

We define some commonly used string constants in an interner. For these strings, we don’t allocate memory on the heap to reduce the overhead of memory allocation and reference counting.

Deref

JsString implements Deref<Target = [u16]>, inheriting all of [u16]’s methods.

Fields
ptr: TaggedJsString
Implementations

impl JsString

pub fn as_slice(&self) -> &[u16]

Obtains the underlying &[u16] slice of a JsString

pub fn concat(x: &[u16], y: &[u16]) -> Self

Creates a new JsString from the concatenation of x and y.

pub fn concat_array(strings: &[&[u16]]) -> Self

Creates a new JsString from the concatenation of every element of strings.

pub fn to_std_string_escaped(&self) -> String

Decodes a JsString into a String, replacing invalid data with its escaped representation in 4 digit hexadecimal.

pub fn to_std_string(&self) -> Result<String, FromUtf16Error>

Decodes a JsString into a String, returning FromUtf16Error if it contains any invalid data.

pub fn code_points(&self) -> impl Iterator<Item = CodePoint> + '_

Gets an iterator of all the Unicode codepoints of a JsString.

pub(crate) fn index_of(
&self,
search_value: &[u16],
from_index: usize
) -> Option<usize>

Abstract operation StringIndexOf ( string, searchValue, fromIndex )

Note: Instead of returning an isize with -1 as the “not found” value, we make use of the type system and return Option<usize> with None as the “not found” value.

More information:

ECMAScript reference

pub(crate) fn code_point_at(&self, position: usize) -> CodePoint

Abstract operation CodePointAt( string, position ).

The abstract operation CodePointAt takes arguments string (a String) and position (a non-negative integer) and returns a Record with fields [[CodePoint]] (a code point), [[CodeUnitCount]] (a positive integer), and [[IsUnpairedSurrogate]] (a Boolean). It interprets string as a sequence of UTF-16 encoded code points, as described in 6.1.4, and reads from it a single code point starting with the code unit at index position.

More information:

ECMAScript reference

pub(crate) fn to_number(&self) -> f64

Abstract operation StringToNumber ( str )

More information:

ECMAScript reference

pub(crate) fn to_big_int(&self) -> Option<JsBigInt>

Abstract operation StringToBigInt ( str )

More information:

ECMAScript reference

fn ptr(&self) -> JsStringPtrKind

Returns the inner pointer data, unwrapping its tagged data if the pointer contains a static index for COMMON_STRINGS.

fn allocate_inner(str_len: usize) -> NonNull<RawJsString>

Allocates a new RawJsString with an internal capacity of str_len chars.

Panics

Panics if try_allocate_inner returns Err.

fn try_allocate_inner(
str_len: usize
) -> Result<NonNull<RawJsString>, Option<Layout>>

Allocates a new RawJsString with an internal capacity of str_len chars.

Errors

Returns Err(None) on integer overflows usize::MAX. Returns Err(Some(Layout)) on allocation error.

fn from_slice_skip_interning(string: &[u16]) -> Self

Creates a new JsString from data, without checking if the string is in the interner.