Struct boa_engine::NativeFunction

pub struct NativeFunction {
    inner: Inner,
}

A callable Rust function that can be invoked by the engine.

NativeFunction functions are divided in two:

Function pointers a.k.a common functions (see NativeFunctionPointer).

Closure functions that can capture the current environment.

Caveats

By limitations of the Rust language, the garbage collector currently cannot inspect closures in order to trace their captured variables. This means that only Copy closures are 100% safe to use. All other closures can also be stored in a NativeFunction, albeit by using an unsafe API, but note that passing closures implicitly capturing traceable types could cause Undefined Behaviour.

Fields
inner: Inner
Implementations

impl NativeFunction

pub fn from_fn_ptr(function: NativeFunctionPointer) -> Self

Creates a NativeFunction from a function pointer.

pub fn from_async_fn<Fut>( f: fn(_: &JsValue, _: &[JsValue], _: &mut Context<'_>) -> Fut ) -> Selfwhere Fut: Future<Output = JsResult<JsValue>> + 'static,

Creates a NativeFunction from a function returning a Future.

The returned NativeFunction will return an ECMAScript Promise that will be fulfilled or rejected when the returned Future completes.

Caveats

Consider the next snippet:

async fn test(
    _this: &JsValue,
    args: &[JsValue],
    _context: &mut Context<'_>,
) -> JsResult<JsValue> {
    let arg = args.get(0).cloned();
    std::future::ready(()).await;
    drop(arg);
    Ok(JsValue::null())
}
NativeFunction::from_async_fn(test);

Seems like a perfectly fine code, right? args is not used after the await point, which in theory should make the whole future 'static … in theory …

This code unfortunately fails to compile at the moment. This is because rustc currently cannot determine that args can be dropped before the await point, which would trivially make the future 'static. Track this issue for more information.

In the meantime, a manual desugaring of the async function does the trick:

fn test(
    _this: &JsValue,
    args: &[JsValue],
    _context: &mut Context<'_>,
) -> impl Future<Output = JsResult<JsValue>> {
    let arg = args.get(0).cloned();
    async move {
        std::future::ready(()).await;
        drop(arg);
        Ok(JsValue::null())
    }
}
NativeFunction::from_async_fn(test);

pub fn from_copy_closure<F>(closure: F) -> Selfwhere F: Fn(&JsValue, &[JsValue], &mut Context<'_>) -> JsResult<JsValue> + Copy + 'static,

Creates a NativeFunction from a Copy closure.

pub fn from_copy_closure_with_captures<F, T>(closure: F, captures: T) -> Selfwhere F: Fn(&JsValue, &[JsValue], &T, &mut Context<'_>) -> JsResult<JsValue> + Copy + 'static, T: Trace + 'static,

Creates a NativeFunction from a Copy closure and a list of traceable captures.

pub unsafe fn from_closure<F>(closure: F) -> Selfwhere F: Fn(&JsValue, &[JsValue], &mut Context<'_>) -> JsResult<JsValue> + 'static,

Creates a new NativeFunction from a closure.

Safety

Passing a closure that contains a captured variable that needs to be traced by the garbage collector could cause an use after free, memory corruption or other kinds of Undefined Behaviour. See https://github.com/Manishearth/rust-gc/issues/50 for a technical explanation on why that is the case.

pub unsafe fn from_closure_with_captures<F, T>(closure: F, captures: T) -> Selfwhere F: Fn(&JsValue, &[JsValue], &T, &mut Context<'_>) -> JsResult<JsValue> + 'static, T: Trace + 'static,

Create a new NativeFunction from a closure and a list of traceable captures.

Safety

Passing a closure that contains a captured variable that needs to be traced by the garbage collector could cause an use after free, memory corruption or other kinds of Undefined Behaviour. See https://github.com/Manishearth/rust-gc/issues/50 for a technical explanation on why that is the case.

pub fn call( &self, this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Calls this NativeFunction, forwarding the arguments to the corresponding function.