Module boa_engine::syntax::ast

The Javascript Abstract Syntax Tree.

Re-exports

pub use self::declaration::Declaration;

pub use self::expression::Expression;

pub use self::keyword::Keyword;

pub use self::position::Position;

pub use self::position::Span;

pub use self::punctuator::Punctuator;

pub use self::statement::Statement;

pub use self::statement_list::StatementList;

pub use self::statement_list::StatementListItem;

Modules

declaration

expression

function

keyword

This module implements the Keyword structure, which represents reserved words of the JavaScript language.

pattern

position

This module implements the Pos structure, which represents a position in the source code.

property

punctuator

This module implements the Punctuator, which represents all punctuators used in JavaScript

statement

This module implements the Statement node.

statement_list

Statement list node.

Enums

ContainsSymbol 🔒

Represents the possible symbols that can be use the the contains function.

Functions

block_to_string 🔒

Displays the body of a block or statement list.

join_nodes 🔒

Utility to join multiple Nodes into a single string.