Module boa_engine::syntax::ast::expression::access

Property access expressions, as defined by the spec.

Property access expressions provide two ways to access properties of an object: dot notation and bracket notation.

Dot notation is mostly used when the name of the property is static, and a valid Javascript identifier e.g. obj.prop, arr.$val.

Bracket notation is used when the name of the property is either variable, not a valid identifier or a symbol e.g. arr[var], arr[5], arr[Symbol.iterator].

A property access expression can be represented by a SimplePropertyAccess (x.y), a PrivatePropertyAccess (x.#y) or a SuperPropertyAccess (super["y"]), each of them with slightly different semantics overall.

Structs

PrivatePropertyAccess

An access expression to a class object’s private fields.

SimplePropertyAccess

A simple property access, where the target object is an Expression.

SuperPropertyAccess

A property access of an object’s parent, as defined by the spec.

Enums

PropertyAccess

A property access expression.

PropertyAccessField

A property access field.