Enum boa_engine::syntax::ast::keyword::Keyword

pub enum Keyword {

    Await,
    Async,
    Break,
    Case,
    Catch,
    Class,
    Continue,
    Const,
    Debugger,
    Default,
    Delete,
    Do,
    Else,
    Enum,
    Export,
    Extends,
    False,
    Finally,
    For,
    Function,
    If,
    In,
    InstanceOf,
    Import,
    Let,
    New,
    Null,
    Of,
    Return,
    Super,
    Switch,
    This,
    Throw,
    True,
    Try,
    TypeOf,
    Var,
    Void,
    While,
    With,
    Yield,

}

Keywords are tokens that have special meaning in JavaScript.

In JavaScript you cannot use these reserved words as variables, labels, or function names.

More information:

ECMAScript reference

MDN documentation

Variants

Await

The await keyword.

More information:

ECMAScript reference

MDN documentation

Async

The async keyword.

More information:

ECMAScript reference

MDN documentation

Break

The break keyword.

More information:

break Node documentation

ECMAScript reference

MDN documentation

Case

The case keyword.

More information:

switch Node documentation

ECMAScript reference

MDN documentation

Catch

The catch keyword.

More information:

try Node documentation

ECMAScript reference

MDN documentation

Class

The class keyword.

More information:

ECMAScript reference

MDN documentation

Continue

The continue keyword.

More information:

continue Node documentation

ECMAScript reference

MDN documentation

Const

The const keyword.

More information:

const Node documentation

ECMAScript reference

MDN documentation

Debugger

The debugger keyword.

More information:

ECMAScript reference

MDN documentation

Default

The default keyword.

More information:

switch Node documentation

ECMAScript reference default clause

ECMAScript reference default export

MDN documentation

Delete

The delete keyword.

More information:

delete UnaryOp documentation

ECMAScript reference

MDN documentation

Do

The do keyword.

More information:

ECMAScript reference

MDN documentation

Else

The else keyword.

More information:

if Node documentation

ECMAScript reference

MDN documentation

Enum

The enum keyword.

Future reserved keyword.

Export

The export keyword.

More information:

ECMAScript reference

MDN documentation

Extends

The extends keyword.

More information:

ECMAScript reference

MDN documentation

False

The false keyword.

More information:

ECMAScript reference

MDN documentation

Finally

The finally keyword.

More information:

try Node documentation

ECMAScript reference

MDN documentation

For

The for keyword.

More information:

for loop Node documentation

ECMAScript reference

MDN documentation

Function

The function keyword.

More information:

function Node documentation

ECMAScript reference

MDN documentation

If

The if keyword.

More information:

if Node documentation

ECMAScript reference

MDN documentation

In

The in keyword.

More information:

ECMAScript reference

MDN documentation

InstanceOf

The instanceof keyword.

More information:

ECMAScript reference

MDN documentation

Import

The import keyword.

More information:

ECMAScript reference

MDN documentation

Let

The let keyword.

More information:

let Node documentation

ECMAScript reference

MDN documentation

New

The new keyword.

More information:

new Node documentation

ECMAScript reference

MDN documentation

Null

The null keyword.

More information:

ECMAScript reference

MDN documentation

Of

The of keyword.

More information:

ECMAScript reference

MDN documentation

Return

The return keyword

More information:

return Node documentation

ECMAScript reference

MDN documentation

Super

The super keyword

More information:

ECMAScript reference

MDN documentation

Switch

The switch keyword.

More information:

switch Node documentation

ECMAScript reference

MDN documentation

This

The this keyword.

More information:

this Node documentation

ECMAScript reference

MDN documentation

Throw

The throw keyword.

More information:

throw Node documentation

ECMAScript reference

MDN documentation

True

The true keyword

More information:

ECMAScript reference

MDN documentation

Try

The try keyword.

More information:

try Node documentation

ECMAScript reference

MDN documentation

TypeOf

The typeof keyword.

More information:

typeof UnaryOp documentation

ECMAScript reference

MDN documentation

Var

The var keyword.

More information:

var Node documentation

ECMAScript reference

MDN documentation

Void

The void keyword.

More information:

void UnaryOp documentation

ECMAScript reference

MDN documentation

While

The while keyword.

More information:

while Node documentation

ECMAScript reference

MDN documentation

With

The with keyword.

More information:

ECMAScript reference

MDN documentation

Yield

The ‘yield’ keyword.

More information:

ECMAScript reference

MDN documentation

Implementations

impl Keyword

pub fn as_binop(self) -> Option<BinOp>

Gets the keyword as a binary operation, if this keyword is the in keyword.

pub fn as_str(self) -> &'static str

Gets the keyword as a string.

pub fn to_sym(self, interner: &mut Interner) -> Sym

Converts the keyword to a symbol in the given interner.