Enum boa_engine::syntax::ast::node::Node

pub enum Node {

    ArrayDecl(ArrayDecl),
    ArrowFunctionDecl(ArrowFunctionDecl),
    Assign(Assign),
    AsyncFunctionDecl(AsyncFunctionDecl),
    AsyncFunctionExpr(AsyncFunctionExpr),
    AsyncGeneratorExpr(AsyncGeneratorExpr),
    AsyncGeneratorDecl(AsyncGeneratorDecl),
    AwaitExpr(AwaitExpr),
    BinOp(BinOp),
    Block(Block),
    Break(Break),
    Call(Call),
    ConditionalOp(ConditionalOp),
    Const(Const),
    ConstDeclList(DeclarationList),
    Continue(Continue),
    DoWhileLoop(DoWhileLoop),
    FunctionDecl(FunctionDecl),
    FunctionExpr(FunctionExpr),
    GetConstField(GetConstField),
    GetPrivateField(GetPrivateField),
    GetField(GetField),
    ForLoop(ForLoop),
    ForInLoop(ForInLoop),
    ForOfLoop(ForOfLoop),
    If(If),
    LetDeclList(DeclarationList),
    Identifier(Identifier),
    New(New),
    Object(Object),
    Return(Return),
    Switch(Switch),
    Spread(Spread),
    TaggedTemplate(Box<TaggedTemplate>),
    TemplateLit(TemplateLit),
    Throw(Throw),
    Try(Box<Try>),
    This,
    UnaryOp(UnaryOp),
    VarDeclList(DeclarationList),
    WhileLoop(WhileLoop),
    Empty,
    Yield(Yield),
    GeneratorDecl(GeneratorDecl),
    GeneratorExpr(GeneratorExpr),
    ClassDecl(Class),
    ClassExpr(Class),

}

Variants

ArrayDecl(ArrayDecl)

Array declaration node. More information.

ArrowFunctionDecl(ArrowFunctionDecl)

An arrow function expression node. More information.

Assign(Assign)

An assignment operator node. More information.

AsyncFunctionDecl(AsyncFunctionDecl)

An async function declaration node. More information.

AsyncFunctionExpr(AsyncFunctionExpr)

An async function expression node. More information.

AsyncGeneratorExpr(AsyncGeneratorExpr)

An async generator expression node.

AsyncGeneratorDecl(AsyncGeneratorDecl)

An async generator declaration node.

AwaitExpr(AwaitExpr)

An await expression node. More information.

BinOp(BinOp)

A binary operator node. More information.

Block(Block)

A Block node. More information.

Break(Break)

A break node. More information.

Call(Call)

A function call. More information.

ConditionalOp(ConditionalOp)

A javascript conditional operand ( x ? y : z ). More information.

Const(Const)

Literals represent values in JavaScript.

These are fixed values not variables that you literally provide in your script.

More information:

ECMAScript reference

MDN documentation

ConstDeclList(DeclarationList)

A constant declaration list. More information.

Continue(Continue)

A continue statement. More information.

DoWhileLoop(DoWhileLoop)

A do … while statement. More information.

FunctionDecl(FunctionDecl)

A function declaration node. More information.

FunctionExpr(FunctionExpr)

A function expression node. More information.

GetConstField(GetConstField)

Provides access to an object types’ constant properties. More information.

GetPrivateField(GetPrivateField)

Provides access to an object types’ private properties. More information.

GetField(GetField)

Provides access to object fields. More information.

ForLoop(ForLoop)

A for statement. More information.

ForInLoop(ForInLoop)

A for...of or for..in statement. More information.

ForOfLoop(ForOfLoop)

A for...of statement. More information.

If(If)

An ‘if’ statement. More information.

LetDeclList(DeclarationList)

A let declaration list. More information.

Identifier(Identifier)

A local identifier node. More information.

New(New)

A new expression. More information.

Object(Object)

An object. More information.

Return(Return)

A return statement. More information.

Switch(Switch)

A switch {case} statement. More information.

Spread(Spread)

A spread (…x) statement. More information.

TaggedTemplate(Box<TaggedTemplate>)

A tagged template. More information.

TemplateLit(TemplateLit)

A template literal. More information.

Throw(Throw)

A throw statement. More information.

Try(Box<Try>)

A try...catch node. More information.

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

UnaryOp(UnaryOp)

Unary operation node. More information

VarDeclList(DeclarationList)

Array declaration node. More information.

WhileLoop(WhileLoop)

A ‘while {…}’ node. More information.

Empty

A empty node.

Empty statement do nothing, just return undefined.

More information:

ECMAScript reference

MDN documentation

Yield(Yield)

A yield node. More information.

GeneratorDecl(GeneratorDecl)

A generator function declaration node. More information.

GeneratorExpr(GeneratorExpr)

A generator function expression node. More information.

ClassDecl(Class)

A class declaration. More information.

ClassExpr(Class)

A class declaration. More information.

Implementations

impl Node

pub(crate) fn hoistable_order(a: &Self, b: &Self) -> Ordering

Returns a node ordering based on the hoistability of each node.

pub fn this() -> Self

Creates a This AST node.

fn to_indented_string(&self, interner: &Interner, indentation: usize) -> String

Creates a string of the value of the node with the given indentation. For example, an indent level of 2 would produce this:

        function hello() {
            console.log("hello");
        }
        hello();
        a = 2;

fn to_no_indent_string(&self, interner: &Interner, indentation: usize) -> String

Implements the display formatting with indentation.

This will not prefix the value with any indentation. If you want to prefix this with proper indents, use to_indented_string().

pub(crate) fn var_declared_names(&self, vars: &mut FxHashSet<Sym>)

pub(crate) fn contains_arguments(&self) -> bool

Returns true if the node contains a identifier reference named ‘arguments’.

More information:

ECMAScript specification