Module boa_engine::syntax::ast::op

This module implements various structure for logic handling.

Enums

AssignOp

An assignment operator assigns a value to its left operand based on the value of its right operand.

BinOp

This represents a binary operation between two values.

BitOp

A bitwise operator is an operator used to perform bitwise operations on bit patterns or binary numerals that involve the manipulation of individual bits.

CompOp

A comparison operator compares its operands and returns a logical value based on whether the comparison is true.

LogOp

Logical operators are typically used with Boolean (logical) values; when they are, they return a Boolean value.

NumOp

Arithmetic operators take numerical values (either literals or variables) as their operands and return a single numerical value.

UnaryOp

A unary operator is one that takes a single operand/argument and performs an operation.