Struct boa_engine::syntax::ast::statement::Case

pub struct Case {
    condition: Expression,
    body: StatementList,
}

A case clause inside a Switch statement, as defined by the spec.

Even though every Case body is a StatementList, it doesn’t create a new lexical environment. This means any variable declared in a Case will be considered as part of the lexical environment of the parent Switch block.

Fields
condition: Expressionbody: StatementList
Implementations

impl Case

pub fn new(condition: Expression, body: StatementList) -> Self

Creates a Case AST node.

pub fn condition(&self) -> &Expression

Gets the condition of the case.

pub fn body(&self) -> &StatementList

Gets the statement listin the body of the case.

pub(crate) fn contains_arguments(&self) -> bool

pub(crate) fn contains(&self, symbol: ContainsSymbol) -> bool