Macro boa_engine::syntax::lexer::operator::vop

macro_rules! vop {
    ($cursor : ident, $assign_op : expr, $op : expr) => { ... };
    ($cursor : ident, $assign_op : expr, $op : expr,
 { $($case : pat => $block : expr), + }) => { ... };
    ($cursor : ident, $op : expr, { $($case : pat => $block : expr), + }) => { ... };
}

vop tests the next token to see if we’re on an assign operation of just a plain binary operation.

If the next value is not an assignment operation it will pattern match the provided values and return the corresponding token.