Struct boa_engine::syntax::parser::cursor::buffered_lexer::BufferedLexer

pub(super) struct BufferedLexer<R> {
    lexer: Lexer<R>,
    peeked: [Option<Token>; 9],
    read_index: usize,
    write_index: usize,
}

Fields
lexer: Lexer<R>peeked: [Option<Token>; 9]read_index: usizewrite_index: usize
Implementations

impl<R> BufferedLexer<R> where
R: Read,

pub(super) fn set_goal(&mut self, elm: InputElement)

Sets the goal symbol for the lexer.

pub(super) fn lex_regex(
&mut self,
start: Position,
interner: &mut Interner
) -> Result<Token, ParseError>

Lexes the next tokens as a regex assuming that the starting ‘/’ has already been consumed.

pub(super) fn lex_template(
&mut self,
start: Position,
interner: &mut Interner
) -> Result<Token, ParseError>

Lexes the next tokens as template middle or template tail assuming that the starting ‘}’ has already been consumed.

pub(super) fn strict_mode(&self) -> bool

pub(super) fn set_strict_mode(&mut self, strict_mode: bool)

fn fill(&mut self, interner: &mut Interner) -> Result<(), ParseError>

Fills the peeking buffer with the next token.

It will not fill two line terminators one after the other.

pub(super) fn next(
&mut self,
skip_line_terminators: bool,
interner: &mut Interner
) -> Result<Option<Token>, ParseError>

Moves the cursor to the next token and returns the token.

If skip_line_terminators is true then line terminators will be discarded.

This follows iterator semantics in that a peek(0, false) followed by a next(false) will return the same value. Note that because a peek(n, false) may return a line terminator a

pub(super) fn peek(
&mut self,
skip_n: usize,
skip_line_terminators: bool,
interner: &mut Interner
) -> Result<Option<&Token>, ParseError>

Peeks the nth token after the next token.

Note: n must be in the range [0, 3]. i.e. if there are tokens A, B, C, D, E and peek(0, false) returns A then:

peek(1, false) == peek(1, true) == B.

peek(2, false) will return C. where A, B, C, D and E are tokens but not line terminators.

If skip_line_terminators is true then line terminators will be discarded. i.e. If there are tokens A, \n, B and peek(0, false) is A then the following will hold:

peek(0, true) == A

peek(0, false) == A

peek(1, true) == B

peek(1, false) == \n

peek(2, true) == None (End of stream)

peek(2, false) == B