Module boa_engine::syntax::parser::expression

Expression parsing.

More information:

MDN documentation

ECMAScript specification

Modules

assignment 🔒

Assignment operator parsing.

await_expr 🔒

Await expression parsing.

identifiers 🔒

Identifiers parsing.

left_hand_side 🔒

Left hand side expression parsing.

primary 🔒

Primary expression parsing.

unary 🔒

Unary operator parsing.

update 🔒

Update expression parsing.

Macros

expression 🔒

Generates an expression parser for a number of expressions whose production rules are of the following pattern.

Structs

AdditiveExpression 🔒

Parses an additive expression.

BitwiseANDExpression 🔒

Parses a bitwise AND expression.

BitwiseORExpression 🔒

Parses a bitwise OR expression.

BitwiseXORExpression 🔒

Parses a bitwise XOR expression.

EqualityExpression 🔒

Parses an equality expression.

Expression 🔒

Expression parsing.

MultiplicativeExpression 🔒

Parses a multiplicative expression.

RelationalExpression 🔒

Parses a relational expression.

ShiftExpression 🔒

Parses a bitwise shift expression.

ShortCircuitExpression 🔒

Parses a logical expression expression.

Enums

PreviousExpr 🔒