Trait boa_engine::syntax::parser::TokenParser

trait TokenParser<R>: Sized where
    R: Read, {
    type Output;

    fn parse(
        self,
        cursor: &mut Cursor<R>,
        interner: &mut Interner
    ) -> ParseResult<Self::Output>;
}

Trait implemented by parsers.

This makes it possible to abstract over the underlying implementation of a parser.

Required Associated Types

type Output

Output type for the parser.

Required Methods

fn parse(
self,
cursor: &mut Cursor<R>,
interner: &mut Interner
) -> ParseResult<Self::Output>

Parses the token stream using the current parser.

This method needs to be provided by the implementor type.