Function boa_engine::value::display::format_rational

fn format_rational(v: f64, f: &mut Formatter<'_>) -> Result

This is different from the ECMAScript compliant number to string, in the printing of -0.

This function prints -0 as -0 instead of positive 0 as the specification says. This is done to make it easer for the user of the REPL to identify what is a -0 vs 0, since the REPL is not bound to the ECMAScript specification we can do this.