Module boa_engine::vm::call_frame

CallFrame

This module will provides everything needed to implement the CallFrame

Structs

CallFrame

A CallFrame holds the state of a function call.

CatchAddresses 🔒

Tracks the address that should be jumped to when an error is caught. Additionally the address of a finally block is tracked, to allow for special handling if it exists.

TryStackEntry 🔒

Tracks the number of environments in the current try-catch-finally block.

Enums

FinallyReturn 🔒

Indicates if a function should return or throw at the end of a finally block.

GeneratorResumeKind 🔒

Indicates how a generator function that has been called/resumed should return.