Struct boa_engine::vm::call_frame::CallFrame

pub struct CallFrame {
    pub(crate) code_block: Gc<CodeBlock>,
    pub(crate) pc: u32,
    pub(crate) rp: u32,
    pub(crate) argument_count: u32,
    pub(crate) env_fp: u32,
    pub(crate) iterators: ThinVec<IteratorRecord>,
    pub(crate) binding_stack: Vec<BindingLocator>,
    pub(crate) loop_iteration_count: u64,
    pub(crate) active_runnable: Option<ActiveRunnable>,
    pub(crate) environments: EnvironmentStack,
    pub(crate) realm: Realm,
    pub(crate) flags: CallFrameFlags,
}

A CallFrame holds the state of a function call.

Fields
code_block: Gc<CodeBlock>pc: u32rp: u32

The register pointer, points to the first register in the stack.

argument_count: u32env_fp: u32iterators: ThinVec<IteratorRecord>binding_stack: Vec<BindingLocator>loop_iteration_count: u64

How many iterations a loop has done.

active_runnable: Option<ActiveRunnable>

[[ScriptOrModule]]

environments: EnvironmentStack

[[Environment]]

realm: Realm

[[Realm]]

flags: CallFrameFlags
Implementations

impl CallFrame

–– CallFrame public API ––

pub const fn code_block(&self) -> &Gc<CodeBlock>

Retrieves the CodeBlock of this call frame.

impl CallFrame

–– CallFrame creation methods ––

pub(crate) const FUNCTION_PROLOGUE: u32 = 2u32

This is the size of the function prologue.

The position of the elements are relative to the CallFrame::fp (register pointer).

                     Setup by the caller
  ┌─────────────────────────────────────────────────────────┐ ┌───── register pointer
  ▼                                                         ▼ ▼
| -(2 + N): this | -(1 + N): func | -N: arg1 | ... | -1: argN | 0: local1 | ... | K: localK |
  ▲                              ▲   ▲                      ▲   ▲                         ▲
  └──────────────────────────────┘   └──────────────────────┘   └─────────────────────────┘
        function prologue                    arguments              Setup by the callee
  ▲
  └─ Frame pointer

Example

The following function calls, generate the following stack:

function x(a) {
}
function y(b, c) {
    return x(b + c)
}

y(1, 2)

    caller prologue    caller arguments   callee prologue   callee arguments
  ┌─────────────────┐   ┌─────────┐   ┌─────────────────┐  ┌──────┐
  ▼                 ▼   ▼         ▼   │                 ▼  ▼      ▼
| 0: undefined | 1: y | 2: 1 | 3: 2 | 4: undefined | 5: x | 6:  3  |
▲                                   ▲                            ▲
│       caller register pointer ────┤                            │
│                                   │                callee register pointer
│                             callee frame pointer
│
└─────  caller frame pointer

Some questions:

Who is responsible for cleaning up the stack after a call? The rust caller.

pub(crate) const THIS_POSITION: u32 = 2u32

pub(crate) const FUNCTION_POSITION: u32 = 1u32

pub(crate) const PROMISE_CAPABILITY_PROMISE_REGISTER_INDEX: u32 = 0u32

pub(crate) const PROMISE_CAPABILITY_RESOLVE_REGISTER_INDEX: u32 = 1u32

pub(crate) const PROMISE_CAPABILITY_REJECT_REGISTER_INDEX: u32 = 2u32

pub(crate) const ASYNC_GENERATOR_OBJECT_REGISTER_INDEX: u32 = 3u32

pub(crate) fn new( code_block: Gc<CodeBlock>, active_runnable: Option<ActiveRunnable>, environments: EnvironmentStack, realm: Realm ) -> Self

Creates a new CallFrame with the provided CodeBlock.

pub(crate) fn with_argument_count(self, count: u32) -> Self

Updates a CallFrame’s argument_count field with the value provided.

pub(crate) fn with_env_fp(self, env_fp: u32) -> Self

Updates a CallFrame’s env_fp field with the value provided.

pub(crate) fn with_flags(self, flags: CallFrameFlags) -> Self

Updates a CallFrame’s flags field with the value provided.

pub(crate) fn this(&self, vm: &Vm) -> JsValue

pub(crate) fn function(&self, vm: &Vm) -> Option<JsObject>

pub(crate) fn arguments<'stack>(&self, vm: &'stack Vm) -> &'stack [JsValue]

pub(crate) fn argument<'stack>( &self, index: usize, vm: &'stack Vm ) -> Option<&'stack JsValue>

pub(crate) fn fp(&self) -> u32

pub(crate) fn restore_stack(&self, vm: &mut Vm)

pub(crate) fn async_generator_object( &self, stack: &[JsValue] ) -> Option<JsObject>

Returns the async generator object, if the function that this CallFrame is from an async generator, None otherwise.

pub(crate) fn promise_capability( &self, stack: &[JsValue] ) -> Option<PromiseCapability>

pub(crate) fn set_promise_capability( &self, stack: &mut [JsValue], promise_capability: Option<&PromiseCapability> )

pub(crate) fn register<'stack>( &self, index: u32, stack: &'stack [JsValue] ) -> &'stack JsValue

Returns the register at the given index.

Panics

If the index is out of bounds.

pub(crate) fn set_register( &self, index: u32, value: JsValue, stack: &mut [JsValue] )

Sets the register at the given index.

Panics

If the index is out of bounds.

pub(crate) fn exit_early(&self) -> bool

Does this have the CallFrameFlags::EXIT_EARLY flag.

pub(crate) fn set_exit_early(&mut self, early_exit: bool)

Set the CallFrameFlags::EXIT_EARLY flag.

pub(crate) fn construct(&self) -> bool

Does this have the CallFrameFlags::CONSTRUCT flag.

pub(crate) fn registers_already_pushed(&self) -> bool

Does this CallFrame need to push registers on Vm::push_frame().

impl CallFrame

–– CallFrame stack methods ––

pub(crate) fn set_register_pointer(&mut self, pointer: u32)