Function boa_engine::vm::code_block::create_function_object

pub(crate) fn create_function_object(
    code: Gc<CodeBlock>,
    async: bool,
    prototype: JsObject,
    context: &mut Context<'_>
) -> JsObject

Creates a new function object.

This is used in cases that the prototype is not known if it’s None or Some.

If the prototype given is None it will use create_function_object_fast. Otherwise it will construct the function from template objects that have all the fields except the prototype, and will perform a prototype transition change to set the prototype.

This is slower than direct object template construction that is done in create_function_object_fast.