Struct boa_engine::vm::code_block::CodeBlockFlags

pub(crate) struct CodeBlockFlags(<CodeBlockFlags as PublicFlags>::Internal);

Flags for CodeBlock.

Tuple Fields
0: <CodeBlockFlags as PublicFlags>::Internal
Implementations

impl CodeBlockFlags

pub const STRICT: Self = _

Is this function in strict mode.

pub const HAS_BINDING_IDENTIFIER: Self = _

Indicates if the function is an expression and has a binding identifier.

pub const IS_CLASS_CONSTRUCTOR: Self = _

The [[IsClassConstructor]] internal slot.

pub const PARAMETERS_ENV_BINDINGS: Self = _

Does this function have a parameters environment.

pub const NEEDS_ARGUMENTS_OBJECT: Self = _

Does this function need a "arguments" object.

The "arguments" binding is the first binding.

pub const IN_CLASS_FIELD_INITIALIZER: Self = _

The [[ClassFieldInitializerName]] internal slot.

pub const TRACEABLE: Self = _

Trace instruction execution to stdout.

impl CodeBlockFlags

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u8

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u8) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u8) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const fn from_bits_retain(bits: u8) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

pub fn from_name(name: &str) -> Option<Self>

Get the value for a flag from its stringified name.

Names are case-sensitive, so must correspond exactly to the identifier given to the flag.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the โbit-clearโ operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl CodeBlockFlags

pub const fn iter(&self) -> Iter<CodeBlockFlags>

Iterate over enabled flag values.

pub const fn iter_names(&self) -> IterNames<CodeBlockFlags>

Iterate over enabled flag values with their stringified names.