Enum boa_engine::vm::Opcode

#[repr(u8)]
pub enum Opcode {

    Pop = 0,
    Dup = 1,
    Swap = 2,
    RotateLeft = 3,
    RotateRight = 4,
    PushZero = 5,
    PushOne = 6,
    PushInt8 = 7,
    PushInt16 = 8,
    PushInt32 = 9,
    PushRational = 10,
    PushNaN = 11,
    PushPositiveInfinity = 12,
    PushNegativeInfinity = 13,
    PushNull = 14,
    PushTrue = 15,
    PushFalse = 16,
    PushUndefined = 17,
    PushLiteral = 18,
    PushEmptyObject = 19,
    PushClassPrototype = 20,
    SetClassPrototype = 21,
    SetHomeObject = 22,
    SetHomeObjectClass = 23,
    SetPrototype = 24,
    PushNewArray = 25,
    PushValueToArray = 26,
    PushElisionToArray = 27,
    PushIteratorToArray = 28,
    Add = 29,
    Sub = 30,
    Div = 31,
    Mul = 32,
    Mod = 33,
    Pow = 34,
    ShiftRight = 35,
    ShiftLeft = 36,
    UnsignedShiftRight = 37,
    BitOr = 38,
    BitAnd = 39,
    BitXor = 40,
    BitNot = 41,
    In = 42,
    InPrivate = 43,
    Eq = 44,
    StrictEq = 45,
    NotEq = 46,
    StrictNotEq = 47,
    GreaterThan = 48,
    GreaterThanOrEq = 49,
    LessThan = 50,
    LessThanOrEq = 51,
    InstanceOf = 52,
    LogicalAnd = 53,
    LogicalOr = 54,
    Coalesce = 55,
    TypeOf = 56,
    Void = 57,
    LogicalNot = 58,
    Pos = 59,
    Neg = 60,
    Inc = 61,
    IncPost = 62,
    Dec = 63,
    DecPost = 64,
    DefVar = 65,
    DefInitVar = 66,
    PutLexicalValue = 67,
    ThrowMutateImmutable = 68,
    GetName = 69,
    GetLocator = 70,
    GetNameAndLocator = 71,
    GetNameOrUndefined = 72,
    SetName = 73,
    SetNameByLocator = 74,
    DeleteName = 75,
    GetPropertyByName = 76,
    GetMethod = 77,
    GetPropertyByValue = 78,
    GetPropertyByValuePush = 79,
    SetPropertyByName = 80,
    SetFunctionName = 81,
    DefineOwnPropertyByName = 82,
    DefineClassStaticMethodByName = 83,
    DefineClassMethodByName = 84,
    SetPropertyByValue = 85,
    DefineOwnPropertyByValue = 86,
    DefineClassStaticMethodByValue = 87,
    DefineClassMethodByValue = 88,
    SetPropertyGetterByName = 89,
    DefineClassStaticGetterByName = 90,
    DefineClassGetterByName = 91,
    SetPropertyGetterByValue = 92,
    DefineClassStaticGetterByValue = 93,
    DefineClassGetterByValue = 94,
    SetPropertySetterByName = 95,
    DefineClassStaticSetterByName = 96,
    DefineClassSetterByName = 97,
    SetPropertySetterByValue = 98,
    DefineClassStaticSetterByValue = 99,
    DefineClassSetterByValue = 100,
    SetPrivateField = 101,
    DefinePrivateField = 102,
    SetPrivateMethod = 103,
    SetPrivateSetter = 104,
    SetPrivateGetter = 105,
    GetPrivateField = 106,
    PushClassField = 107,
    PushClassFieldPrivate = 108,
    PushClassPrivateGetter = 109,
    PushClassPrivateSetter = 110,
    PushClassPrivateMethod = 111,
    DeletePropertyByName = 112,
    DeletePropertyByValue = 113,
    DeleteSuperThrow = 114,
    CopyDataProperties = 115,
    ToPropertyKey = 116,
    Jump = 117,
    JumpIfTrue = 118,
    JumpIfFalse = 119,
    JumpIfNotUndefined = 120,
    JumpIfNullOrUndefined = 121,
    Throw = 122,
    ThrowNewTypeError = 123,
    TryStart = 124,
    TryEnd = 125,
    FinallyStart = 126,
    FinallyEnd = 127,
    Break = 128,
    Continue = 129,
    ToBoolean = 130,
    This = 131,
    Super = 132,
    SuperCallPrepare = 133,
    SuperCall = 134,
    SuperCallSpread = 135,
    SuperCallDerived = 136,
    ImportCall = 137,
    Case = 138,
    Default = 139,
    GetArrowFunction = 140,
    GetAsyncArrowFunction = 141,
    GetFunction = 142,
    GetFunctionAsync = 143,
    GetGenerator = 144,
    GetGeneratorAsync = 145,
    CallEval = 146,
    CallEvalSpread = 147,
    Call = 148,
    CallSpread = 149,
    New = 150,
    NewSpread = 151,
    Return = 152,
    GetReturnValue = 153,
    SetReturnValue = 154,
    PushDeclarativeEnvironment = 155,
    PushObjectEnvironment = 156,
    PushFunctionEnvironment = 157,
    PopEnvironment = 158,
    LoopStart = 159,
    LoopContinue = 160,
    LoopEnd = 161,
    LabelledStart = 162,
    LabelledEnd = 163,
    CreateForInIterator = 164,
    IteratorLoopStart = 165,
    GetIterator = 166,
    GetAsyncIterator = 167,
    IteratorNext = 168,
    IteratorDone = 169,
    IteratorFinishAsyncNext = 170,
    IteratorValue = 171,
    IteratorResult = 172,
    IteratorToArray = 173,
    IteratorPop = 174,
    IteratorStackEmpty = 175,
    CreateIteratorResult = 176,
    IteratorReturn = 177,
    ConcatToString = 178,
    RequireObjectCoercible = 179,
    ValueNotNullOrUndefined = 180,
    RestParameterInit = 181,
    RestParameterPop = 182,
    GeneratorYield = 183,
    GeneratorNext = 184,
    GeneratorResumeReturn = 185,
    AsyncGeneratorYield = 186,
    GeneratorJumpOnResumeKind = 187,
    GeneratorSetReturn = 188,
    GeneratorDelegateNext = 189,
    GeneratorDelegateResume = 190,
    Await = 191,
    NewTarget = 192,
    ImportMeta = 193,
    IsObject = 194,
    TemplateLookup = 195,
    TemplateCreate = 196,
    PushPrivateEnvironment = 197,
    PopPrivateEnvironment = 198,
    Nop = 199,
    Reserved1 = 200,
    Reserved2 = 201,
    Reserved3 = 202,
    Reserved4 = 203,
    Reserved5 = 204,
    Reserved6 = 205,
    Reserved7 = 206,
    Reserved8 = 207,
    Reserved9 = 208,
    Reserved10 = 209,
    Reserved11 = 210,
    Reserved12 = 211,
    Reserved13 = 212,
    Reserved14 = 213,
    Reserved15 = 214,
    Reserved16 = 215,
    Reserved17 = 216,
    Reserved18 = 217,
    Reserved19 = 218,
    Reserved20 = 219,
    Reserved21 = 220,
    Reserved22 = 221,
    Reserved23 = 222,
    Reserved24 = 223,
    Reserved25 = 224,
    Reserved26 = 225,
    Reserved27 = 226,
    Reserved28 = 227,
    Reserved29 = 228,
    Reserved30 = 229,
    Reserved31 = 230,
    Reserved32 = 231,
    Reserved33 = 232,
    Reserved34 = 233,
    Reserved35 = 234,
    Reserved36 = 235,
    Reserved37 = 236,
    Reserved38 = 237,
    Reserved39 = 238,
    Reserved40 = 239,
    Reserved41 = 240,
    Reserved42 = 241,
    Reserved43 = 242,
    Reserved44 = 243,
    Reserved45 = 244,
    Reserved46 = 245,
    Reserved47 = 246,
    Reserved48 = 247,
    Reserved49 = 248,
    Reserved50 = 249,
    Reserved51 = 250,
    Reserved52 = 251,
    Reserved53 = 252,
    Reserved54 = 253,
    Reserved55 = 254,
    Reserved56 = 255,

}

The opcodes of the vm.

Variants

Pop = 0

Pop the top value from the stack.

Operands:

Stack: value =>

Dup = 1

Push a copy of the top value on the stack.

Operands:

Stack: value => value, value

Swap = 2

Swap the top two values on the stack.

Operands:

Stack: second, first => first, second

RotateLeft = 3

Rotates the top n values of the stack to the left by 1.

Equivalent to calling slice::rotate_left with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[n-1], … , v[1], v[0], v[n]

RotateRight = 4

Rotates the top n values of the stack to the right by 1.

Equivalent to calling slice::rotate_right with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[0], v[n], v[n-1], … , v[1]

PushZero = 5

Push integer 0 on the stack.

Operands:

Stack: => 0

PushOne = 6

Push integer 1 on the stack.

Operands:

Stack: => 1

PushInt8 = 7

Push i8 value on the stack.

Operands: value: i8

Stack: => value

PushInt16 = 8

Push i16 value on the stack.

Operands: value: i16

Stack: => value

PushInt32 = 9

Push i32 value on the stack.

Operands: value: i32

Stack: => value

PushRational = 10

Push f64 value on the stack.

Operands: value: f64

Stack: => value

PushNaN = 11

Push NaN integer on the stack.

Operands:

Stack: => NaN

PushPositiveInfinity = 12

Push Infinity value on the stack.

Operands:

Stack: => Infinity

PushNegativeInfinity = 13

Push -Infinity value on the stack.

Operands:

Stack: => -Infinity

PushNull = 14

Push null value on the stack.

Operands:

Stack: => null

PushTrue = 15

Push true value on the stack.

Operands:

Stack: => true

PushFalse = 16

Push false value on the stack.

Operands:

Stack: => false

PushUndefined = 17

Push undefined value on the stack.

Operands:

Stack: => undefined

PushLiteral = 18

Push literal value on the stack.

Like strings and bigints. The index operand is used to index into the literals array to get the value.

Operands: index: u32

Stack: => (literals[index])

PushEmptyObject = 19

Push empty object {} value on the stack.

Operands:

Stack: => {}

PushClassPrototype = 20

Get the prototype of a superclass and push it on the stack.

Operands:

Stack: class, superclass => class, superclass.prototype

SetClassPrototype = 21

Set the prototype of a class object.

Operands:

Stack: class, prototype => class.prototype

SetHomeObject = 22

Set home object internal slot of an object literal method.

Operands:

Stack: home, function => home, function

SetHomeObjectClass = 23

Set home object internal slot of a class method.

Operands:

Stack: home, function => home, function

SetPrototype = 24

Set the prototype of an object if the value is an object or null.

Operands:

Stack: object, value =>

PushNewArray = 25

Push an empty array value on the stack.

Operands:

Stack: => []

PushValueToArray = 26

Push a value to an array.

Operands:

Stack: array, value => array

PushElisionToArray = 27

Push an empty element/hole to an array.

Operands:

Stack: array => array

PushIteratorToArray = 28

Push all iterator values to an array.

Operands:

Stack: array, iterator, next_method => array

Add = 29

Binary + operator.

Operands:

Stack: lhs, rhs => (lhs + rhs)

Sub = 30

Binary - operator.

Operands:

Stack: lhs, rhs => (lhs - rhs)

Div = 31

Binary / operator.

Operands:

Stack: lhs, rhs => (lhs / rhs)

Mul = 32

Binary * operator.

Operands:

Stack: lhs, rhs => (lhs * rhs)

Mod = 33

Binary % operator.

Operands:

Stack: lhs, rhs => (lhs % rhs)

Pow = 34

Binary ** operator.

Operands:

Stack: lhs, rhs => (lhs ** rhs)

ShiftRight = 35

Binary >> operator.

Operands:

Stack: lhs, rhs => (lhs >> rhs)

ShiftLeft = 36

Binary << operator.

Operands:

Stack: lhs, rhs => (lhs << rhs)

UnsignedShiftRight = 37

Binary >>> operator.

Operands:

Stack: lhs, rhs => (lhs >>> rhs)

BitOr = 38

Binary bitwise | operator.

Operands:

Stack: lhs, rhs => (lhs | rhs)

BitAnd = 39

Binary bitwise & operator.

Operands:

Stack: lhs, rhs => (lhs & rhs)

BitXor = 40

Binary bitwise ^ operator.

Operands:

Stack: lhs, rhs => (lhs ^ rhs)

BitNot = 41

Unary bitwise ~ operator.

Operands:

Stack: value => ~value

In = 42

Binary in operator.

Operands:

Stack: lhs, rhs => (lhs in rhs)

InPrivate = 43

Binary in operator for private names.

Operands: private_name_index: u32

Stack: rhs => (private_name in rhs)

Eq = 44

Binary == operator.

Operands:

Stack: lhs, rhs => (lhs == rhs)

StrictEq = 45

Binary === operator.

Operands:

Stack: lhs, rhs => (lhs === rhs)

NotEq = 46

Binary != operator.

Operands:

Stack: lhs, rhs => (lhs != rhs)

StrictNotEq = 47

Binary !== operator.

Operands:

Stack: lhs, rhs => (lhs !== rhs)

GreaterThan = 48

Binary > operator.

Operands:

Stack: lhs, rhs => (lhs > rhs)

GreaterThanOrEq = 49

Binary >= operator.

Operands:

Stack: lhs, rhs => (lhs >= rhs)

LessThan = 50

Binary < operator.

Operands:

Stack: lhs, rhs => (lhs < rhs)

LessThanOrEq = 51

Binary <= operator.

Operands:

Stack: lhs, rhs => (lhs <= rhs)

InstanceOf = 52

Binary instanceof operator.

Operands:

Stack: lhs, rhs => (lhs instanceof rhs)

LogicalAnd = 53

Binary logical && operator.

This is a short-circuit operator, if the lhs value is false, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs && rhs)

LogicalOr = 54

Binary logical || operator.

This is a short-circuit operator, if the lhs value is true, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs || rhs)

Coalesce = 55

Binary ?? operator.

This is a short-circuit operator, if the lhs value is not null or undefined, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs ?? rhs)

TypeOf = 56

Unary typeof operator.

Operands:

Stack: value => (typeof value)

Void = 57

Unary void operator.

Operands:

Stack: value => undefined

LogicalNot = 58

Unary logical ! operator.

Operands:

Stack: value => (!value)

Pos = 59

Unary + operator.

Operands:

Stack: value => (+value)

Neg = 60

Unary - operator.

Operands:

Stack: value => (-value)

Inc = 61

Unary ++ operator.

Operands:

Stack: value => (value + 1)

IncPost = 62

Unary postfix ++ operator.

Operands:

Stack: value => (ToNumeric(value)), (value + 1)

Dec = 63

Unary -- operator.

Operands:

Stack: value => (value - 1)

DecPost = 64

Unary postfix -- operator.

Operands:

Stack: value => (ToNumeric(value)), (value - 1)

DefVar = 65

Declare var type variable.

Operands: name_index: u32

Stack: =>

DefInitVar = 66

Declare and initialize var type variable.

Operands: name_index: u32

Stack: value =>

PutLexicalValue = 67

Initialize a lexical binding.

Operands: name_index: u32

Stack: value =>

ThrowMutateImmutable = 68

Throws an error because the binding access is illegal.

Operands: binding_index: u32

Stack: =>

GetName = 69

Find a binding on the environment chain and push its value.

Operands: name_index: u32

Stack: => value

GetLocator = 70

Find a binding on the environment and set the current_binding of the current frame.

Operands: name_index: u32

Stack: =>

GetNameAndLocator = 71

Find a binding on the environment chain and push its value to the stack and its BindingLocator to the bindings_stack.

Operands: name_index: u32

Stack: => value

GetNameOrUndefined = 72

Find a binding on the environment chain and push its value. If the binding does not exist push undefined.

Operands: name_index: u32

Stack: => value

SetName = 73

Find a binding on the environment chain and assign its value.

Operands: name_index: u32

Stack: value =>

SetNameByLocator = 74

Assigns a value to the binding pointed by the top of the bindings_stack.

Stack: value =>

DeleteName = 75

Deletes a property of the global object.

Operands: name_index: u32

Stack: => deleted

GetPropertyByName = 76

Get a property by name from an object an push it on the stack.

Like object.name

Operands: name_index: u32

Stack: object, receiver => value

GetMethod = 77

Get a property method or undefined if the property is null or undefined.

Throws if the method is not a callable object.

Operands: name_index: u32 Stack: object => object, method

GetPropertyByValue = 78

Get a property by value from an object an push it on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => value

GetPropertyByValuePush = 79

Get a property by value from an object an push the key and value on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => key, value

SetPropertyByName = 80

Sets a property by name of an object.

Like object.name = value

Operands: name_index: u32

Stack: object, receiver, value => value

SetFunctionName = 81

Sets the name of a function object.

This operation is corresponds to the SetFunctionName abstract operation in the spec.

The prefix operand is mapped as follows:

0 -> no prefix

1 -> “get “

2 -> “set “

Operands: prefix: u8

Stack: name, function => function

DefineOwnPropertyByName = 82

Defines a own property of an object by name.

Operands: name_index: u32

Stack: object, value =>

DefineClassStaticMethodByName = 83

Defines a static class method by name.

Operands: name_index: u32

Stack: class, function =>

DefineClassMethodByName = 84

Defines a class method by name.

Operands: name_index: u32

Stack: class_proto, function =>

SetPropertyByValue = 85

Sets a property by value of an object.

Like object[key] = value

Operands:

Stack: object, receiver, key, value => value

DefineOwnPropertyByValue = 86

Defines a own property of an object by value.

Operands:

Stack: object, key, value =>

DefineClassStaticMethodByValue = 87

Defines a static class method by value.

Operands:

Stack: class, key, function =>

DefineClassMethodByValue = 88

Defines a class method by value.

Operands:

Stack: class_proto, key, function =>

SetPropertyGetterByName = 89

Sets a getter property by name of an object.

Like get name() value

Operands: name_index: u32

Stack: object, value =>

DefineClassStaticGetterByName = 90

Defines a static getter class method by name.

Like static get name() value

Operands: name_index: u32

Stack: class, function =>

DefineClassGetterByName = 91

Defines a getter class method by name.

Like get name() value

Operands: name_index: u32

Stack: class_proto, function => class

SetPropertyGetterByValue = 92

Sets a getter property by value of an object.

Like get [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticGetterByValue = 93

Defines a static getter class method by value.

Like static get [key]() value

Operands:

Stack: class, key, function =>

DefineClassGetterByValue = 94

Defines a getter class method by value.

Like get [key]() value

Operands:

Stack: class_proto, key, function =>

SetPropertySetterByName = 95

Sets a setter property by name of an object.

Like set name() value

Operands: name_index: u32

Stack: object, value =>

DefineClassStaticSetterByName = 96

Defines a static setter class method by name.

Like static set name() value

Operands: name_index: u32

Stack: class, function =>

DefineClassSetterByName = 97

Defines a setter class method by name.

Like set name() value

Operands: name_index: u32

Stack: class_proto, function =>

SetPropertySetterByValue = 98

Sets a setter property by value of an object.

Like set [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticSetterByValue = 99

Defines a static setter class method by value.

Like static set [key]() value

Operands:

Stack: class, key, function =>

DefineClassSetterByValue = 100

Defines a setter class method by value.

Like set [key]() value

Operands:

Stack: class_proto, key, function =>

SetPrivateField = 101

Set the value of a private property of an object by it’s name.

Like obj.#name = value

Operands: private_name_index: u32

Stack: object, value => value

DefinePrivateField = 102

Define a private property of a class constructor by it’s name.

Like #name = value

Operands: private_name_index: u32

Stack: object, value =>

SetPrivateMethod = 103

Set a private method of a class constructor by it’s name.

Like #name() {}

Operands: private_name_index: u32

Stack: object, value =>

SetPrivateSetter = 104

Set a private setter property of a class constructor by it’s name.

Like set #name() {}

Operands: private_name_index: u32

Stack: object, value =>

SetPrivateGetter = 105

Set a private getter property of a class constructor by it’s name.

Like get #name() {}

Operands: private_name_index: u32

Stack: object, value =>

GetPrivateField = 106

Get a private property by name from an object an push it on the stack.

Like object.#name

Operands: private_name_index: u32

Stack: object => value

PushClassField = 107

Push a field to a class.

Operands:

Stack: class, field_name, field_function =>

PushClassFieldPrivate = 108

Push a private field to the class.

Operands: private_name_index: u32

Stack: class, field_function =>

PushClassPrivateGetter = 109

Push a private getter to the class.

Operands: private_name_index: u32

Stack: class, getter =>

PushClassPrivateSetter = 110

Push a private setter to the class.

Operands: private_name_index: u32

Stack: class, setter =>

PushClassPrivateMethod = 111

Push a private method to the class.

Operands: private_name_index: u32

Stack: class, method =>

DeletePropertyByName = 112

Deletes a property by name of an object.

Like delete object.key

Operands: name_index: u32

Stack: object =>

DeletePropertyByValue = 113

Deletes a property by value of an object.

Like delete object[key]

Operands:

Stack: object, key =>

DeleteSuperThrow = 114

Throws an error when trying to delete a property of super

Operands:

Stack: =>

CopyDataProperties = 115

Copy all properties of one object to another object.

Operands: excluded_key_count: u32, excluded_key_count_computed: u32

Stack: excluded_key_computed_0 … excluded_key_computed_n, source, value, excluded_key_0 … excluded_key_n => value

ToPropertyKey = 116

Call ToPropertyKey on the value on the stack.

Operands:

Stack: value => key

Jump = 117

Unconditional jump to address.

Operands: address: u32

Stack: =>

JumpIfTrue = 118

Conditional jump to address.

If the value popped is truthy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfFalse = 119

Conditional jump to address.

If the value popped is falsy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfNotUndefined = 120

Conditional jump to address.

If the value popped is not undefined jump to address.

Operands: address: u32

Stack: value => value

JumpIfNullOrUndefined = 121

Conditional jump to address.

If the value popped is undefined jump to address.

Operands: address: u32

Stack: value => value

Throw = 122

Throw exception

Operands:

Stack: value =>

ThrowNewTypeError = 123

Throw a new TypeError exception

Operands: message: u32

Stack: =>

TryStart = 124

Start of a try block.

Operands: next_address: u32, finally_address: u32

Stack: =>

TryEnd = 125

End of a try block.

Operands:

Stack: =>

FinallyStart = 126

Start of a finally block.

Operands:

Stack: =>

FinallyEnd = 127

End of a finally block.

Operands:

Stack: =>

Break = 128

Jumps to a target location and pops the environments involved.

Operands: Jump Address: u32, Target address: u32

Stack: loop_return_value =>

Continue = 129

Sets the AbruptCompletionRecord for a delayed continue

Operands: Jump Address: u32, Target address: u32,

Stack: loop_return_value =>

ToBoolean = 130

Pops value converts it to boolean and pushes it back.

Operands:

Stack: value => (ToBoolean(value))

This = 131

Pushes this value

Operands:

Stack: => this

Super = 132

Pushes the current super value to the stack.

Operands:

Stack: => super

SuperCallPrepare = 133

Get the super constructor and the new target of the current environment.

Operands:

Stack: => super_constructor, new_target

SuperCall = 134

Execute the super() method.

Operands: argument_count: u32

Stack: super_constructor, new_target, argument_1, … argument_n =>

SuperCallSpread = 135

Execute the super() method where the arguments contain spreads.

Operands:

Stack: super_constructor, new_target, arguments_array =>

SuperCallDerived = 136

Execute the super() method when no constructor of the class is defined.

Operands:

Stack: argument_n, … argument_1 =>

ImportCall = 137

Dynamically import a module.

Operands:

Stack: specifier => promise

Case = 138

Pop the two values of the stack, strict equal compares the two values, if true jumps to address, otherwise push the second pop’ed value.

Operands: address: u32

Stack: value, cond => cond (if cond !== value).

Default = 139

Pops the top of stack and jump to address.

Operands: address: u32

Stack: value =>

GetArrowFunction = 140

Get arrow function from the pre-compiled inner functions.

Operands: address: u32, method: u8

Stack: => func

GetAsyncArrowFunction = 141

Get async arrow function from the pre-compiled inner functions.

Operands: address: u32, method: u8

Stack: => func

GetFunction = 142

Get function from the pre-compiled inner functions.

Operands: address: u32, method: u8

Stack: => func

GetFunctionAsync = 143

Get async function from the pre-compiled inner functions.

Operands: address: u32, method: u8

Stack: => func

GetGenerator = 144

Get generator function from the pre-compiled inner functions.

Operands: address: u32,

Stack: => func

GetGeneratorAsync = 145

Get async generator function from the pre-compiled inner functions.

Operands: address: u32,

Stack: => func

CallEval = 146

Call a function named “eval”.

Operands: argument_count: u32

Stack: this, func, argument_1, … argument_n => result

CallEvalSpread = 147

Call a function named “eval” where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

Call = 148

Call a function.

Operands: argument_count: u32

Stack: this, func, argument_1, … argument_n => result

CallSpread = 149

Call a function where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

New = 150

Call construct on a function.

Operands: argument_count: u32

Stack: func, argument_1, … argument_n => result

NewSpread = 151

Call construct on a function where the arguments contain spreads.

Operands:

Stack: arguments_array, func => result

Return = 152

Return from a function.

Operands:

Stack: =>

GetReturnValue = 153

Get return value of a function.

Operands:

Stack: => value

SetReturnValue = 154

Set return value of a function.

Operands:

Stack: value =>

PushDeclarativeEnvironment = 155

Push a declarative environment.

Operands: compile_environments_index: u32

Stack: =>

PushObjectEnvironment = 156

Push an object environment.

Operands:

Stack: object =>

PushFunctionEnvironment = 157

Push a function environment.

Operands: compile_environments_index: u32

Stack: =>

PopEnvironment = 158

Pop the current environment.

Operands:

Stack: =>

LoopStart = 159

Push loop start marker.

Operands:

start: u32

exit: u32

Stack: =>

LoopContinue = 160

Clean up environments when a loop continues.

Operands:

Stack: =>

LoopEnd = 161

Clean up environments at the end of a loop and return it’s value.

Operands:

Stack: => value

LabelledStart = 162

Push labelled start marker.

Operands: Exit Address: u32,

Stack: =>

LabelledEnd = 163

Clean up environments at the end of a labelled block.

Operands:

Stack: =>

CreateForInIterator = 164

Creates the ForInIterator of an object.

Stack: object =>

Iterator Stack: iterator

IteratorLoopStart = 165

Push iterator loop start marker.

Operands:

start: u32

exit: u32

Stack: =>

GetIterator = 166

Gets the iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

GetAsyncIterator = 167

Gets the async iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

IteratorNext = 168

Calls the next method of iterator, updating its record with the next value.

Operands:

Iterator Stack: iterator => iterator

IteratorDone = 169

Returns true if the current iterator is done, or false otherwise

Stack: => done

Iterator Stack: iterator => iterator

IteratorFinishAsyncNext = 170

Finishes the call to Opcode::IteratorNext within a for await loop by setting the current result of the current iterator.

Operands:

Stack: next_result =>

Iterator Stack: iterator => iterator

IteratorValue = 171

Gets the value property of the current iterator record.

Stack: => value

Iterator Stack: iterator => iterator

IteratorResult = 172

Gets the last iteration result of the current iterator record.

Stack: => result

Iterator Stack: iterator => iterator

IteratorToArray = 173

Consume the iterator and construct and array with all the values.

Operands:

Stack: => array

Iterator Stack: iterator => iterator

IteratorPop = 174

Pop an iterator from the call frame close iterator stack.

Iterator Stack:

iterator =>

IteratorStackEmpty = 175

Pushes true to the stack if the iterator stack is empty.

Stack:

=> is_empty

Iterator Stack:

=>

CreateIteratorResult = 176

Creates a new iterator result object.

Operands:

done: bool (codified as u8 with 0 -> false and !0 -> true)

Stack:

value =>

IteratorReturn = 177

Calls return on the current iterator and returns the result.

Stack: => return_val (if return is a method), is_return_method

Iterator Stack: iterator =>

ConcatToString = 178

Concat multiple stack objects into a string.

Operands: value_count: u32

Stack: value_1,…value_n => string

RequireObjectCoercible = 179

Call RequireObjectCoercible on the stack value.

Operands:

Stack: value => value

ValueNotNullOrUndefined = 180

Require the stack value to be neither null nor undefined.

Operands:

Stack: value => value

RestParameterInit = 181

Initialize the rest parameter value of a function from the remaining arguments.

Operands:

Stack: argument_1 .. argument_n => array

RestParameterPop = 182

Pop the remaining arguments of a function.

Operands:

Stack: argument_1 .. argument_n =>

GeneratorYield = 183

Yields from the current generator execution.

Operands:

Stack: value => received

GeneratorNext = 184

Resumes the current generator function.

Operands:

Stack: received =>

GeneratorResumeReturn = 185

Resumes a generator with a return completion.

Operands:

Stack: =>

AsyncGeneratorYield = 186

Yields from the current async generator execution.

Operands:

Stack: value => received

GeneratorJumpOnResumeKind = 187

Jumps to the specified instruction for each resume kind.

Operands:

normal: u32,

throw: u32,

return: u32,

Stack:

GeneratorSetReturn = 188

Sets the current generator resume kind to Return.

Operands:

Stack:

GeneratorDelegateNext = 189

Delegates the current async generator function to another iterator.

Operands: throw_method_undefined: u32, return_method_undefined: u32

Stack: received => result

GeneratorDelegateResume = 190

Resume the async generator with yield delegate logic after it awaits a value.

Operands: return: u32, exit: u32

Stack: is_return, received => value

Await = 191

Stops the current async function and schedules it to resume later.

Operands:

Stack: promise => received

NewTarget = 192

Push the current new target to the stack.

Operands:

Stack: => new.target

ImportMeta = 193

Push the current import.meta to the stack.

Operands:

Stack: => import.meta

IsObject = 194

Pushes true to the stack if the top stack value is an object, or false otherwise.

Operands:

Stack: value => is_object

TemplateLookup = 195

Lookup if a tagged template object is cached and skip the creation if it is.

Operands: jump: u32, site: u64

Stack: => template (if cached)

TemplateCreate = 196

Create a new tagged template object and cache it.

Operands: count: u32, site: u64

Stack: count * (cooked_value, raw_value) => template

PushPrivateEnvironment = 197

Push a private environment.

Operands: count: u32, count * private_name_index: u32

Stack: class => class

PopPrivateEnvironment = 198

Pop a private environment.

Operands:

Stack: =>

Nop = 199

No-operation instruction, does nothing.

Operands:

Stack: =>

Reserved1 = 200

Reserved Opcode.

Reserved2 = 201

Reserved Opcode.

Reserved3 = 202

Reserved Opcode.

Reserved4 = 203

Reserved Opcode.

Reserved5 = 204

Reserved Opcode.

Reserved6 = 205

Reserved Opcode.

Reserved7 = 206

Reserved Opcode.

Reserved8 = 207

Reserved Opcode.

Reserved9 = 208

Reserved Opcode.

Reserved10 = 209

Reserved Opcode.

Reserved11 = 210

Reserved Opcode.

Reserved12 = 211

Reserved Opcode.

Reserved13 = 212

Reserved Opcode.

Reserved14 = 213

Reserved Opcode.

Reserved15 = 214

Reserved Opcode.

Reserved16 = 215

Reserved Opcode.

Reserved17 = 216

Reserved Opcode.

Reserved18 = 217

Reserved Opcode.

Reserved19 = 218

Reserved Opcode.

Reserved20 = 219

Reserved Opcode.

Reserved21 = 220

Reserved Opcode.

Reserved22 = 221

Reserved Opcode.

Reserved23 = 222

Reserved Opcode.

Reserved24 = 223

Reserved Opcode.

Reserved25 = 224

Reserved Opcode.

Reserved26 = 225

Reserved Opcode.

Reserved27 = 226

Reserved Opcode.

Reserved28 = 227

Reserved Opcode.

Reserved29 = 228

Reserved Opcode.

Reserved30 = 229

Reserved Opcode.

Reserved31 = 230

Reserved Opcode.

Reserved32 = 231

Reserved Opcode.

Reserved33 = 232

Reserved Opcode.

Reserved34 = 233

Reserved Opcode.

Reserved35 = 234

Reserved Opcode.

Reserved36 = 235

Reserved Opcode.

Reserved37 = 236

Reserved Opcode.

Reserved38 = 237

Reserved Opcode.

Reserved39 = 238

Reserved Opcode.

Reserved40 = 239

Reserved Opcode.

Reserved41 = 240

Reserved Opcode.

Reserved42 = 241

Reserved Opcode.

Reserved43 = 242

Reserved Opcode.

Reserved44 = 243

Reserved Opcode.

Reserved45 = 244

Reserved Opcode.

Reserved46 = 245

Reserved Opcode.

Reserved47 = 246

Reserved Opcode.

Reserved48 = 247

Reserved Opcode.

Reserved49 = 248

Reserved Opcode.

Reserved50 = 249

Reserved Opcode.

Reserved51 = 250

Reserved Opcode.

Reserved52 = 251

Reserved Opcode.

Reserved53 = 252

Reserved Opcode.

Reserved54 = 253

Reserved Opcode.

Reserved55 = 254

Reserved Opcode.

Reserved56 = 255

Reserved Opcode.

Implementations

impl Opcode

pub const fn as_str(self) -> &'static str

Name of this opcode.

pub const fn as_instruction_str(self) -> &'static str

Name of the profiler event for this opcode.