Module boa_engine::vm::opcode

Modules

await_stm ğŸ”’

binary_ops ğŸ”’

call ğŸ”’

concat ğŸ”’

control_flow ğŸ”’

copy ğŸ”’

define ğŸ”’

delete ğŸ”’

dup ğŸ”’

environment ğŸ”’

generator ğŸ”’

get ğŸ”’

iteration ğŸ”’

jump ğŸ”’

new ğŸ”’

nop ğŸ”’

pop ğŸ”’

push ğŸ”’

require ğŸ”’

rest_parameter ğŸ”’

set ğŸ”’

swap ğŸ”’

switch ğŸ”’

to ğŸ”’

unary_ops ğŸ”’

value ğŸ”’

Macros

generate_impl ğŸ”’

Enums

BindingOpcode ğŸ”’

Specific opcodes for bindings.

Opcode

The opcodes of the vm.

Constants

Add__num_enum_0__ ğŸ”’

AsyncGeneratorNext__num_enum_0__ ğŸ”’

Await__num_enum_0__ ğŸ”’

BitAnd__num_enum_0__ ğŸ”’

BitNot__num_enum_0__ ğŸ”’

BitOr__num_enum_0__ ğŸ”’

BitXor__num_enum_0__ ğŸ”’

Break__num_enum_0__ ğŸ”’

CallEvalSpread__num_enum_0__ ğŸ”’

CallEval__num_enum_0__ ğŸ”’

CallSpread__num_enum_0__ ğŸ”’

Call__num_enum_0__ ğŸ”’

Case__num_enum_0__ ğŸ”’

CatchEnd2__num_enum_0__ ğŸ”’

CatchEnd__num_enum_0__ ğŸ”’

CatchStart__num_enum_0__ ğŸ”’

Coalesce__num_enum_0__ ğŸ”’

ConcatToString__num_enum_0__ ğŸ”’

Continue__num_enum_0__ ğŸ”’

CopyDataProperties__num_enum_0__ ğŸ”’

CreateForInIterator__num_enum_0__ ğŸ”’

DecPost__num_enum_0__ ğŸ”’

Dec__num_enum_0__ ğŸ”’

DefInitVar__num_enum_0__ ğŸ”’

DefVar__num_enum_0__ ğŸ”’

Default__num_enum_0__ ğŸ”’

DefineClassGetterByName__num_enum_0__ ğŸ”’

DefineClassGetterByValue__num_enum_0__ ğŸ”’

DefineClassMethodByName__num_enum_0__ ğŸ”’

DefineClassMethodByValue__num_enum_0__ ğŸ”’

DefineClassSetterByName__num_enum_0__ ğŸ”’

DefineClassSetterByValue__num_enum_0__ ğŸ”’

DefineClassStaticGetterByName__num_enum_0__ ğŸ”’

DefineClassStaticGetterByValue__num_enum_0__ ğŸ”’

DefineClassStaticMethodByName__num_enum_0__ ğŸ”’

DefineClassStaticMethodByValue__num_enum_0__ ğŸ”’

DefineClassStaticSetterByName__num_enum_0__ ğŸ”’

DefineClassStaticSetterByValue__num_enum_0__ ğŸ”’

DefineOwnPropertyByName__num_enum_0__ ğŸ”’

DefineOwnPropertyByValue__num_enum_0__ ğŸ”’

DefinePrivateField__num_enum_0__ ğŸ”’

DeleteName__num_enum_0__ ğŸ”’

DeletePropertyByName__num_enum_0__ ğŸ”’

DeletePropertyByValue__num_enum_0__ ğŸ”’

DeleteSuperThrow__num_enum_0__ ğŸ”’

Div__num_enum_0__ ğŸ”’

Dup__num_enum_0__ ğŸ”’

Eq__num_enum_0__ ğŸ”’

FinallyEnd__num_enum_0__ ğŸ”’

FinallyStart__num_enum_0__ ğŸ”’

GeneratorAsyncDelegateNext__num_enum_0__ ğŸ”’

GeneratorAsyncDelegateResume__num_enum_0__ ğŸ”’

GeneratorAsyncResumeYield__num_enum_0__ ğŸ”’

GeneratorNextDelegate__num_enum_0__ ğŸ”’

GeneratorNext__num_enum_0__ ğŸ”’

GeneratorResumeReturn__num_enum_0__ ğŸ”’

GetArrowFunction__num_enum_0__ ğŸ”’

GetAsyncArrowFunction__num_enum_0__ ğŸ”’

GetAsyncIterator__num_enum_0__ ğŸ”’

GetFunctionAsync__num_enum_0__ ğŸ”’

GetFunction__num_enum_0__ ğŸ”’

GetGeneratorAsync__num_enum_0__ ğŸ”’

GetGenerator__num_enum_0__ ğŸ”’

GetIterator__num_enum_0__ ğŸ”’

GetLocator__num_enum_0__ ğŸ”’

GetMethod__num_enum_0__ ğŸ”’

GetNameAndLocator__num_enum_0__ ğŸ”’

GetNameOrUndefined__num_enum_0__ ğŸ”’

GetName__num_enum_0__ ğŸ”’

GetPrivateField__num_enum_0__ ğŸ”’

GetPropertyByName__num_enum_0__ ğŸ”’

GetPropertyByValuePush__num_enum_0__ ğŸ”’

GetPropertyByValue__num_enum_0__ ğŸ”’

GreaterThanOrEq__num_enum_0__ ğŸ”’

GreaterThan__num_enum_0__ ğŸ”’

InPrivate__num_enum_0__ ğŸ”’

In__num_enum_0__ ğŸ”’

IncPost__num_enum_0__ ğŸ”’

Inc__num_enum_0__ ğŸ”’

InstanceOf__num_enum_0__ ğŸ”’

IsObject__num_enum_0__ ğŸ”’

IteratorClosePop__num_enum_0__ ğŸ”’

IteratorClosePush__num_enum_0__ ğŸ”’

IteratorNextSetDone__num_enum_0__ ğŸ”’

IteratorNext__num_enum_0__ ğŸ”’

IteratorToArray__num_enum_0__ ğŸ”’

IteratorUnwrapNextOrJump__num_enum_0__ ğŸ”’

IteratorUnwrapNext__num_enum_0__ ğŸ”’

IteratorUnwrapValue__num_enum_0__ ğŸ”’

JumpIfFalse__num_enum_0__ ğŸ”’

JumpIfNotUndefined__num_enum_0__ ğŸ”’

JumpIfNullOrUndefined__num_enum_0__ ğŸ”’

JumpIfTrue__num_enum_0__ ğŸ”’

Jump__num_enum_0__ ğŸ”’

LabelledEnd__num_enum_0__ ğŸ”’

LabelledStart__num_enum_0__ ğŸ”’

LessThanOrEq__num_enum_0__ ğŸ”’

LessThan__num_enum_0__ ğŸ”’

LogicalAnd__num_enum_0__ ğŸ”’

LogicalNot__num_enum_0__ ğŸ”’

LogicalOr__num_enum_0__ ğŸ”’

LoopContinue__num_enum_0__ ğŸ”’

LoopEnd__num_enum_0__ ğŸ”’

LoopStart__num_enum_0__ ğŸ”’

Mod__num_enum_0__ ğŸ”’

Mul__num_enum_0__ ğŸ”’

Neg__num_enum_0__ ğŸ”’

NewSpread__num_enum_0__ ğŸ”’

New__num_enum_0__ ğŸ”’

Nop__num_enum_0__ ğŸ”’

NotEq__num_enum_0__ ğŸ”’

PopEnvironment__num_enum_0__ ğŸ”’

PopIfThrown__num_enum_0__ ğŸ”’

PopOnReturnAdd__num_enum_0__ ğŸ”’

PopOnReturnSub__num_enum_0__ ğŸ”’

Pop__num_enum_0__ ğŸ”’

Pos__num_enum_0__ ğŸ”’

Pow__num_enum_0__ ğŸ”’

PushClassFieldPrivate__num_enum_0__ ğŸ”’

PushClassField__num_enum_0__ ğŸ”’

PushClassPrivateGetter__num_enum_0__ ğŸ”’

PushClassPrivateMethod__num_enum_0__ ğŸ”’

PushClassPrivateSetter__num_enum_0__ ğŸ”’

PushClassPrototype__num_enum_0__ ğŸ”’

PushDeclarativeEnvironment__num_enum_0__ ğŸ”’

PushElisionToArray__num_enum_0__ ğŸ”’

PushEmptyObject__num_enum_0__ ğŸ”’

PushFalse__num_enum_0__ ğŸ”’

PushFunctionEnvironment__num_enum_0__ ğŸ”’

PushInt8__num_enum_0__ ğŸ”’

PushInt16__num_enum_0__ ğŸ”’

PushInt32__num_enum_0__ ğŸ”’

PushIteratorToArray__num_enum_0__ ğŸ”’

PushLiteral__num_enum_0__ ğŸ”’

PushNaN__num_enum_0__ ğŸ”’

PushNegativeInfinity__num_enum_0__ ğŸ”’

PushNewArray__num_enum_0__ ğŸ”’

PushNewTarget__num_enum_0__ ğŸ”’

PushNull__num_enum_0__ ğŸ”’

PushObjectEnvironment__num_enum_0__ ğŸ”’

PushOne__num_enum_0__ ğŸ”’

PushPositiveInfinity__num_enum_0__ ğŸ”’

PushRational__num_enum_0__ ğŸ”’

PushTrue__num_enum_0__ ğŸ”’

PushUndefined__num_enum_0__ ğŸ”’

PushValueToArray__num_enum_0__ ğŸ”’

PushZero__num_enum_0__ ğŸ”’

PutLexicalValue__num_enum_0__ ğŸ”’

RequireObjectCoercible__num_enum_0__ ğŸ”’

RestParameterInit__num_enum_0__ ğŸ”’

RestParameterPop__num_enum_0__ ğŸ”’

Return__num_enum_0__ ğŸ”’

RotateLeft__num_enum_0__ ğŸ”’

RotateRight__num_enum_0__ ğŸ”’

SetClassPrototype__num_enum_0__ ğŸ”’

SetFunctionName__num_enum_0__ ğŸ”’

SetHomeObject__num_enum_0__ ğŸ”’

SetNameByLocator__num_enum_0__ ğŸ”’

SetName__num_enum_0__ ğŸ”’

SetPrivateField__num_enum_0__ ğŸ”’

SetPrivateGetter__num_enum_0__ ğŸ”’

SetPrivateMethod__num_enum_0__ ğŸ”’

SetPrivateSetter__num_enum_0__ ğŸ”’

SetPropertyByName__num_enum_0__ ğŸ”’

SetPropertyByValue__num_enum_0__ ğŸ”’

SetPropertyGetterByName__num_enum_0__ ğŸ”’

SetPropertyGetterByValue__num_enum_0__ ğŸ”’

SetPropertySetterByName__num_enum_0__ ğŸ”’

SetPropertySetterByValue__num_enum_0__ ğŸ”’

SetPrototype__num_enum_0__ ğŸ”’

ShiftLeft__num_enum_0__ ğŸ”’

ShiftRight__num_enum_0__ ğŸ”’

StrictEq__num_enum_0__ ğŸ”’

StrictNotEq__num_enum_0__ ğŸ”’

Sub__num_enum_0__ ğŸ”’

SuperCallDerived__num_enum_0__ ğŸ”’

SuperCallPrepare__num_enum_0__ ğŸ”’

SuperCallSpread__num_enum_0__ ğŸ”’

SuperCall__num_enum_0__ ğŸ”’

Super__num_enum_0__ ğŸ”’

Swap__num_enum_0__ ğŸ”’

This__num_enum_0__ ğŸ”’

ThrowNewTypeError__num_enum_0__ ğŸ”’

Throw__num_enum_0__ ğŸ”’

ToBoolean__num_enum_0__ ğŸ”’

ToPropertyKey__num_enum_0__ ğŸ”’

TryEnd__num_enum_0__ ğŸ”’

TryStart__num_enum_0__ ğŸ”’

TypeOf__num_enum_0__ ğŸ”’

UnsignedShiftRight__num_enum_0__ ğŸ”’

ValueNotNullOrUndefined__num_enum_0__ ğŸ”’

Void__num_enum_0__ ğŸ”’

Yield__num_enum_0__ ğŸ”’

Traits

Operation ğŸ”’

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.