Enum boa_engine::vm::opcode::Instruction

#[repr(u8)]
pub(crate) enum Instruction {

    Pop,
    Dup,
    Swap,
    RotateLeft {
        n: u8,
    },
    RotateRight {
        n: u8,
    },
    PushZero,
    PushOne,
    PushInt8 {
        value: i8,
    },
    PushInt16 {
        value: i16,
    },
    PushInt32 {
        value: i32,
    },
    PushFloat {
        value: f32,
    },
    PushDouble {
        value: f64,
    },
    PushNaN,
    PushPositiveInfinity,
    PushNegativeInfinity,
    PushNull,
    PushTrue,
    PushFalse,
    PushUndefined,
    PushLiteral {
        index: VaryingOperand,
    },
    PushRegExp {
        pattern_index: VaryingOperand,
        flags_index: VaryingOperand,
    },
    PushEmptyObject,
    PushClassPrototype,
    SetClassPrototype,
    SetHomeObject,
    SetHomeObjectClass,
    SetPrototype,
    PushNewArray,
    PushValueToArray,
    PushElisionToArray,
    PushIteratorToArray,
    Add,
    Sub,
    Div,
    Mul,
    Mod,
    Pow,
    ShiftRight,
    ShiftLeft,
    UnsignedShiftRight,
    BitOr,
    BitAnd,
    BitXor,
    BitNot,
    In,
    InPrivate {
        index: VaryingOperand,
    },
    Eq,
    StrictEq,
    NotEq,
    StrictNotEq,
    GreaterThan,
    GreaterThanOrEq,
    LessThan,
    LessThanOrEq,
    InstanceOf,
    LogicalAnd {
        exit: u32,
    },
    LogicalOr {
        exit: u32,
    },
    Coalesce {
        exit: u32,
    },
    TypeOf,
    Void,
    LogicalNot,
    Pos,
    Neg,
    Inc,
    IncPost,
    Dec,
    DecPost,
    DefVar {
        index: VaryingOperand,
    },
    DefInitVar {
        index: VaryingOperand,
    },
    PutLexicalValue {
        index: VaryingOperand,
    },
    ThrowMutateImmutable {
        index: VaryingOperand,
    },
    GetName {
        index: VaryingOperand,
    },
    GetLocator {
        index: VaryingOperand,
    },
    GetNameAndLocator {
        index: VaryingOperand,
    },
    GetNameOrUndefined {
        index: VaryingOperand,
    },
    SetName {
        index: VaryingOperand,
    },
    SetNameByLocator,
    DeleteName {
        index: VaryingOperand,
    },
    GetPropertyByName {
        index: VaryingOperand,
    },
    GetPropertyByValue,
    GetPropertyByValuePush,
    SetPropertyByName {
        index: VaryingOperand,
    },
    SetFunctionName {
        prefix: u8,
    },
    DefineOwnPropertyByName {
        index: VaryingOperand,
    },
    DefineClassStaticMethodByName {
        index: VaryingOperand,
    },
    DefineClassMethodByName {
        index: VaryingOperand,
    },
    SetPropertyByValue,
    DefineOwnPropertyByValue,
    DefineClassStaticMethodByValue,
    DefineClassMethodByValue,
    SetPropertyGetterByName {
        index: VaryingOperand,
    },
    DefineClassStaticGetterByName {
        index: VaryingOperand,
    },
    DefineClassGetterByName {
        index: VaryingOperand,
    },
    SetPropertyGetterByValue,
    DefineClassStaticGetterByValue,
    DefineClassGetterByValue,
    SetPropertySetterByName {
        index: VaryingOperand,
    },
    DefineClassStaticSetterByName {
        index: VaryingOperand,
    },
    DefineClassSetterByName {
        index: VaryingOperand,
    },
    SetPropertySetterByValue,
    DefineClassStaticSetterByValue,
    DefineClassSetterByValue,
    SetPrivateField {
        index: VaryingOperand,
    },
    DefinePrivateField {
        index: VaryingOperand,
    },
    SetPrivateMethod {
        index: VaryingOperand,
    },
    SetPrivateSetter {
        index: VaryingOperand,
    },
    SetPrivateGetter {
        index: VaryingOperand,
    },
    GetPrivateField {
        index: VaryingOperand,
    },
    PushClassField,
    PushClassFieldPrivate {
        index: VaryingOperand,
    },
    PushClassPrivateGetter {
        index: VaryingOperand,
    },
    PushClassPrivateSetter {
        index: VaryingOperand,
    },
    PushClassPrivateMethod {
        index: VaryingOperand,
    },
    DeletePropertyByName {
        index: VaryingOperand,
    },
    DeletePropertyByValue,
    DeleteSuperThrow,
    CopyDataProperties {
        excluded_key_count: VaryingOperand,
        excluded_key_count_computed: VaryingOperand,
    },
    ToPropertyKey,
    Jump {
        address: u32,
    },
    JumpIfTrue {
        address: u32,
    },
    JumpIfFalse {
        address: u32,
    },
    JumpIfNotUndefined {
        address: u32,
    },
    JumpIfNullOrUndefined {
        address: u32,
    },
    JumpTable {
        default: u32,
        addresses: ThinVec<u32>,
    },
    Throw,
    ReThrow,
    Exception,
    MaybeException,
    ThrowNewTypeError {
        message: VaryingOperand,
    },
    ToBoolean,
    This,
    Super,
    SuperCallPrepare,
    SuperCall {
        argument_count: VaryingOperand,
    },
    SuperCallSpread,
    SuperCallDerived,
    ImportCall,
    Case {
        address: u32,
    },
    Default {
        address: u32,
    },
    GetArrowFunction {
        index: VaryingOperand,
    },
    GetAsyncArrowFunction {
        index: VaryingOperand,
    },
    GetFunction {
        index: VaryingOperand,
        method: bool,
    },
    GetFunctionAsync {
        index: VaryingOperand,
        method: bool,
    },
    GetGenerator {
        index: VaryingOperand,
    },
    GetGeneratorAsync {
        index: VaryingOperand,
    },
    CallEval {
        argument_count: VaryingOperand,
    },
    CallEvalSpread,
    Call {
        argument_count: VaryingOperand,
    },
    CallSpread,
    New {
        argument_count: VaryingOperand,
    },
    NewSpread,
    Return,
    AsyncGeneratorClose,
    Generator {
        async: bool,
    },
    GetReturnValue,
    SetReturnValue,
    PushDeclarativeEnvironment {
        compile_environments_index: u32,
    },
    PushObjectEnvironment,
    PushFunctionEnvironment {
        compile_environments_index: u32,
    },
    PopEnvironment,
    IncrementLoopIteration,
    CreateForInIterator,
    GetIterator,
    GetAsyncIterator,
    IteratorNext,
    IteratorNextWithoutPop,
    IteratorDone,
    IteratorFinishAsyncNext,
    IteratorValue,
    IteratorValueWithoutPop,
    IteratorResult,
    IteratorToArray,
    IteratorPop,
    IteratorStackEmpty,
    CreateIteratorResult {
        done: bool,
    },
    IteratorReturn,
    ConcatToString {
        value_count: VaryingOperand,
    },
    RequireObjectCoercible,
    ValueNotNullOrUndefined,
    RestParameterInit,
    RestParameterPop,
    GeneratorYield,
    GeneratorNext,
    AsyncGeneratorYield,
    CreatePromiseCapability,
    CompletePromiseCapability,
    JumpIfNotResumeKind {
        exit: u32,
        resume_kind: GeneratorResumeKind,
    },
    GeneratorDelegateNext {
        throw_method_undefined: u32,
        return_method_undefined: u32,
    },
    GeneratorDelegateResume {
        return: u32,
        exit: u32,
    },
    Await,
    NewTarget,
    ImportMeta,
    IsObject,
    TemplateLookup {
        exit: u32,
        site: u64,
    },
    TemplateCreate {
        count: VaryingOperand,
        site: u64,
    },
    PushPrivateEnvironment {
        name_indices: ThinVec<u32>,
    },
    PopPrivateEnvironment,
    Nop,
    U16Operands,
    U32Operands,
    Reserved1,
    Reserved2,
    Reserved3,
    Reserved4,
    Reserved5,
    Reserved6,
    Reserved7,
    Reserved8,
    Reserved9,
    Reserved10,
    Reserved11,
    Reserved12,
    Reserved13,
    Reserved14,
    Reserved15,
    Reserved16,
    Reserved17,
    Reserved18,
    Reserved19,
    Reserved20,
    Reserved21,
    Reserved22,
    Reserved23,
    Reserved24,
    Reserved25,
    Reserved26,
    Reserved27,
    Reserved28,
    Reserved29,
    Reserved30,
    Reserved31,
    Reserved32,
    Reserved33,
    Reserved34,
    Reserved35,
    Reserved36,
    Reserved37,
    Reserved38,
    Reserved39,
    Reserved40,
    Reserved41,
    Reserved42,
    Reserved43,
    Reserved44,
    Reserved45,
    Reserved46,
    Reserved47,
    Reserved48,
    Reserved49,
    Reserved50,
    Reserved51,
    Reserved52,
    Reserved53,
    Reserved54,
    Reserved55,
    Reserved56,

}

This represents a VM instruction, it contains both opcode and operands.

Variants

Pop

Pop the top value from the stack.

Operands:

Stack: value =>

Dup

Push a copy of the top value on the stack.

Operands:

Stack: value => value, value

Swap

Swap the top two values on the stack.

Operands:

Stack: second, first => first, second

RotateLeft

Fields

n: u8

Rotates the top n values of the stack to the left by 1.

Equivalent to calling slice::rotate_left with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[n-1], … , v[1], v[0], v[n]

RotateRight

Fields

n: u8

Rotates the top n values of the stack to the right by 1.

Equivalent to calling slice::rotate_right with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[0], v[n], v[n-1], … , v[1]

PushZero

Push integer 0 on the stack.

Operands:

Stack: => 0

PushOne

Push integer 1 on the stack.

Operands:

Stack: => 1

PushInt8

Fields

value: i8

Push i8 value on the stack.

Operands: value: i8

Stack: => value

PushInt16

Fields

value: i16

Push i16 value on the stack.

Operands: value: i16

Stack: => value

PushInt32

Fields

value: i32

Push i32 value on the stack.

Operands: value: i32

Stack: => value

PushFloat

Fields

value: f32

Push f32 value on the stack.

Operands: value: f32

Stack: => value

PushDouble

Fields

value: f64

Push f64 value on the stack.

Operands: value: f64

Stack: => value

PushNaN

Push NaN integer on the stack.

Operands:

Stack: => NaN

PushPositiveInfinity

Push Infinity value on the stack.

Operands:

Stack: => Infinity

PushNegativeInfinity

Push -Infinity value on the stack.

Operands:

Stack: => -Infinity

PushNull

Push null value on the stack.

Operands:

Stack: => null

PushTrue

Push true value on the stack.

Operands:

Stack: => true

PushFalse

Push false value on the stack.

Operands:

Stack: => false

PushUndefined

Push undefined value on the stack.

Operands:

Stack: => undefined

PushLiteral

Fields

index: VaryingOperand

Push literal value on the stack.

Like strings and bigints. The index operand is used to index into the literals array to get the value.

Operands: index: VaryingOperand

Stack: => (literals[index])

PushRegExp

Fields

pattern_index: VaryingOperand

flags_index: VaryingOperand

Push regexp value on the stack.

Operands: pattern_index: VaryingOperand, flags: VaryingOperand

Stack: => regexp

PushEmptyObject

Push empty object {} value on the stack.

Operands:

Stack: => {}

PushClassPrototype

Get the prototype of a superclass and push it on the stack.

Additionally this sets the [[prototype]] of the class and the DERIVED flag.

Operands:

Stack: class, superclass => class, superclass.prototype

SetClassPrototype

Set the prototype of a class object.

Operands:

Stack: class, prototype => class.prototype

SetHomeObject

Set home object internal slot of an object literal method.

Operands:

Stack: home, function => home, function

SetHomeObjectClass

Set home object internal slot of a class method.

Operands:

Stack: home, function => home, function

SetPrototype

Set the prototype of an object if the value is an object or null.

Operands:

Stack: object, value =>

PushNewArray

Push an empty array value on the stack.

Operands:

Stack: => []

PushValueToArray

Push a value to an array.

Operands:

Stack: array, value => array

PushElisionToArray

Push an empty element/hole to an array.

Operands:

Stack: array => array

PushIteratorToArray

Push all iterator values to an array.

Operands:

Stack: array, iterator, next_method => array

Add

Binary + operator.

Operands:

Stack: lhs, rhs => (lhs + rhs)

Sub

Binary - operator.

Operands:

Stack: lhs, rhs => (lhs - rhs)

Div

Binary / operator.

Operands:

Stack: lhs, rhs => (lhs / rhs)

Mul

Binary * operator.

Operands:

Stack: lhs, rhs => (lhs * rhs)

Mod

Binary % operator.

Operands:

Stack: lhs, rhs => (lhs % rhs)

Pow

Binary ** operator.

Operands:

Stack: lhs, rhs => (lhs ** rhs)

ShiftRight

Binary >> operator.

Operands:

Stack: lhs, rhs => (lhs >> rhs)

ShiftLeft

Binary << operator.

Operands:

Stack: lhs, rhs => (lhs << rhs)

UnsignedShiftRight

Binary >>> operator.

Operands:

Stack: lhs, rhs => (lhs >>> rhs)

BitOr

Binary bitwise | operator.

Operands:

Stack: lhs, rhs => (lhs | rhs)

BitAnd

Binary bitwise & operator.

Operands:

Stack: lhs, rhs => (lhs & rhs)

BitXor

Binary bitwise ^ operator.

Operands:

Stack: lhs, rhs => (lhs ^ rhs)

BitNot

Unary bitwise ~ operator.

Operands:

Stack: value => ~value

In

Binary in operator.

Operands:

Stack: lhs, rhs => (lhs in rhs)

InPrivate

Fields

index: VaryingOperand

Binary in operator for private names.

Operands: index: u32

Stack: rhs => (private_name in rhs)

Eq

Binary == operator.

Operands:

Stack: lhs, rhs => (lhs == rhs)

StrictEq

Binary === operator.

Operands:

Stack: lhs, rhs => (lhs === rhs)

NotEq

Binary != operator.

Operands:

Stack: lhs, rhs => (lhs != rhs)

StrictNotEq

Binary !== operator.

Operands:

Stack: lhs, rhs => (lhs !== rhs)

GreaterThan

Binary > operator.

Operands:

Stack: lhs, rhs => (lhs > rhs)

GreaterThanOrEq

Binary >= operator.

Operands:

Stack: lhs, rhs => (lhs >= rhs)

LessThan

Binary < operator.

Operands:

Stack: lhs, rhs => (lhs < rhs)

LessThanOrEq

Binary <= operator.

Operands:

Stack: lhs, rhs => (lhs <= rhs)

InstanceOf

Binary instanceof operator.

Operands:

Stack: lhs, rhs => (lhs instanceof rhs)

LogicalAnd

Fields

exit: u32

Binary logical && operator.

This is a short-circuit operator, if the lhs value is false, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs && rhs)

LogicalOr

Fields

exit: u32

Binary logical || operator.

This is a short-circuit operator, if the lhs value is true, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs || rhs)

Coalesce

Fields

exit: u32

Binary ?? operator.

This is a short-circuit operator, if the lhs value is not null or undefined, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs ?? rhs)

TypeOf

Unary typeof operator.

Operands:

Stack: value => (typeof value)

Void

Unary void operator.

Operands:

Stack: value => undefined

LogicalNot

Unary logical ! operator.

Operands:

Stack: value => (!value)

Pos

Unary + operator.

Operands:

Stack: value => (+value)

Neg

Unary - operator.

Operands:

Stack: value => (-value)

Inc

Unary ++ operator.

Operands:

Stack: value => (value + 1)

IncPost

Unary postfix ++ operator.

Operands:

Stack: value => (ToNumeric(value)), (value + 1)

Dec

Unary -- operator.

Operands:

Stack: value => (value - 1)

DecPost

Unary postfix -- operator.

Operands:

Stack: value => (ToNumeric(value)), (value - 1)

DefVar

Fields

index: VaryingOperand

Declare var type variable.

Operands: index: u32

Stack: =>

DefInitVar

Fields

index: VaryingOperand

Declare and initialize var type variable.

Operands: index: u32

Stack: value =>

PutLexicalValue

Fields

index: VaryingOperand

Initialize a lexical binding.

Operands: index: u32

Stack: value =>

ThrowMutateImmutable

Fields

index: VaryingOperand

Throws an error because the binding access is illegal.

Operands: index: u32

Stack: =>

GetName

Fields

index: VaryingOperand

Find a binding on the environment chain and push its value.

Operands: index: u32

Stack: => value

GetLocator

Fields

index: VaryingOperand

Find a binding on the environment and set the current_binding of the current frame.

Operands: index: u32

Stack: =>

GetNameAndLocator

Fields

index: VaryingOperand

Find a binding on the environment chain and push its value to the stack and its BindingLocator to the bindings_stack.

Operands: index: u32

Stack: => value

GetNameOrUndefined

Fields

index: VaryingOperand

Find a binding on the environment chain and push its value. If the binding does not exist push undefined.

Operands: index: u32

Stack: => value

SetName

Fields

index: VaryingOperand

Find a binding on the environment chain and assign its value.

Operands: index: u32

Stack: value =>

SetNameByLocator

Assigns a value to the binding pointed by the top of the bindings_stack.

Stack: value =>

DeleteName

Fields

index: VaryingOperand

Deletes a property of the global object.

Operands: index: u32

Stack: => deleted

GetPropertyByName

Fields

index: VaryingOperand

Get a property by name from an object an push it on the stack.

Like object.name

Operands: index: u32

Stack: object, receiver => value

GetPropertyByValue

Get a property by value from an object an push it on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => value

GetPropertyByValuePush

Get a property by value from an object an push the key and value on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => key, value

SetPropertyByName

Fields

index: VaryingOperand

Sets a property by name of an object.

Like object.name = value

Operands: index: u32

Stack: object, receiver, value => value

SetFunctionName

Fields

prefix: u8

Sets the name of a function object.

This operation is corresponds to the SetFunctionName abstract operation in the spec.

The prefix operand is mapped as follows:

0 -> no prefix

1 -> “get “

2 -> “set “

Operands: prefix: u8

Stack: name, function => function

DefineOwnPropertyByName

Fields

index: VaryingOperand

Defines a own property of an object by name.

Operands: index: u32

Stack: object, value =>

DefineClassStaticMethodByName

Fields

index: VaryingOperand

Defines a static class method by name.

Operands: index: u32

Stack: class, function =>

DefineClassMethodByName

Fields

index: VaryingOperand

Defines a class method by name.

Operands: index: u32

Stack: class_proto, function =>

SetPropertyByValue

Sets a property by value of an object.

Like object[key] = value

Operands:

Stack: object, receiver, key, value => value

DefineOwnPropertyByValue

Defines a own property of an object by value.

Operands:

Stack: object, key, value =>

DefineClassStaticMethodByValue

Defines a static class method by value.

Operands:

Stack: class, key, function =>

DefineClassMethodByValue

Defines a class method by value.

Operands:

Stack: class_proto, key, function =>

SetPropertyGetterByName

Fields

index: VaryingOperand

Sets a getter property by name of an object.

Like get name() value

Operands: index: u32

Stack: object, value =>

DefineClassStaticGetterByName

Fields

index: VaryingOperand

Defines a static getter class method by name.

Like static get name() value

Operands: index: u32

Stack: class, binding_function =>

DefineClassGetterByName

Fields

index: VaryingOperand

Defines a getter class method by name.

Like get name() value

Operands: index: u32

Stack: class_proto, function => class

SetPropertyGetterByValue

Sets a getter property by value of an object.

Like get [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticGetterByValue

Defines a static getter class method by value.

Like static get [key]() value

Operands:

Stack: class, key, function =>

DefineClassGetterByValue

Defines a getter class method by value.

Like get [key]() value

Operands:

Stack: class_proto, key, function =>

SetPropertySetterByName

Fields

index: VaryingOperand

Sets a setter property by name of an object.

Like set name() value

Operands: index: u32

Stack: object, value =>

DefineClassStaticSetterByName

Fields

index: VaryingOperand

Defines a static setter class method by name.

Like static set name() value

Operands: index: u32

Stack: class, function =>

DefineClassSetterByName

Fields

index: VaryingOperand

Defines a setter class method by name.

Like set name() value

Operands: index: u32

Stack: class_proto, function =>

SetPropertySetterByValue

Sets a setter property by value of an object.

Like set [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticSetterByValue

Defines a static setter class method by value.

Like static set [key]() value

Operands:

Stack: class, key, function =>

DefineClassSetterByValue

Defines a setter class method by value.

Like set [key]() value

Operands:

Stack: class_proto, key, function =>

SetPrivateField

Fields

index: VaryingOperand

Set the value of a private property of an object by it’s name.

Like obj.#name = value

Operands: index: u32

Stack: object, value => value

DefinePrivateField

Fields

index: VaryingOperand

Define a private property of a class constructor by it’s name.

Like #name = value

Operands: index: u32

Stack: object, value =>

SetPrivateMethod

Fields

index: VaryingOperand

Set a private method of a class constructor by it’s name.

Like #name() {}

Operands: index: u32

Stack: object, value =>

SetPrivateSetter

Fields

index: VaryingOperand

Set a private setter property of a class constructor by it’s name.

Like set #name() {}

Operands: index: u32

Stack: object, value =>

SetPrivateGetter

Fields

index: VaryingOperand

Set a private getter property of a class constructor by it’s name.

Like get #name() {}

Operands: index: u32

Stack: object, value =>

GetPrivateField

Fields

index: VaryingOperand

Get a private property by name from an object an push it on the stack.

Like object.#name

Operands: index: u32

Stack: object => value

PushClassField

Push a field to a class.

Operands:

Stack: class, field_name, field_function =>

PushClassFieldPrivate

Fields

index: VaryingOperand

Push a private field to the class.

Operands: index: u32

Stack: class, field_function =>

PushClassPrivateGetter

Fields

index: VaryingOperand

Push a private getter to the class.

Operands: index: u32

Stack: class, getter =>

PushClassPrivateSetter

Fields

index: VaryingOperand

Push a private setter to the class.

Operands: index: u32

Stack: class, setter =>

PushClassPrivateMethod

Fields

index: VaryingOperand

Push a private method to the class.

Operands: index: u32

Stack: class, method =>

DeletePropertyByName

Fields

index: VaryingOperand

Deletes a property by name of an object.

Like delete object.key

Operands: index: u32

Stack: object =>

DeletePropertyByValue

Deletes a property by value of an object.

Like delete object[key]

Operands:

Stack: object, key =>

DeleteSuperThrow

Throws an error when trying to delete a property of super

Operands:

Stack: =>

CopyDataProperties

Fields

excluded_key_count: VaryingOperand

excluded_key_count_computed: VaryingOperand

Copy all properties of one object to another object.

Operands: excluded_key_count: VaryingOperand, excluded_key_count_computed: VaryingOperand

Stack: excluded_key_computed_0 … excluded_key_computed_n, source, value, excluded_key_0 … excluded_key_n => value

ToPropertyKey

Call ToPropertyKey on the value on the stack.

Operands:

Stack: value => key

Jump

Fields

address: u32

Unconditional jump to address.

Operands: address: u32

Stack: =>

JumpIfTrue

Fields

address: u32

Conditional jump to address.

If the value popped is truthy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfFalse

Fields

address: u32

Conditional jump to address.

If the value popped is falsy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfNotUndefined

Fields

address: u32

Conditional jump to address.

If the value popped is not undefined jump to address.

Operands: address: u32

Stack: value => value

JumpIfNullOrUndefined

Fields

address: u32

Conditional jump to address.

If the value popped is undefined jump to address.

Operands: address: u32

Stack: value => value

JumpTable

Fields

default: u32

addresses: ThinVec<u32>

Jump table that jumps depending on top value of the stack.

This is used to handle special cases when we call continue, break or return in a try block, that has finally block.

Operands: default: u32, count: u32, address: u32 * count

Stack: value: i32 =>

Throw

Throw exception.

This sets pending exception and searches for an exception handler.

Operands:

Stack: value =>

ReThrow

Rethrow thrown exception.

This is also used to handle generator return() call, we throw an empty exception, by setting pending exception to None, propagating it and calling finally code until there is no exception handler left, in that case we consume the empty exception and return from the generator.

Operands:

Stack: =>

Exception

Get the thrown pending exception (if it’s set) and push on the stack.

If there is no pending exception, which can happend if we are handling return() call on generator, then we rethrow the empty exception. See Opcode::ReThrow.

Operands:

Stack: => exception

MaybeException

Get the thrown pending exception if it’s set and push true, otherwise push only false.

Operands:

Stack: => (true, exception) or false

ThrowNewTypeError

Fields

message: VaryingOperand

Throw a new TypeError exception

Operands: message: u32

Stack: =>

ToBoolean

Pops value converts it to boolean and pushes it back.

Operands:

Stack: value => (ToBoolean(value))

This

Pushes this value

Operands:

Stack: => this

Super

Pushes the current super value to the stack.

Operands:

Stack: => super

SuperCallPrepare

Get the super constructor and the new target of the current environment.

Operands:

Stack: => super_constructor, new_target

SuperCall

Fields

argument_count: VaryingOperand

Execute the super() method.

Operands: argument_count: u32

Stack: super_constructor, new_target, argument_1, … argument_n =>

SuperCallSpread

Execute the super() method where the arguments contain spreads.

Operands:

Stack: super_constructor, new_target, arguments_array =>

SuperCallDerived

Execute the super() method when no constructor of the class is defined.

Operands:

Stack: argument_n, … argument_1 =>

ImportCall

Dynamically import a module.

Operands:

Stack: specifier => promise

Case

Fields

address: u32

Pop the two values of the stack, strict equal compares the two values, if true jumps to address, otherwise push the second pop’ed value.

Operands: address: u32

Stack: value, cond => cond (if cond !== value).

Default

Fields

address: u32

Pops the top of stack and jump to address.

Operands: address: u32

Stack: value =>

GetArrowFunction

Fields

index: VaryingOperand

Get arrow function from the pre-compiled inner functions.

Operands: index: u32

Stack: => func

GetAsyncArrowFunction

Fields

index: VaryingOperand

Get async arrow function from the pre-compiled inner functions.

Operands: index: VaryingOperand

Stack: => func

GetFunction

Fields

index: VaryingOperand

method: bool

Get function from the pre-compiled inner functions.

Operands: index: VaryingOperand, is_method: u8

Stack: => func

GetFunctionAsync

Fields

index: VaryingOperand

method: bool

Get async function from the pre-compiled inner functions.

Operands: index: VaryingOperand, method: u8

Stack: => func

GetGenerator

Fields

index: VaryingOperand

Get generator function from the pre-compiled inner functions.

Operands: index: VaryingOperand,

Stack: => func

GetGeneratorAsync

Fields

index: VaryingOperand

Get async generator function from the pre-compiled inner functions.

Operands: index: VaryingOperand,

Stack: => func

CallEval

Fields

argument_count: VaryingOperand

Call a function named “eval”.

Operands: argument_count: VaryingOperand

Stack: this, func, argument_1, … argument_n => result

CallEvalSpread

Call a function named “eval” where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

Call

Fields

argument_count: VaryingOperand

Call a function.

Operands: argument_count: u32

Stack: this, func, argument_1, … argument_n => result

CallSpread

Call a function where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

New

Fields

argument_count: VaryingOperand

Call construct on a function.

Operands: argument_count: u32

Stack: func, argument_1, … argument_n => result

NewSpread

Call construct on a function where the arguments contain spreads.

Operands:

Stack: arguments_array, func => result

Return

Return from a function.

Operands:

Stack: =>

AsyncGeneratorClose

Close an async generator function.

Operands:

Stack: =>

Generator

Fields

async: bool

Creates the generator object and yields.

Operands: async: u8

Stack: => resume_kind

GetReturnValue

Get return value of a function.

Operands:

Stack: => value

SetReturnValue

Set return value of a function.

Operands:

Stack: value =>

PushDeclarativeEnvironment

Fields

compile_environments_index: u32

Push a declarative environment.

Operands: compile_environments_index: u32

Stack: =>

PushObjectEnvironment

Push an object environment.

Operands:

Stack: object =>

PushFunctionEnvironment

Fields

compile_environments_index: u32

Push a function environment.

Operands: compile_environments_index: u32

Stack: =>

PopEnvironment

Pop the current environment.

Operands:

Stack: =>

IncrementLoopIteration

Increment loop itearation count.

Used for limiting the loop iteration.

Operands:

Stack: =>

CreateForInIterator

Creates the ForInIterator of an object.

Stack: object =>

Iterator Stack: iterator

GetIterator

Gets the iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

GetAsyncIterator

Gets the async iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

IteratorNext

Calls the next method of iterator, updating its record with the next value.

Operands:

Iterator Stack: iterator => iterator

IteratorNextWithoutPop

Calls the next method of iterator, updating its record with the next value.

Operands:

Iterator Stack: iterator => iterator

IteratorDone

Returns true if the current iterator is done, or false otherwise

Stack: => done

Iterator Stack: iterator => iterator

IteratorFinishAsyncNext

Finishes the call to Opcode::IteratorNext within a for await loop by setting the current result of the current iterator.

Operands:

Stack: next_result, resume_kind => resume_kind

Iterator Stack: iterator => iterator

IteratorValue

Gets the value property of the current iterator record.

Stack: => value

Iterator Stack: iterator => iterator

IteratorValueWithoutPop

Gets the value property of the current iterator record.

Stack: => value

Iterator Stack: iterator => iterator

IteratorResult

Gets the last iteration result of the current iterator record.

Stack: => result

Iterator Stack: iterator => iterator

IteratorToArray

Consume the iterator and construct and array with all the values.

Operands:

Stack: => array

Iterator Stack: iterator => iterator

IteratorPop

Pop an iterator from the call frame close iterator stack.

Iterator Stack:

iterator =>

IteratorStackEmpty

Pushes true to the stack if the iterator stack is empty.

Stack:

=> is_empty

Iterator Stack:

=>

CreateIteratorResult

Fields

done: bool

Creates a new iterator result object.

Operands:

done: bool (codified as u8 with 0 -> false and !0 -> true)

Stack:

value =>

IteratorReturn

Calls return on the current iterator and returns the result.

Stack: => return_val (if return is a method), is_return_method

Iterator Stack: iterator =>

ConcatToString

Fields

value_count: VaryingOperand

Concat multiple stack objects into a string.

Operands: value_count: u32

Stack: value_1,…value_n => string

RequireObjectCoercible

Call RequireObjectCoercible on the stack value.

Operands:

Stack: value => value

ValueNotNullOrUndefined

Require the stack value to be neither null nor undefined.

Operands:

Stack: value => value

RestParameterInit

Initialize the rest parameter value of a function from the remaining arguments.

Operands:

Stack: argument_1 .. argument_n => array

RestParameterPop

Pop the remaining arguments of a function.

Operands:

Stack: argument_1 .. argument_n =>

GeneratorYield

Yields from the current generator execution.

Operands:

Stack: => resume_kind, received

GeneratorNext

Resumes the current generator function.

If the resume_kind is Throw, then the value is poped and thrown, otherwise if Return we pop the value, set it as the return value and throw and empty exception. See Opcode::ReThrow.

Operands:

Stack: resume_kind, value => value

AsyncGeneratorYield

Yields from the current async generator execution.

Operands:

Stack: value => received

CreatePromiseCapability

Create a promise capacity for an async function, if not already set.

Operands:

Stack: =>

CompletePromiseCapability

Resolves or rejects the promise capability of an async function.

If the pending exception is set, reject and rethrow the exception, otherwise resolve.

Operands:

Stack: =>

JumpIfNotResumeKind

Fields

exit: u32

resume_kind: GeneratorResumeKind

Jumps to the specified address if the resume kind is not equal.

Operands: exit: u32, resume_kind: u8.

Stack: resume_kind => resume_kind

GeneratorDelegateNext

Fields

throw_method_undefined: u32

return_method_undefined: u32

Delegates the current async generator function to another iterator.

Operands: throw_method_undefined: u32, return_method_undefined: u32

Stack: received => result

GeneratorDelegateResume

Fields

return: u32

exit: u32

Resume the async generator with yield delegate logic after it awaits a value.

Operands: return: u32, exit: u32

Stack: is_return, received => value

Await

Stops the current async function and schedules it to resume later.

Operands:

Stack: promise => received

NewTarget

Push the current new target to the stack.

Operands:

Stack: => new.target

ImportMeta

Push the current import.meta to the stack.

Operands:

Stack: => import.meta

IsObject

Pushes true to the stack if the top stack value is an object, or false otherwise.

Operands:

Stack: value => is_object

TemplateLookup

Fields

exit: u32

site: u64

Lookup if a tagged template object is cached and skip the creation if it is.

Operands: exit: u32, site: u64

Stack: => template (if cached)

TemplateCreate

Fields

count: VaryingOperand

site: u64

Create a new tagged template object and cache it.

Operands: count: VaryingOperand, site: u64

Stack: count * (cooked_value, raw_value) => template

PushPrivateEnvironment

Fields

name_indices: ThinVec<u32>

Push a private environment.

Operands: count: u32, count * name_index: u32

Stack: class => class

PopPrivateEnvironment

Pop a private environment.

Operands:

Stack: =>

Nop

No-operation instruction, does nothing.

Operands:

Stack: =>

U16Operands

Opcode prefix modifier, makes all VaryingOperands of an instruction u16 sized.

Operands: opcode (operands if any).

Stack: The stack changes based on the opcode that is being prefixed.

U32Operands

Opcode prefix modifier, Opcode prefix operand modifier, makes all VaryingOperands of an instruction u32 sized.

Operands: opcode (operands if any).

Stack: The stack changes based on the opcode that is being prefixed.

Reserved1

Reserved Opcode.

Reserved2

Reserved Opcode.

Reserved3

Reserved Opcode.

Reserved4

Reserved Opcode.

Reserved5

Reserved Opcode.

Reserved6

Reserved Opcode.

Reserved7

Reserved Opcode.

Reserved8

Reserved Opcode.

Reserved9

Reserved Opcode.

Reserved10

Reserved Opcode.

Reserved11

Reserved Opcode.

Reserved12

Reserved Opcode.

Reserved13

Reserved Opcode.

Reserved14

Reserved Opcode.

Reserved15

Reserved Opcode.

Reserved16

Reserved Opcode.

Reserved17

Reserved Opcode.

Reserved18

Reserved Opcode.

Reserved19

Reserved Opcode.

Reserved20

Reserved Opcode.

Reserved21

Reserved Opcode.

Reserved22

Reserved Opcode.

Reserved23

Reserved Opcode.

Reserved24

Reserved Opcode.

Reserved25

Reserved Opcode.

Reserved26

Reserved Opcode.

Reserved27

Reserved Opcode.

Reserved28

Reserved Opcode.

Reserved29

Reserved Opcode.

Reserved30

Reserved Opcode.

Reserved31

Reserved Opcode.

Reserved32

Reserved Opcode.

Reserved33

Reserved Opcode.

Reserved34

Reserved Opcode.

Reserved35

Reserved Opcode.

Reserved36

Reserved Opcode.

Reserved37

Reserved Opcode.

Reserved38

Reserved Opcode.

Reserved39

Reserved Opcode.

Reserved40

Reserved Opcode.

Reserved41

Reserved Opcode.

Reserved42

Reserved Opcode.

Reserved43

Reserved Opcode.

Reserved44

Reserved Opcode.

Reserved45

Reserved Opcode.

Reserved46

Reserved Opcode.

Reserved47

Reserved Opcode.

Reserved48

Reserved Opcode.

Reserved49

Reserved Opcode.

Reserved50

Reserved Opcode.

Reserved51

Reserved Opcode.

Reserved52

Reserved Opcode.

Reserved53

Reserved Opcode.

Reserved54

Reserved Opcode.

Reserved55

Reserved Opcode.

Reserved56

Reserved Opcode.

Implementations

impl Instruction

pub(crate) fn to_bytecode(&self, bytes: &mut Vec<u8>)

Convert Instruction to compact bytecode.

pub(crate) fn from_bytecode( bytes: &[u8], pc: &mut usize, varying_kind: VaryingOperandKind ) -> Self

Convert compact bytecode to Instruction.

Panics

If the provided bytecode is not valid.

pub(crate) const fn opcode(&self) -> Opcode

Get the Opcode of the Instruction.