Enum boa_engine::vm::opcode::Opcode

#[repr(u8)]
pub(crate) enum Opcode {

    Pop = 0,
    Dup = 1,
    Swap = 2,
    RotateLeft = 3,
    RotateRight = 4,
    PushZero = 5,
    PushOne = 6,
    PushInt8 = 7,
    PushInt16 = 8,
    PushInt32 = 9,
    PushFloat = 10,
    PushDouble = 11,
    PushNaN = 12,
    PushPositiveInfinity = 13,
    PushNegativeInfinity = 14,
    PushNull = 15,
    PushTrue = 16,
    PushFalse = 17,
    PushUndefined = 18,
    PushLiteral = 19,
    PushRegExp = 20,
    PushEmptyObject = 21,
    PushClassPrototype = 22,
    SetClassPrototype = 23,
    SetHomeObject = 24,
    SetPrototype = 25,
    PushNewArray = 26,
    PushValueToArray = 27,
    PushElisionToArray = 28,
    PushIteratorToArray = 29,
    Add = 30,
    Sub = 31,
    Div = 32,
    Mul = 33,
    Mod = 34,
    Pow = 35,
    ShiftRight = 36,
    ShiftLeft = 37,
    UnsignedShiftRight = 38,
    BitOr = 39,
    BitAnd = 40,
    BitXor = 41,
    BitNot = 42,
    In = 43,
    InPrivate = 44,
    Eq = 45,
    StrictEq = 46,
    NotEq = 47,
    StrictNotEq = 48,
    GreaterThan = 49,
    GreaterThanOrEq = 50,
    LessThan = 51,
    LessThanOrEq = 52,
    InstanceOf = 53,
    LogicalAnd = 54,
    LogicalOr = 55,
    Coalesce = 56,
    TypeOf = 57,
    Void = 58,
    LogicalNot = 59,
    Pos = 60,
    Neg = 61,
    Inc = 62,
    IncPost = 63,
    Dec = 64,
    DecPost = 65,
    DefVar = 66,
    DefInitVar = 67,
    PutLexicalValue = 68,
    ThrowMutateImmutable = 69,
    GetArgument = 70,
    GetName = 71,
    GetLocator = 72,
    GetNameAndLocator = 73,
    GetNameOrUndefined = 74,
    SetName = 75,
    SetNameByLocator = 76,
    DeleteName = 77,
    GetPropertyByName = 78,
    GetPropertyByValue = 79,
    GetPropertyByValuePush = 80,
    SetPropertyByName = 81,
    SetFunctionName = 82,
    DefineOwnPropertyByName = 83,
    DefineClassStaticMethodByName = 84,
    DefineClassMethodByName = 85,
    SetPropertyByValue = 86,
    DefineOwnPropertyByValue = 87,
    DefineClassStaticMethodByValue = 88,
    DefineClassMethodByValue = 89,
    SetPropertyGetterByName = 90,
    DefineClassStaticGetterByName = 91,
    DefineClassGetterByName = 92,
    SetPropertyGetterByValue = 93,
    DefineClassStaticGetterByValue = 94,
    DefineClassGetterByValue = 95,
    SetPropertySetterByName = 96,
    DefineClassStaticSetterByName = 97,
    DefineClassSetterByName = 98,
    SetPropertySetterByValue = 99,
    DefineClassStaticSetterByValue = 100,
    DefineClassSetterByValue = 101,
    SetPrivateField = 102,
    DefinePrivateField = 103,
    SetPrivateMethod = 104,
    SetPrivateSetter = 105,
    SetPrivateGetter = 106,
    GetPrivateField = 107,
    PushClassField = 108,
    PushClassFieldPrivate = 109,
    PushClassPrivateGetter = 110,
    PushClassPrivateSetter = 111,
    PushClassPrivateMethod = 112,
    DeletePropertyByName = 113,
    DeletePropertyByValue = 114,
    DeleteSuperThrow = 115,
    CopyDataProperties = 116,
    ToPropertyKey = 117,
    Jump = 118,
    JumpIfTrue = 119,
    JumpIfFalse = 120,
    JumpIfNotUndefined = 121,
    JumpIfNullOrUndefined = 122,
    JumpTable = 123,
    Throw = 124,
    ReThrow = 125,
    Exception = 126,
    MaybeException = 127,
    ThrowNewTypeError = 128,
    ToBoolean = 129,
    This = 130,
    Super = 131,
    SuperCallPrepare = 132,
    SuperCall = 133,
    SuperCallSpread = 134,
    SuperCallDerived = 135,
    BindThisValue = 136,
    ImportCall = 137,
    Case = 138,
    Default = 139,
    GetFunction = 140,
    CallEval = 141,
    CallEvalSpread = 142,
    Call = 143,
    CallSpread = 144,
    New = 145,
    NewSpread = 146,
    CheckReturn = 147,
    Return = 148,
    AsyncGeneratorClose = 149,
    Generator = 150,
    GetReturnValue = 151,
    SetReturnValue = 152,
    PushDeclarativeEnvironment = 153,
    PushObjectEnvironment = 154,
    PopEnvironment = 155,
    IncrementLoopIteration = 156,
    CreateForInIterator = 157,
    GetIterator = 158,
    GetAsyncIterator = 159,
    IteratorNext = 160,
    IteratorNextWithoutPop = 161,
    IteratorDone = 162,
    IteratorFinishAsyncNext = 163,
    IteratorValue = 164,
    IteratorValueWithoutPop = 165,
    IteratorResult = 166,
    IteratorToArray = 167,
    IteratorPop = 168,
    IteratorStackEmpty = 169,
    CreateIteratorResult = 170,
    IteratorReturn = 171,
    ConcatToString = 172,
    RequireObjectCoercible = 173,
    ValueNotNullOrUndefined = 174,
    RestParameterInit = 175,
    GeneratorYield = 176,
    GeneratorNext = 177,
    AsyncGeneratorYield = 178,
    CreatePromiseCapability = 179,
    CompletePromiseCapability = 180,
    JumpIfNotResumeKind = 181,
    GeneratorDelegateNext = 182,
    GeneratorDelegateResume = 183,
    Await = 184,
    NewTarget = 185,
    ImportMeta = 186,
    IsObject = 187,
    TemplateLookup = 188,
    TemplateCreate = 189,
    PushPrivateEnvironment = 190,
    PopPrivateEnvironment = 191,
    CreateMappedArgumentsObject = 192,
    CreateUnmappedArgumentsObject = 193,
    CreateGlobalFunctionBinding = 194,
    Nop = 195,
    U16Operands = 196,
    U32Operands = 197,
    Reserved1 = 198,
    Reserved2 = 199,
    Reserved3 = 200,
    Reserved4 = 201,
    Reserved5 = 202,
    Reserved6 = 203,
    Reserved7 = 204,
    Reserved8 = 205,
    Reserved9 = 206,
    Reserved10 = 207,
    Reserved11 = 208,
    Reserved12 = 209,
    Reserved13 = 210,
    Reserved14 = 211,
    Reserved15 = 212,
    Reserved16 = 213,
    Reserved17 = 214,
    Reserved18 = 215,
    Reserved19 = 216,
    Reserved20 = 217,
    Reserved21 = 218,
    Reserved22 = 219,
    Reserved23 = 220,
    Reserved24 = 221,
    Reserved25 = 222,
    Reserved26 = 223,
    Reserved27 = 224,
    Reserved28 = 225,
    Reserved29 = 226,
    Reserved30 = 227,
    Reserved31 = 228,
    Reserved32 = 229,
    Reserved33 = 230,
    Reserved34 = 231,
    Reserved35 = 232,
    Reserved36 = 233,
    Reserved37 = 234,
    Reserved38 = 235,
    Reserved39 = 236,
    Reserved40 = 237,
    Reserved41 = 238,
    Reserved42 = 239,
    Reserved43 = 240,
    Reserved44 = 241,
    Reserved45 = 242,
    Reserved46 = 243,
    Reserved47 = 244,
    Reserved48 = 245,
    Reserved49 = 246,
    Reserved50 = 247,
    Reserved51 = 248,
    Reserved52 = 249,
    Reserved53 = 250,
    Reserved54 = 251,
    Reserved55 = 252,
    Reserved56 = 253,
    Reserved57 = 254,
    Reserved58 = 255,

}

The opcodes of the vm.

Variants

Pop = 0

Pop the top value from the stack.

Operands:

Stack: value =>

Dup = 1

Push a copy of the top value on the stack.

Operands:

Stack: value => value, value

Swap = 2

Swap the top two values on the stack.

Operands:

Stack: second, first => first, second

RotateLeft = 3

Rotates the top n values of the stack to the left by 1.

Equivalent to calling slice::rotate_left with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[n-1], … , v[1], v[0], v[n]

RotateRight = 4

Rotates the top n values of the stack to the right by 1.

Equivalent to calling slice::rotate_right with argument 1 on the top n values of the stack.

Operands: n: u8

Stack: v[n], v[n-1], … , v[1], v[0] => v[0], v[n], v[n-1], … , v[1]

PushZero = 5

Push integer 0 on the stack.

Operands:

Stack: => 0

PushOne = 6

Push integer 1 on the stack.

Operands:

Stack: => 1

PushInt8 = 7

Push i8 value on the stack.

Operands: value: i8

Stack: => value

PushInt16 = 8

Push i16 value on the stack.

Operands: value: i16

Stack: => value

PushInt32 = 9

Push i32 value on the stack.

Operands: value: i32

Stack: => value

PushFloat = 10

Push f32 value on the stack.

Operands: value: f32

Stack: => value

PushDouble = 11

Push f64 value on the stack.

Operands: value: f64

Stack: => value

PushNaN = 12

Push NaN integer on the stack.

Operands:

Stack: => NaN

PushPositiveInfinity = 13

Push Infinity value on the stack.

Operands:

Stack: => Infinity

PushNegativeInfinity = 14

Push -Infinity value on the stack.

Operands:

Stack: => -Infinity

PushNull = 15

Push null value on the stack.

Operands:

Stack: => null

PushTrue = 16

Push true value on the stack.

Operands:

Stack: => true

PushFalse = 17

Push false value on the stack.

Operands:

Stack: => false

PushUndefined = 18

Push undefined value on the stack.

Operands:

Stack: => undefined

PushLiteral = 19

Push literal value on the stack.

Like strings and bigints. The index operand is used to index into the literals array to get the value.

Operands: index: VaryingOperand

Stack: => (literals[index])

PushRegExp = 20

Push regexp value on the stack.

Operands: pattern_index: VaryingOperand, flags: VaryingOperand

Stack: => regexp

PushEmptyObject = 21

Push empty object {} value on the stack.

Operands:

Stack: => {}

PushClassPrototype = 22

Get the prototype of a superclass and push it on the stack.

Additionally this sets the [[prototype]] of the class and the DERIVED flag.

Operands:

Stack: class, superclass => class, superclass.prototype

SetClassPrototype = 23

Set the prototype of a class object.

Operands:

Stack: class, prototype => class.prototype

SetHomeObject = 24

Set home object internal slot of an object literal method.

Operands:

Stack: home, function => home, function

SetPrototype = 25

Set the prototype of an object if the value is an object or null.

Operands:

Stack: object, value =>

PushNewArray = 26

Push an empty array value on the stack.

Operands:

Stack: => []

PushValueToArray = 27

Push a value to an array.

Operands:

Stack: array, value => array

PushElisionToArray = 28

Push an empty element/hole to an array.

Operands:

Stack: array => array

PushIteratorToArray = 29

Push all iterator values to an array.

Operands:

Stack: array, iterator, next_method => array

Add = 30

Binary + operator.

Operands:

Stack: lhs, rhs => (lhs + rhs)

Sub = 31

Binary - operator.

Operands:

Stack: lhs, rhs => (lhs - rhs)

Div = 32

Binary / operator.

Operands:

Stack: lhs, rhs => (lhs / rhs)

Mul = 33

Binary * operator.

Operands:

Stack: lhs, rhs => (lhs * rhs)

Mod = 34

Binary % operator.

Operands:

Stack: lhs, rhs => (lhs % rhs)

Pow = 35

Binary ** operator.

Operands:

Stack: lhs, rhs => (lhs ** rhs)

ShiftRight = 36

Binary >> operator.

Operands:

Stack: lhs, rhs => (lhs >> rhs)

ShiftLeft = 37

Binary << operator.

Operands:

Stack: lhs, rhs => (lhs << rhs)

UnsignedShiftRight = 38

Binary >>> operator.

Operands:

Stack: lhs, rhs => (lhs >>> rhs)

BitOr = 39

Binary bitwise | operator.

Operands:

Stack: lhs, rhs => (lhs | rhs)

BitAnd = 40

Binary bitwise & operator.

Operands:

Stack: lhs, rhs => (lhs & rhs)

BitXor = 41

Binary bitwise ^ operator.

Operands:

Stack: lhs, rhs => (lhs ^ rhs)

BitNot = 42

Unary bitwise ~ operator.

Operands:

Stack: value => ~value

In = 43

Binary in operator.

Operands:

Stack: lhs, rhs => (lhs in rhs)

InPrivate = 44

Binary in operator for private names.

Operands: index: u32

Stack: rhs => (private_name in rhs)

Eq = 45

Binary == operator.

Operands:

Stack: lhs, rhs => (lhs == rhs)

StrictEq = 46

Binary === operator.

Operands:

Stack: lhs, rhs => (lhs === rhs)

NotEq = 47

Binary != operator.

Operands:

Stack: lhs, rhs => (lhs != rhs)

StrictNotEq = 48

Binary !== operator.

Operands:

Stack: lhs, rhs => (lhs !== rhs)

GreaterThan = 49

Binary > operator.

Operands:

Stack: lhs, rhs => (lhs > rhs)

GreaterThanOrEq = 50

Binary >= operator.

Operands:

Stack: lhs, rhs => (lhs >= rhs)

LessThan = 51

Binary < operator.

Operands:

Stack: lhs, rhs => (lhs < rhs)

LessThanOrEq = 52

Binary <= operator.

Operands:

Stack: lhs, rhs => (lhs <= rhs)

InstanceOf = 53

Binary instanceof operator.

Operands:

Stack: lhs, rhs => (lhs instanceof rhs)

LogicalAnd = 54

Binary logical && operator.

This is a short-circuit operator, if the lhs value is false, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs && rhs)

LogicalOr = 55

Binary logical || operator.

This is a short-circuit operator, if the lhs value is true, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs || rhs)

Coalesce = 56

Binary ?? operator.

This is a short-circuit operator, if the lhs value is not null or undefined, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs ?? rhs)

TypeOf = 57

Unary typeof operator.

Operands:

Stack: value => (typeof value)

Void = 58

Unary void operator.

Operands:

Stack: value => undefined

LogicalNot = 59

Unary logical ! operator.

Operands:

Stack: value => (!value)

Pos = 60

Unary + operator.

Operands:

Stack: value => (+value)

Neg = 61

Unary - operator.

Operands:

Stack: value => (-value)

Inc = 62

Unary ++ operator.

Operands:

Stack: value => (value + 1)

IncPost = 63

Unary postfix ++ operator.

Operands:

Stack: value => (ToNumeric(value)), (value + 1)

Dec = 64

Unary -- operator.

Operands:

Stack: value => (value - 1)

DecPost = 65

Unary postfix -- operator.

Operands:

Stack: value => (ToNumeric(value)), (value - 1)

DefVar = 66

Declare var type variable.

Operands: index: u32

Stack: =>

DefInitVar = 67

Declare and initialize var type variable.

Operands: index: u32

Stack: value =>

PutLexicalValue = 68

Initialize a lexical binding.

Operands: index: u32

Stack: value =>

ThrowMutateImmutable = 69

Throws an error because the binding access is illegal.

Operands: index: u32

Stack: =>

GetArgument = 70

Get i-th argument of the current frame.

Returns undefined if arguments.len() < index.

Operands: index: u32

Stack: => value

GetName = 71

Find a binding on the environment chain and push its value.

Operands: index: u32

Stack: => value

GetLocator = 72

Find a binding on the environment and set the current_binding of the current frame.

Operands: index: u32

Stack: =>

GetNameAndLocator = 73

Find a binding on the environment chain and push its value to the stack and its BindingLocator to the bindings_stack.

Operands: index: u32

Stack: => value

GetNameOrUndefined = 74

Find a binding on the environment chain and push its value. If the binding does not exist push undefined.

Operands: index: u32

Stack: => value

SetName = 75

Find a binding on the environment chain and assign its value.

Operands: index: u32

Stack: value =>

SetNameByLocator = 76

Assigns a value to the binding pointed by the top of the bindings_stack.

Stack: value =>

DeleteName = 77

Deletes a property of the global object.

Operands: index: u32

Stack: => deleted

GetPropertyByName = 78

Get a property by name from an object an push it on the stack.

Like object.name

Operands: index: u32

Stack: object, receiver => value

GetPropertyByValue = 79

Get a property by value from an object an push it on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => value

GetPropertyByValuePush = 80

Get a property by value from an object an push the key and value on the stack.

Like object[key]

Operands:

Stack: object, receiver, key => key, value

SetPropertyByName = 81

Sets a property by name of an object.

Like object.name = value

Operands: index: u32

Stack: object, receiver, value => value

SetFunctionName = 82

Sets the name of a function object.

This operation is corresponds to the SetFunctionName abstract operation in the spec.

The prefix operand is mapped as follows:

0 -> no prefix

1 -> “get “

2 -> “set “

Operands: prefix: u8

Stack: name, function => function

DefineOwnPropertyByName = 83

Defines a own property of an object by name.

Operands: index: u32

Stack: object, value =>

DefineClassStaticMethodByName = 84

Defines a static class method by name.

Operands: index: u32

Stack: class, function =>

DefineClassMethodByName = 85

Defines a class method by name.

Operands: index: u32

Stack: class_proto, function =>

SetPropertyByValue = 86

Sets a property by value of an object.

Like object[key] = value

Operands:

Stack: object, receiver, key, value => value

DefineOwnPropertyByValue = 87

Defines a own property of an object by value.

Operands:

Stack: object, key, value =>

DefineClassStaticMethodByValue = 88

Defines a static class method by value.

Operands:

Stack: class, key, function =>

DefineClassMethodByValue = 89

Defines a class method by value.

Operands:

Stack: class_proto, key, function =>

SetPropertyGetterByName = 90

Sets a getter property by name of an object.

Like get name() value

Operands: index: u32

Stack: object, value =>

DefineClassStaticGetterByName = 91

Defines a static getter class method by name.

Like static get name() value

Operands: index: u32

Stack: class, binding_function =>

DefineClassGetterByName = 92

Defines a getter class method by name.

Like get name() value

Operands: index: u32

Stack: class_proto, function => class

SetPropertyGetterByValue = 93

Sets a getter property by value of an object.

Like get [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticGetterByValue = 94

Defines a static getter class method by value.

Like static get [key]() value

Operands:

Stack: class, key, function =>

DefineClassGetterByValue = 95

Defines a getter class method by value.

Like get [key]() value

Operands:

Stack: class_proto, key, function =>

SetPropertySetterByName = 96

Sets a setter property by name of an object.

Like set name() value

Operands: index: u32

Stack: object, value =>

DefineClassStaticSetterByName = 97

Defines a static setter class method by name.

Like static set name() value

Operands: index: u32

Stack: class, function =>

DefineClassSetterByName = 98

Defines a setter class method by name.

Like set name() value

Operands: index: u32

Stack: class_proto, function =>

SetPropertySetterByValue = 99

Sets a setter property by value of an object.

Like set [key]() value

Operands:

Stack: object, key, value =>

DefineClassStaticSetterByValue = 100

Defines a static setter class method by value.

Like static set [key]() value

Operands:

Stack: class, key, function =>

DefineClassSetterByValue = 101

Defines a setter class method by value.

Like set [key]() value

Operands:

Stack: class_proto, key, function =>

SetPrivateField = 102

Set the value of a private property of an object by it’s name.

Like obj.#name = value

Operands: index: u32

Stack: object, value => value

DefinePrivateField = 103

Define a private property of a class constructor by it’s name.

Like #name = value

Operands: index: u32

Stack: object, value =>

SetPrivateMethod = 104

Set a private method of a class constructor by it’s name.

Like #name() {}

Operands: index: u32

Stack: object, value =>

SetPrivateSetter = 105

Set a private setter property of a class constructor by it’s name.

Like set #name() {}

Operands: index: u32

Stack: object, value =>

SetPrivateGetter = 106

Set a private getter property of a class constructor by it’s name.

Like get #name() {}

Operands: index: u32

Stack: object, value =>

GetPrivateField = 107

Get a private property by name from an object an push it on the stack.

Like object.#name

Operands: index: u32

Stack: object => value

PushClassField = 108

Push a field to a class.

Operands:

Stack: class, field_name, field_function =>

PushClassFieldPrivate = 109

Push a private field to the class.

Operands: index: u32

Stack: class, field_function =>

PushClassPrivateGetter = 110

Push a private getter to the class.

Operands: index: u32

Stack: class, getter =>

PushClassPrivateSetter = 111

Push a private setter to the class.

Operands: index: u32

Stack: class, setter =>

PushClassPrivateMethod = 112

Push a private method to the class.

Operands: index: u32

Stack: class, method =>

DeletePropertyByName = 113

Deletes a property by name of an object.

Like delete object.key

Operands: index: u32

Stack: object =>

DeletePropertyByValue = 114

Deletes a property by value of an object.

Like delete object[key]

Operands:

Stack: object, key =>

DeleteSuperThrow = 115

Throws an error when trying to delete a property of super

Operands:

Stack: =>

CopyDataProperties = 116

Copy all properties of one object to another object.

Operands: excluded_key_count: VaryingOperand, excluded_key_count_computed: VaryingOperand

Stack: excluded_key_computed_0 … excluded_key_computed_n, source, value, excluded_key_0 … excluded_key_n => value

ToPropertyKey = 117

Call ToPropertyKey on the value on the stack.

Operands:

Stack: value => key

Jump = 118

Unconditional jump to address.

Operands: address: u32

Stack: =>

JumpIfTrue = 119

Conditional jump to address.

If the value popped is truthy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfFalse = 120

Conditional jump to address.

If the value popped is falsy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfNotUndefined = 121

Conditional jump to address.

If the value popped is not undefined jump to address.

Operands: address: u32

Stack: value => value

JumpIfNullOrUndefined = 122

Conditional jump to address.

If the value popped is undefined jump to address.

Operands: address: u32

Stack: value => value

JumpTable = 123

Jump table that jumps depending on top value of the stack.

This is used to handle special cases when we call continue, break or return in a try block, that has finally block.

Operands: default: u32, count: u32, address: u32 * count

Stack: value: i32 =>

Throw = 124

Throw exception.

This sets pending exception and searches for an exception handler.

Operands:

Stack: value =>

ReThrow = 125

Rethrow thrown exception.

This is also used to handle generator return() call, we throw an empty exception, by setting pending exception to None, propagating it and calling finally code until there is no exception handler left, in that case we consume the empty exception and return from the generator.

Operands:

Stack: =>

Exception = 126

Get the thrown pending exception (if it’s set) and push on the stack.

If there is no pending exception, which can happend if we are handling return() call on generator, then we rethrow the empty exception. See Opcode::ReThrow.

Operands:

Stack: => exception

MaybeException = 127

Get the thrown pending exception if it’s set and push true, otherwise push only false.

Operands:

Stack: => (true, exception) or false

ThrowNewTypeError = 128

Throw a new TypeError exception

Operands: message: u32

Stack: =>

ToBoolean = 129

Pops value converts it to boolean and pushes it back.

Operands:

Stack: value => (ToBoolean(value))

This = 130

Pushes this value

Operands:

Stack: => this

Super = 131

Pushes the current super value to the stack.

Operands:

Stack: => super

SuperCallPrepare = 132

Get the super constructor and the new target of the current environment.

Operands:

Stack: => super_constructor

SuperCall = 133

Execute the super() method.

Operands: argument_count: u32

Stack: super_constructor, argument_1, … argument_n =>

SuperCallSpread = 134

Execute the super() method where the arguments contain spreads.

Operands:

Stack: super_constructor, arguments_array =>

SuperCallDerived = 135

Execute the super() method when no constructor of the class is defined.

Operands:

Stack: argument_n, … argument_1 =>

BindThisValue = 136

Binds this value and initializes the instance elements.

Performs steps 7-12 of SuperCall: super Arguments

Operands:

Stack: result => result

ImportCall = 137

Dynamically import a module.

Operands:

Stack: specifier => promise

Case = 138

Pop the two values of the stack, strict equal compares the two values, if true jumps to address, otherwise push the second pop’ed value.

Operands: address: u32

Stack: value, cond => cond (if cond !== value).

Default = 139

Pops the top of stack and jump to address.

Operands: address: u32

Stack: value =>

GetFunction = 140

Get function from the pre-compiled inner functions.

Operands: index: VaryingOperand

Stack: => func

CallEval = 141

Call a function named “eval”.

Operands: argument_count: VaryingOperand

Stack: this, func, argument_1, … argument_n => result

CallEvalSpread = 142

Call a function named “eval” where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

Call = 143

Call a function.

Operands: argument_count: u32

Stack: this, func, argument_1, … argument_n => result

CallSpread = 144

Call a function where the arguments contain spreads.

Operands:

Stack: this, func, arguments_array => result

New = 145

Call construct on a function.

Operands: argument_count: u32

Stack: func, argument_1, … argument_n => result

NewSpread = 146

Call construct on a function where the arguments contain spreads.

Operands:

Stack: arguments_array, func => result

CheckReturn = 147

Check return from a function.

Operands:

Stack: =>

Return = 148

Return from a function.

Operands:

Stack: =>

AsyncGeneratorClose = 149

Close an async generator function.

Operands:

Stack: =>

Generator = 150

Creates the generator object and yields.

Operands: async: u8

Stack: => resume_kind

GetReturnValue = 151

Get return value of a function.

Operands:

Stack: => value

SetReturnValue = 152

Set return value of a function.

Operands:

Stack: value =>

PushDeclarativeEnvironment = 153

Push a declarative environment.

Operands: compile_environments_index: VaryingOperand

Stack: =>

PushObjectEnvironment = 154

Push an object environment.

Operands:

Stack: object =>

PopEnvironment = 155

Pop the current environment.

Operands:

Stack: =>

IncrementLoopIteration = 156

Increment loop itearation count.

Used for limiting the loop iteration.

Operands:

Stack: =>

CreateForInIterator = 157

Creates the ForInIterator of an object.

Stack: object =>

Iterator Stack: iterator

GetIterator = 158

Gets the iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

GetAsyncIterator = 159

Gets the async iterator of an object.

Operands:

Stack: object =>

Iterator Stack: iterator

IteratorNext = 160

Calls the next method of iterator, updating its record with the next value.

Operands:

Iterator Stack: iterator => iterator

IteratorNextWithoutPop = 161

Calls the next method of iterator, updating its record with the next value.

Operands:

Iterator Stack: iterator => iterator

IteratorDone = 162

Returns true if the current iterator is done, or false otherwise

Stack: => done

Iterator Stack: iterator => iterator

IteratorFinishAsyncNext = 163

Finishes the call to Opcode::IteratorNext within a for await loop by setting the current result of the current iterator.

Operands:

Stack: next_result, resume_kind => resume_kind

Iterator Stack: iterator => iterator

IteratorValue = 164

Gets the value property of the current iterator record.

Stack: => value

Iterator Stack: iterator => iterator

IteratorValueWithoutPop = 165

Gets the value property of the current iterator record.

Stack: => value

Iterator Stack: iterator => iterator

IteratorResult = 166

Gets the last iteration result of the current iterator record.

Stack: => result

Iterator Stack: iterator => iterator

IteratorToArray = 167

Consume the iterator and construct and array with all the values.

Operands:

Stack: => array

Iterator Stack: iterator => iterator

IteratorPop = 168

Pop an iterator from the call frame close iterator stack.

Iterator Stack:

iterator =>

IteratorStackEmpty = 169

Pushes true to the stack if the iterator stack is empty.

Stack:

=> is_empty

Iterator Stack:

=>

CreateIteratorResult = 170

Creates a new iterator result object.

Operands:

done: bool (codified as u8 with 0 -> false and !0 -> true)

Stack:

value =>

IteratorReturn = 171

Calls return on the current iterator and returns the result.

Stack: => return_val (if return is a method), is_return_method

Iterator Stack: iterator =>

ConcatToString = 172

Concat multiple stack objects into a string.

Operands: value_count: u32

Stack: value_1,…value_n => string

RequireObjectCoercible = 173

Call RequireObjectCoercible on the stack value.

Operands:

Stack: value => value

ValueNotNullOrUndefined = 174

Require the stack value to be neither null nor undefined.

Operands:

Stack: value => value

RestParameterInit = 175

Initialize the rest parameter value of a function from the remaining arguments.

Operands:

Stack: argument_1 .. argument_n => array

GeneratorYield = 176

Yields from the current generator execution.

Operands:

Stack: value => resume_kind, received

GeneratorNext = 177

Resumes the current generator function.

If the resume_kind is Throw, then the value is poped and thrown, otherwise if Return we pop the value, set it as the return value and throw and empty exception. See Opcode::ReThrow.

Operands:

Stack: resume_kind, value => value

AsyncGeneratorYield = 178

Yields from the current async generator execution.

Operands:

Stack: value => received

CreatePromiseCapability = 179

Create a promise capacity for an async function, if not already set.

Operands:

Stack: =>

CompletePromiseCapability = 180

Resolves or rejects the promise capability of an async function.

If the pending exception is set, reject and rethrow the exception, otherwise resolve.

Operands:

Stack: =>

JumpIfNotResumeKind = 181

Jumps to the specified address if the resume kind is not equal.

Operands: exit: u32, resume_kind: u8.

Stack: resume_kind => resume_kind

GeneratorDelegateNext = 182

Delegates the current async generator function to another iterator.

Operands: throw_method_undefined: u32, return_method_undefined: u32

Stack: received => result

GeneratorDelegateResume = 183

Resume the async generator with yield delegate logic after it awaits a value.

Operands: return: u32, exit: u32

Stack: is_return, received => value

Await = 184

Stops the current async function and schedules it to resume later.

Operands:

Stack: promise => received

NewTarget = 185

Push the current new target to the stack.

Operands:

Stack: => new.target

ImportMeta = 186

Push the current import.meta to the stack.

Operands:

Stack: => import.meta

IsObject = 187

Pushes true to the stack if the top stack value is an object, or false otherwise.

Operands:

Stack: value => is_object

TemplateLookup = 188

Lookup if a tagged template object is cached and skip the creation if it is.

Operands: exit: u32, site: u64

Stack: => template (if cached)

TemplateCreate = 189

Create a new tagged template object and cache it.

Operands: count: VaryingOperand, site: u64

Stack: count * (cooked_value, raw_value) => template

PushPrivateEnvironment = 190

Push a private environment.

Operands: count: u32, count * name_index: u32

Stack: class => class

PopPrivateEnvironment = 191

Pop a private environment.

Operands:

Stack: =>

CreateMappedArgumentsObject = 192

Creates a mapped arguments object.

Performs [10.4.4.7 CreateMappedArgumentsObject ( func, formals, argumentsList, env )]

Operands:

Stack: => arguments

CreateUnmappedArgumentsObject = 193

Creates an unmapped arguments object.

Performs: [10.4.4.6 CreateUnmappedArgumentsObject ( argumentsList )]

Stack: => arguments

CreateGlobalFunctionBinding = 194

Performs CreateGlobalFunctionBinding ( N, V, D )

Operands: configurable: bool, name_index: VaryingOperand

Stack: value =>

Nop = 195

No-operation instruction, does nothing.

Operands:

Stack: =>

U16Operands = 196

Opcode prefix modifier, makes all VaryingOperands of an instruction u16 sized.

Operands: opcode (operands if any).

Stack: The stack changes based on the opcode that is being prefixed.

U32Operands = 197

Opcode prefix modifier, Opcode prefix operand modifier, makes all VaryingOperands of an instruction u32 sized.

Operands: opcode (operands if any).

Stack: The stack changes based on the opcode that is being prefixed.

Reserved1 = 198

Reserved Opcode.

Reserved2 = 199

Reserved Opcode.

Reserved3 = 200

Reserved Opcode.

Reserved4 = 201

Reserved Opcode.

Reserved5 = 202

Reserved Opcode.

Reserved6 = 203

Reserved Opcode.

Reserved7 = 204

Reserved Opcode.

Reserved8 = 205

Reserved Opcode.

Reserved9 = 206

Reserved Opcode.

Reserved10 = 207

Reserved Opcode.

Reserved11 = 208

Reserved Opcode.

Reserved12 = 209

Reserved Opcode.

Reserved13 = 210

Reserved Opcode.

Reserved14 = 211

Reserved Opcode.

Reserved15 = 212

Reserved Opcode.

Reserved16 = 213

Reserved Opcode.

Reserved17 = 214

Reserved Opcode.

Reserved18 = 215

Reserved Opcode.

Reserved19 = 216

Reserved Opcode.

Reserved20 = 217

Reserved Opcode.

Reserved21 = 218

Reserved Opcode.

Reserved22 = 219

Reserved Opcode.

Reserved23 = 220

Reserved Opcode.

Reserved24 = 221

Reserved Opcode.

Reserved25 = 222

Reserved Opcode.

Reserved26 = 223

Reserved Opcode.

Reserved27 = 224

Reserved Opcode.

Reserved28 = 225

Reserved Opcode.

Reserved29 = 226

Reserved Opcode.

Reserved30 = 227

Reserved Opcode.

Reserved31 = 228

Reserved Opcode.

Reserved32 = 229

Reserved Opcode.

Reserved33 = 230

Reserved Opcode.

Reserved34 = 231

Reserved Opcode.

Reserved35 = 232

Reserved Opcode.

Reserved36 = 233

Reserved Opcode.

Reserved37 = 234

Reserved Opcode.

Reserved38 = 235

Reserved Opcode.

Reserved39 = 236

Reserved Opcode.

Reserved40 = 237

Reserved Opcode.

Reserved41 = 238

Reserved Opcode.

Reserved42 = 239

Reserved Opcode.

Reserved43 = 240

Reserved Opcode.

Reserved44 = 241

Reserved Opcode.

Reserved45 = 242

Reserved Opcode.

Reserved46 = 243

Reserved Opcode.

Reserved47 = 244

Reserved Opcode.

Reserved48 = 245

Reserved Opcode.

Reserved49 = 246

Reserved Opcode.

Reserved50 = 247

Reserved Opcode.

Reserved51 = 248

Reserved Opcode.

Reserved52 = 249

Reserved Opcode.

Reserved53 = 250

Reserved Opcode.

Reserved54 = 251

Reserved Opcode.

Reserved55 = 252

Reserved Opcode.

Reserved56 = 253

Reserved Opcode.

Reserved57 = 254

Reserved Opcode.

Reserved58 = 255

Reserved Opcode.

Implementations

impl Opcode

const MAX: usize = 256usize

const NAMES: [&'static str; 768] = _

pub(crate) const fn as_str(self) -> &'static str

Name of this opcode.

const INSTRUCTIONS: [&'static str; 768] = _

pub(crate) const fn as_instruction_str(self) -> &'static str

Name of the profiler event for this opcode.

const SPEND_FNS: [fn(_: &mut Context, _: &mut u32) -> JsResult<CompletionType>; 256] = _

pub(super) fn spend_budget_and_execute( self, context: &mut Context, budget: &mut u32 ) -> JsResult<CompletionType>

Spends the cost of this opcode into the provided budget and executes it.

const COSTS: [u8; 256] = _

pub(super) const fn cost(self) -> u8

Return the cost of this opcode.

const EXECUTE_FNS: [fn(_: &mut Context) -> JsResult<CompletionType>; 768] = _

pub(super) fn execute(self, context: &mut Context) -> JsResult<CompletionType>