Trait boa_engine::vm::opcode::Operation

pub(crate) trait Operation {
    const NAME: &'static str;
    const INSTRUCTION: &'static str;

    // Required method
    fn execute(context: &mut Context<'_>) -> JsResult<CompletionType>;

    // Provided methods
    fn execute_with_u16_operands(
        context: &mut Context<'_>
    ) -> JsResult<CompletionType> { ... }
    fn execute_with_u32_operands(
        context: &mut Context<'_>
    ) -> JsResult<CompletionType> { ... }
}

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.

This trait should be implemented for a struct that corresponds with any arm of the OpCode enum.

Required Associated Constants

const NAME: &'static str

const INSTRUCTION: &'static str

Required Methods

fn execute(context: &mut Context<'_>) -> JsResult<CompletionType>

Execute opcode with VaryingOperandKind::U8 sized VaryingOperands.

Provided Methods

fn execute_with_u16_operands( context: &mut Context<'_> ) -> JsResult<CompletionType>

Execute opcode with VaryingOperandKind::U16 sized VaryingOperands.

By default if not implemented will call [Reserved::u16_execute()] which panics.

fn execute_with_u32_operands( context: &mut Context<'_> ) -> JsResult<CompletionType>

Execute opcode with VaryingOperandKind::U32 sized VaryingOperands.

By default if not implemented will call [Reserved::u32_execute()] which panics.