Struct boa_engine::vm::CodeBlock

pub struct CodeBlock {

    pub(crate) name: Sym,
    pub(crate) length: u32,
    pub(crate) strict: bool,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) code: Vec<u8>,
    pub(crate) literals: Vec<JsValue>,
    pub(crate) names: Vec<Sym>,
    pub(crate) bindings: Vec<BindingLocator>,
    pub(crate) num_bindings: usize,
    pub(crate) functions: Vec<Gc<Self>>,
    pub(crate) arguments_binding: Option<BindingLocator>,
    pub(crate) compile_environments: Vec<Gc<Cell<CompileTimeEnvironment>>>,
    pub(crate) is_class_constructor: bool,

}

The internal representation of a JavaScript function.

A CodeBlock is generated for each function compiled by the ByteCompiler. It stores the bytecode and the other attributes of the function.

Fields
name: Sym

Name of this function

length: u32

The number of arguments expected.

strict: bool

Is this function in strict mode.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

code: Vec<u8>

Bytecode

literals: Vec<JsValue>

Literals

names: Vec<Sym>

Property field names.

bindings: Vec<BindingLocator>

Locators for all bindings in the codeblock.

num_bindings: usize

Number of binding for the function environment.

functions: Vec<Gc<Self>>

Functions inside this function

arguments_binding: Option<BindingLocator>

The arguments binding location of the function, if set.

compile_environments: Vec<Gc<Cell<CompileTimeEnvironment>>>

Compile time environments in this function.

is_class_constructor: bool

The [[IsClassConstructor]] internal slot.

Implementations

impl CodeBlock

pub fn new(name: Sym, length: u32, strict: bool) -> Self

Constructs a new CodeBlock.

pub(crate) unsafe fn read_unchecked<T>(&self, offset: usize) -> T where
T: Readable,

Read type T from code.

Safety

Does not check if read happens out-of-bounds.

pub(crate) fn read<T>(&self, offset: usize) -> T where
T: Readable,

Read type T from code.

pub(crate) fn instruction_operands(
&self,
pc: &mut usize,
interner: &Interner
) -> String

Get the operands after the Opcode pointed to by pc as a String. Modifies the pc to point to the next instruction.

Returns an empty String if no operands are present.