Struct boa_engine::vm::Vm

pub struct Vm {
    pub(crate) frames: Vec<CallFrame>,
    pub(crate) stack: Vec<JsValue>,
    pub(crate) return_value: JsValue,
    pub(crate) pending_exception: Option<JsError>,
    pub(crate) environments: EnvironmentStack,
    pub(crate) runtime_limits: RuntimeLimits,
    pub(crate) active_function: Option<JsObject>,
    pub(crate) active_runnable: Option<ActiveRunnable>,
    pub(crate) trace: bool,
}

Virtual Machine.

Fields
frames: Vec<CallFrame>stack: Vec<JsValue>return_value: JsValuepending_exception: Option<JsError>

When an error is thrown, the pending exception is set.

If we throw an empty exception (None), this means that return() was called on a generator, propagating though the exception handlers and executing the finally code (if any).

See ReThrow and ReThrow opcodes.

This is also used to eliminates crate::JsNativeError to opaque conversion if not needed.

environments: EnvironmentStackruntime_limits: RuntimeLimitsactive_function: Option<JsObject>active_runnable: Option<ActiveRunnable>trace: bool
Implementations

impl Vm

pub(crate) fn new(global: Gc<DeclarativeEnvironment>) -> Self

Creates a new virtual machine.

pub(crate) fn push<T>(&mut self, value: T)where T: Into<JsValue>,

Push a value on the stack.

pub(crate) fn pop(&mut self) -> JsValue

Pop a value off the stack.

Panics

If there is nothing to pop, then this will panic.

pub(crate) fn read<T: Readable>(&mut self) -> T

pub(crate) fn frame(&self) -> &CallFrame

Retrieves the VM frame

Panics

If there is no frame, then this will panic.

pub(crate) fn frame_mut(&mut self) -> &mut CallFrame

Retrieves the VM frame mutably

Panics

If there is no frame, then this will panic.

pub(crate) fn push_frame(&mut self, frame: CallFrame)

pub(crate) fn pop_frame(&mut self) -> Option<CallFrame>

pub(crate) fn handle_exception_at(&mut self, pc: u32) -> bool

Handles an exception thrown at position pc.

Returns true if the exception was handled, false otherwise.

pub(crate) fn get_return_value(&self) -> JsValue

pub(crate) fn set_return_value(&mut self, value: JsValue)