Crate boa_engine

Boa’s boa_engine crate implements ECMAScript’s standard library of builtin objects and an ECMAScript context, bytecompiler, and virtual machine for code execution.

Crate Features

serde - Enables serialization and deserialization of the AST (Abstract Syntax Tree).

console - Enables boa’s WHATWG console object implementation.

profiler - Enables profiling with measureme (this is mostly internal).

intl - Enables boa’s ECMA-402 Internationalization API (Intl object)

About Boa

Boa is an open-source, experimental ECMAScript Engine written in Rust for lexing, parsing and executing ECMAScript/JavaScript. Currently, Boa supports some of the language. More information can be viewed at Boa’s website.

Try out the most recent release with Boa’s live demo playground.

Boa Crates

boa_ast - Boa’s ECMAScript Abstract Syntax Tree.

boa_engine - Boa’s implementation of ECMAScript builtin objects and execution.

boa_gc - Boa’s garbage collector.

boa_interner - Boa’s string interner.

boa_parser - Boa’s lexer and parser.

boa_profiler - Boa’s code profiler.

boa_unicode - Boa’s Unicode identifier.

boa_icu_provider - Boa’s ICU4X data provider.

Modules

bigint

Boa’s implementation of ECMAScript’s bigint primitive type.

builtins

Boa’s ECMAScript built-in object implementations, e.g. Object, String, Math, Array, etc.

bytecompiler

This module contains the bytecode compiler.

class

Traits and structs for implementing native classes.

context

The ECMAScript context.

environments

Boa’s implementation of ECMAScript’s Environment Records.

error

Error-related types and conversions.

job

Data structures for the microtask job queue.

native_function

Boa’s wrappers for native Rust functions to be compatible with ECMAScript calls.

object

Boa’s representation of a JavaScript object and builtin object wrappers

prelude

A convenience module that re-exports the most commonly-used Boa APIs

property

Boa’s implementation of ECMAScript’s Property Descriptor.

realm

Boa’s implementation of ECMAScript’s Realm Records

string

A UTF-16–encoded, reference counted, immutable string.

symbol

Boa’s implementation of ECMAScript’s global Symbol object.

value

Boa’s ECMAScript Value implementation.

vm

Boa’s ECMAScript Virtual Machine

Macros

js_string

Utility macro to create a JsString.

Structs

Context

ECMAScript context. It is the primary way to interact with the runtime.

JsBigInt

JavaScript bigint primitive rust type.

JsError

The error type returned by all operations related to the execution of Javascript code.

JsNativeError

Native representation of an ideal Error object from Javascript.

JsString

A UTF-16–encoded, reference counted, immutable string.

JsSymbol

This represents a JavaScript symbol primitive.

Enums

JsNativeErrorKind

The list of possible error types a JsNativeError can be.

JsValue

A Javascript value

Type Definitions

JsResult

The result of a Javascript expression is represented like this so it can succeed (Ok) or fail (Err)