Module boa_engine::builtins

Boa’s ECMAScript built-in object implementations, e.g. Object, String, Math, Array, etc.

This module also contains a JavaScript Console implementation.

Modules

array

Boa’s implementation of ECMAScript’s global Array object.

array_buffer

Boa’s implementation of ECMAScript’s global ArrayBuffer object.

async_function

Boa’s implementation of ECMAScript’s global AsyncFunction object.

async_generator

Boa’s implementation of ECMAScript’s global AsyncGenerator object.

async_generator_function

Boa’s implementation of ECMAScript’s AsyncGeneratorFunction object.

bigint

Boa’s implementation of ECMAScript’s global BigInt object.

boolean

Boa’s implementation of ECMAScript’s global Boolean object.

console

Boa’s implementation of JavaScript’s console Web API object.

dataview

Boa’s implementation of ECMAScript’s global DataView object.

date

Boa’s implementation of ECMAScript’s Date object.

error

Boa’s implementation of ECMAScript’s global Error object.

eval

Boa’s implementation of ECMAScript’s global eval function.

function

Boa’s implementation of ECMAScript’s global Function object and Native Functions.

generator

Boa’s implementation of ECMAScript’s global Generator object.

generator_function

Boa’s implementation of ECMAScript’s global GeneratorFunction object.

global_this

Boa’s implementation of ECMAScript’s global globalThis property.

infinity

Boa’s implementation of ECMAScript’s global Infinity property.

intl

Boa’s implementation of ECMAScript’s global Intl object.

iterable

Boa’s implementation of ECMAScript’s IteratorRecord and iterator prototype objects.

json

Boa’s implementation of ECMAScript’s global JSON object.

map

Boa’s implementation of ECMAScript’s global Map object.

math

Boa’s implementation of ECMAScript’s global Math object.

nan

Boa’s implementation of ECMAScript’s global NaN property.

number

Boa’s implementation of ECMAScript’s global Number object.

object

Boa’s implementation of ECMAScript’s global Object object.

promise

Boa’s implementation of ECMAScript’s global Promise object.

proxy

Boa’s implementation of ECMAScript’s global Proxy object.

reflect

Boa’s implementation of ECMAScript’s global Reflect object.

regexp

Boa’s implementation of ECMAScript’s global RegExp object.

set

Boa’s implementation of ECMAScript’s global Set object.

string

Boa’s implementation of ECMAScript’s global String object.

symbol

Boa’s implementation of ECMAScript’s global Symbol object.

typed_array

Boa’s implementation of ECMAScript’s global TypedArray objects.

undefined

Boa’s implementation of ECMAScript’s global undefined property.

uri

Boa’s implementation of ECMAScript’s URI Handling Functions.

weak

Boa’s implementation of ECMAScript’s WeakRef object.

Traits

BuiltIn 🔒

Trait representing a global built-in object such as Math, Object or String.

JsArgs

A utility trait to make working with function arguments easier.

Functions

init

Initializes built-in objects and functions

init_builtin 🔒

Utility function that checks if a type implements BuiltIn before initializing it as a global built-in.