Module boa_engine::builtins

Builtins live here, such as Object, String, Math, etc.

Modules

array

This module implements the global Array object.

array_buffer

async_function

This module implements the global AsyncFunction object.

async_generator

This module implements the global AsyncGenerator object.

async_generator_function

This module implements the AsyncGeneratorFunction object.

bigint

This module implements the global BigInt object.

boolean

This module implements the global Boolean object.

console

This module implements the global console object.

dataview

date

error

This module implements the global Error object.

eval

This module implements the global eval function.

function

This module implements the global Function object as well as creates Native Functions.

generator

This module implements the global Generator object.

generator_function

This module implements the global GeneratorFunction object.

global_this

This module implements the global globalThis property.

infinity

This module implements the global Infinity property.

intl

This module implements the global Intl object.

iterable

json

This module implements the global JSON object.

map

This module implements the global Map object.

math

This module implements the global Math object.

nan

This module implements the global NaN property.

number

This module implements the global Number object.

object

This module implements the global Object object.

promise

This module implements the global Promise object.

proxy

This module implements the global Proxy object.

reflect

This module implements the global Reflect object.

regexp

This module implements the global RegExp object.

set

This module implements the global Set object.

string

This module implements the global String object.

symbol

This module implements the global Symbol object.

typed_array

This module implements the global TypedArray objects.

undefined

This module implements the global undefined property.

uri

URI Handling Functions

weak

Traits

BuiltIn 🔒

Trait representing a global built-in object such as Math, Object or String.

JsArgs

Functions

init

Initializes built-in objects and functions

init_builtin 🔒

Utility function that checks if a type implements BuiltIn before initializing it as a global built-in.