Module boa_engine::builtins::date

Boa’s implementation of ECMAScript’s Date object.

More information:

ECMAScript reference

MDN documentation

Modules

utils 🔒

Macros

get_mut_date 🔒

Gets a mutable reference to the inner Date object of val and stores it on var, or returns a TypeError if val is not a Date object.

some_or_nan 🔒

Extracts Some from an Option<T> or returns NaN if the object contains None.

Structs

Date

The internal representation of a Date object.

Functions

this_time_value 🔒

Abstract operation thisTimeValue.