Module boa_engine::builtins::error::type

This module implements the global TypeError object.

The TypeError object represents an error when an operation could not be performed, typically (but not exclusively) when a value is not of the expected type.

A TypeError may be thrown when:

an operand or argument passed to a function is incompatible with the type expected by that operator or function.

when attempting to modify a value that cannot be changed.

when attempting to use a value in an inappropriate way.

More information:

MDN documentation

ECMAScript reference

Structs

TypeError

JavaScript TypeError implementation.

Functions

create_throw_type_error