Module boa_engine::builtins::function

This module implements the global Function object as well as creates Native Functions.

Objects wrap Functions and expose them via call/construct slots.

The Function object is used for matching text with a pattern.

More information:

ECMAScript reference

MDN documentation

Modules

arguments 🔒

sealed 🔒

Structs

BoundFunction

Binds a Function Object when bind is called.

BuiltInFunctionObject

Captures

Wrapper for Gc<GcCell<dyn NativeObject>> that allows passing additional captures through a Copy closure.

Enums

ClassFieldDefinition

Record containing the field definition of classes.

ConstructorKind

Represents the [[ConstructorKind]] internal slot of function objects.

Function

Boa representation of a Function Object.

ThisMode

Traits

ClosureFunctionSignature

Trait representing a native built-in closure.

DynCopy

Functions

make_builtin_fn 🔒

Creates a new member function of a Object or prototype.

set_function_name 🔒

Abstract operation SetFunctionName

Type Definitions

NativeFunctionSignature

Type representing a native built-in function a.k.a. function pointer.