Module boa_engine::builtins::intl

Boa’s implementation of ECMAScript’s global Intl object.

Intl is a built-in object that has properties and methods for i18n. It’s not a function object.

More information:

ECMAScript reference

Modules

collator 🔒

date_time_format 🔒

This module implements the global Intl.DateTimeFormat object.

list_format 🔒

locale 🔒

options 🔒

segmenter 🔒

Structs

Intl 🔒

JavaScript Intl object.

Traits

Service 🔒

A service component that is part of the Intl API.