Module boa_engine::builtins::intl

This module implements the global Intl object.

Intl is a built-in object that has properties and methods for i18n. It’s not a function object.

More information:

ECMAScript reference

Modules

date_time_format

This module implements the global Intl.DateTimeFormat object.

Structs

DateTimeFormatRecord 🔒

DateTimeFormatRecord type aggregates locale_matcher selector and properties map.

Intl 🔒

JavaScript Intl object.

Keyword 🔒

Keyword structure is a pair of keyword key and keyword value.

MatcherRecord 🔒

MatcherRecord type aggregates unicode locale string and unicode locale extension.

ResolveLocaleRecord 🔒

ResolveLocaleRecord type consists of unicode locale string, data_locale string and properties map.

UniExtRecord 🔒

UniExtRecord structure represents unicode extension records.

Enums

GetOptionType 🔒

Functions

best_available_locale 🔒

Abstract operation BestAvailableLocale ( availableLocales, locale )

best_fit_matcher 🔒

Abstract operation BestFitMatcher ( availableLocales, requestedLocales )

canonicalize_locale_list 🔒

Abstract operation CanonicalizeLocaleList ( locales )

canonicalize_unicode_locale_id 🔒

Abstract operation CanonicalizeUnicodeLocaleId ( locale ).

default_locale 🔒

Abstract operation DefaultLocale ( )

default_number_option 🔒

Abstract operation DefaultNumberOption ( value, minimum, maximum, fallback )

get_number_option 🔒

Abstract operation GetNumberOption ( options, property, minimum, maximum, fallback )

get_option 🔒

Abstract operation GetOption ( options, property, type, values, fallback )

insert_unicode_extension_and_canonicalize 🔒

Abstract operation InsertUnicodeExtensionAndCanonicalize ( locale, extension )

lookup_matcher 🔒

Abstract operation LookupMatcher ( availableLocales, requestedLocales )

resolve_locale 🔒

Abstract operation ResolveLocale ( availableLocales, requestedLocales, options, relevantExtensionKeys, localeData )

unicode_extension_components 🔒

Abstract operation UnicodeExtensionComponents ( extension )

Type Definitions

LocaleDataRecord 🔒

LocaleDataRecord is the type of locale_data argument in resolve_locale subroutine.