Function boa_engine::builtins::intl::locale::utils::resolve_locale

pub(in builtins::intl) fn resolve_locale<'provider, S>(
    requested_locales: &[Locale],
    options: &mut IntlOptions<S::LocaleOptions>,
    icu: &Icu<'provider>
) -> Localewhere
    S: Service,
    BoaProvider<'provider>: DataProvider<S::LangMarker>,

Abstract operation ResolveLocale ( availableLocales, requestedLocales, options, relevantExtensionKeys, localeData )

Compares a BCP 47 language priority list requestedLocales against the locales in availableLocales and determines the best available language to meet the request. availableLocales, requestedLocales, and relevantExtensionKeys must be provided as List values, options and localeData as Records.

More information:

ECMAScript reference