Function boa_engine::builtins::intl::options::get_option

pub(super) fn get_option<T: OptionType>(
    options: &JsObject,
    property: &str,
    required: bool,
    context: &mut Context<'_>
) -> JsResult<Option<T>>

Abstract operation GetOption ( options, property, type, values, fallback )

Extracts the value of the property named property from the provided options object, converts it to the required type and checks whether it is one of a List of allowed values. If values is undefined, there is no fixed set of values and any is permitted. If the value is undefined, required determines if the function should return None or an Err. Use Option::unwrap_or and friends to manage the default value.

This is a safer alternative to GetOption, which tries to parse from the provided property a valid variant of the provided type T. It doesn’t accept a type parameter since the type can specify in its implementation of TryFrom whether it wants to parse from a str or convert directly from a boolean or number.