Trait boa_engine::builtins::intl::Service

trait Service {
    type LangMarker: KeyedDataMarker;
    type LocaleOptions;

    fn resolve(
        _locale: &mut Locale,
        _options: &mut Self::LocaleOptions,
        _provider: BoaProvider<'_>
    ) { ... }
}

A service component that is part of the Intl API.

This needs to be implemented for every Intl service in order to use the functions defined in locale::utils, such as locale resolution and selection.

Required Associated Types

type LangMarker: KeyedDataMarker

The data marker used by resolve_locale to decide which locales are supported by this service.

type LocaleOptions

The set of options used in the Service::resolve method to resolve the provided locale.

Provided Methods

fn resolve(
_locale: &mut Locale,
_options: &mut Self::LocaleOptions,
_provider: BoaProvider<'_>
)

Resolves the final value of locale from a set of options.

The provided options will also be modified with the final values, in case there were changes in the resolution algorithm.

Note

A correct implementation must ensure locale and options are both written with the new final values.

If the implementor service doesn’t contain any [[RelevantExtensionKeys]], this can be skipped.