Struct boa_engine::builtins::regexp::RegExp

pub struct RegExp {
    matcher: Regex,
    flags: RegExpFlags,
    original_source: JsString,
    original_flags: JsString,
}

The internal representation on a RegExp object.

Fields
matcher: Regex

Regex matcher.

flags: RegExpFlagsoriginal_source: JsStringoriginal_flags: JsString
Implementations

impl RegExp

pub(crate) const NAME: &'static str = "RegExp"

The name of the object.

pub(crate) const LENGTH: usize = 2usize

The amount of arguments this function object takes.

pub(crate) fn constructor(
new_target: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

22.2.3.1 RegExp ( pattern, flags )

More information:

ECMAScript reference

pub(crate) fn alloc(
new_target: &JsValue,
context: &mut Context
) -> JsResult<JsObject>

22.2.3.2.1 RegExpAlloc ( newTarget )

More information:

ECMAScript reference

pub(crate) fn initialize(
obj: JsObject,
pattern: &JsValue,
flags: &JsValue,
context: &mut Context
) -> JsResult<JsValue>

22.2.3.2.2 RegExpInitialize ( obj, pattern, flags )

More information:

ECMAScript reference

pub(crate) fn create(
p: &JsValue,
f: &JsValue,
context: &mut Context
) -> JsResult<JsValue>

22.2.3.2.4 RegExpCreate ( P, F )

More information:

ECMAScript reference

fn get_species(
this: &JsValue,
_: &[JsValue],
_: &mut Context
) -> JsResult<JsValue>

get RegExp [ @@species ]

The RegExp [ @@species ] accessor property returns the RegExp constructor.

More information:

ECMAScript reference

MDN documentation

fn regexp_has_flag(
this: &JsValue,
flag: u8,
context: &mut Context
) -> JsResult<JsValue>

pub(crate) fn get_has_indices(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.hasIndices

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_global(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.global

The global property indicates whether or not the “g” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_ignore_case(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.ignoreCase

The ignoreCase property indicates whether or not the “i” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_multiline(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.multiline

The multiline property indicates whether or not the “m” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_dot_all(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.dotAll

The dotAll property indicates whether or not the “s” flag is used with the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_unicode(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.unicode

The unicode property indicates whether or not the “u” flag is used with a regular expression. unicode is a read-only property of an individual regular expression instance.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_sticky(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.sticky

This flag indicates that it matches only from the index indicated by the lastIndex property of this regular expression in the target string (and does not attempt to match from any later indexes).

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_flags(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.flags

The flags property returns a string consisting of the flags of the current regular expression object.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn get_source(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

get RegExp.prototype.source

The source property returns a String containing the source text of the regexp object, and it doesn’t contain the two forward slashes on both sides and any flags.

More information:

ECMAScript reference

MDN documentation

fn escape_pattern(src: &JsString, _flags: &JsString) -> JsValue

22.2.3.2.5 EscapeRegExpPattern ( P, F )

More information:

ECMAScript reference

pub(crate) fn test(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype.test( string )

The test() method executes a search for a match between a regular expression and a specified string.

Returns true or false.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn exec(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype.exec( string )

The exec() method executes a search for a match in a specified string.

Returns a result array, or null.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn abstract_exec(
this: &JsObject,
input: JsString,
context: &mut Context
) -> JsResult<Option<JsObject>>

22.2.5.2.1 RegExpExec ( R, S )

More information:

ECMAScript reference

pub(crate) fn abstract_builtin_exec(
this: &JsObject,
input: &JsString,
context: &mut Context
) -> JsResult<Option<JsObject>>

22.2.5.2.2 RegExpBuiltinExec ( R, S )

More information:

ECMAScript reference

pub(crate) fn match(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype[ @@match ]( string )

This method retrieves the matches when matching a string against a regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn to_string(
this: &JsValue,
_: &[JsValue],
_: &mut Context
) -> JsResult<JsValue>

RegExp.prototype.toString()

Return a string representing the regular expression.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn match_all(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype[ @@matchAll ]( string )

The [@@matchAll] method returns all matches of the regular expression against a string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn replace(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype [ @@replace ] ( string, replaceValue )

The @@replace method replaces some or all matches of a this pattern in a string by a replacement, and returns the result of the replacement as a new string. The replacement can be a string or a function to be called for each match.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn search(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype[ @@search ]( string )

This method executes a search for a match between a this regular expression and a string.

More information:

ECMAScript reference

MDN documentation

pub(crate) fn split(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

RegExp.prototype [ @@split ] ( string, limit )

The @@split method splits a String object into an array of strings by separating the string into substrings.

More information:

ECMAScript reference

MDN documentation