Trait boa_engine::builtins::BuiltIn

pub(crate) trait BuiltIn {
    const NAME: &'static str;
    const ATTRIBUTE: Attribute = _;

    fn init(context: &mut Context) -> Option<JsValue>;
}

Trait representing a global built-in object such as Math, Object or String.

This trait must be implemented for any global built-in accessible from JavaScript.

Required Associated Constants

const NAME: &'static str

Binding name of the built-in inside the global object.

E.g. If you want access the properties of a Complex built-in with the name Cplx you must assign "Cplx" to this constant, making any property inside it accessible from Javascript as Cplx.prop

Provided Associated Constants

const ATTRIBUTE: Attribute = _

Property attribute flags of the built-in. Check Attribute for more information.

Required Methods

fn init(context: &mut Context) -> Option<JsValue>

Initialization code for the built-in.

This is where the methods, properties, static methods and the constructor of a built-in must be initialized to be accessible from Javascript.

Note

A return value of None indicates that the value must not be added as a global attribute for the global object.