Trait boa_engine::builtins::BuiltInConstructor

pub(crate) trait BuiltInConstructor: BuiltInObject {
    const LENGTH: usize;
    const STANDARD_CONSTRUCTOR: fn(_: &StandardConstructors) -> &StandardConstructor;

    // Required method
    fn constructor(
        new_target: &JsValue,
        args: &[JsValue],
        context: &mut Context<'_>
    ) -> JsResult<JsValue>;
}

A built-in object that is also a constructor.

This trait must be implemented for any global built-in that can also be called with new to construct an object instance e.g. Array, Map or Object.

Required Associated Constants

const LENGTH: usize

The amount of arguments this function object takes.

const STANDARD_CONSTRUCTOR: fn(_: &StandardConstructors) -> &StandardConstructor

The corresponding standard constructor of this constructor.

Required Methods

fn constructor( new_target: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

The native constructor function.