Trait boa_engine::builtins::BuiltInObject

pub(crate) trait BuiltInObject: IntrinsicObject {
    const NAME: JsString;
    const ATTRIBUTE: Attribute = _;
}

A built-in object.

This trait must be implemented for any global built-in that lives in the global context of a script.

Required Associated Constants

const NAME: JsString

Binding name of the builtin inside the global object.

E.g. If you want access the properties of a Complex built-in with the name Cplx you must assign "Cplx" to this constant, making any property inside it accessible from ECMAScript as Cplx.prop

Provided Associated Constants

const ATTRIBUTE: Attribute = _

Property attribute flags of the built-in. Check Attribute for more information.