Module boa_engine::builtins::typed_array

This module implements the global TypedArray objects.

A TypedArray object describes an array-like view of an underlying binary data buffer. There is no global property named TypedArray, nor is there a directly visible TypedArray constructor. Instead, there are a number of different global properties, whose values are typed array constructors for specific element types.

More information:

ECMAScript reference

MDN documentation

Modules

integer_indexed_object

This module implements the Integer-Indexed exotic object.

Macros

typed_array 🔒

Structs

BigInt64Array

JavaScript BigInt64Array built-in implementation.

BigUint64Array

JavaScript BigUint64Array built-in implementation.

Float32Array

JavaScript Float32Array built-in implementation.

Float64Array

JavaScript Float64Array built-in implementation.

Int8Array

JavaScript Int8Array built-in implementation.

Int16Array

JavaScript Int16Array built-in implementation.

Int32Array

JavaScript Int32Array built-in implementation.

TypedArray 🔒

The JavaScript %TypedArray% object.

Uint8Array

JavaScript Uint8Array built-in implementation.

Uint8ClampedArray

JavaScript Uint8ClampedArray built-in implementation.

Uint16Array

JavaScript Uint16Array built-in implementation.

Uint32Array

JavaScript Uint32Array built-in implementation.

Enums

TypedArrayKind 🔒

Names of all the typed arrays.