Trait boa_engine::builtins::typed_array::element::Element

pub(crate) trait Element: Sized + Into<TypedArrayElement> + NoUninit + AnyBitPattern {
    type Atomic: Atomic;

    // Required methods
    fn from_js_value(value: &JsValue, context: &mut Context) -> JsResult<Self>;
    fn from_plain(bytes: <Self::Atomic as Atomic>::Plain) -> Self;
    fn to_plain(self) -> <Self::Atomic as Atomic>::Plain;
    fn to_little_endian(self) -> Self;
    fn to_big_endian(self) -> Self;
    unsafe fn read(buffer: SliceRef<'_>) -> ElementRef<'_, Self>;
    unsafe fn read_mut(buffer: SliceRefMut<'_>) -> ElementRefMut<'_, Self>;
}

A native element that can be inside a TypedArray.

Required Associated Types

type Atomic: Atomic

The atomic type used for shared array buffers.

Required Methods

fn from_js_value(value: &JsValue, context: &mut Context) -> JsResult<Self>

Converts a JsValue into the native element Self.

fn from_plain(bytes: <Self::Atomic as Atomic>::Plain) -> Self

Converts from the plain type of an atomic to Self.

fn to_plain(self) -> <Self::Atomic as Atomic>::Plain

Converts from Self to the plain type of an atomic.

fn to_little_endian(self) -> Self

Gets the little endian representation of Self.

fn to_big_endian(self) -> Self

Gets the big endian representation of Self.

unsafe fn read(buffer: SliceRef<'_>) -> ElementRef<'_, Self>

Reads Self from the buffer.

This will always read values in the native endianness of the target architecture.

Safety

buffer must be aligned to the native alignment of Self.

buffer must contain enough bytes to read std::sizeof::<Self> bytes.

unsafe fn read_mut(buffer: SliceRefMut<'_>) -> ElementRefMut<'_, Self>

Writes the bytes of this element into buffer.

This will always write values in the native endianness of the target architecture.

Safety

buffer must be aligned to the native alignment of Self.

buffer must contain enough bytes to store std::sizeof::<Self> bytes.