Trait boa_engine::builtins::typed_array::element::Element

pub(crate) trait Element: Sized + Into<TypedArrayElement> + NoUninit + AnyBitPattern {
    // Required methods
    fn from_js_value(
        value: &JsValue,
        context: &mut Context<'_>
    ) -> JsResult<Self>;
    fn to_little_endian(self) -> Self;
    fn to_big_endian(self) -> Self;
    unsafe fn read_from_buffer(buffer: SliceRef<'_>, order: Ordering) -> Self;
    unsafe fn write_to_buffer(
        buffer: SliceRefMut<'_>,
        value: Self,
        order: Ordering
    );
}

Required Methods

fn from_js_value(value: &JsValue, context: &mut Context<'_>) -> JsResult<Self>

fn to_little_endian(self) -> Self

Gets the little endian representation of Self.

fn to_big_endian(self) -> Self

Gets the big endian representation of Self.

unsafe fn read_from_buffer(buffer: SliceRef<'_>, order: Ordering) -> Self

Reads Self from the buffer.

This will always read values in the native endianness of the target architecture.

Safety

buffer must be aligned to the native alignment of Self.

buffer must contain enough bytes to read std::sizeof::<Self> bytes.

unsafe fn write_to_buffer(buffer: SliceRefMut<'_>, value: Self, order: Ordering)

Writes the bytes of this element into buffer.

This will always write values in the native endianness of the target architecture.

Safety

buffer must be aligned to the native alignment of Self.

buffer must contain enough bytes to store std::sizeof::<Self> bytes.