Struct boa_engine::builtins::typed_array::TypedArray

pub(crate) struct TypedArray;

The JavaScript %TypedArray% object.

https://tc39.es/ecma262/#sec-%typedarray%-intrinsic-object

Implementations

impl TypedArray

const LENGTH: usize

fn constructor(
_new_target: &JsValue,
_args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.1.1 %TypedArray% ( )

More information:

ECMAScript reference

fn from(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.2.1 %TypedArray%.from ( source [ , mapfn [ , thisArg ] ] )

More information:

ECMAScript reference

fn of(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.2.2 %TypedArray%.of ( ...items )

More information:

ECMAScript reference

fn get_species(
this: &JsValue,
_: &[JsValue],
_: &mut Context
) -> JsResult<JsValue>

23.2.2.4 get %TypedArray% [ @@species ]

More information:

ECMAScript reference

fn at(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.1 %TypedArray%.prototype.at ( index )

More information:

ECMAScript reference

fn buffer(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.2 get %TypedArray%.prototype.buffer

More information:

ECMAScript reference

fn byte_length(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.3 get %TypedArray%.prototype.byteLength

More information:

ECMAScript reference

fn byte_offset(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.4 get %TypedArray%.prototype.byteOffset

More information:

ECMAScript reference

fn copy_within(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.6 %TypedArray%.prototype.copyWithin ( target, start [ , end ] )

More information:

ECMAScript reference

fn entries(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.7 %TypedArray%.prototype.entries ( )

More information:

ECMAScript reference

fn every(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.8 %TypedArray%.prototype.every ( callbackfn [ , thisArg ] )

More information:

ECMAScript reference

fn fill(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.9 %TypedArray%.prototype.fill ( value [ , start [ , end ] ] )

More information:

ECMAScript reference

fn filter(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.10 %TypedArray%.prototype.filter ( callbackfn [ , thisArg ] )

More information:

ECMAScript reference

fn find(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.11 %TypedArray%.prototype.find ( predicate [ , thisArg ] )

More information:

ECMAScript reference

fn findindex(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.12 %TypedArray%.prototype.findIndex ( predicate [ , thisArg ] )

More information:

ECMAScript reference

fn foreach(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.13 %TypedArray%.prototype.forEach ( callbackfn [ , thisArg ] )

More information:

ECMAScript reference

fn includes(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.14 %TypedArray%.prototype.includes ( searchElement [ , fromIndex ] )

More information:

ECMAScript reference

fn index_of(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.15 %TypedArray%.prototype.indexOf ( searchElement [ , fromIndex ] )

More information:

ECMAScript reference

fn join(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.16 %TypedArray%.prototype.join ( separator )

More information:

ECMAScript reference

fn keys(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.17 %TypedArray%.prototype.keys ( )

More information:

ECMAScript reference

fn last_index_of(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.18 %TypedArray%.prototype.lastIndexOf ( searchElement [ , fromIndex ] )

More information:

ECMAScript reference

fn length(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.19 get %TypedArray%.prototype.length

More information:

ECMAScript reference

fn map(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.20 %TypedArray%.prototype.map ( callbackfn [ , thisArg ] )

More information:

ECMAScript reference

fn reduce(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.21 %TypedArray%.prototype.reduce ( callbackfn [ , initialValue ] )

More information:

ECMAScript reference

fn reduceright(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.22 %TypedArray%.prototype.reduceRight ( callbackfn [ , initialValue ] )

More information:

ECMAScript reference

fn reverse(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.23 %TypedArray%.prototype.reverse ( )

More information:

ECMAScript reference

fn set(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.24 %TypedArray%.prototype.set ( source [ , offset ] )

More information:

ECMAScript reference

fn set_typed_array_from_typed_array(
target: &JsObject,
target_offset: IntegerOrInfinity,
source: &JsObject,
context: &mut Context
) -> JsResult<()>

3.2.3.24.1 SetTypedArrayFromTypedArray ( target, targetOffset, source )

More information:

ECMAScript reference

fn set_typed_array_from_array_like(
target: &JsObject,
target_offset: IntegerOrInfinity,
source: &JsValue,
context: &mut Context
) -> JsResult<()>

23.2.3.24.2 SetTypedArrayFromArrayLike ( target, targetOffset, source )

More information:

ECMAScript reference

fn slice(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.25 %TypedArray%.prototype.slice ( start, end )

More information:

ECMAScript reference

fn some(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.26 %TypedArray%.prototype.some ( callbackfn [ , thisArg ] )

More information:

ECMAScript reference

fn sort(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.27 %TypedArray%.prototype.sort ( comparefn )

More information:

ECMAScript reference

fn subarray(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.28 %TypedArray%.prototype.subarray ( begin, end )

More information:

ECMAScript reference

fn values(
this: &JsValue,
_: &[JsValue],
context: &mut Context
) -> JsResult<JsValue>

23.2.3.31 %TypedArray%.prototype.values ( )

More information:

ECMAScript reference

fn to_string_tag(
this: &JsValue,
_: &[JsValue],
_: &mut Context
) -> JsResult<JsValue>

23.2.3.33 get %TypedArray%.prototype [ @@toStringTag ]

More information:

ECMAScript reference

fn species_create(
exemplar: &JsObject,
typed_array_name: TypedArrayKind,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsObject>

23.2.4.1 TypedArraySpeciesCreate ( exemplar, argumentList )

More information:

ECMAScript reference

fn create(
constructor: &JsObject,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsObject>

23.2.4.2 TypedArrayCreate ( constructor, argumentList )

More information:

ECMAScript reference

fn allocate_buffer(
indexed: &mut IntegerIndexed,
length: usize,
context: &mut Context
) -> JsResult<()>

https://tc39.es/ecma262/#sec-allocatetypedarraybuffer

fn initialize_from_list(
o: &JsObject,
values: Vec<JsValue>,
context: &mut Context
) -> JsResult<()>

https://tc39.es/ecma262/#sec-initializetypedarrayfromlist

fn allocate<P>(
constructor_name: TypedArrayKind,
new_target: &JsValue,
default_proto: P,
length: Option<usize>,
context: &mut Context
) -> JsResult<JsObject> where
P: FnOnce(&StandardConstructors) -> &StandardConstructor,

AllocateTypedArray ( constructorName, newTarget, defaultProto [ , length ] )

It is used to validate and create an instance of a TypedArray constructor. If the length argument is passed, an ArrayBuffer of that length is also allocated and associated with the new TypedArray instance. AllocateTypedArray provides common semantics that is used by TypedArray.

For more information, check the spec.

fn initialize_from_typed_array(
o: &JsObject,
src_array: &JsObject,
context: &mut Context
) -> JsResult<()>

23.2.5.1.2 InitializeTypedArrayFromTypedArray ( O, srcArray )

More information:

ECMAScript reference

fn initialize_from_array_buffer(
o: &JsObject,
buffer: JsObject,
byte_offset: &JsValue,
length: &JsValue,
context: &mut Context
) -> JsResult<()>

23.2.5.1.3 InitializeTypedArrayFromArrayBuffer ( O, buffer, byteOffset, length )

More information:

ECMAScript reference

fn initialize_from_array_like(
o: &JsObject,
array_like: &JsObject,
context: &mut Context
) -> JsResult<()>

23.2.5.1.5 InitializeTypedArrayFromArrayLike ( O, arrayLike )

More information:

ECMAScript reference