Module boa_engine::builtins::uri::consts

URI handling function constants

This module contains a few constants used to handle decoding and encoding for URI handling functions. They make it easier and more performant to compare different ranges and code points.

Constants

DECIMAL_DIGIT 🔒

A range containing all the DecimalDigit code points.

NUMBER_SIGN 🔒

The number sign (#) symbol as a UTF-16 code potint.

URI_ALPHA_LOWER 🔒

A range containing all the lowercase uriAlpha code points.

URI_ALPHA_UPPER 🔒

A range containing all the uppercase uriAlpha code points.

URI_MARK 🔒

An array containing all the uriMark code points.

URI_RESERVED 🔒

An array containing all the uriReserved code points.

Functions

is_uri_reserved_or_number_sign 🔒

Constant with all the reserved URI characters, plus the number sign symbol (#).

is_uri_reserved_or_uri_unescaped_or_number_sign 🔒

Constant with all the reserved and unescaped URI characters, plus the number sign symbol (#).

is_uri_unescaped 🔒

Constant with all the unescaped URI characters.