Module boa_engine::bytecompiler::jump_control

JumpControlInfo tracks relevant jump information used during compilation.

Primarily, jump control tracks information related to the compilation of iteration statements, switch statements, try statements, and labelled statements.

Structs

JumpControlInfo 🔒

Boa’s ByteCompiler jump information tracking struct.

JumpControlInfoFlags 🔒

A bitflag that contains the type flags and relevant booleans for JumpControlInfo.

JumpRecord 🔒

This represents a local control flow handling. See JumpRecordKind for types.

Enums

JumpRecordAction 🔒

An actions to be performed for the local control flow.

JumpRecordKind 🔒

Local Control flow type.