Struct boa_engine::bytecompiler::ByteCompiler

pub struct ByteCompiler<'ctx, 'host> {

    pub(crate) function_name: Sym,
    pub(crate) length: u32,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) bytecode: Vec<u8>,
    pub(crate) literals: Vec<JsValue>,
    pub(crate) names: Vec<JsString>,
    pub(crate) bindings: Vec<BindingLocator>,
    pub(crate) functions: Vec<Gc<CodeBlock>>,
    pub(crate) compile_environments: Vec<Rc<CompileTimeEnvironment>>,
    pub(crate) current_environment: Rc<CompileTimeEnvironment>,
    current_open_environments_count: u32,
    current_stack_value_count: u32,
    code_block_flags: CodeBlockFlags,
    handlers: ThinVec<Handler>,
    literals_map: FxHashMap<Literal, u32>,
    names_map: FxHashMap<Identifier, u32>,
    bindings_map: FxHashMap<BindingLocator, u32>,
    jump_info: Vec<JumpControlInfo>,
    pub(crate) in_async: bool,
    in_generator: bool,
    pub(crate) async_handler: Option<u32>,
    json_parse: bool,
    context: &'ctx mut Context<'host>,
    annex_b_function_names: Vec<Identifier>,

}

The ByteCompiler is used to compile ECMAScript AST from boa_ast to bytecode.

Fields
function_name: Sym

Name of this function.

length: u32

The number of arguments expected.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

bytecode: Vec<u8>

Bytecode

literals: Vec<JsValue>

Literals

names: Vec<JsString>

Property field names and private name [[Description]]s.

bindings: Vec<BindingLocator>

Locators for all bindings in the codeblock.

functions: Vec<Gc<CodeBlock>>

Functions inside this function

compile_environments: Vec<Rc<CompileTimeEnvironment>>

Compile time environments in this function.

current_environment: Rc<CompileTimeEnvironment>

The environment that is currently active.

current_open_environments_count: u32current_stack_value_count: u32code_block_flags: CodeBlockFlagshandlers: ThinVec<Handler>literals_map: FxHashMap<Literal, u32>names_map: FxHashMap<Identifier, u32>bindings_map: FxHashMap<BindingLocator, u32>jump_info: Vec<JumpControlInfo>in_async: boolin_generator: boolasync_handler: Option<u32>

Used to handle exception throws that escape the async function types.

Async functions and async generator functions, need to be closed and resolved.

json_parse: boolcontext: &'ctx mut Context<'host>annex_b_function_names: Vec<Identifier>
Implementations

impl ByteCompiler<'_, '_>

pub(crate) fn compile_class(&mut self, class: &Class, expression: bool)

This function compiles a class declaration or expression.

The compilation of a class declaration and expression is mostly equal. A class declaration binds the resulting class object to it’s identifier. A class expression leaves the resulting class object on the stack for following operations.

impl ByteCompiler<'_, '_>

pub(crate) fn compile_declaration_pattern_impl( &mut self, pattern: &Pattern, def: BindingOpcode )

fn compile_array_pattern_element( &mut self, element: &ArrayPatternElement, def: BindingOpcode )

impl ByteCompiler<'_, '_>

pub(crate) fn global_declaration_instantiation( &mut self, script: &Script ) -> JsResult<()>

GlobalDeclarationInstantiation ( script, env )

More information:

ECMAScript reference

pub(crate) fn block_declaration_instantiation<'a, N>(&mut self, block: &'a N)where &'a N: Into<NodeRef<'a>>,

BlockDeclarationInstantiation ( code, env )

More information:

ECMAScript reference

pub(crate) fn eval_declaration_instantiation( &mut self, body: &Script, strict: bool ) -> JsResult<()>

EvalDeclarationInstantiation ( body, varEnv, lexEnv, privateEnv, strict )

More information:

ECMAScript reference

pub(crate) fn function_declaration_instantiation( &mut self, body: &FunctionBody, formals: &FormalParameterList, arrow: bool, strict: bool, generator: bool ) -> (Option<Label>, bool)

FunctionDeclarationInstantiation ( func, argumentsList )

More information:

ECMAScript reference

impl ByteCompiler<'_, '_>

pub(crate) fn push_compile_environment(&mut self, function_scope: bool)

Push either a new declarative or function environment on the compile time environment stack.

pub(crate) fn pop_compile_environment(&mut self) -> u32

Pops the top compile time environment and returns its index in the compile time environments array.

pub(crate) fn get_binding_value(&self, name: Identifier) -> BindingLocator

Get the binding locator of the binding at bytecode compile time.

pub(crate) fn has_binding(&self, name: Identifier) -> bool

Return if a declarative binding exists at bytecode compile time. This does not include bindings on the global object.

pub(crate) fn has_binding_eval(&self, name: Identifier, strict: bool) -> bool

Check if a binding name exists in a environment. If strict is false check until a function scope is reached.

pub(crate) fn has_binding_until_var(&self, name: Identifier) -> bool

Check if a binding name exists in a environment. Stop when a function scope is reached.

pub(crate) fn create_mutable_binding( &mut self, name: Identifier, function_scope: bool )

Create a mutable binding at bytecode compile time. This function returns a syntax error, if the binding is a redeclaration.

Panics

Panics if the global environment is not function scoped.

pub(crate) fn initialize_mutable_binding( &self, name: Identifier, function_scope: bool ) -> BindingLocator

Initialize a mutable binding at bytecode compile time and return its binding locator.

pub(crate) fn create_immutable_binding( &mut self, name: Identifier, strict: bool )

Create an immutable binding at bytecode compile time. This function returns a syntax error, if the binding is a redeclaration.

Panics

Panics if the global environment does not exist.

pub(crate) fn initialize_immutable_binding( &self, name: Identifier ) -> BindingLocator

Initialize an immutable binding at bytecode compile time and return it’s binding locator.

Panics

Panics if the global environment does not exist or a the binding was not created on the current environment.

pub(crate) fn set_mutable_binding( &self, name: Identifier ) -> Result<BindingLocator, BindingLocatorError>

Return the binding locator for a set operation on an existing binding.

pub(crate) fn set_mutable_binding_var( &self, name: Identifier ) -> Result<BindingLocator, BindingLocatorError>

Return the binding locator for a set operation on an existing var binding.

impl ByteCompiler<'_, '_>

pub(crate) fn compile_assign(&mut self, assign: &Assign, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_binary(&mut self, binary: &Binary, use_expr: bool)

pub(crate) fn compile_binary_in_private( &mut self, binary: &BinaryInPrivate, use_expr: bool )

impl ByteCompiler<'_, '_>

pub(crate) fn compile_object_literal( &mut self, object: &ObjectLiteral, use_expr: bool )

fn compile_object_literal_computed_method( &mut self, name: &Expression, function: FunctionSpec<'_>, kind: MethodKind )

impl ByteCompiler<'_, '_>

pub(crate) fn compile_unary(&mut self, unary: &Unary, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_update(&mut self, update: &Update, use_expr: bool)

impl ByteCompiler<'_, '_>

fn compile_literal(&mut self, lit: &AstLiteral, use_expr: bool)

fn compile_conditional(&mut self, op: &Conditional, use_expr: bool)

fn compile_template_literal( &mut self, template_literal: &TemplateLiteral, use_expr: bool )

pub(crate) fn compile_expr_impl(&mut self, expr: &Expression, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn push_empty_loop_jump_control(&mut self, use_expr: bool)

Pushes a generic JumpControlInfo onto ByteCompiler

Default JumpControlInfoKind is JumpControlInfoKind::Loop

pub(crate) fn current_jump_control_mut( &mut self ) -> Option<&mut JumpControlInfo>

pub(crate) fn push_contol_info(&mut self, info: JumpControlInfo, use_expr: bool)

pub(crate) fn push_handler(&mut self) -> u32

Pushes an exception Handler.

Must be patched with Self::patch_handler().

pub(crate) fn patch_handler(&mut self, handler_index: u32)

pub(crate) fn jump_control_info_has_use_expr(&self) -> bool

Does the jump control info have the use_expr flag set to true.

See JumpControlInfoFlags.

pub(crate) fn push_labelled_control_info( &mut self, label: Sym, start_address: u32, use_expr: bool )

Pushes a LabelledStatement’s JumpControlInfo onto the jump_info stack.

pub(crate) fn pop_labelled_control_info(&mut self)

Pops and handles the info for a label’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a LabelledStatement.

pub(crate) fn push_loop_control_info( &mut self, label: Option<Sym>, start_address: u32, use_expr: bool )

Pushes an WhileStatement, ForStatement or DoWhileStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn push_loop_control_info_for_of_in_loop( &mut self, label: Option<Sym>, start_address: u32, use_expr: bool )

Pushes a ForInOfStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn push_loop_control_info_for_await_of_loop( &mut self, label: Option<Sym>, start_address: u32, use_expr: bool )

pub(crate) fn pop_loop_control_info(&mut self)

Pops and handles the info for a loop control block’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a loop block.

pub(crate) fn push_switch_control_info( &mut self, label: Option<Sym>, start_address: u32, use_expr: bool )

Pushes a SwitchStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn pop_switch_control_info(&mut self)

Pops and handles the info for a switch block’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a switch block.

pub(crate) fn push_try_with_finally_control_info(&mut self, use_expr: bool)

Pushes a TryStatement’s JumpControlInfo onto the jump_info stack.

pub(crate) fn pop_try_with_finally_control_info(&mut self, finally_start: u32)

Pops and handles the info for a try statement with a finally block.

Panic

Will panic if popped JumpControlInfo is not for a try block.

pub(crate) fn jump_info_open_environment_count(&self, index: usize) -> u32

impl ByteCompiler<'_, '_>

pub fn compile_module_item_list(&mut self, list: &ModuleItemList)

Compiles a ModuleItemList.

pub fn compile_module_item(&mut self, item: &ModuleItem)

Compiles a ModuleItem.

impl ByteCompiler<'_, '_>

pub(crate) fn compile_block(&mut self, block: &Block, use_expr: bool)

Compile a Block boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_break(&mut self, node: Break, _use_expr: bool)

Compile a Break boa_ast node

fn break_jump_record_actions(&self, node: Break) -> Vec<JumpRecordAction>

impl ByteCompiler<'_, '_>

pub(crate) fn compile_continue(&mut self, node: Continue, _use_expr: bool)

fn continue_jump_record_actions(&self, node: Continue) -> Vec<JumpRecordAction>

impl ByteCompiler<'_, '_>

pub(crate) fn compile_if(&mut self, node: &If, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_labelled(&mut self, labelled: &Labelled, use_expr: bool)

Compile a Labelled boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_for_loop( &mut self, for_loop: &ForLoop, label: Option<Sym>, use_expr: bool )

pub(crate) fn compile_for_in_loop( &mut self, for_in_loop: &ForInLoop, label: Option<Sym>, use_expr: bool )

pub(crate) fn compile_for_of_loop( &mut self, for_of_loop: &ForOfLoop, label: Option<Sym>, use_expr: bool )

pub(crate) fn compile_while_loop( &mut self, while_loop: &WhileLoop, label: Option<Sym>, use_expr: bool )

pub(crate) fn compile_do_while_loop( &mut self, do_while_loop: &DoWhileLoop, label: Option<Sym>, use_expr: bool )

impl ByteCompiler<'_, '_>

pub(crate) fn compile_switch(&mut self, switch: &Switch, use_expr: bool)

Compile a Switch boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_try(&mut self, t: &Try, use_expr: bool)

Compile try statement.

pub(crate) fn compile_catch_stmt( &mut self, catch: &Catch, _has_finally: bool, use_expr: bool )

pub(crate) fn compile_finally_stmt( &mut self, finally: &Finally, has_catch: bool )

impl ByteCompiler<'_, '_>

pub(crate) fn compile_with(&mut self, with: &With, use_expr: bool)

Compile a With boa_ast node

impl ByteCompiler<'_, '_>

pub fn compile_stmt( &mut self, node: &Statement, use_expr: bool, root_statement: bool )

Compiles a Statement boa_ast node.

pub(crate) fn return(&mut self)

fn return_jump_record_actions(&self) -> Vec<JumpRecordAction>

impl ByteCompiler<'_, '_>

pub(super) fn iterator_close(&mut self, async_: bool)

Closes an iterator

This is equivalent to the IteratorClose and AsyncIteratorClose operations.

Iterator Stack:

iterator => <empty>

pub(super) fn close_active_iterators(&mut self)

Closes all active iterators in the current CallFrame.

pub(super) fn yield(&mut self)

Yields from the current generator.

This is equivalent to the Yield ( value ) operation from the spec.

stack:

value => received

pub(super) fn async_generator_yield(&mut self)

Yields from the current async generator.

This is equivalent to the AsyncGeneratorYield ( value ) operation from the spec.

stack:

value => received

impl<'ctx, 'host> ByteCompiler<'ctx, 'host>

const DUMMY_ADDRESS: u32 = 4_294_967_295u32

Represents a placeholder address that will be patched later.

const DUMMY_LABEL: Label = _

pub(crate) fn new( name: Sym, strict: bool, json_parse: bool, current_environment: Rc<CompileTimeEnvironment>, context: &'ctx mut Context<'host> ) -> ByteCompiler<'ctx, 'host>

Creates a new ByteCompiler.

pub(crate) const fn strict(&self) -> bool

pub(crate) const fn in_async(&self) -> bool

pub(crate) const fn in_generator(&self) -> bool

pub(crate) const fn in_async_generator(&self) -> bool

pub(crate) fn interner(&self) -> &Interner

fn get_or_insert_literal(&mut self, literal: Literal) -> u32

fn get_or_insert_name(&mut self, name: Identifier) -> u32

fn get_or_insert_private_name(&mut self, name: PrivateName) -> u32

pub(crate) fn get_or_insert_binding(&mut self, binding: BindingLocator) -> u32

fn emit_binding(&mut self, opcode: BindingOpcode, name: Identifier)

fn next_opcode_location(&mut self) -> u32

pub(crate) fn emit(&mut self, opcode: Opcode, operands: &[u32])

fn emit_u64(&mut self, value: u64)

fn emit_u32(&mut self, value: u32)

fn emit_u16(&mut self, value: u16)

pub(crate) fn emit_opcode(&mut self, opcode: Opcode)

fn emit_u8(&mut self, value: u8)

fn emit_push_integer(&mut self, value: i32)

fn emit_push_literal(&mut self, literal: Literal)

fn emit_push_rational(&mut self, value: f64)

fn jump(&mut self) -> Label

fn jump_if_true(&mut self) -> Label

fn jump_if_false(&mut self) -> Label

fn jump_if_null_or_undefined(&mut self) -> Label

fn jump_if_not_resume_kind(&mut self, resume_kind: GeneratorResumeKind) -> Label

fn jump_table(&mut self, count: u32) -> (Vec<Label>, Label)

Push a jump table with count of entries.

Returns the jump label entries and the default label.

pub(crate) fn emit_opcode_with_operand(&mut self, opcode: Opcode) -> Label

Emit an opcode with a dummy operand. Return the Label of the operand.

pub(crate) fn emit_opcode_with_two_operands( &mut self, opcode: Opcode ) -> (Label, Label)

Emit an opcode with two dummy operands. Return the Labels of the two operands.

pub(crate) fn patch_jump_with_target(&mut self, label: Label, target: u32)

fn patch_jump(&mut self, label: Label)

fn access_get(&mut self, access: Access<'_>, use_expr: bool)

fn access_set_top_of_stack_expr_fn( compiler: &mut ByteCompiler<'_, '_>, level: u8 )

fn access_set<F, R>(&mut self, access: Access<'_>, use_expr: bool, expr_fn: F)where F: FnOnce(&mut ByteCompiler<'_, '_>, u8) -> R,

fn access_delete(&mut self, access: Access<'_>)

pub fn compile_statement_list( &mut self, list: &StatementList, use_expr: bool, block: bool )

Compile a StatementList.

pub fn compile_expr(&mut self, expr: &Expression, use_expr: bool)

Compile an Expression.

fn compile_access_preserve_this(&mut self, access: &PropertyAccess)

Compile a property access expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the property access becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the access, and value is the final result of the access.

This is mostly useful for optional chains with calls (a.b?.()) and for regular chains with calls (a.b()), since both of them must have a be the value of this for the function call b(), but a regular compilation of the access would lose the this value after accessing b.

fn compile_optional_preserve_this(&mut self, optional: &Optional)

Compile an optional chain expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the optional chain becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the chain, and value is the result of the chain.

This is mostly useful for inner optional chains with external calls ((a?.b)()), because the external call is not in the optional chain, and compiling an optional chain in the usual way would only return the result of the chain without preserving the this value. In other words, this would be set to undefined for that call, which is incorrect since a should be the this value of the call.

fn compile_optional_item_kind(&mut self, kind: &OptionalOperationKind)

Compile a single operation in an optional chain.

On successful compilation, the state of the stack on execution will become ...rest, this, value, where this is the target of the property access (undefined on calls), and value is the result of executing the action. For example, in the expression a?.b.c(), after compiling and executing:

a?.b, the state of the stack will become ...rest, a, b.

b.c, the state of the stack will become ...rest, b, c.

c(), the state of the stack will become ...rest, undefined, c().

Requirements

This should only be called after verifying that the previous value of the chain is not null or undefined (if the operator ?. was used).

This assumes that the state of the stack before compiling is ...rest, this, value, since the operation compiled by this function could be a call.

fn compile_var_decl(&mut self, decl: &VarDeclaration)

Compile a VarDeclaration.

fn compile_lexical_decl(&mut self, decl: &LexicalDeclaration)

Compile a LexicalDeclaration.

fn compile_stmt_list_item( &mut self, item: &StatementListItem, use_expr: bool, block: bool )

Compile a StatementListItem.

pub fn compile_decl(&mut self, decl: &Declaration, block: bool)

Compile a Declaration.

pub(crate) fn function(&mut self, function: FunctionSpec<'_>) -> u32

Compiles a function AST Node into bytecode, and returns its index into the functions array.

pub(crate) fn function_with_binding( &mut self, function: FunctionSpec<'_>, node_kind: NodeKind, use_expr: bool )

Compiles a function AST Node into bytecode, setting its corresponding binding or pushing it to the stack if necessary.

pub(crate) fn object_method(&mut self, function: FunctionSpec<'_>)

Compile an object method AST Node into bytecode.

fn method(&mut self, function: FunctionSpec<'_>, class_name: Sym)

Compile a class method AST Node into bytecode.

fn call(&mut self, callable: Callable<'_>, use_expr: bool)

pub fn finish(self) -> CodeBlock

Finish compiling code with the ByteCompiler and return the generated CodeBlock.

fn compile_declaration_pattern(&mut self, pattern: &Pattern, def: BindingOpcode)

fn class(&mut self, class: &Class, expression: bool)