Struct boa_engine::bytecompiler::ByteCompiler

pub struct ByteCompiler<'b, 'host> {

    pub(crate) function_name: Sym,
    pub(crate) has_binding_identifier: bool,
    pub(crate) length: u32,
    pub(crate) strict: bool,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) bytecode: Vec<u8>,
    pub(crate) literals: Vec<JsValue>,
    pub(crate) names: Vec<Identifier>,
    pub(crate) private_names: Vec<PrivateName>,
    pub(crate) bindings: Vec<BindingLocator>,
    pub(crate) num_bindings: usize,
    pub(crate) functions: Vec<Gc<CodeBlock>>,
    pub(crate) arguments_binding: Option<BindingLocator>,
    pub(crate) compile_environments: Vec<Gc<GcRefCell<CompileTimeEnvironment>>>,
    pub(crate) is_class_constructor: bool,
    pub(crate) class_field_initializer_name: Option<Sym>,
    pub(crate) function_environment_push_location: u32,
    literals_map: FxHashMap<Literal, u32>,
    names_map: FxHashMap<Identifier, u32>,
    private_names_map: FxHashMap<PrivateName, u32>,
    bindings_map: FxHashMap<BindingLocator, u32>,
    jump_info: Vec<JumpControlInfo>,
    in_async_generator: bool,
    json_parse: bool,
    context: &'b mut Context<'host>,

}

The ByteCompiler is used to compile ECMAScript AST from [boa_ast] to bytecode.

Fields
function_name: Sym

Name of this function.

has_binding_identifier: bool

Indicates if the function is an expression and has a binding identifier.

length: u32

The number of arguments expected.

strict: bool

Is this function in strict mode.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

bytecode: Vec<u8>

Bytecode

literals: Vec<JsValue>

Literals

names: Vec<Identifier>

Property field names.

private_names: Vec<PrivateName>

Private names.

bindings: Vec<BindingLocator>

Locators for all bindings in the codeblock.

num_bindings: usize

Number of binding for the function environment.

functions: Vec<Gc<CodeBlock>>

Functions inside this function

arguments_binding: Option<BindingLocator>

The arguments binding location of the function, if set.

compile_environments: Vec<Gc<GcRefCell<CompileTimeEnvironment>>>

Compile time environments in this function.

is_class_constructor: bool

The [[IsClassConstructor]] internal slot.

class_field_initializer_name: Option<Sym>

The [[ClassFieldInitializerName]] internal slot.

function_environment_push_location: u32

Marks the location in the code where the function environment in pushed. This is only relevant for functions with expressions in the parameters. We execute the parameter expressions in the function code and push the function environment afterward. When the execution of the parameter expressions throws an error, we do not need to pop the function environment.

literals_map: FxHashMap<Literal, u32>names_map: FxHashMap<Identifier, u32>private_names_map: FxHashMap<PrivateName, u32>bindings_map: FxHashMap<BindingLocator, u32>jump_info: Vec<JumpControlInfo>in_async_generator: booljson_parse: boolcontext: &'b mut Context<'host>
Implementations

impl ByteCompiler<'_, '_>

pub(crate) fn compile_class(&mut self, class: &Class, expression: bool)

This function compiles a class declaration or expression.

The compilation of a class declaration and expression is mostly equal. A class declaration binds the resulting class object to it’s identifier. A class expression leaves the resulting class object on the stack for following operations.

impl ByteCompiler<'_, '_>

pub(crate) fn compile_declaration_pattern_impl(
&mut self,
pattern: &Pattern,
def: BindingOpcode
)

fn compile_array_pattern_element(
&mut self,
element: &ArrayPatternElement,
def: BindingOpcode,
with_done: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_assign(&mut self, assign: &Assign, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_binary(&mut self, binary: &Binary, use_expr: bool)

pub(crate) fn compile_binary_in_private(
&mut self,
binary: &BinaryInPrivate,
use_expr: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_object_literal(
&mut self,
object: &ObjectLiteral,
use_expr: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_unary(&mut self, unary: &Unary, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_update(&mut self, update: &Update, use_expr: bool)

impl ByteCompiler<'_, '_>

fn compile_literal(&mut self, lit: &AstLiteral, use_expr: bool)

fn compile_conditional(&mut self, op: &Conditional, use_expr: bool)

fn compile_template_literal(
&mut self,
template_literal: &TemplateLiteral,
use_expr: bool
)

pub(crate) fn compile_expr_impl(&mut self, expr: &Expression, use_expr: bool)

impl ByteCompiler<'_, '_>

pub(crate) fn push_empty_loop_jump_control(&mut self)

Pushes a generic JumpControlInfo onto ByteCompiler

Default JumpControlInfoKind is JumpControlInfoKind::Loop

pub(crate) fn current_jump_control_mut(
&mut self
) -> Option<&mut JumpControlInfo>

pub(crate) fn set_jump_control_start_address(&mut self, start_address: u32)

pub(crate) fn set_jump_control_in_finally(&mut self, value: bool)

pub(crate) fn set_jump_control_in_catch(&mut self, value: bool)

pub(crate) fn push_labelled_control_info(
&mut self,
label: Sym,
start_address: u32
)

Pushes a LabelledStatement’s JumpControlInfo onto the jump_info stack.

pub(crate) fn pop_labelled_control_info(&mut self)

Pops and handles the info for a label’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a LabelledStatement.

pub(crate) fn push_loop_control_info(
&mut self,
label: Option<Sym>,
start_address: u32
)

Pushes an WhileStatement, ForStatement or DoWhileStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn push_loop_control_info_for_of_in_loop(
&mut self,
label: Option<Sym>,
start_address: u32
)

Pushes a ForInOfStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn pop_loop_control_info(&mut self)

Pops and handles the info for a loop control block’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a loop block.

pub(crate) fn push_switch_control_info(
&mut self,
label: Option<Sym>,
start_address: u32
)

Pushes a SwitchStatement’s JumpControlInfo on to the jump_info stack.

pub(crate) fn pop_switch_control_info(&mut self)

Pops and handles the info for a switch block’s JumpControlInfo

Panic

Will panic if jump_info stack is empty.

Will panic if popped JumpControlInfo is not for a switch block.

pub(crate) fn push_try_control_info(
&mut self,
has_finally: bool,
start_address: u32
)

Pushes a TryStatement’s JumpControlInfo onto the jump_info stack.

pub(crate) fn pop_try_control_info(&mut self, try_end: u32)

Pops and handles the info for a try block’s JumpControlInfo

Panic

Will panic if jump_info is empty.

Will panic if popped JumpControlInfo is not for a try block.

pub(crate) fn push_init_finally_control_info(&mut self)

Pushes a TryStatement’s Finally block JumpControlInfo onto the jump_info stack.

pub(crate) fn pop_finally_control_info(&mut self)

impl ByteCompiler<'_, '_>

pub fn compile_module_item_list(
&mut self,
list: &ModuleItemList,
configurable_globals: bool
)

Compiles a [ModuleItemList].

pub fn compile_module_item(
&mut self,
item: &ModuleItem,
configurable_globals: bool
)

Compiles a [ModuleItem].

pub(crate) fn create_module_decls(
&mut self,
stmt_list: &ModuleItemList,
configurable_globals: bool
)

Creates the declarations for a module.

pub(crate) fn create_decls_from_module_item(
&mut self,
item: &ModuleItem,
configurable_globals: bool
) -> bool

Creates the declarations from a [ModuleItem].

impl ByteCompiler<'_, '_>

pub(crate) fn compile_block(
&mut self,
block: &Block,
use_expr: bool,
configurable_globals: bool
)

Compile a [Block] boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_break(&mut self, node: Break)

Compile a [Break] boa_ast node

fn search_jump_info_label(&mut self, address: Label, node_label: Sym)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_continue(&mut self, node: Continue)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_if(
&mut self,
node: &If,
use_expr: bool,
configurable_globals: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_labelled(
&mut self,
labelled: &Labelled,
use_expr: bool,
configurable_globals: bool
)

Compile a [Labelled] boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_for_loop(
&mut self,
for_loop: &ForLoop,
label: Option<Sym>,
configurable_globals: bool
)

pub(crate) fn compile_for_in_loop(
&mut self,
for_in_loop: &ForInLoop,
label: Option<Sym>,
configurable_globals: bool
)

pub(crate) fn compile_for_of_loop(
&mut self,
for_of_loop: &ForOfLoop,
label: Option<Sym>,
configurable_globals: bool
)

pub(crate) fn compile_while_loop(
&mut self,
while_loop: &WhileLoop,
label: Option<Sym>,
configurable_globals: bool
)

pub(crate) fn compile_do_while_loop(
&mut self,
do_while_loop: &DoWhileLoop,
label: Option<Sym>,
configurable_globals: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_switch(
&mut self,
switch: &Switch,
configurable_globals: bool
)

Compile a [Switch] boa_ast node

impl ByteCompiler<'_, '_>

pub(crate) fn compile_try(
&mut self,
t: &Try,
use_expr: bool,
configurable_globals: bool
)

pub(crate) fn compile_catch_stmt(
&mut self,
parent_try: &Try,
use_expr: bool,
configurable_globals: bool
)

pub(crate) fn compile_finally_stmt(
&mut self,
finally: &Finally,
finally_end_label: Label,
configurable_globals: bool
)

impl ByteCompiler<'_, '_>

pub(crate) fn compile_with(&mut self, with: &With, configurable_globals: bool)

Compile a [With] boa_ast node

impl ByteCompiler<'_, '_>

pub fn compile_stmt(
&mut self,
node: &Statement,
use_expr: bool,
configurable_globals: bool
)

Compiles a [Statement] boa_ast node.

impl ByteCompiler<'_, '_>

pub(super) fn iterator_close(&mut self, async_: bool)

Closes an iterator

This is equivalent to the IteratorClose and AsyncIteratorClose operations.

Stack:

iterator, next_method, done => <empty>

impl<'b, 'host> ByteCompiler<'b, 'host>

const DUMMY_ADDRESS: u32 = 4_294_967_295u32

Represents a placeholder address that will be patched later.

pub fn new(
name: Sym,
strict: bool,
json_parse: bool,
context: &'b mut Context<'host>
) -> ByteCompiler<'b, 'host>

Creates a new ByteCompiler.

fn interner(&self) -> &Interner

fn push_compile_environment(
&mut self,
environment: Gc<GcRefCell<CompileTimeEnvironment>>
) -> usize

Push a compile time environment to the current CodeBlock and return it’s index.

fn get_or_insert_literal(&mut self, literal: Literal) -> u32

fn get_or_insert_name(&mut self, name: Identifier) -> u32

fn get_or_insert_private_name(&mut self, name: PrivateName) -> u32

fn get_or_insert_binding(&mut self, binding: BindingLocator) -> u32

fn emit_binding(&mut self, opcode: BindingOpcode, name: Identifier)

fn next_opcode_location(&mut self) -> u32

fn emit(&mut self, opcode: Opcode, operands: &[u32])

fn emit_u64(&mut self, value: u64)

fn emit_u32(&mut self, value: u32)

fn emit_u16(&mut self, value: u16)

fn emit_opcode(&mut self, opcode: Opcode)

fn emit_u8(&mut self, value: u8)

fn emit_push_integer(&mut self, value: i32)

fn emit_push_literal(&mut self, literal: Literal)

fn emit_push_rational(&mut self, value: f64)

fn jump(&mut self) -> Label

fn jump_if_true(&mut self) -> Label

fn jump_if_false(&mut self) -> Label

fn jump_if_not_undefined(&mut self) -> Label

fn jump_if_null_or_undefined(&mut self) -> Label

fn emit_opcode_with_operand(&mut self, opcode: Opcode) -> Label

Emit an opcode with a dummy operand. Return the Label of the operand.

fn emit_opcode_with_two_operands(&mut self, opcode: Opcode) -> (Label, Label)

Emit an opcode with two dummy operands. Return the Labels of the two operands.

fn patch_jump_with_target(&mut self, label: Label, target: u32)

fn patch_jump(&mut self, label: Label)

fn access_get(&mut self, access: Access<'_>, use_expr: bool)

fn access_set_top_of_stack_expr_fn(
compiler: &mut ByteCompiler<'_, '_>,
level: u8
)

fn access_set<F, R>(&mut self, access: Access<'_>, use_expr: bool, expr_fn: F)where
F: FnOnce(&mut ByteCompiler<'_, '_>, u8) -> R,

fn access_delete(&mut self, access: Access<'_>)

pub fn compile_statement_list(
&mut self,
list: &StatementList,
use_expr: bool,
configurable_globals: bool
)

Compile a [StatementList].

pub(crate) fn compile_statement_list_with_new_declarative(
&mut self,
list: &StatementList,
use_expr: bool,
strict: bool
)

Compile a statement list in a new declarative environment.

pub fn compile_expr(&mut self, expr: &Expression, use_expr: bool)

Compile an [Expression].

fn compile_access_preserve_this(&mut self, access: &PropertyAccess)

Compile a property access expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the property access becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the access, and value is the final result of the access.

This is mostly useful for optional chains with calls (a.b?.()) and for regular chains with calls (a.b()), since both of them must have a be the value of this for the function call b(), but a regular compilation of the access would lose the this value after accessing b.

fn compile_optional_preserve_this(&mut self, optional: &Optional)

Compile an optional chain expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the optional chain becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the chain, and value is the result of the chain.

This is mostly useful for inner optional chains with external calls ((a?.b)()), because the external call is not in the optional chain, and compiling an optional chain in the usual way would only return the result of the chain without preserving the this value. In other words, this would be set to undefined for that call, which is incorrect since a should be the this value of the call.

fn compile_optional_item_kind(&mut self, kind: &OptionalOperationKind)

Compile a single operation in an optional chain.

On successful compilation, the state of the stack on execution will become ...rest, this, value, where this is the target of the property access (undefined on calls), and value is the result of executing the action. For example, in the expression a?.b.c(), after compiling and executing:

a?.b, the state of the stack will become ...rest, a, b.

b.c, the state of the stack will become ...rest, b, c.

c(), the state of the stack will become ...rest, undefined, c().

Requirements

This should only be called after verifying that the previous value of the chain is not null or undefined (if the operator ?. was used).

This assumes that the state of the stack before compiling is ...rest, this, value, since the operation compiled by this function could be a call.

fn compile_var_decl(&mut self, decl: &VarDeclaration)

Compile a [VarDeclaration].

fn compile_lexical_decl(&mut self, decl: &LexicalDeclaration)

Compile a [LexicalDeclaration].

fn compile_stmt_list_item(
&mut self,
item: &StatementListItem,
use_expr: bool,
configurable_globals: bool
)

Compile a [StatementListItem].

pub fn compile_decl(&mut self, decl: &Declaration)

Compile a [Declaration].

fn function(
&mut self,
function: FunctionSpec<'_>,
node_kind: NodeKind,
use_expr: bool
)

Compile a function AST Node into bytecode.

fn method(
&mut self,
function: FunctionSpec<'_>,
node_kind: NodeKind,
class_name: Sym,
use_expr: bool
)

Compile a class method AST Node into bytecode.

fn call(&mut self, callable: Callable<'_>, use_expr: bool)

pub fn finish(self) -> CodeBlock

Finish compiling code with the ByteCompiler and return the generated CodeBlock.

fn compile_declaration_pattern(&mut self, pattern: &Pattern, def: BindingOpcode)

pub(crate) fn create_script_decls(
&mut self,
stmt_list: &StatementList,
configurable_globals: bool
)

Creates the declarations for a sript.

pub(crate) fn create_decls_from_var_decl(
&mut self,
list: &VarDeclaration,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_lexical_decl(
&mut self,
list: &LexicalDeclaration
) -> bool

pub(crate) fn create_decls_from_decl(
&mut self,
declaration: &Declaration,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_stmt(
&mut self,
statement: &Statement,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_stmt_list_item(
&mut self,
item: &StatementListItem,
configurable_globals: bool
) -> bool

fn class(&mut self, class: &Class, expression: bool)