Struct boa_engine::bytecompiler::ByteCompiler

pub struct ByteCompiler<'b> {
    code_block: CodeBlock,
    literals_map: FxHashMap<Literal, u32>,
    names_map: FxHashMap<Identifier, u32>,
    bindings_map: FxHashMap<BindingLocator, u32>,
    jump_info: Vec<JumpControlInfo>,
    in_async_generator: bool,
    json_parse: bool,
    context: &'b mut Context,
}

Fields
code_block: CodeBlockliterals_map: FxHashMap<Literal, u32>names_map: FxHashMap<Identifier, u32>bindings_map: FxHashMap<BindingLocator, u32>jump_info: Vec<JumpControlInfo>in_async_generator: booljson_parse: boolcontext: &'b mut Context
Implementations

impl<'b> ByteCompiler<'b>

const DUMMY_ADDRESS: u32 = 4_294_967_295u32

Represents a placeholder address that will be patched later.

pub fn new(
name: Sym,
strict: bool,
json_parse: bool,
context: &'b mut Context
) -> Self

fn interner(&self) -> &Interner

fn push_compile_environment(
&mut self,
environment: Gc<GcCell<CompileTimeEnvironment>>
) -> usize

Push a compile time environment to the current CodeBlock and return it’s index.

fn get_or_insert_literal(&mut self, literal: Literal) -> u32

fn get_or_insert_name(&mut self, name: Identifier) -> u32

fn get_or_insert_binding(&mut self, binding: BindingLocator) -> u32

fn emit_binding(&mut self, opcode: BindingOpcode, name: Identifier)

fn next_opcode_location(&mut self) -> u32

fn emit(&mut self, opcode: Opcode, operands: &[u32])

fn emit_u64(&mut self, value: u64)

fn emit_u32(&mut self, value: u32)

fn emit_u16(&mut self, value: u16)

fn emit_opcode(&mut self, opcode: Opcode)

fn emit_u8(&mut self, value: u8)

fn emit_push_integer(&mut self, value: i32)

fn emit_push_literal(&mut self, literal: Literal)

fn emit_push_rational(&mut self, value: f64)

fn jump(&mut self) -> Label

fn jump_if_false(&mut self) -> Label

fn jump_if_null_or_undefined(&mut self) -> Label

fn emit_opcode_with_operand(&mut self, opcode: Opcode) -> Label

Emit an opcode with a dummy operand. Return the Label of the operand.

fn emit_opcode_with_two_operands(&mut self, opcode: Opcode) -> (Label, Label)

Emit an opcode with two dummy operands. Return the Labels of the two operands.

fn patch_jump_with_target(&mut self, label: Label, target: u32)

fn patch_jump(&mut self, label: Label)

fn push_loop_control_info(&mut self, label: Option<Sym>, start_address: u32)

fn push_loop_control_info_for_of_in_loop(
&mut self,
label: Option<Sym>,
start_address: u32
)

fn pop_loop_control_info(&mut self)

fn push_switch_control_info(&mut self, label: Option<Sym>, start_address: u32)

fn pop_switch_control_info(&mut self)

fn push_try_control_info(&mut self, has_finally: bool)

fn push_try_control_info_catch_start(&mut self)

fn push_try_control_info_finally_start(&mut self, start: Label)

fn pop_try_control_info(&mut self, finally_start_address: Option<u32>)

fn push_labelled_block_control_info(&mut self, label: Sym, start_address: u32)

fn pop_labelled_block_control_info(&mut self)

fn access_get(&mut self, access: Access<'_>, use_expr: bool) -> JsResult<()>

fn access_set_top_of_stack_expr_fn(
compiler: &mut ByteCompiler<'_>,
level: u8
) -> JsResult<()>

fn access_set<F, R>(
&mut self,
access: Access<'_>,
use_expr: bool,
expr_fn: F
) -> JsResult<R>where
F: FnOnce(&mut ByteCompiler<'_>, u8) -> JsResult<R>,

fn access_delete(&mut self, access: Access<'_>) -> JsResult<()>

pub fn compile_statement_list(
&mut self,
list: &StatementList,
use_expr: bool,
configurable_globals: bool
) -> JsResult<()>

pub(crate) fn compile_statement_list_with_new_declarative(
&mut self,
list: &StatementList,
use_expr: bool,
strict: bool
) -> JsResult<()>

Compile a statement list in a new declarative environment.

pub fn compile_expr(&mut self, expr: &Expression, use_expr: bool) -> JsResult<()>

fn compile_access_preserve_this(
&mut self,
access: &PropertyAccess
) -> JsResult<()>

Compile a property access expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the property access becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the access, and value is the final result of the access.

This is mostly useful for optional chains with calls (a.b?.()) and for regular chains with calls (a.b()), since both of them must have a be the value of this for the function call b(), but a regular compilation of the access would lose the this value after accessing b.

fn compile_optional_preserve_this(&mut self, optional: &Optional) -> JsResult<()>

Compile an optional chain expression, prepending this to the property value in the stack.

This compiles the access in a way that the state of the stack after executing the optional chain becomes ...rest, this, value. where ...rest is the rest of the stack, this is the this value of the chain, and value is the result of the chain.

This is mostly useful for inner optional chains with external calls ((a?.b)()), because the external call is not in the optional chain, and compiling an optional chain in the usual way would only return the result of the chain without preserving the this value. In other words, this would be set to undefined for that call, which is incorrect since a should be the this value of the call.

fn compile_optional_item_kind(
&mut self,
kind: &OptionalOperationKind
) -> JsResult<()>

Compile a single operation in an optional chain.

On successful compilation, the state of the stack on execution will become ...rest, this, value, where this is the target of the property access (undefined on calls), and value is the result of executing the action. For example, in the expression a?.b.c(), after compiling and executing:

a?.b, the state of the stack will become ...rest, a, b.

b.c, the state of the stack will become ...rest, b, c.

c(), the state of the stack will become ...rest, undefined, c().

Requirements

This should only be called after verifying that the previous value of the chain is not null or undefined (if the operator ?. was used).

This assumes that the state of the stack before compiling is ...rest, this, value, since the operation compiled by this function could be a call.

pub fn compile_var_decl(&mut self, decl: &VarDeclaration) -> JsResult<()>

pub fn compile_lexical_decl(&mut self, decl: &LexicalDeclaration) -> JsResult<()>

pub fn compile_stmt_list_item(
&mut self,
item: &StatementListItem,
use_expr: bool,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_decl(&mut self, decl: &Declaration) -> JsResult<()>

pub fn compile_for_loop(
&mut self,
for_loop: &ForLoop,
label: Option<Sym>,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_for_in_loop(
&mut self,
for_in_loop: &ForInLoop,
label: Option<Sym>,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_for_of_loop(
&mut self,
for_of_loop: &ForOfLoop,
label: Option<Sym>,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_while_loop(
&mut self,
while_loop: &WhileLoop,
label: Option<Sym>,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_do_while_loop(
&mut self,
do_while_loop: &DoWhileLoop,
label: Option<Sym>,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_block(
&mut self,
block: &Block,
label: Option<Sym>,
use_expr: bool,
configurable_globals: bool
) -> JsResult<()>

pub fn compile_stmt(
&mut self,
node: &Statement,
use_expr: bool,
configurable_globals: bool
) -> JsResult<()>

fn function(
&mut self,
function: FunctionSpec<'_>,
node_kind: NodeKind,
use_expr: bool
) -> JsResult<()>

Compile a function AST Node into bytecode.

fn call(&mut self, callable: Callable<'_>, use_expr: bool) -> JsResult<()>

pub fn finish(self) -> CodeBlock

fn compile_declaration_pattern(
&mut self,
pattern: &Pattern,
def: BindingOpcode
) -> JsResult<()>

pub(crate) fn create_decls(
&mut self,
stmt_list: &StatementList,
configurable_globals: bool
)

pub(crate) fn create_decls_from_var_decl(
&mut self,
list: &VarDeclaration,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_lexical_decl(
&mut self,
list: &LexicalDeclaration
) -> bool

pub(crate) fn create_decls_from_decl(
&mut self,
declaration: &Declaration,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_stmt(
&mut self,
statement: &Statement,
configurable_globals: bool
) -> bool

pub(crate) fn create_decls_from_stmt_list_item(
&mut self,
item: &StatementListItem,
configurable_globals: bool
) -> bool

fn class(&mut self, class: &Class, expression: bool) -> JsResult<()>

This function compiles a class declaration or expression.

The compilation of a class declaration and expression is mostly equal. A class declaration binds the resulting class object to it’s identifier. A class expression leaves the resulting class object on the stack for following operations.