Struct boa_engine::bytecompiler::ByteCompiler

pub struct ByteCompiler<'b> {
    code_block: CodeBlock,
    literals_map: FxHashMap<Literal, u32>,
    names_map: FxHashMap<Sym, u32>,
    bindings_map: FxHashMap<BindingLocator, u32>,
    jump_info: Vec<JumpControlInfo>,
    context: &'b mut Context,
}

Fields
code_block: CodeBlockliterals_map: FxHashMap<Literal, u32>names_map: FxHashMap<Sym, u32>bindings_map: FxHashMap<BindingLocator, u32>jump_info: Vec<JumpControlInfo>context: &'b mut Context
Implementations

impl<'b> ByteCompiler<'b>

const DUMMY_ADDRESS: u32 = 4_294_967_295u32

Represents a placeholder address that will be patched later.

pub fn new(name: Sym, strict: bool, context: &'b mut Context) -> Self

fn interner(&self) -> &Interner

fn push_compile_environment(
&mut self,
environment: Gc<Cell<CompileTimeEnvironment>>
) -> usize

Push a compile time environment to the current CodeBlock and return it’s index.

fn get_or_insert_literal(&mut self, literal: Literal) -> u32

fn get_or_insert_name(&mut self, name: Sym) -> u32

fn get_or_insert_binding(&mut self, binding: BindingLocator) -> u32

fn emit_binding(&mut self, opcode: BindingOpcode, name: Sym)

fn next_opcode_location(&mut self) -> u32

fn emit(&mut self, opcode: Opcode, operands: &[u32])

fn emit_u64(&mut self, value: u64)

fn emit_u32(&mut self, value: u32)

fn emit_u16(&mut self, value: u16)

fn emit_opcode(&mut self, opcode: Opcode)

fn emit_u8(&mut self, value: u8)

fn emit_push_integer(&mut self, value: i32)

fn emit_push_literal(&mut self, literal: Literal)

fn emit_push_rational(&mut self, value: f64)

fn jump(&mut self) -> Label

fn jump_if_false(&mut self) -> Label

fn emit_opcode_with_operand(&mut self, opcode: Opcode) -> Label

Emit an opcode with a dummy operand. Return the Label of the operand.

fn emit_opcode_with_two_operands(&mut self, opcode: Opcode) -> (Label, Label)

Emit an opcode with two dummy operands. Return the Labels of the two operands.

fn patch_jump_with_target(&mut self, label: Label, target: u32)

fn patch_jump(&mut self, label: Label)

fn push_loop_control_info(&mut self, label: Option<Sym>, start_address: u32)

fn push_loop_control_info_for_of_in_loop(
&mut self,
label: Option<Sym>,
start_address: u32
)

fn pop_loop_control_info(&mut self)

fn push_switch_control_info(&mut self, label: Option<Sym>, start_address: u32)

fn pop_switch_control_info(&mut self)

fn push_try_control_info(&mut self, has_finally: bool)

fn push_try_control_info_catch_start(&mut self)

fn push_try_control_info_finally_start(&mut self, start: Label)

fn pop_try_control_info(&mut self, finally_start_address: Option<u32>)

fn compile_access(node: &Node) -> Option<Access<'_>>

fn access_get(&mut self, access: Access<'_>, use_expr: bool) -> JsResult<()>

fn access_set(
&mut self,
access: Access<'_>,
expr: Option<&Node>,
use_expr: bool
) -> JsResult<()>

pub fn compile_statement_list(
&mut self,
list: &[Node],
use_expr: bool
) -> JsResult<()>

pub(crate) fn compile_statement_list_with_new_declarative(
&mut self,
list: &[Node],
use_expr: bool,
strict: bool
) -> JsResult<()>

Compile a statement list in a new declarative environment.

pub fn compile_expr(&mut self, expr: &Node, use_expr: bool) -> JsResult<()>

pub fn compile_stmt(&mut self, node: &Node, use_expr: bool) -> JsResult<()>

pub(crate) fn compile_function_code(
kind: FunctionKind,
name: Option<Sym>,
parameters: &FormalParameterList,
body: &StatementList,
generator: bool,
strict: bool,
context: &mut Context
) -> JsResult<Gc<CodeBlock>>

Compile a function statement list and it’s parameters into bytecode.

pub(crate) fn function(
&mut self,
function: &Node,
use_expr: bool
) -> JsResult<()>

Compile a function AST Node into bytecode.

pub(crate) fn call(&mut self, node: &Node, use_expr: bool) -> JsResult<()>

pub fn finish(self) -> CodeBlock

fn compile_declaration_pattern(
&mut self,
pattern: &DeclarationPattern,
def: BindingOpcode
) -> JsResult<()>

pub(crate) fn create_declarations(&mut self, stmt_list: &[Node]) -> JsResult<()>

pub(crate) fn create_decls_from_stmt(&mut self, node: &Node) -> JsResult<bool>

fn class(&mut self, class: &Class, expression: bool) -> JsResult<()>

This function compiles a class declaration or expression.

The compilation of a class declaration and expression is mostly equal. A class declaration binds the resulting class object to it’s identifier. A class expression leaves the resulting class object on the stack for following operations.