Trait boa_engine::class::ClassConstructor

pub trait ClassConstructor: Class {
    fn raw_constructor(
        this: &JsValue, 
        args: &[JsValue], 
        context: &mut Context
    ) -> JsResult<JsValue>
    where
        Self: Sized;
}

This is a wrapper around Class::constructor that sets the internal data of a class.

This is automatically implemented, when a type implements Class.

Required methods

fn raw_constructor(
this: &JsValue,
args: &[JsValue],
context: &mut Context
) -> JsResult<JsValue> where
Self: Sized,

The raw constructor that matches the NativeFunction signature.