Module boa_engine::context

The ECMAScript context.

Modules

hooks 🔒

icu 🔒

intrinsics

Data structures that contain intrinsic objects and constructors.

Structs

Context

ECMAScript context. It is the primary way to interact with the runtime.

ContextBuilder

Builder for the Context type.

Enums

BoaProvider

ICU4X data provider used in boa.

Traits

HostHooks

Host Hooks customizable by the host code or engine.