Module boa_engine::context

The ECMAScript context.

Modules

hooks 🔒

icu 🔒

intrinsics

Data structures that contain intrinsic objects and constructors.

maybe_shared 🔒

Structs

Context

ECMAScript context. It is the primary way to interact with the runtime.

ContextBuilder

Builder for the Context type.

ContextCleanupGuard 🔒

A cleanup guard for a Context that is executed when dropped.

DefaultHooks

Default implementation of HostHooks, which doesn’t carry any state.

Enums

BoaProvider

ICU4X data provider used in boa.

IcuError

Error thrown when the engine cannot initialize the ICU tools from a data provider.

MaybeShared

A Cow-like pointer where the Owned variant is an Rc.

Traits

HostHooks

Host Hooks customizable by the host code or engine.