Struct boa_engine::context::Context

pub struct Context<'host> {

    realm: Realm,
    interner: Interner,
    strict: bool,
    pub(crate) instructions_remaining: usize,
    pub(crate) vm: Vm,
    pub(crate) kept_alive: Vec<JsObject>,
    icu: Icu<'host>,
    host_hooks: MaybeShared<'host, dyn HostHooks>,
    job_queue: MaybeShared<'host, dyn JobQueue>,
    module_loader: MaybeShared<'host, dyn ModuleLoader>,
    optimizer_options: OptimizerOptions,
    root_shape: RootShape,
    parser_identifier: u32,

}

ECMAScript context. It is the primary way to interact with the runtime.

Contexts constructed in a thread share the same runtime, therefore it is possible to share objects from one context to another context, but they have to be in the same thread.

Examples

Execute Function of Script File

use boa_engine::{
    object::ObjectInitializer,
    property::{Attribute, PropertyDescriptor},
    Context,
    Source
};

let script = r#"
function test(arg1) {
    if(arg1 != null) {
        return arg1.x;
    }
    return 112233;
}
"#;

let mut context = Context::default();

// Populate the script definition to the context.
context.eval_script(Source::from_bytes(script)).unwrap();

// Create an object that can be used in eval calls.
let arg = ObjectInitializer::new(&mut context)
    .property("x", 12, Attribute::READONLY)
    .build();
context.register_global_property("arg", arg, Attribute::all());

let value = context.eval_script(Source::from_bytes("test(arg)")).unwrap();

assert_eq!(value.as_number(), Some(12.0))

Fields
realm: Realm

realm holds both the global object and the environment

interner: Interner

String interner in the context.

strict: bool

Execute in strict mode,

instructions_remaining: usize

Number of instructions remaining before a forced exit

vm: Vmkept_alive: Vec<JsObject>icu: Icu<'host>

ICU related utilities

host_hooks: MaybeShared<'host, dyn HostHooks>job_queue: MaybeShared<'host, dyn JobQueue>module_loader: MaybeShared<'host, dyn ModuleLoader>optimizer_options: OptimizerOptionsroot_shape: RootShapeparser_identifier: u32

Unique identifier for each parser instance used during the context lifetime.

Implementations

impl<'host> Context<'host>

pub fn builder() -> ContextBuilder<'static, 'static, 'static, 'static>

Create a new ContextBuilder to specify the Interner and/or the icu data provider.

pub fn eval_script<R: Read>(&mut self, src: Source<'_, R>) -> JsResult<JsValue>

Evaluates the given script src by compiling down to bytecode, then interpreting the bytecode into a value.

Examples

let mut context = Context::default();

let source = Source::from_bytes("1 + 3");
let value = context.eval_script(source).unwrap();

assert!(value.is_number());
assert_eq!(value.as_number().unwrap(), 4.0);

Note that this won’t run any scheduled promise jobs; you need to call Context::run_jobs on the context or JobQueue::run_jobs on the provided queue to run them.

pub fn optimize_statement_list( &mut self, statement_list: &mut StatementList ) -> OptimizerStatistics

Applies optimizations to the StatementList inplace.

pub fn parse_script<R: Read>( &mut self, src: Source<'_, R> ) -> JsResult<StatementList>

Parse the given source script.

pub fn compile_script( &mut self, statement_list: &StatementList ) -> JsResult<Gc<CodeBlock>>

Compile the script AST into a CodeBlock ready to be executed by the VM.

pub fn execute(&mut self, code_block: Gc<CodeBlock>) -> JsResult<JsValue>

Call the VM with a CodeBlock and return the result.

Since this function receives a Gc<CodeBlock>, cloning the code is very cheap, since it’s just a pointer copy. Therefore, if you’d like to execute the same CodeBlock multiple times, there is no need to re-compile it, and you can just call clone() on the Gc<CodeBlock> returned by the Context::compile_script function.

Note that this won’t run any scheduled promise jobs; you need to call Context::run_jobs on the context or JobQueue::run_jobs on the provided queue to run them.

pub fn register_global_property<K, V>( &mut self, key: K, value: V, attribute: Attribute ) -> JsResult<()>where K: Into<PropertyKey>, V: Into<JsValue>,

Register a global property.

It will return an error if the property is already defined.

Example

use boa_engine::{
    object::ObjectInitializer,
    property::{Attribute, PropertyDescriptor},
    Context,
};

let mut context = Context::default();

context
    .register_global_property("myPrimitiveProperty", 10, Attribute::all())
    .expect("property shouldn't exist");

let object = ObjectInitializer::new(&mut context)
    .property("x", 0, Attribute::all())
    .property("y", 1, Attribute::all())
    .build();
context
    .register_global_property("myObjectProperty", object, Attribute::all())
    .expect("property shouldn't exist");

pub fn register_global_callable( &mut self, name: &str, length: usize, body: NativeFunction ) -> JsResult<()>

Register a global native callable.

The function will be both constructable (call with new <name>()) and callable (call with <name>()).

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note

If you wish to only create the function object without binding it to the global object, you can use the FunctionObjectBuilder API.

pub fn register_global_builtin_callable( &mut self, name: &str, length: usize, body: NativeFunction ) -> JsResult<()>

Register a global native function that is not a constructor.

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note

The difference to Context::register_global_callable is, that the function will not be constructable. Usage of the function as a constructor will produce a TypeError.

pub fn register_global_class<T>(&mut self) -> JsResult<()>where T: Class,

Register a global class of type T, where T implements Class.

It will return an error if the global property is already defined.

Example

#[derive(Debug, Trace, Finalize)]
struct MyClass;

impl Class for MyClass {
   // ...
}

context.register_global_class::<MyClass>();

pub const fn interner(&self) -> &Interner

Gets the string interner.

pub fn interner_mut(&mut self) -> &mut Interner

Gets a mutable reference to the string interner.

pub fn global_object(&self) -> JsObject

Returns the global object.

pub fn intrinsics(&self) -> &Intrinsics

Returns the currently active intrinsic constructors and objects.

pub const fn realm(&self) -> &Realm

Returns the currently active realm.

pub fn set_trace(&mut self, trace: bool)

Set the value of trace on the context

pub const fn optimizer_options(&self) -> OptimizerOptions

Get optimizer options.

pub fn set_optimizer_options(&mut self, optimizer_options: OptimizerOptions)

Enable or disable optimizations

pub fn strict(&mut self, strict: bool)

Changes the strictness mode of the context.

pub fn enqueue_job(&mut self, job: NativeJob)

Enqueues a NativeJob on the JobQueue.

pub fn run_jobs(&mut self)

Runs all the jobs in the job queue.

pub fn clear_kept_objects(&mut self)

Abstract operation ClearKeptObjects.

Clears all objects maintained alive by calls to the AddToKeptObjects abstract operation, used within the WeakRef constructor.

pub fn stack_trace(&mut self) -> impl Iterator<Item = &CallFrame>

Retrieves the current stack trace of the context.

pub fn enter_realm(&mut self, realm: Realm) -> Realm

Replaces the currently active realm with realm, and returns the old realm.

pub const fn root_shape(&self) -> &RootShape

Get the RootShape.

pub fn host_hooks(&self) -> MaybeShared<'host, dyn HostHooks>

Gets the host hooks.

pub fn job_queue(&self) -> MaybeShared<'host, dyn JobQueue>

Gets the job queue.

pub fn module_loader(&self) -> MaybeShared<'host, dyn ModuleLoader>

Gets the module loader.

pub const fn runtime_limits(&self) -> RuntimeLimits

Get the RuntimeLimits.

pub fn set_runtime_limits(&mut self, runtime_limits: RuntimeLimits)

Set the RuntimeLimits.

pub fn runtime_limits_mut(&mut self) -> &mut RuntimeLimits

Get a mutable reference to the RuntimeLimits.

impl Context<'_>

pub(crate) fn swap_realm(&mut self, realm: &mut Realm)

Swaps the currently active realm with realm.

pub(crate) fn next_parser_identifier(&mut self) -> u32

Increment and get the parser identifier.

pub(crate) fn can_declare_global_function( &mut self, name: Identifier ) -> JsResult<bool>

CanDeclareGlobalFunction ( N )

More information:

ECMAScript reference

pub(crate) fn can_declare_global_var( &mut self, name: Identifier ) -> JsResult<bool>

CanDeclareGlobalVar ( N )

More information:

ECMAScript reference

pub(crate) fn create_global_var_binding( &mut self, name: Identifier, configurable: bool ) -> JsResult<()>

CreateGlobalVarBinding ( N, D )

More information:

ECMAScript reference

pub(crate) fn create_global_function_binding( &mut self, name: Identifier, function: JsObject, configurable: bool ) -> JsResult<()>

CreateGlobalFunctionBinding ( N, V, D )

More information:

ECMAScript reference

pub(crate) fn has_restricted_global_property( &mut self, name: Identifier ) -> JsResult<bool>

HasRestrictedGlobalProperty ( N )

More information:

ECMAScript reference

impl<'host> Context<'host>

pub(crate) fn guard<F>( &mut self, cleanup: F ) -> ContextCleanupGuard<'_, 'host, F>where F: FnOnce(&mut Context<'_>) + 'static,

Creates a ContextCleanupGuard that executes some cleanup after being dropped.

pub(crate) const fn icu(&self) -> &Icu<'host>

Get the ICU related utilities

impl Context<'_>

pub(crate) fn find_runtime_binding( &mut self, locator: &mut BindingLocator ) -> JsResult<()>

Gets the corresponding runtime binding of the provided BindingLocator, modifying its indexes in place.

This readjusts a BindingLocator to the correct binding if a with environment or eval call modified the compile-time bindings.

Only use if the binding origin is unknown or comes from a var declaration. Lexical bindings are completely removed of runtime checks because the specification guarantees that runtime semantics cannot add or remove lexical bindings.

pub(crate) fn is_initialized_binding( &mut self, locator: &BindingLocator ) -> JsResult<bool>

Checks if the binding pointed by locator is initialized.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn get_binding( &mut self, locator: BindingLocator ) -> JsResult<Option<JsValue>>

Get the value of a binding.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn set_binding( &mut self, locator: BindingLocator, value: JsValue, strict: bool ) -> JsResult<()>

Sets the value of a binding.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn delete_binding( &mut self, locator: BindingLocator ) -> JsResult<bool>

Deletes a binding if it exists.

Returns true if the binding was deleted.

Panics

Panics if the environment or binding index are out of range.

pub(crate) fn environment_expect(&self, index: u32) -> &Environment

Return the environment at the given index. Panics if the index is out of range.

impl Context<'_>

fn execute_instruction(&mut self) -> JsResult<CompletionType>

pub(crate) fn run(&mut self) -> CompletionRecord