Struct boa_engine::context::Context

pub struct Context<'host> {
    pub(crate) realm: Realm,
    interner: Interner,
    strict: bool,
    pub(crate) instructions_remaining: usize,
    pub(crate) vm: Vm,
    pub(crate) kept_alive: Vec<JsObject>,
    icu: Icu<'host>,
    host_hooks: &'host dyn HostHooks,
    job_queue: &'host dyn JobQueue,
}

ECMAScript context. It is the primary way to interact with the runtime.

Contexts constructed in a thread share the same runtime, therefore it is possible to share objects from one context to another context, but they have to be in the same thread.

Examples

Execute Function of Script File

use boa_engine::{
    object::ObjectInitializer,
    property::{Attribute, PropertyDescriptor},
    Context,
    Source
};

let script = r#"
function test(arg1) {
    if(arg1 != null) {
        return arg1.x;
    }
    return 112233;
}
"#;

let mut context = Context::default();

// Populate the script definition to the context.
context.eval_script(Source::from_bytes(script)).unwrap();

// Create an object that can be used in eval calls.
let arg = ObjectInitializer::new(&mut context)
    .property("x", 12, Attribute::READONLY)
    .build();
context.register_global_property("arg", arg, Attribute::all());

let value = context.eval_script(Source::from_bytes("test(arg)")).unwrap();

assert_eq!(value.as_number(), Some(12.0))

Fields
realm: Realm

realm holds both the global object and the environment

interner: Interner

String interner in the context.

strict: bool

Execute in strict mode,

instructions_remaining: usize

Number of instructions remaining before a forced exit

vm: Vmkept_alive: Vec<JsObject>icu: Icu<'host>

ICU related utilities

host_hooks: &'host dyn HostHooksjob_queue: &'host dyn JobQueue
Implementations

impl Context<'_>

pub fn builder() -> ContextBuilder<'static, 'static, 'static>

Create a new ContextBuilder to specify the [Interner] and/or the icu data provider.

pub fn eval_script<R: Read>(&mut self, src: Source<'_, R>) -> JsResult<JsValue>

Evaluates the given script src by compiling down to bytecode, then interpreting the bytecode into a value.

Examples

let mut context = Context::default();

let source = Source::from_bytes("1 + 3");
let value = context.eval_script(source).unwrap();

assert!(value.is_number());
assert_eq!(value.as_number().unwrap(), 4.0);

Note that this won’t run any scheduled promise jobs; you need to call Context::run_jobs on the context or JobQueue::run_jobs on the provided queue to run them.

pub fn eval_module<R: Read>(&mut self, src: Source<'_, R>) -> JsResult<JsValue>

Evaluates the given module src by compiling down to bytecode, then interpreting the bytecode into a value.

Examples

let mut context = Context::default();

let source = Source::from_bytes("1 + 3");

let value = context.eval_module(source).unwrap();

assert!(value.is_number());
assert_eq!(value.as_number().unwrap(), 4.0);

pub fn parse_script<R: Read>(
&mut self,
src: Source<'_, R>
) -> Result<StatementList, ParseError>

Parse the given source script.

pub fn parse_module<R: Read>(
&mut self,
src: Source<'_, R>
) -> Result<ModuleItemList, ParseError>

Parse the given source script.

pub fn compile_script(
&mut self,
statement_list: &StatementList
) -> JsResult<Gc<CodeBlock>>

Compile the script AST into a CodeBlock ready to be executed by the VM.

pub fn compile_module(
&mut self,
statement_list: &ModuleItemList
) -> JsResult<Gc<CodeBlock>>

Compile the module AST into a CodeBlock ready to be executed by the VM.

pub fn execute(&mut self, code_block: Gc<CodeBlock>) -> JsResult<JsValue>

Call the VM with a CodeBlock and return the result.

Since this function receives a Gc<CodeBlock>, cloning the code is very cheap, since it’s just a pointer copy. Therefore, if you’d like to execute the same CodeBlock multiple times, there is no need to re-compile it, and you can just call clone() on the Gc<CodeBlock> returned by the Context::compile_script or Context::compile_module functions.

Note that this won’t run any scheduled promise jobs; you need to call Context::run_jobs on the context or JobQueue::run_jobs on the provided queue to run them.

pub fn register_global_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
)where
K: Into<PropertyKey>,
V: Into<JsValue>,

Register a global property.

Example

use boa_engine::{
    object::ObjectInitializer,
    property::{Attribute, PropertyDescriptor},
    Context,
};

let mut context = Context::default();

context.register_global_property("myPrimitiveProperty", 10, Attribute::all());

let object = ObjectInitializer::new(&mut context)
    .property("x", 0, Attribute::all())
    .property("y", 1, Attribute::all())
    .build();
context.register_global_property("myObjectProperty", object, Attribute::all());

pub fn register_global_callable(
&mut self,
name: &str,
length: usize,
body: NativeFunction
)

Register a global native callable.

The function will be both constructable (call with new <name>()) and callable (call with <name>()).

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note

If you wish to only create the function object without binding it to the global object, you can use the FunctionObjectBuilder API.

pub fn register_global_builtin_callable(
&mut self,
name: &str,
length: usize,
body: NativeFunction
)

Register a global native function that is not a constructor.

The function will be bound to the global object with writable, non-enumerable and configurable attributes. The same as when you create a function in JavaScript.

Note

The difference to Context::register_global_callable is, that the function will not be constructable. Usage of the function as a constructor will produce a TypeError.

pub fn register_global_class<T>(&mut self) -> JsResult<()>where
T: Class,

Register a global class of type T, where T implements Class.

Example

#[derive(Debug, Trace, Finalize)]
struct MyClass;

impl Class for MyClass {
   // ...
}

context.register_global_class::<MyClass>();

pub const fn interner(&self) -> &Interner

Gets the string interner.

pub fn interner_mut(&mut self) -> &mut Interner

Gets a mutable reference to the string interner.

pub const fn global_object(&self) -> &JsObject

Return the global object.

pub const fn intrinsics(&self) -> &Intrinsics

Return the intrinsic constructors and objects.

pub fn set_trace(&mut self, trace: bool)

Set the value of trace on the context

pub fn strict(&mut self, strict: bool)

Changes the strictness mode of the context.

pub fn enqueue_job(&mut self, job: NativeJob)

Enqueues a NativeJob on the JobQueue.

pub fn run_jobs(&mut self)

Runs all the jobs in the job queue.

pub fn clear_kept_objects(&mut self)

Abstract operation ClearKeptObjects.

Clears all objects maintained alive by calls to the AddToKeptObjects abstract operation, used within the WeakRef constructor.

impl Context<'_>

pub(crate) fn global_bindings_mut(
&mut self
) -> &mut IndexMap<JsString, PropertyDescriptor, BuildHasherDefault<FxHasher>>

Return a mutable reference to the global object string bindings.

pub(crate) fn compile_json_parse(
&mut self,
statement_list: &StatementList
) -> Gc<CodeBlock>

Compile the AST into a CodeBlock ready to be executed by the VM in a JSON.parse context.

pub(crate) fn compile_with_new_declarative(
&mut self,
statement_list: &StatementList,
strict: bool
) -> Gc<CodeBlock>

Compile the AST into a CodeBlock with an additional declarative environment.

impl<'host> Context<'host>

pub fn host_hooks(&self) -> &'host dyn HostHooks

Get the host hooks.

pub fn job_queue(&mut self) -> &'host dyn JobQueue

Get the job queue.

pub(crate) const fn icu(&self) -> &Icu<'host>

Get the ICU related utilities

impl Context<'_>

pub(crate) fn push_compile_time_environment(&mut self, function_scope: bool)

Push either a new declarative or function environment on the compile time environment stack.

Note: This function only works at bytecode compile time!

pub(crate) fn pop_compile_time_environment(
&mut self
) -> (usize, Gc<GcRefCell<CompileTimeEnvironment>>)

Pop the last compile time environment from the stack.

Note: This function only works at bytecode compile time!

Panics

Panics if there are no more environments that can be pop’ed.

pub(crate) fn get_binding_number(&self) -> usize

Get the number of bindings for the current compile time environment.

Note: This function only works at bytecode compile time!

Panics

Panics if there are no environments on the compile time environment stack.

pub(crate) fn get_binding_value(&self, name: Identifier) -> BindingLocator

Get the binding locator of the binding at bytecode compile time.

Note: This function only works at bytecode compile time!

pub(crate) fn has_binding(&self, name: Identifier) -> bool

Return if a declarative binding exists at bytecode compile time. This does not include bindings on the global object.

Note: This function only works at bytecode compile time!

pub(crate) fn create_mutable_binding(
&mut self,
name: Identifier,
function_scope: bool,
configurable: bool
)

Create a mutable binding at bytecode compile time. This function returns a syntax error, if the binding is a redeclaration.

Note: This function only works at bytecode compile time!

Panics

Panics if the global environment is not function scoped.

pub(crate) fn initialize_mutable_binding(
&self,
name: Identifier,
function_scope: bool
) -> BindingLocator

Initialize a mutable binding at bytecode compile time and return it’s binding locator.

Note: This function only works at bytecode compile time!

pub(crate) fn create_immutable_binding(&mut self, name: Identifier, strict: bool)

Create an immutable binding at bytecode compile time. This function returns a syntax error, if the binding is a redeclaration.

Note: This function only works at bytecode compile time!

Panics

Panics if the global environment does not exist.

pub(crate) fn initialize_immutable_binding(
&self,
name: Identifier
) -> BindingLocator

Initialize an immutable binding at bytecode compile time and return it’s binding locator.

Note: This function only works at bytecode compile time!

Panics

Panics if the global environment does not exist or a the binding was not created on the current environment.

pub(crate) fn set_mutable_binding(&self, name: Identifier) -> BindingLocator

Return the binding locator for a set operation on an existing binding.

Note: This function only works at bytecode compile time!

impl Context<'_>

fn execute_instruction(&mut self) -> JsResult<ShouldExit>

pub(crate) fn run(&mut self) -> JsResult<(JsValue, ReturnType)>