Struct boa_engine::context::ContextCleanupGuard

pub(crate) struct ContextCleanupGuard<'a, 'host, F>where
    F: FnOnce(&mut Context<'_>) + 'static,{
    context: &'a mut Context<'host>,
    cleanup: Option<F>,
}

A cleanup guard for a Context that is executed when dropped.

Fields
context: &'a mut Context<'host>cleanup: Option<F>
Implementations

impl<'a, 'host, F> ContextCleanupGuard<'a, 'host, F>where F: FnOnce(&mut Context<'_>) + 'static,

pub(crate) fn new(context: &'a mut Context<'host>, cleanup: F) -> Self

Creates a new ContextCleanupGuard from the current context and its cleanup operation.