Trait boa_engine::context::HostHooks

pub trait HostHooks {
    fn make_job_callback(
        &self,
        callback: JsFunction,
        _context: &mut Context<'_>
    ) -> JobCallback { ... }
    fn call_job_callback(
        &self,
        job: JobCallback,
        this: &JsValue,
        args: &[JsValue],
        context: &mut Context<'_>
    ) -> JsResult<JsValue> { ... }
    fn promise_rejection_tracker(
        &self,
        _promise: &JsObject,
        _operation: OperationType,
        _context: &mut Context<'_>
    ) { ... }
    fn ensure_can_compile_strings(
        &self,
        _context: &mut Context<'_>
    ) -> JsResult<()> { ... }
    fn has_source_text_available(
        &self,
        _function: &JsFunction,
        _context: &mut Context<'_>
    ) -> bool { ... }
    fn ensure_can_add_private_element(
        &self,
        _o: &JsObject,
        _context: &mut Context<'_>
    ) -> JsResult<()> { ... }
}

Host Hooks customizable by the host code or engine.

Every hook contains on its Requirements section the spec requirements that the hook must abide to for spec compliance.

Usage

Implement the trait for a custom struct (maybe with additional state), overriding the methods that need to be redefined:

use boa_engine::{JsNativeError, JsResult, context::{Context, ContextBuilder, HostHooks}};

struct Hooks;

impl HostHooks for Hooks {
    fn ensure_can_compile_strings(
        &self,
        context: &mut Context<'_>,
    ) -> JsResult<()> {
        Err(JsNativeError::typ().with_message("eval calls not available").into())
    }
}
let hooks = Hooks; // Can have additional state.
let context = &mut ContextBuilder::new().host_hooks(&hooks).build();
let result = context.eval(r#"eval("let a = 5")"#);
assert_eq!(result.unwrap_err().to_string(), "TypeError: eval calls not available");

Provided Methods

fn make_job_callback(
&self,
callback: JsFunction,
_context: &mut Context<'_>
) -> JobCallback

HostMakeJobCallback ( callback )

Requirements

It must return a JobCallback Record whose [[Callback]] field is callback.

fn call_job_callback(
&self,
job: JobCallback,
this: &JsValue,
args: &[JsValue],
context: &mut Context<'_>
) -> JsResult<JsValue>

HostCallJobCallback ( jobCallback, V, argumentsList )

Requirements

It must perform and return the result of Call(jobCallback.[[Callback]], V, argumentsList).

fn promise_rejection_tracker(
&self,
_promise: &JsObject,
_operation: OperationType,
_context: &mut Context<'_>
)

HostPromiseRejectionTracker ( promise, operation )

Requirements

It must complete normally (i.e. not return an abrupt completion). This is already ensured by the return type.

fn ensure_can_compile_strings(&self, _context: &mut Context<'_>) -> JsResult<()>

HostEnsureCanCompileStrings ( calleeRealm )

Requirements

If the returned Completion Record is a normal completion, it must be a normal completion containing unused. This is already ensured by the return type.

fn has_source_text_available(
&self,
_function: &JsFunction,
_context: &mut Context<'_>
) -> bool

HostHasSourceTextAvailable ( func )

Requirements

It must be deterministic with respect to its parameters. Each time it is called with a specific func as its argument, it must return the same result.

fn ensure_can_add_private_element(
&self,
_o: &JsObject,
_context: &mut Context<'_>
) -> JsResult<()>

HostEnsureCanAddPrivateElement ( O )

Requirements

If O is not a host-defined exotic object, this abstract operation must return NormalCompletion(unused) and perform no other steps.

Any two calls of this abstract operation with the same argument must return the same kind of Completion Record.

This abstract operation should only be overriden by ECMAScript hosts that are web browsers.