Enum boa_engine::JsValue

pub enum JsValue {
    Null,
    Undefined,
    Boolean(bool),
    String(JsString),
    Rational(f64),
    Integer(i32),
    BigInt(JsBigInt),
    Object(JsObject),
    Symbol(JsSymbol),
}

A Javascript value

Variants

Null

null - A null value, for when a value doesn’t exist.

Undefined

undefined - An undefined value, for when a field or index doesn’t exist.

Boolean(bool)

boolean - A true / false value, for if a certain criteria is met.

String(JsString)

String - A UTF-16 string, such as "Hello, world".

Rational(f64)

Number - A 64-bit floating point number, such as 3.1415

Integer(i32)

Number - A 32-bit integer, such as 42.

BigInt(JsBigInt)

BigInt - holds any arbitrary large signed integer.

Object(JsObject)

Object - An object, such as Math, represented by a binary tree of string keys to Javascript values.

Symbol(JsSymbol)

Symbol - A Symbol Primitive type.

Implementations

impl JsValue

pub fn get_iterator(
&self,
context: &mut Context,
hint: Option<IteratorHint>,
method: Option<Self>
) -> JsResult<IteratorRecord>

GetIterator ( obj [ , hint [ , method ] ] )

More information:

ECMA reference

impl JsValue

pub(crate) fn get_v<K>(&self, key: K, context: &mut Context) -> JsResult<Self>where
K: Into<PropertyKey>,

Abstract operation GetV ( V, P ).

Retrieves the value of a specific property of an ECMAScript language value. If the value is not an object, the property lookup is performed using a wrapper object appropriate for the type of the value.

More information:

ECMAScript reference

pub(crate) fn get_method<K>(
&self,
key: K,
context: &mut Context
) -> JsResult<Option<JsObject>>where
K: Into<PropertyKey>,

Abstract operation GetMethod ( V, P )

Retrieves the value of a specific property, when the value of the property is expected to be a function.

More information:

ECMAScript reference

pub(crate) fn create_list_from_array_like(
&self,
element_types: &[Type],
context: &mut Context
) -> JsResult<Vec<Self>>

It is used to create List value whose elements are provided by the indexed properties of self.

More information:

ECMAScript reference

pub(crate) fn invoke<K>(
&self,
key: K,
args: &[Self],
context: &mut Context
) -> JsResult<Self>where
K: Into<PropertyKey>,

Abstract operation ( V, P [ , argumentsList ] )

Calls a method property of an ECMAScript language value.

More information:

ECMAScript reference

pub fn ordinary_has_instance(
function: &Self,
object: &Self,
context: &mut Context
) -> JsResult<bool>

Abstract operation OrdinaryHasInstance ( C, O )

More information:

ECMAScript reference

impl JsValue

pub fn strict_equals(&self, other: &Self) -> bool

Strict equality comparison.

This method is executed when doing strict equality comparisons with the === operator. For more information, check https://tc39.es/ecma262/#sec-strict-equality-comparison.

pub fn equals(&self, other: &Self, context: &mut Context) -> JsResult<bool>

Abstract equality comparison.

This method is executed when doing abstract equality comparisons with the == operator. For more information, check https://tc39.es/ecma262/#sec-abstract-equality-comparison

pub fn same_value(x: &Self, y: &Self) -> bool

The internal comparison abstract operation SameValue(x, y), where x and y are ECMAScript language values, produces true or false.

More information:

ECMAScript

pub fn same_value_zero(x: &Self, y: &Self) -> bool

The internal comparison abstract operation SameValueZero(x, y), where x and y are ECMAScript language values, produces true or false.

SameValueZero differs from SameValue only in its treatment of +0 and -0.

More information:

ECMAScript

fn same_value_non_numeric(x: &Self, y: &Self) -> bool

impl JsValue

pub fn add(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn sub(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn mul(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn div(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn rem(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn pow(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn bitand(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn bitor(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn bitxor(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn shl(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn shr(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn ushr(&self, other: &Self, context: &mut Context) -> JsResult<Self>

pub fn instance_of(&self, target: &Self, context: &mut Context) -> JsResult<bool>

Abstract operation InstanceofOperator ( V, target )

More information:

ECMAScript reference

pub fn neg(&self, context: &mut Context) -> JsResult<Self>

pub fn not(&self, _: &mut Context) -> JsResult<bool>

pub fn abstract_relation(
&self,
other: &Self,
left_first: bool,
context: &mut Context
) -> JsResult<AbstractRelation>

Abstract relational comparison

The comparison x < y, where x and y are values, produces true, false, or undefined (which indicates that at least one operand is NaN).

In addition to x and y the algorithm takes a Boolean flag named LeftFirst as a parameter. The flag is used to control the order in which operations with potentially visible side-effects are performed upon x and y. It is necessary because ECMAScript specifies left to right evaluation of expressions. The default value of LeftFirst is true and indicates that the x parameter corresponds to an expression that occurs to the left of the y parameter’s corresponding expression.

If LeftFirst is false, the reverse is the case and operations must be performed upon y before x.

More Information:

ECMAScript reference

pub fn lt(&self, other: &Self, context: &mut Context) -> JsResult<bool>

The less than operator (<) returns true if the left operand is less than the right operand, and false otherwise.

More Information:

MDN documentation

ECMAScript reference

pub fn le(&self, other: &Self, context: &mut Context) -> JsResult<bool>

The less than or equal operator (<=) returns true if the left operand is less than or equal to the right operand, and false otherwise.

More Information:

MDN documentation

ECMAScript reference

pub fn gt(&self, other: &Self, context: &mut Context) -> JsResult<bool>

The greater than operator (>) returns true if the left operand is greater than the right operand, and false otherwise.

More Information:

MDN documentation

ECMAScript reference

pub fn ge(&self, other: &Self, context: &mut Context) -> JsResult<bool>

The greater than or equal operator (>=) returns true if the left operand is greater than or equal to the right operand, and false otherwise.

More Information:

MDN documentation

ECMAScript reference

impl JsValue

pub fn from_json(json: &Value, context: &mut Context) -> JsResult<Self>

Converts a serde_json::Value to a JsValue.

Example

use boa_engine::{Context, JsValue};

let data = r#"
    {
        "name": "John Doe",
        "age": 43,
        "phones": [
            "+44 1234567",
            "+44 2345678"
        ]
     }"#;

let json: serde_json::Value = serde_json::from_str(data).unwrap();

let mut context = Context::default();
let value = JsValue::from_json(&json, &mut context).unwrap();

pub fn to_json(&self, context: &mut Context) -> JsResult<Value>

Converts the JsValue to a serde_json::Value.

Example

use boa_engine::{Context, JsValue};

let data = r#"
    {
        "name": "John Doe",
        "age": 43,
        "phones": [
            "+44 1234567",
            "+44 2345678"
        ]
     }"#;

let json: serde_json::Value = serde_json::from_str(data).unwrap();

let mut context = Context::default();
let value = JsValue::from_json(&json, &mut context).unwrap();

let back_to_json = value.to_json(&mut context).unwrap();

impl JsValue

pub fn get_type(&self) -> Type

Get the type of a value

This is the abstract operation Type(v), as described in https://tc39.es/ecma262/multipage/ecmascript-data-types-and-values.html#sec-ecmascript-language-types.

Check JsValue::type_of if you need to call the typeof operator.

impl JsValue

pub fn new<T>(value: T) -> Selfwhere
T: Into<Self>,

Create a new JsValue.

pub fn undefined() -> Self

Creates a new undefined value.

pub fn null() -> Self

Creates a new null value.

pub fn nan() -> Self

Creates a new number with NaN value.

pub fn positive_infinity() -> Self

Creates a new number with Infinity value.

pub fn negative_infinity() -> Self

Creates a new number with -Infinity value.

pub fn is_object(&self) -> bool

Returns true if the value is an object

pub fn as_object(&self) -> Option<&JsObject>

pub fn is_callable(&self) -> bool

It determines if the value is a callable function with a [[Call]] internal method.

More information:

ECMAScript reference

pub fn as_callable(&self) -> Option<&JsObject>

pub fn is_constructor(&self) -> bool

Returns true if the value is a constructor object.

pub fn as_constructor(&self) -> Option<&JsObject>

pub fn is_promise(&self) -> bool

Returns true if the value is a promise object.

pub fn as_promise(&self) -> Option<&JsObject>

pub fn is_symbol(&self) -> bool

Returns true if the value is a symbol.

pub fn as_symbol(&self) -> Option<JsSymbol>

pub fn is_undefined(&self) -> bool

Returns true if the value is undefined.

pub fn is_null(&self) -> bool

Returns true if the value is null.

pub fn is_null_or_undefined(&self) -> bool

Returns true if the value is null or undefined.

pub fn is_double(&self) -> bool

Returns true if the value is a 64-bit floating-point number.

pub fn is_integer(&self) -> bool

Returns true if the value is integer.

pub fn is_number(&self) -> bool

Returns true if the value is a number.

pub fn as_number(&self) -> Option<f64>

pub fn is_string(&self) -> bool

Returns true if the value is a string.

pub fn as_string(&self) -> Option<&JsString>

Returns the string if the values is a string, otherwise None.

pub fn is_boolean(&self) -> bool

Returns true if the value is a boolean.

pub fn as_boolean(&self) -> Option<bool>

pub fn is_bigint(&self) -> bool

Returns true if the value is a bigint.

pub fn as_bigint(&self) -> Option<&JsBigInt>

Returns an optional reference to a BigInt if the value is a BigInt primitive.

pub fn to_boolean(&self) -> bool

Converts the value to a bool type.

More information:

ECMAScript

pub(crate) fn get_property<Key>(&self, key: Key) -> Option<PropertyDescriptor>where
Key: Into<PropertyKey>,

Resolve the property in the object.

A copy of the Property is returned.

pub fn set_data(&self, data: ObjectData)

Set the kind of an object.

pub fn to_primitive(
&self,
context: &mut Context,
preferred_type: PreferredType
) -> JsResult<Self>

The abstract operation ToPrimitive takes an input argument and an optional argumenPreferredTypepe.

https://tc39.es/ecma262/#sec-toprimitive

pub fn to_bigint(&self, context: &mut Context) -> JsResult<JsBigInt>

7.1.13 ToBigInt ( argument )

More information:

ECMAScript reference

pub fn display(&self) -> ValueDisplay<'_>

Returns an object that implements Display.

By default the internals are not shown, but they can be toggled with ValueDisplay::internals method.

Examples

use boa_engine::JsValue;

let value = JsValue::new(3);

println!("{}", value.display());

pub fn to_string(&self, context: &mut Context) -> JsResult<JsString>

Converts the value to a string.

This function is equivalent to String(value) in JavaScript.

pub fn to_object(&self, context: &mut Context) -> JsResult<JsObject>

Converts the value to an Object.

This function is equivalent to Object(value) in JavaScript.

See: https://tc39.es/ecma262/#sec-toobject

pub fn to_property_key(&self, context: &mut Context) -> JsResult<PropertyKey>

Converts the value to a PropertyKey, that can be used as a key for properties.

See https://tc39.es/ecma262/#sec-topropertykey

pub fn to_numeric(&self, context: &mut Context) -> JsResult<Numeric>

It returns value converted to a numeric value of type Number or BigInt.

See: https://tc39.es/ecma262/#sec-tonumeric

pub fn to_u32(&self, context: &mut Context) -> JsResult<u32>

Converts a value to an integral 32 bit unsigned integer.

This function is equivalent to value | 0 in JavaScript

See: https://tc39.es/ecma262/#sec-touint32

pub fn to_i32(&self, context: &mut Context) -> JsResult<i32>

Converts a value to an integral 32 bit signed integer.

See: https://tc39.es/ecma262/#sec-toint32

pub fn to_int8(&self, context: &mut Context) -> JsResult<i8>

7.1.10 ToInt8 ( argument )

More information:

ECMAScript reference

pub fn to_uint8(&self, context: &mut Context) -> JsResult<u8>

7.1.11 ToUint8 ( argument )

More information:

ECMAScript reference

pub fn to_uint8_clamp(&self, context: &mut Context) -> JsResult<u8>

7.1.12 ToUint8Clamp ( argument )

More information:

ECMAScript reference

pub fn to_int16(&self, context: &mut Context) -> JsResult<i16>

7.1.8 ToInt16 ( argument )

More information:

ECMAScript reference

pub fn to_uint16(&self, context: &mut Context) -> JsResult<u16>

7.1.9 ToUint16 ( argument )

More information:

ECMAScript reference

pub fn to_big_int64(&self, context: &mut Context) -> JsResult<BigInt>

7.1.15 ToBigInt64 ( argument )

More information:

ECMAScript reference

pub fn to_big_uint64(&self, context: &mut Context) -> JsResult<BigInt>

7.1.16 ToBigUint64 ( argument )

More information:

ECMAScript reference

pub fn to_index(&self, context: &mut Context) -> JsResult<u64>

Converts a value to a non-negative integer if it is a valid integer index value.

See: https://tc39.es/ecma262/#sec-toindex

pub fn to_length(&self, context: &mut Context) -> JsResult<u64>

Converts argument to an integer suitable for use as the length of an array-like object.

See: https://tc39.es/ecma262/#sec-tolength

pub fn to_integer_or_infinity(
&self,
context: &mut Context
) -> JsResult<IntegerOrInfinity>

Abstract operation ToIntegerOrInfinity ( argument )

This method converts a Value to an integer representing its Number value with fractional part truncated, or to +∞ or -∞ when that Number value is infinite.

More information:

ECMAScript reference

pub fn to_number(&self, context: &mut Context) -> JsResult<f64>

Converts a value to a double precision floating point.

This function is equivalent to the unary + operator (+value) in JavaScript

See: https://tc39.es/ecma262/#sec-tonumber

pub fn to_numeric_number(&self, context: &mut Context) -> JsResult<f64>

This is a more specialized version of to_numeric, including BigInt.

This function is equivalent to Number(value) in JavaScript

See: https://tc39.es/ecma262/#sec-tonumeric

pub fn require_object_coercible(&self) -> JsResult<&Self>

Check if the Value can be converted to an Object

The abstract operation RequireObjectCoercible takes argument argument. It throws an error if argument is a value that cannot be converted to an Object using ToObject. It is defined by Table 15

More information:

ECMAScript reference

pub fn to_property_descriptor(
&self,
context: &mut Context
) -> JsResult<PropertyDescriptor>

pub fn type_of(&self) -> &'static str

typeof operator. Returns a string representing the type of the given ECMA Value.

More information:

ECMAScript reference

pub(crate) fn is_array(&self) -> JsResult<bool>

Abstract operation IsArray ( argument )

Check if a value is an array.

More information:

ECMAScript reference