Struct boa_engine::environments::runtime::BindingLocator

pub(crate) struct BindingLocator {
    name: Identifier,
    environment_index: usize,
    binding_index: usize,
    global: bool,
    mutate_immutable: bool,
    silent: bool,
}

A binding locator contains all information about a binding that is needed to resolve it at runtime.

Binding locators get created at bytecode compile time and are accessible at runtime via the crate::vm::CodeBlock.

Fields
name: Identifierenvironment_index: usizebinding_index: usizeglobal: boolmutate_immutable: boolsilent: bool
Implementations

impl BindingLocator

pub(super) fn declarative(
name: Identifier,
environment_index: usize,
binding_index: usize
) -> Self

Creates a new declarative binding locator that has knows indices.

pub(super) fn global(name: Identifier) -> Self

Creates a binding locator that indicates that the binding is on the global object.

pub(super) fn mutate_immutable(name: Identifier) -> Self

Creates a binding locator that indicates that it was attempted to mutate an immutable binding. At runtime this should always produce a type error.

pub(super) fn silent(name: Identifier) -> Self

Creates a binding locator that indicates that any action is silently ignored.

pub(crate) fn name(&self) -> Identifier

Returns the name of the binding.

pub(crate) fn is_global(&self) -> bool

Returns if the binding is located on the global object.

pub(crate) fn environment_index(&self) -> usize

Returns the environment index of the binding.

pub(crate) fn binding_index(&self) -> usize

Returns the binding index of the binding.

pub(crate) fn is_silent(&self) -> bool

Returns if the binding is a silent operation.

pub(crate) fn throw_mutate_immutable(
&self,
context: &mut Context
) -> Result<(), JsNativeError>

Helper method to throws an error if the binding access is illegal.