Module boa_engine::error

Error-related types and conversions.

Structs

JsError

The error type returned by all operations related to the execution of Javascript code.

JsNativeError

Native representation of an ideal Error object from Javascript.

Enums

JsNativeErrorKind

The list of possible error types a JsNativeError can be.

Repr 🔒

Internal representation of a JsError.

TryNativeError

The error type returned by the JsError::try_native method.