Struct boa_engine::error::JsNativeError

pub struct JsNativeError {
    pub kind: JsNativeErrorKind,
    message: Box<str>,
    cause: Option<Box<JsError>>,
    realm: Option<Realm>,
}

Native representation of an ideal Error object from Javascript.

This representation is more space efficient than its JsObject equivalent, since it doesn’t need to create a whole new JsObject to be instantiated. Prefer using this over JsError when you don’t need to throw plain JsValues as errors, or when you need to inspect the error type of a JsError.

Examples

let native_error = JsNativeError::uri().with_message("cannot decode uri");

match native_error.kind {
    JsNativeErrorKind::Uri => { /* handle URI error*/ }
    _ => unreachable!(),
}

assert_eq!(native_error.message(), "cannot decode uri");

Fields
kind: JsNativeErrorKind

The kind of native error (e.g. TypeError, SyntaxError, etc.)

message: Box<str>cause: Option<Box<JsError>>realm: Option<Realm>
Implementations

impl JsNativeError

fn new( kind: JsNativeErrorKind, message: Box<str>, cause: Option<Box<JsError>> ) -> Self

Creates a new JsNativeError from its kind, message and (optionally) its cause.

pub fn aggregate(errors: Vec<JsError>) -> Self

Creates a new JsNativeError of kind AggregateError from a list of JsErrors, with empty message and undefined cause.

Examples

let inner_errors = vec![
    JsNativeError::typ().into(),
    JsNativeError::syntax().into()
];
let error = JsNativeError::aggregate(inner_errors);

assert!(matches!(
    error.kind,
    JsNativeErrorKind::Aggregate(ref errors) if errors.len() == 2
));

pub const fn is_aggregate(&self) -> bool

Check if it’s a JsNativeErrorKind::Aggregate.

pub fn error() -> Self

Creates a new JsNativeError of kind Error, with empty message and undefined cause.

Examples

let error = JsNativeError::error();

assert!(matches!(error.kind, JsNativeErrorKind::Error));

pub const fn is_error(&self) -> bool

Check if it’s a JsNativeErrorKind::Error.

pub fn eval() -> Self

Creates a new JsNativeError of kind EvalError, with empty message and undefined cause.

Examples

let error = JsNativeError::eval();

assert!(matches!(error.kind, JsNativeErrorKind::Eval));

pub const fn is_eval(&self) -> bool

Check if it’s a JsNativeErrorKind::Eval.

pub fn range() -> Self

Creates a new JsNativeError of kind RangeError, with empty message and undefined cause.

Examples

let error = JsNativeError::range();

assert!(matches!(error.kind, JsNativeErrorKind::Range));

pub const fn is_range(&self) -> bool

Check if it’s a JsNativeErrorKind::Range.

pub fn reference() -> Self

Creates a new JsNativeError of kind ReferenceError, with empty message and undefined cause.

Examples

let error = JsNativeError::reference();

assert!(matches!(error.kind, JsNativeErrorKind::Reference));

pub const fn is_reference(&self) -> bool

Check if it’s a JsNativeErrorKind::Reference.

pub fn syntax() -> Self

Creates a new JsNativeError of kind SyntaxError, with empty message and undefined cause.

Examples

let error = JsNativeError::syntax();

assert!(matches!(error.kind, JsNativeErrorKind::Syntax));

pub const fn is_syntax(&self) -> bool

Check if it’s a JsNativeErrorKind::Syntax.

pub fn typ() -> Self

Creates a new JsNativeError of kind TypeError, with empty message and undefined cause.

Examples

let error = JsNativeError::typ();

assert!(matches!(error.kind, JsNativeErrorKind::Type));

pub const fn is_type(&self) -> bool

Check if it’s a JsNativeErrorKind::Type.

pub fn uri() -> Self

Creates a new JsNativeError of kind UriError, with empty message and undefined cause.

Examples

let error = JsNativeError::uri();

assert!(matches!(error.kind, JsNativeErrorKind::Uri));

pub const fn is_uri(&self) -> bool

Check if it’s a JsNativeErrorKind::Uri.

pub fn no_instructions_remain() -> Self

Creates a new JsNativeError that indicates that the context hit its execution limit. This is only used in a fuzzing context.

pub const fn is_no_instructions_remain(&self) -> bool

Check if it’s a JsNativeErrorKind::NoInstructionsRemain.

pub fn runtime_limit() -> Self

Creates a new JsNativeError that indicates that the context exceeded the runtime limits.

pub const fn is_runtime_limit(&self) -> bool

Check if it’s a JsNativeErrorKind::RuntimeLimit.

pub fn with_message<S>(self, message: S) -> Selfwhere S: Into<Box<str>>,

Sets the message of this error.

Examples

let error = JsNativeError::range().with_message("number too large");

assert_eq!(error.message(), "number too large");

pub fn with_cause<V>(self, cause: V) -> Selfwhere V: Into<JsError>,

Sets the cause of this error.

Examples

let cause = JsNativeError::syntax();
let error = JsNativeError::error().with_cause(cause);

assert!(error.cause().unwrap().as_native().is_some());

pub const fn message(&self) -> &str

Gets the message of this error.

This is equivalent to the NativeError.prototype.message property.

Examples

let error = JsNativeError::range().with_message("number too large");

assert_eq!(error.message(), "number too large");

pub fn cause(&self) -> Option<&JsError>

Gets the cause of this error.

This is equivalent to the NativeError.prototype.cause property.

Examples

let cause = JsNativeError::syntax();
let error = JsNativeError::error().with_cause(cause);

assert!(error.cause().unwrap().as_native().is_some());

pub fn to_opaque(&self, context: &mut Context<'_>) -> JsObject

Converts this native error to its opaque representation as a JsObject.

Examples

let context = &mut Context::default();

let error = JsNativeError::error().with_message("error!");
let error_obj = error.to_opaque(context);

assert!(error_obj.borrow().is_error());
assert_eq!(error_obj.get("message", context).unwrap(), "error!".into())

Panics

If converting a JsNativeErrorKind::RuntimeLimit to an opaque object.

pub(crate) fn with_realm(self, realm: Realm) -> Self

Sets the realm of this error.