Trait boa_engine::job::JobQueue

pub trait JobQueue {
    fn enqueue_promise_job(&self, job: NativeJob, context: &mut Context<'_>);
    fn run_jobs(&self, context: &mut Context<'_>);
    fn enqueue_future_job(&self, future: FutureJob, context: &mut Context<'_>);
}

A queue of ECMAscript Jobs.

This is the main API that allows creating custom event loops with custom job queues.

Required Methods

fn enqueue_promise_job(&self, job: NativeJob, context: &mut Context<'_>)

HostEnqueuePromiseJob ( job, realm ).

Enqueues a NativeJob on the job queue. Note that host-defined Jobs need to satisfy a set of requirements for them to be spec-compliant.

fn run_jobs(&self, context: &mut Context<'_>)

Runs all jobs in the queue.

Running a job could enqueue more jobs in the queue. The implementor of the trait determines if the method should loop until there are no more queued jobs or if it should only run one iteration of the queue.

fn enqueue_future_job(&self, future: FutureJob, context: &mut Context<'_>)

Enqueues a new Future job on the job queue.

On completion, future returns a new NativeJob that needs to be enqueued into the job queue to update the state of the inner Promise, which is what ECMAScript sees. Failing to do this will leave the inner Promise in the pending state, which won’t call any then or catch handlers, even if future was already completed.