Type Alias boa_engine::native_function::NativeFunctionPointer

pub type NativeFunctionPointer = fn(_: &JsValue, _: &[JsValue], _: &mut Context) -> JsResult<JsValue>;

The required signature for all native built-in function pointers.

Arguments

The first argument represents the this variable of every ECMAScript function.

The second argument represents the list of all arguments passed to the function.

The last argument is the engine Context.