Module boa_engine::object

Boa’s representation of a JavaScript object and builtin object wrappers

For the builtin object wrappers, please see object::builtins for implementors.

Modules

builtins

All Rust API wrappers for Boa’s ECMAScript objects.

internal_methods 🔒

This module defines the object internal methods.

jsobject 🔒

This module implements the JsObject structure.

operations 🔒

property_map 🔒

shape

Implements object shapes.

Structs

BorrowError

An error returned by JsObject::try_borrow.

BorrowMutError

An error returned by JsObject::try_borrow_mut.

ConstructorBuilder

Builder for creating constructors objects, like Array.

FunctionBinding

The functions binding.

FunctionObjectBuilder

Builder for creating native function objects

Iter

An iterator over the property entries of an Object

JsObject

Garbage collected Object.

Object

The internal representation of a JavaScript object.

ObjectData

Defines the kind of an object and its internal methods

ObjectInitializer

Builder for creating objects with properties.

PrivateName

A Private Name.

PropertyMap

A PropertyMap contains all the properties of an object.

RecursionLimiter

Prevents infinite recursion during Debug::fmt, JSON.stringify, and other conversions. This uses a thread local, so is not safe to use where the object graph will be traversed by multiple threads!

VTableObject

An Object that has an additional vtable with its internal methods.

Enums

IndexProperties

An iterator over the indexed property entries of an Object.

IndexPropertyKeys

An iterator over the index keys (u32) of an Object.

IndexPropertyValues

An iterator over the index values (Property) of an Object.

IntegrityLevel

Object integrity level.

ObjectKind

Defines the different types of objects.

PrivateElement

The representation of private object elements.

Constants

CONSTRUCTOR

Const constructor, usually set on prototypes as a key to point to their respective constructor object.

PROTOTYPE

Const prototype, usually set on constructors as a key to point to their respective prototype object.

Traits

JsObjectType 🔒

NativeObject

This trait allows Rust types to be passed around as objects.

Type Definitions

JsPrototype

Common field names. A type alias for an object prototype.

ObjectStorage 🔒

The internal storage of an object’s property values.

Ref

A wrapper type for an immutably borrowed type T.

RefMut

A wrapper type for a mutably borrowed type T.