Module boa_engine::object

This module implements the Rust representation of a JavaScript object.

Modules

internal_methods 🔒

This module defines the object internal methods.

jsarray 🔒

jsfunction 🔒

jsobject 🔒

This module implements the JsObject structure.

jsproxy 🔒

jstypedarray 🔒

operations 🔒

property_map 🔒

Structs

ConstructorBuilder

Builder for creating constructors objects, like Array.

FunctionBinding

The functions binding.

FunctionBuilder

Builder for creating native function objects

IndexProperties

An iterator over the indexed property entries of an Object

IndexPropertyKeys

An iterator over the index keys (u32) of an Object.

IndexPropertyValues

An iterator over the index values (Property) of an Object.

Iter

An iterator over the property entries of an Object

JsArray

JavaScript Array rust object.

JsFloat32Array

JavaScript Float32Array rust object.

JsFloat64Array

JavaScript Float64Array rust object.

JsFunction

JavaScript Function rust object.

JsInt8Array

JavaScript Int8Array rust object.

JsInt16Array

JavaScript Int16Array rust object.

JsInt32Array

JavaScript Int32Array rust object.

JsObject

Garbage collected Object.

JsProxy

JavaScript Proxy rust object.

JsProxyBuilder

Utility builder to create JsProxy objects from native functions.

JsRevocableProxy

JavaScript Proxy rust object that can be disabled.

JsTypedArray

JavaScript TypedArray rust object.

JsUint8Array

JavaScript Uint8Array rust object.

JsUint16Array

JavaScript Uint16Array rust object.

JsUint32Array

JavaScript Uint32Array rust object.

Keys

An iterator over the keys (PropertyKey) of an Object.

Object

The internal representation of a JavaScript object.

ObjectData

Defines the kind of an object and its internal methods

ObjectInitializer

Builder for creating objects with properties.

PropertyMap

RecursionLimiter

Prevents infinite recursion during Debug::fmt, JSON.stringify, and other conversions. This uses a thread local, so is not safe to use where the object graph will be traversed by multiple threads!

StringProperties

An iterator over the String property entries of an Object

StringPropertyKeys

An iterator over the string keys (RcString) of an Object.

StringPropertyValues

An iterator over the string values (Property) of an Object.

SymbolProperties

An iterator over the Symbol property entries of an Object

SymbolPropertyKeys

An iterator over the keys (RcSymbol) of an Object.

SymbolPropertyValues

An iterator over the Symbol values (Property) of an Object.

Values

An iterator over the values (Property) of an Object.

Enums

IntegrityLevel

Object integrity level.

ObjectKind

Defines the different types of objects.

PrivateElement

The representation of private object elements.

Statics

PROTOTYPE

Static prototype, usually set on constructors as a key to point to their respective prototype object.

Traits

JsObjectType 🔒

NativeObject

This trait allows Rust types to be passed around as objects.

Type Definitions

JsPrototype

Ref

A wrapper type for an immutably borrowed type T.

RefMut

A wrapper type for a mutably borrowed type T.