Module boa_engine::object::internal_methods

This module defines the object internal methods.

More information:

ECMAScript reference

Modules

arguments 🔒

array 🔒

bound_function 🔒

function 🔒

global 🔒

integer_indexed 🔒

proxy 🔒

string 🔒

Structs

InternalObjectMethods 🔒

The internal representation of the internal methods of a JsObject.

Statics

ORDINARY_INTERNAL_METHODS 🔒

Definitions of the internal object methods for ordinary objects.

Functions

get_prototype_from_constructor 🔒

Abstract operation GetPrototypeFromConstructor

is_compatible_property_descriptor 🔒

Abstract operation IsCompatiblePropertyDescriptor

ordinary_define_own_property 🔒

Abstract operation OrdinaryDefineOwnProperty.

ordinary_delete 🔒

Abstract operation OrdinaryDelete.

ordinary_get 🔒

Abstract operation OrdinaryGet.

ordinary_get_own_property 🔒

Abstract operation OrdinaryGetOwnProperty.

ordinary_get_prototype_of 🔒

Abstract operation OrdinaryGetPrototypeOf.

ordinary_has_property 🔒

Abstract operation OrdinaryHasProperty.

ordinary_is_extensible 🔒

Abstract operation OrdinaryIsExtensible.

ordinary_own_property_keys 🔒

Abstract operation OrdinaryOwnPropertyKeys.

ordinary_prevent_extensions 🔒

Abstract operation OrdinaryPreventExtensions.

ordinary_set 🔒

Abstract operation OrdinarySet.

ordinary_set_prototype_of 🔒

Abstract operation OrdinarySetPrototypeOf.

validate_and_apply_property_descriptor 🔒

Abstract operation ValidateAndApplyPropertyDescriptor