Module boa_engine::object::jsobject

This module implements the JsObject structure.

The JsObject is a garbage collected Object.

Structs

BorrowError

An error returned by JsObject::try_borrow.

BorrowMutError

An error returned by JsObject::try_borrow_mut.

JsObject

Garbage collected Object.

RecursionLimiter

Prevents infinite recursion during Debug::fmt, JSON.stringify, and other conversions. This uses a thread local, so is not safe to use where the object graph will be traversed by multiple threads!

VTableObject

An Object that has an additional vtable with its internal methods.

Enums

RecursionValueState 🔒

Constants

SEEN 🔒

The map of pointers to JsObject that have been visited during the current Debug::fmt graph, and the current state of their RecursionLimiter (dropped or live – see RecursionValueState)

Type Definitions

Ref

A wrapper type for an immutably borrowed type T.

RefMut

A wrapper type for a mutably borrowed type T.