Struct boa_engine::object::ConstructorBuilder

pub struct ConstructorBuilder<'context> {
    context: &'context mut Context,
    function: NativeFunctionSignature,
    object: JsObject,
    has_prototype_property: bool,
    prototype: JsObject,
    name: JsString,
    length: usize,
    callable: bool,
    constructor: Option<ConstructorKind>,
    inherit: Option<JsPrototype>,
    custom_prototype: Option<JsPrototype>,
}

Builder for creating constructors objects, like Array.

Fields
context: &'context mut Contextfunction: NativeFunctionSignatureobject: JsObjecthas_prototype_property: boolprototype: JsObjectname: JsStringlength: usizecallable: boolconstructor: Option<ConstructorKind>inherit: Option<JsPrototype>custom_prototype: Option<JsPrototype>
Implementations

impl<'context> ConstructorBuilder<'context>

pub fn new(
context: &'context mut Context,
function: NativeFunctionSignature
) -> Self

Create a new ConstructorBuilder.

pub(crate) fn with_standard_constructor(
context: &'context mut Context,
function: NativeFunctionSignature,
standard_constructor: StandardConstructor
) -> Self

pub fn method<B>(
&mut self,
function: NativeFunctionSignature,
binding: B,
length: usize
) -> &mut Selfwhere
B: Into<FunctionBinding>,

Add new method to the constructors prototype.

pub fn static_method<B>(
&mut self,
function: NativeFunctionSignature,
binding: B,
length: usize
) -> &mut Selfwhere
B: Into<FunctionBinding>,

Add new static method to the constructors object itself.

pub fn property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Selfwhere
K: Into<PropertyKey>,
V: Into<JsValue>,

Add new data property to the constructor’s prototype.

pub fn static_property<K, V>(
&mut self,
key: K,
value: V,
attribute: Attribute
) -> &mut Selfwhere
K: Into<PropertyKey>,
V: Into<JsValue>,

Add new static data property to the constructor object itself.

pub fn accessor<K>(
&mut self,
key: K,
get: Option<JsFunction>,
set: Option<JsFunction>,
attribute: Attribute
) -> &mut Selfwhere
K: Into<PropertyKey>,

Add new accessor property to the constructor’s prototype.

pub fn static_accessor<K>(
&mut self,
key: K,
get: Option<JsFunction>,
set: Option<JsFunction>,
attribute: Attribute
) -> &mut Selfwhere
K: Into<PropertyKey>,

Add new static accessor property to the constructor object itself.

pub fn property_descriptor<K, P>(&mut self, key: K, property: P) -> &mut Selfwhere
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Add new property to the constructor’s prototype.

pub fn static_property_descriptor<K, P>(
&mut self,
key: K,
property: P
) -> &mut Selfwhere
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Add new static property to the constructor object itself.

pub fn length(&mut self, length: usize) -> &mut Self

Specify how many arguments the constructor function takes.

Default is 0.

pub fn name<N>(&mut self, name: N) -> &mut Selfwhere
N: AsRef<str>,

Specify the name of the constructor function.

Default is "[object]"

pub fn callable(&mut self, callable: bool) -> &mut Self

Specify whether the constructor function can be called.

Default is true

pub fn constructor(&mut self, constructor: bool) -> &mut Self

Specify whether the constructor function can be called with new keyword.

Default is true

pub fn inherit<O: Into<JsPrototype>>(&mut self, prototype: O) -> &mut Self

Specify the parent prototype which objects created by this constructor inherit from.

Default is Object.prototype

pub fn custom_prototype<O: Into<JsPrototype>>(
&mut self,
prototype: O
) -> &mut Self

Specify the [[Prototype]] internal field of this constructor.

Default is Function.prototype

pub fn has_prototype_property(
&mut self,
has_prototype_property: bool
) -> &mut Self

Specify whether the constructor function has a ‘prototype’ property.

Default is true

pub fn context(&mut self) -> &mut Context

Return the current context.

pub fn build(&mut self) -> JsFunction

Build the constructor function object.