Struct boa_engine::object::JsObject

pub struct JsObject {
    inner: Gc<VTableObject>,
}

Garbage collected Object.

Fields
inner: Gc<VTableObject>
Implementations

impl JsObject

pub(crate) fn __get_prototype_of__( &self, context: &mut Context<'_> ) -> JsResult<JsPrototype>

Internal method [[GetPrototypeOf]]

Return either the prototype of this object or null.

More information:

ECMAScript reference

pub(crate) fn __set_prototype_of__( &self, val: JsPrototype, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[SetPrototypeOf]]

Set the property of a specified object to another object or null.

More information:

ECMAScript reference

pub(crate) fn __is_extensible__( &self, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[IsExtensible]]

Check if the object is extensible.

More information:

ECMAScript reference

pub(crate) fn __prevent_extensions__( &self, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[PreventExtensions]]

Disable extensibility for this object.

More information:

ECMAScript reference

pub(crate) fn __get_own_property__( &self, key: &PropertyKey, context: &mut Context<'_> ) -> JsResult<Option<PropertyDescriptor>>

Internal method [[GetOwnProperty]]

Get the specified property of this object.

More information:

ECMAScript reference

pub(crate) fn __define_own_property__( &self, key: &PropertyKey, desc: PropertyDescriptor, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[DefineOwnProperty]]

Define a new property of this object.

More information:

ECMAScript reference

pub(crate) fn __has_property__( &self, key: &PropertyKey, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[hasProperty]].

Check if the object or its prototype has the required property.

More information:

ECMAScript reference

pub(crate) fn __get__( &self, key: &PropertyKey, receiver: JsValue, context: &mut Context<'_> ) -> JsResult<JsValue>

Internal method [[Get]]

Get the specified property of this object or its prototype.

More information:

ECMAScript reference

pub(crate) fn __set__( &self, key: PropertyKey, value: JsValue, receiver: JsValue, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[Set]]

Set the specified property of this object or its prototype to the provided value.

More information:

ECMAScript reference

pub(crate) fn __delete__( &self, key: &PropertyKey, context: &mut Context<'_> ) -> JsResult<bool>

Internal method [[Delete]]

Delete the specified own property of this object.

More information:

ECMAScript reference

pub(crate) fn __own_property_keys__( &self, context: &mut Context<'_> ) -> JsResult<Vec<PropertyKey>>

Internal method [[OwnPropertyKeys]]

Get all the keys of the properties of this object.

More information:

ECMAScript reference

pub(crate) fn __call__( &self, this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Internal method [[Call]]

Call this object if it has a [[Call]] internal method.

More information:

ECMAScript reference

pub(crate) fn __construct__( &self, args: &[JsValue], new_target: &Self, context: &mut Context<'_> ) -> JsResult<Self>

Internal method [[Construct]]

Construct a new instance of this object if this object has a [[Construct]] internal method.

More information:

ECMAScript reference

impl JsObject

pub(crate) fn from_object_and_vtable( object: Object, vtable: &'static InternalObjectMethods ) -> Self

Creates a new JsObject from its inner object and its vtable.

pub fn with_object_proto(intrinsics: &Intrinsics) -> Self

Creates a new ordinary object with its prototype set to the Object prototype.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate(%Object.prototype%).

pub fn with_null_proto() -> Self

Creates a new ordinary object, with its prototype set to null.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate(null).

pub fn from_proto_and_data<O: Into<Option<Self>>>( prototype: O, data: ObjectData ) -> Self

Creates a new object with the provided prototype and object data.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate, with the difference that the additionalInternalSlotsList parameter is automatically set by the ObjectData provided.

pub(crate) fn from_proto_and_data_with_shared_shape<O: Into<Option<Self>>>( root_shape: &RootShape, prototype: O, data: ObjectData ) -> Self

Creates a new object with the provided prototype and object data.

This is equivalent to calling the specification’s abstract operation OrdinaryObjectCreate, with the difference that the additionalInternalSlotsList parameter is automatically set by the ObjectData provided.

pub fn borrow(&self) -> Ref<'_, Object>

Immutably borrows the Object.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the object is currently mutably borrowed.

pub fn borrow_mut(&self) -> RefMut<'_, Object, Object>

Mutably borrows the Object.

The borrow lasts until the returned RefMut exits scope. The object cannot be borrowed while this borrow is active.

Panics

Panics if the object is currently borrowed.

pub fn try_borrow(&self) -> StdResult<Ref<'_, Object>, BorrowError>

Immutably borrows the Object, returning an error if the value is currently mutably borrowed.

The borrow lasts until the returned GcCellRef exits scope. Multiple immutable borrows can be taken out at the same time.

This is the non-panicking variant of borrow.

pub fn try_borrow_mut( &self ) -> StdResult<RefMut<'_, Object, Object>, BorrowMutError>

Mutably borrows the object, returning an error if the value is currently borrowed.

The borrow lasts until the returned GcCellRefMut exits scope. The object be borrowed while this borrow is active.

This is the non-panicking variant of borrow_mut.

pub fn equals(lhs: &Self, rhs: &Self) -> bool

Checks if the garbage collected memory is the same.

pub(crate) fn ordinary_to_primitive( &self, context: &mut Context<'_>, hint: PreferredType ) -> JsResult<JsValue>

Converts an object to a primitive.

Diverges from the spec to prevent a stack overflow when the object is recursive. For example,

let a = [1];
a[1] = a;
console.log(a.toString()); // We print "1,"

The spec doesn’t mention what to do in this situation, but a naive implementation would overflow the stack recursively calling toString(). We follow v8 and SpiderMonkey instead by returning a default value for the given hint – either 0. or "". Example in v8: https://repl.it/repls/IvoryCircularCertification#index.js

More information:

ECMAScript

pub fn is<T>(&self) -> boolwhere T: NativeObject,

Return true if it is a native object and the native type is T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_ref<T>(&self) -> Option<Ref<'_, T>>where T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently mutably borrowed.

pub fn downcast_mut<T>(&self) -> Option<RefMut<'_, Object, T>>where T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

Panics

Panics if the object is currently borrowed.

pub fn prototype(&self) -> JsPrototype

Get the prototype of the object.

Panics

Panics if the object is currently mutably borrowed.

pub(crate) fn extensible(&self) -> bool

Get the extensibility of the object.

Panics

Panics if the object is currently mutably borrowed.

pub fn set_prototype(&self, prototype: JsPrototype) -> bool

Set the prototype of the object.

Panics

Panics if the object is currently mutably borrowed

pub fn is_array(&self) -> bool

Checks if it’s an Array object.

pub fn is_data_view(&self) -> bool

Checks if it’s a DataView object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_array_iterator(&self) -> bool

Checks if it is an ArrayIterator object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_array_buffer(&self) -> bool

Checks if it’s an ArrayBuffer object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_map(&self) -> bool

Checks if it is a Map object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_map_iterator(&self) -> bool

Checks if it’s a MapIterator object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_set(&self) -> bool

Checks if it is a Set object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_set_iterator(&self) -> bool

Checks if it is a SetIterator object

Panics

Panics if the object is currently mutably borrowed.

pub fn is_string(&self) -> bool

Checks if it’s a String object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_function(&self) -> bool

Checks if it’s a Function object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_generator(&self) -> bool

Checks if it’s a Generator object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_symbol(&self) -> bool

Checks if it’s a Symbol object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_error(&self) -> bool

Checks if it’s an Error object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_boolean(&self) -> bool

Checks if it’s a Boolean object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_number(&self) -> bool

Checks if it’s a Number object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_bigint(&self) -> bool

Checks if it’s a BigInt object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_date(&self) -> bool

Checks if it’s a Date object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_regexp(&self) -> bool

Checks if it’s a RegExp object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_array(&self) -> bool

Checks if it’s a TypedArray object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint8_array(&self) -> bool

Checks if it’s a Uint8Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int8_array(&self) -> bool

Checks if it’s a Int8Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint16_array(&self) -> bool

Checks if it’s a Uint16Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int16_array(&self) -> bool

Checks if it’s a Int16Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_uint32_array(&self) -> bool

Checks if it’s a Uint32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_int32_array(&self) -> bool

Checks if it’s a Int32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_float32_array(&self) -> bool

Checks if it’s a Float32Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_typed_float64_array(&self) -> bool

Checks if it’s a Float64Array object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_promise(&self) -> bool

Checks if it’s a Promise object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_ordinary(&self) -> bool

Checks if it’s an ordinary object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_proxy(&self) -> bool

Checks if it’s a proxy object.

Panics

Panics if the object is currently mutably borrowed.

pub fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

Panics

Panics if the object is currently mutably borrowed.

pub fn to_property_descriptor( &self, context: &mut Context<'_> ) -> JsResult<PropertyDescriptor>

The abstract operation ToPropertyDescriptor.

More information:

ECMAScript reference

pub fn copy_data_properties<K>( &self, source: &JsValue, excluded_keys: Vec<K>, context: &mut Context<'_> ) -> JsResult<()>where K: Into<PropertyKey>,

7.3.25 CopyDataProperties ( target, source, excludedItems )

More information:

ECMAScript

pub(crate) fn get_property( &self, key: &PropertyKey ) -> Option<PropertyDescriptor>

pub(crate) fn insert<K, P>(&self, key: K, property: P) -> boolwhere K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Helper function for property insertion.

pub fn insert_property<K, P>(&self, key: K, property: P) -> boolwhere K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name, than true is returned with that field, otherwise false is returned.

pub fn is_callable(&self) -> bool

It determines if Object is a callable function with a [[Call]] internal method.

More information:

ECMAScript reference

pub fn is_constructor(&self) -> bool

It determines if Object is a function object with a [[Construct]] internal method.

More information:

ECMAScript reference

pub fn is_global(&self) -> bool

Returns true if the JsObject is the global for a Realm

pub(crate) fn vtable(&self) -> &'static InternalObjectMethods

pub(crate) const fn inner(&self) -> &Gc<VTableObject>

pub(crate) fn private_name(&self, description: JsString) -> PrivateName

Create a new private name with this object as the unique identifier.

impl JsObject

pub fn is_extensible(&self, context: &mut Context<'_>) -> JsResult<bool>

Check if object is extensible.

More information:

ECMAScript reference

pub fn get<K>(&self, key: K, context: &mut Context<'_>) -> JsResult<JsValue>where K: Into<PropertyKey>,

Get property from object or throw.

More information:

ECMAScript reference

pub fn set<K, V>( &self, key: K, value: V, throw: bool, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>, V: Into<JsValue>,

set property of object or throw if bool flag is passed.

More information:

ECMAScript reference

pub fn create_data_property<K, V>( &self, key: K, value: V, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>, V: Into<JsValue>,

Create data property

More information:

ECMAScript reference

pub fn create_data_property_or_throw<K, V>( &self, key: K, value: V, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>, V: Into<JsValue>,

Create data property or throw

More information:

ECMAScript reference

pub(crate) fn create_non_enumerable_data_property_or_throw<K, V>( &self, key: K, value: V, context: &mut Context<'_> )where K: Into<PropertyKey>, V: Into<JsValue>,

Create non-enumerable data property or throw

More information:

ECMAScript reference

pub fn define_property_or_throw<K, P>( &self, key: K, desc: P, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>, P: Into<PropertyDescriptor>,

Define property or throw.

More information:

ECMAScript reference

pub fn delete_property_or_throw<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>,

Defines the property or throws a TypeError if the operation fails.

More information:

ECMAScript reference

pub fn has_property<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>,

Check if object has property.

More information:

ECMAScript reference

pub fn has_own_property<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<bool>where K: Into<PropertyKey>,

Check if object has an own property.

More information:

ECMAScript reference

pub fn call( &self, this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

Call ( F, V [ , argumentsList ] )

Panics

Panics if the object is currently mutably borrowed.

More information:

ECMAScript reference

pub fn construct( &self, args: &[JsValue], new_target: Option<&Self>, context: &mut Context<'_> ) -> JsResult<Self>

Construct ( F [ , argumentsList [ , newTarget ] ] )

Construct an instance of this object with the specified arguments.

Panics

Panics if the object is currently mutably borrowed.

More information:

ECMAScript reference

pub fn set_integrity_level( &self, level: IntegrityLevel, context: &mut Context<'_> ) -> JsResult<bool>

Make the object sealed or frozen.

More information:

ECMAScript reference

pub fn test_integrity_level( &self, level: IntegrityLevel, context: &mut Context<'_> ) -> JsResult<bool>

Check if the object is sealed or frozen.

More information:

ECMAScript reference

pub(crate) fn length_of_array_like( &self, context: &mut Context<'_> ) -> JsResult<u64>

Abstract operation LengthOfArrayLike ( obj ).

Returns the value of the “length” property of an array-like object.

pub(crate) fn species_constructor<F>( &self, default_constructor: F, context: &mut Context<'_> ) -> JsResult<Self>where F: FnOnce(&StandardConstructors) -> &StandardConstructor,

7.3.22 SpeciesConstructor ( O, defaultConstructor )

The abstract operation SpeciesConstructor takes arguments O (an Object) and defaultConstructor (a constructor). It is used to retrieve the constructor that should be used to create new objects that are derived from O. defaultConstructor is the constructor to use if a constructor @@species property cannot be found starting from O.

More information:

ECMAScript reference

pub(crate) fn enumerable_own_property_names( &self, kind: PropertyNameKind, context: &mut Context<'_> ) -> JsResult<Vec<JsValue>>

It is used to iterate over names of object’s keys.

More information:

ECMAScript reference

pub(crate) fn get_method<K>( &self, key: K, context: &mut Context<'_> ) -> JsResult<Option<Self>>where K: Into<PropertyKey>,

Abstract operation GetMethod ( V, P )

Retrieves the value of a specific property, when the value of the property is expected to be a function.

More information:

ECMAScript reference

pub(crate) fn is_array_abstract(&self) -> JsResult<bool>

Abstract operation IsArray ( argument )

Check if a value is an array.

More information:

ECMAScript reference

pub(crate) fn get_function_realm( &self, context: &mut Context<'_> ) -> JsResult<Realm>

Abstract operation GetFunctionRealm.

pub(crate) fn private_element_find( &self, name: &PrivateName, is_getter: bool, is_setter: bool ) -> Option<PrivateElement>

Abstract operation PrivateElementFind ( O, P )

Get the private element from an object.

More information:

ECMAScript specification

pub(crate) fn private_field_add( &self, name: &PrivateName, value: JsValue, context: &mut Context<'_> ) -> JsResult<()>

Abstract operation PrivateFieldAdd ( O, P, value )

Add private field to an object.

More information:

ECMAScript specification

pub(crate) fn private_method_or_accessor_add( &self, name: &PrivateName, method: &PrivateElement, context: &mut Context<'_> ) -> JsResult<()>

Abstract operation PrivateMethodOrAccessorAdd ( O, method )

Add private method or accessor to an object.

More information:

ECMAScript specification

pub(crate) fn private_get( &self, name: &PrivateName, context: &mut Context<'_> ) -> JsResult<JsValue>

Abstract operation PrivateGet ( O, P )

Get the value of a private element.

More information:

ECMAScript specification

pub(crate) fn private_set( &self, name: &PrivateName, value: JsValue, context: &mut Context<'_> ) -> JsResult<()>

Abstract operation PrivateSet ( O, P, value )

Set the value of a private element.

More information:

ECMAScript specification

pub(crate) fn define_field( &self, field_record: &ClassFieldDefinition, context: &mut Context<'_> ) -> JsResult<()>

Abstract operation DefineField ( receiver, fieldRecord )

Define a field on an object.

More information:

ECMAScript specification

pub(crate) fn initialize_instance_elements( &self, constructor: &Self, context: &mut Context<'_> ) -> JsResult<()>

Abstract operation InitializeInstanceElements ( O, constructor )

Add private methods and fields from a class constructor to an object.

More information:

ECMAScript specification

pub(crate) fn invoke<K>( &self, key: K, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>where K: Into<PropertyKey>,

Abstract operation Invoke ( V, P [ , argumentsList ] )

Calls a method property of an ECMAScript object.

Equivalent to the JsValue::invoke method, but specialized for objects.

More information:

ECMAScript reference

impl JsObject

pub(crate) fn call_internal( &self, this: &JsValue, args: &[JsValue], context: &mut Context<'_> ) -> JsResult<JsValue>

pub(crate) fn construct_internal( &self, args: &[JsValue], this_target: &JsValue, context: &mut Context<'_> ) -> JsResult<Self>