Struct boa_engine::object::Object

pub struct Object {
    pub data: ObjectData,
    properties: PropertyMap,
    prototype: JsPrototype,
    extensible: bool,
    private_elements: Vec<(PrivateName, PrivateElement)>,
}

The internal representation of a JavaScript object.

Fields
data: ObjectData

The type of the object.

properties: PropertyMap

The collection of properties contained in the object

prototype: JsPrototype

Instance prototype __proto__.

extensible: bool

Whether it can have new properties added to it.

private_elements: Vec<(PrivateName, PrivateElement)>

The [[PrivateElements]] internal slot.

Implementations

impl Object

pub const fn kind(&self) -> &ObjectKind

Returns the kind of the object.

pub const fn is_async_from_sync_iterator(&self) -> bool

Checks if it’s an AsyncFromSyncIterator object.

pub const fn as_async_from_sync_iterator(
&self
) -> Option<&AsyncFromSyncIterator>

Returns a reference to the AsyncFromSyncIterator data on the object.

pub const fn is_async_generator(&self) -> bool

Checks if it’s an AsyncGenerator object.

pub const fn as_async_generator(&self) -> Option<&AsyncGenerator>

Returns a reference to the async generator data on the object.

pub fn as_async_generator_mut(&mut self) -> Option<&mut AsyncGenerator>

Returns a mutable reference to the async generator data on the object.

pub const fn is_array(&self) -> bool

Checks if the object is a Array object.

pub(crate) const fn has_viewed_array_buffer(&self) -> bool

pub const fn is_data_view(&self) -> bool

Checks if the object is a DataView object.

pub const fn is_array_buffer(&self) -> bool

Checks if the object is a ArrayBuffer object.

pub const fn as_array_buffer(&self) -> Option<&ArrayBuffer>

Gets the array buffer data if the object is a ArrayBuffer.

pub fn as_array_buffer_mut(&mut self) -> Option<&mut ArrayBuffer>

Gets the mutable array buffer data if the object is a ArrayBuffer.

pub const fn is_array_iterator(&self) -> bool

Checks if the object is a ArrayIterator object.

pub const fn as_array_iterator(&self) -> Option<&ArrayIterator>

Gets the array-iterator data if the object is a ArrayIterator.

pub fn as_array_iterator_mut(&mut self) -> Option<&mut ArrayIterator>

Gets the mutable array-iterator data if the object is a ArrayIterator.

pub fn as_string_iterator_mut(&mut self) -> Option<&mut StringIterator>

Gets the mutable string-iterator data if the object is a StringIterator.

pub fn as_regexp_string_iterator_mut(
&mut self
) -> Option<&mut RegExpStringIterator>

Gets the mutable regexp-string-iterator data if the object is a RegExpStringIterator.

pub const fn as_for_in_iterator(&self) -> Option<&ForInIterator>

Gets the for-in-iterator data if the object is a ForInIterator.

pub fn as_for_in_iterator_mut(&mut self) -> Option<&mut ForInIterator>

Gets the mutable for-in-iterator data if the object is a ForInIterator.

pub const fn is_map(&self) -> bool

Checks if the object is a Map object.

pub const fn as_map(&self) -> Option<&OrderedMap<JsValue>>

Gets the map data if the object is a Map.

pub fn as_map_mut(&mut self) -> Option<&mut OrderedMap<JsValue>>

Gets the mutable map data if the object is a Map.

pub const fn is_map_iterator(&self) -> bool

Checks if the object is a MapIterator object.

pub const fn as_map_iterator_ref(&self) -> Option<&MapIterator>

Gets the map iterator data if the object is a MapIterator.

pub fn as_map_iterator_mut(&mut self) -> Option<&mut MapIterator>

Gets the mutable map iterator data if the object is a MapIterator.

pub const fn is_set(&self) -> bool

Checks if the object is a Set object.

pub const fn as_set(&self) -> Option<&OrderedSet<JsValue>>

Gets the set data if the object is a Set.

pub fn as_set_mut(&mut self) -> Option<&mut OrderedSet<JsValue>>

Gets the mutable set data if the object is a Set.

pub const fn is_set_iterator(&self) -> bool

Checks if the object is a SetIterator object.

pub fn as_set_iterator_mut(&mut self) -> Option<&mut SetIterator>

Gets the mutable set iterator data if the object is a SetIterator.

pub const fn is_string(&self) -> bool

Checks if the object is a String object.

pub fn as_string(&self) -> Option<JsString>

Gets the string data if the object is a String.

pub const fn is_function(&self) -> bool

Checks if the object is a Function object.

pub const fn as_function(&self) -> Option<&Function>

Gets the function data if the object is a Function.

pub fn as_function_mut(&mut self) -> Option<&mut Function>

Gets the mutable function data if the object is a Function.

pub const fn as_bound_function(&self) -> Option<&BoundFunction>

Gets the bound function data if the object is a BoundFunction.

pub const fn is_generator(&self) -> bool

Checks if the object is a Generator object.

pub const fn as_generator(&self) -> Option<&Generator>

Gets the generator data if the object is a Generator.

pub fn as_generator_mut(&mut self) -> Option<&mut Generator>

Gets the mutable generator data if the object is a Generator.

pub const fn is_symbol(&self) -> bool

Checks if the object is a Symbol object.

pub fn as_symbol(&self) -> Option<JsSymbol>

Gets the error data if the object is a Symbol.

pub const fn is_error(&self) -> bool

Checks if the object is a Error object.

pub const fn as_error(&self) -> Option<ErrorKind>

Gets the error data if the object is a Error.

pub const fn is_boolean(&self) -> bool

Checks if the object is a Boolean object.

pub const fn as_boolean(&self) -> Option<bool>

Gets the boolean data if the object is a Boolean.

pub const fn is_number(&self) -> bool

Checks if the object is a Number object.

pub const fn as_number(&self) -> Option<f64>

Gets the number data if the object is a Number.

pub const fn is_bigint(&self) -> bool

Checks if the object is a BigInt object.

pub const fn as_bigint(&self) -> Option<&JsBigInt>

Gets the bigint data if the object is a BigInt.

pub const fn is_date(&self) -> bool

Checks if the object is a Date object.

pub const fn as_date(&self) -> Option<&Date>

Gets the date data if the object is a Date.

pub fn as_date_mut(&mut self) -> Option<&mut Date>

Gets the mutable date data if the object is a Date.

pub const fn is_regexp(&self) -> bool

Checks if it a RegExp object.

pub const fn as_regexp(&self) -> Option<&RegExp>

Gets the regexp data if the object is a regexp.

pub const fn is_typed_array(&self) -> bool

Checks if it a TypedArray object.

pub const fn as_data_view(&self) -> Option<&DataView>

Gets the data view data if the object is a DataView.

pub fn as_data_view_mut(&mut self) -> Option<&mut DataView>

Gets the mutable data view data if the object is a DataView.

pub const fn is_arguments(&self) -> bool

Checks if it is an Arguments object.

pub const fn as_mapped_arguments(&self) -> Option<&ParameterMap>

Gets the mapped arguments data if this is a mapped arguments object.

pub fn as_mapped_arguments_mut(&mut self) -> Option<&mut ParameterMap>

Gets the mutable mapped arguments data if this is a mapped arguments object.

pub const fn as_typed_array(&self) -> Option<&IntegerIndexed>

Gets the typed array data (integer indexed object) if this is a typed array.

pub fn as_typed_array_mut(&mut self) -> Option<&mut IntegerIndexed>

Gets the typed array data (integer indexed object) if this is a typed array.

pub const fn is_ordinary(&self) -> bool

Checks if it an ordinary object.

pub const fn is_proxy(&self) -> bool

Checks if it’s an proxy object.

pub const fn as_proxy(&self) -> Option<&Proxy>

Gets the proxy data if the object is a Proxy.

pub fn as_proxy_mut(&mut self) -> Option<&mut Proxy>

Gets the mutable proxy data if the object is a Proxy.

pub const fn as_weak_map(&self) -> Option<&WeakMap<GcRefCell<Object>, JsValue>>

Gets the weak map data if the object is a WeakMap.

pub fn as_weak_map_mut(
&mut self
) -> Option<&mut WeakMap<GcRefCell<Object>, JsValue>>

Gets the mutable weak map data if the object is a WeakMap.

pub const fn as_weak_set(&self) -> Option<&WeakMap<GcRefCell<Object>, ()>>

Gets the weak set data if the object is a WeakSet.

pub fn as_weak_set_mut(&mut self) -> Option<&mut WeakMap<GcRefCell<Object>, ()>>

Gets the mutable weak set data if the object is a WeakSet.

pub const fn prototype(&self) -> &JsPrototype

Gets the prototype instance of this object.

pub fn set_prototype<O: Into<JsPrototype>>(&mut self, prototype: O) -> bool

Sets the prototype instance of the object.

More information

pub const fn is_native_object(&self) -> bool

Returns true if it holds an Rust type that implements NativeObject.

pub fn as_native_object(&self) -> Option<&dyn NativeObject>

Gets the native object data if the object is a NativeObject.

pub const fn is_promise(&self) -> bool

Checks if it is a Promise object.

pub const fn as_promise(&self) -> Option<&Promise>

Gets the promise data if the object is a Promise.

pub fn as_promise_mut(&mut self) -> Option<&mut Promise>

Gets the mutable promise data if the object is a Promise.

pub const fn as_weak_ref(&self) -> Option<&WeakGc<GcRefCell<Self>>>

Gets the WeakRef data if the object is a WeakRef.

pub const fn as_collator(&self) -> Option<&Collator>

Gets the Collator data if the object is a Collator.

pub fn as_collator_mut(&mut self) -> Option<&mut Collator>

Gets a mutable reference to the Collator data if the object is a Collator.

pub const fn is_locale(&self) -> bool

Checks if it is a Locale object.

pub const fn as_locale(&self) -> Option<&Locale>

Gets the Locale data if the object is a Locale.

pub const fn as_list_format(&self) -> Option<&ListFormat>

Gets the ListFormat data if the object is a ListFormat.

pub fn is<T>(&self) -> boolwhere
T: NativeObject,

Return true if it is a native object and the native type is T.

pub fn downcast_ref<T>(&self) -> Option<&T>where
T: NativeObject,

Downcast a reference to the object, if the object is type native object type T.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T>where
T: NativeObject,

Downcast a mutable reference to the object, if the object is type native object type T.

pub const fn properties(&self) -> &PropertyMap

Returns the properties of the object.

pub(crate) fn properties_mut(&mut self) -> &mut PropertyMap

pub(crate) fn insert<K, P>(
&mut self,
key: K,
property: P
) -> Option<PropertyDescriptor>where
K: Into<PropertyKey>,
P: Into<PropertyDescriptor>,

Inserts a field in the object properties without checking if it’s writable.

If a field was already in the object with the same name, then a Some is returned with that field’s value, otherwise, None is returned.

pub(crate) fn remove(&mut self, key: &PropertyKey) -> Option<PropertyDescriptor>

Helper function for property removal.

pub(crate) fn append_private_element(
&mut self,
name: PrivateName,
element: PrivateElement
)

Append a private element to an object.