Struct boa_engine::object::ObjectData

pub struct ObjectData {
    pub(crate) kind: ObjectKind,
    pub(crate) internal_methods: &'static InternalObjectMethods,
}

Defines the kind of an object and its internal methods

Fields
kind: ObjectKindinternal_methods: &'static InternalObjectMethods
Implementations

impl ObjectData

pub(crate) fn object_prototype() -> Self

Create the immutable %Object.prototype% object data

pub fn async_from_sync_iterator( async_from_sync_iterator: AsyncFromSyncIterator ) -> Self

Create the AsyncFromSyncIterator object data

pub fn async_generator(async_generator: AsyncGenerator) -> Self

Create the AsyncGenerator object data

pub fn async_generator_function(function: OrdinaryFunction) -> Self

Create the AsyncGeneratorFunction object data

pub fn array() -> Self

Create the Array object data and reference its exclusive internal methods

pub fn array_iterator(array_iterator: ArrayIterator) -> Self

Create the ArrayIterator object data

pub fn array_buffer(array_buffer: ArrayBuffer) -> Self

Create the ArrayBuffer object data

pub fn map(map: OrderedMap<JsValue>) -> Self

Create the Map object data

pub fn map_iterator(map_iterator: MapIterator) -> Self

Create the MapIterator object data

pub fn reg_exp(reg_exp: Box<RegExp>) -> Self

Create the RegExp object data

pub fn reg_exp_string_iterator( reg_exp_string_iterator: RegExpStringIterator ) -> Self

Create the RegExpStringIterator object data

pub fn big_int(big_int: JsBigInt) -> Self

Create the BigInt object data

pub fn boolean(boolean: bool) -> Self

Create the Boolean object data

pub fn data_view(data_view: DataView) -> Self

Create the DataView object data

pub fn promise(promise: Promise) -> Self

Create the Promise object data

pub fn for_in_iterator(for_in_iterator: ForInIterator) -> Self

Create the ForInIterator object data

pub fn ordinary_function(function: OrdinaryFunction, constructor: bool) -> Self

Create the ordinary function object data

pub fn native_function( function: NativeFunction, constructor: bool, realm: Realm ) -> Self

Create the native function object data

pub fn bound_function(bound_function: BoundFunction, constructor: bool) -> Self

Create the BoundFunction object data

pub fn generator(generator: Generator) -> Self

Create the Generator object data

pub fn generator_function(function: OrdinaryFunction) -> Self

Create the GeneratorFunction object data

pub fn set(set: OrderedSet) -> Self

Create the Set object data

pub fn set_iterator(set_iterator: SetIterator) -> Self

Create the SetIterator object data

pub fn string(string: JsString) -> Self

Create the String object data and reference its exclusive internal methods

pub fn string_iterator(string_iterator: StringIterator) -> Self

Create the StringIterator object data

pub fn number(number: f64) -> Self

Create the Number object data

pub fn symbol(symbol: JsSymbol) -> Self

Create the Symbol object data

pub(crate) fn error(error: ErrorKind) -> Self

Create the Error object data

pub fn ordinary() -> Self

Create the Ordinary object data

pub fn proxy(proxy: Proxy, call: bool, construct: bool) -> Self

Create the Proxy object data

pub fn date(date: Date) -> Self

Create the Date object data

pub fn arguments(arguments: Arguments) -> Self

Create the Arguments object data

pub fn weak_ref(weak_ref: WeakGc<VTableObject>) -> Self

Creates the WeakRef object data

pub fn weak_map(weak_map: WeakMap<VTableObject, JsValue>) -> Self

Create the WeakMap object data

pub fn weak_set(weak_set: WeakMap<VTableObject, ()>) -> Self

Create the WeakSet object data

pub fn native_object<T: NativeObject>(native_object: T) -> Self

Create the NativeObject object data

pub fn integer_indexed(integer_indexed: IntegerIndexed) -> Self

Creates the IntegerIndexed object data

pub fn module_namespace(namespace: ModuleNamespace) -> Self

Creates the ModuleNamespace object data

pub fn collator(date_time_fmt: Collator) -> Self

Create the Collator object data

pub fn date_time_format(date_time_fmt: Box<DateTimeFormat>) -> Self

Create the DateTimeFormat object data

pub fn list_format(list_format: ListFormat) -> Self

Create the ListFormat object data

pub fn locale(locale: Locale) -> Self

Create the Locale object data

pub fn segmenter(segmenter: Segmenter) -> Self

Create the Segmenter object data

pub fn segments(segments: Segments) -> Self

Create the Segments object data

pub fn segment_iterator(segment_iterator: SegmentIterator) -> Self

Create the SegmentIterator object data

pub fn plural_rules(plural_rules: PluralRules) -> Self

Create the PluralRules object data