Trait boa_engine::object::NativeObject

pub trait NativeObject: Any + Trace {
    fn as_any(&self) -> &dyn Any;
    fn as_mut_any(&mut self) -> &mut dyn Any;
}

This trait allows Rust types to be passed around as objects.

This is automatically implemented when a type implements Any and Trace.

Required Methods

fn as_any(&self) -> &dyn Any

Convert the Rust type which implements NativeObject to a &dyn Any.

fn as_mut_any(&mut self) -> &mut dyn Any

Convert the Rust type which implements NativeObject to a &mut dyn Any.