Module boa_engine::prelude

A convenience module that re-exports the most commonly-used Boa APIs

Re-exports

pub use crate::bigint::JsBigInt;

pub use crate::context::Context;

pub use crate::error::JsError;

pub use crate::error::JsNativeError;

pub use crate::error::JsNativeErrorKind;

pub use crate::module::Module;

pub use crate::native_function::NativeFunction;

pub use crate::object::JsData;

pub use crate::object::JsObject;

pub use crate::object::NativeObject;

pub use crate::script::Script;

pub use crate::string::JsString;

pub use crate::symbol::JsSymbol;

pub use crate::value::JsValue;

Structs

HostDefined

This represents a ECMASCript specification [HostDefined] field.

Source

A source of ECMAScript code.

Traits

Finalize

Substitute for the Drop trait for garbage collected types.

Trace

The Trace trait, which needs to be implemented on garbage-collected objects.

Derive Macros

Finalize

Derive the Finalize trait.

JsData

Derive the JsData trait.

Trace

Derive the Trace trait.