Struct boa_engine::property::Attribute

pub struct Attribute {
    bits: u8,
}

This struct constains the property flags as described in the ECMAScript specification.

It contains the following flags:

[[Writable]] (WRITABLE) - If false, attempts by ECMAScript code to change the property’s [[Value]] attribute using [[Set]] will not succeed.

[[Enumerable]] (ENUMERABLE) - If the property will be enumerated by a for-in enumeration.

[[Configurable]] (CONFIGURABLE) - If false, attempts to delete the property, change the property to be an accessor property, or change its attributes (other than [[Value]], or changing [[Writable]] to false) will fail.

Fields
bits: u8
Implementations

impl Attribute

pub const WRITABLE: Self = Self{bits: 1,}

The Writable attribute decides whether the value associated with the property can be changed or not, from its initial value.

pub const ENUMERABLE: Self = Self{bits: 2,}

If the property can be enumerated by a for-in loop.

pub const CONFIGURABLE: Self = Self{bits: 4,}

If the property descriptor can be changed later.

pub const READONLY: Self = Self{bits: 0,}

The property is not writable.

pub const NON_ENUMERABLE: Self = Self{bits: 0,}

The property can not be enumerated in a for-in loop.

pub const PERMANENT: Self = Self{bits: 0,}

The property descriptor cannot be changed.

pub const fn empty() -> Self

Returns an empty set of flags.

pub const fn all() -> Self

Returns the set containing all flags.

pub const fn bits(&self) -> u8

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u8) -> Option<Self>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u8) -> Self

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u8) -> Self

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: Self) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: Self) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: Self)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: Self)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: Self)

Toggles the specified flags in-place.

pub fn set(&mut self, other: Self, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: Self) -> Self

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: Self) -> Self

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: Self) -> Self

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: Self) -> Self

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> Self

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.

impl Attribute

pub fn clear(&mut self)

Clear all flags.

pub fn set_writable(&mut self, value: bool)

Sets the writable flag.

pub fn writable(self) -> bool

Gets the writable flag.

pub fn set_enumerable(&mut self, value: bool)

Sets the enumerable flag.

pub fn enumerable(self) -> bool

Gets the enumerable flag.

pub fn set_configurable(&mut self, value: bool)

Sets the configurable flag.

pub fn configurable(self) -> bool

Gets the configurable flag.