Struct boa_engine::small_map::SmallMap

pub(crate) struct SmallMap<K, V, const ARRAY_SIZE: usize> {
    inner: Inner<K, V, ARRAY_SIZE>,
}

A map that is initially backed by an inline vec, but changes its backing to a heap map if its number of elements exceeds ARRAY_SIZE.

Fields
inner: Inner<K, V, ARRAY_SIZE>
Implementations

impl<K, V, const ARRAY_SIZE: usize> SmallMap<K, V, ARRAY_SIZE>

pub const fn new() -> Self

Makes a new, empty SmallMap.

pub fn clear(&mut self)

Clears the map, removing all elements.

The current implementation will preserve the heap map allocation if the map has already transitioned to the fallback heap map.

pub fn get<Q>(&self, key: &Q) -> Option<&V>
where K: Borrow<Q> + Ord + Eq, Q: Ord + Eq + ?Sized,

Returns a reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get_key_value<Q>(&self, key: &Q) -> Option<(&K, &V)>
where K: Borrow<Q> + Ord + Eq, Q: Ord + Eq + ?Sized,

Returns the key-value pair corresponding to the supplied key.

The supplied key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn contains_key<Q>(&self, key: &Q) -> bool
where K: Borrow<Q> + Ord + Eq, Q: Ord + Eq + ?Sized,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn get_mut<Q>(&mut self, key: &Q) -> Option<&mut V>
where K: Borrow<Q> + Ord + Eq, Q: Ord + Eq + ?Sized,

Returns a mutable reference to the value corresponding to the key.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn insert(&mut self, key: K, value: V) -> Option<V>
where K: Eq + Ord,

Inserts a key-value pair into the map.

If the map did not have this key present, None is returned.

If the map did have this key present, the value is updated, and the old value is returned. The key is not updated, though; this matters for types that can be == without being identical. See the Insert and complex keys section from the std::collections module documentation for more information.

pub fn remove<Q>(&mut self, key: &Q) -> Option<V>
where K: Borrow<Q> + Ord + Eq, Q: Ord + Eq + ?Sized,

Removes a key from the map, returning the value at the key if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn remove_entry<Q>(&mut self, key: &Q) -> Option<(K, V)>
where K: Borrow<Q> + Ord, Q: Ord + ?Sized,

Removes a key from the map, returning the stored key and value if the key was previously in the map.

The key may be any borrowed form of the map’s key type, but the ordering on the borrowed form must match the ordering on the key type.

pub fn retain<F>(&mut self, f: F)
where K: Ord, F: FnMut(&K, &mut V) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all pairs (k, v) for which f(&k, &mut v) returns false.

pub fn append<const OTHER_SIZE: usize>( &mut self, other: &mut SmallMap<K, V, OTHER_SIZE> )
where K: Ord + Eq,

Moves all elements from other into self, leaving other empty.

If a key from other is already present in self, the respective value from self will be overwritten with the respective value from other.

pub fn entry(&mut self, key: K) -> Entry<'_, K, V, ARRAY_SIZE>
where K: Eq + Ord,

Gets the given key’s corresponding entry in the map for in-place manipulation.

impl<K, V, const ARRAY_SIZE: usize> SmallMap<K, V, ARRAY_SIZE>

pub fn iter(&self) -> Iter<'_, K, V>

Gets an iterator over the entries of the map.

pub fn iter_mut(&mut self) -> IterMut<'_, K, V>

Gets a mutable iterator over the entries of the map.

pub fn len(&self) -> usize

Returns the number of elements in the map.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.