Module boa_engine::string

A UTF-16–encoded, reference counted, immutable string.

This module contains the JsString type, the js_string macro and the utf16 macro.

The js_string macro is used when you need to create a new JsString, and the utf16 macro is used for const conversions of string literals to UTF-16.

Modules

common 🔒

Macros

utf16

Construct a utf-16 array literal from a utf-8 str literal.

Structs

JsString

A UTF-16–encoded, reference counted, immutable string.

RawJsString 🔒

The raw representation of a JsString in the heap.

WideStringDecoderIterator 🔒

Enums

CodePoint

Represents a Unicode codepoint within a JsString, which could be a valid ‘Unicode scalar value’, or an unpaired surrogate.

Constants

DATA_OFFSET 🔒

Traits

ToStringEscaped 🔒

Utility trait that adds a UTF-16 escaped representation to every [u16].

Utf16Trim 🔒

Utility trait that adds trimming functionality to every UTF-16 string.

Functions

alloc_overflow 🔒