Struct boa_engine::string::common::StaticJsStrings

pub(crate) struct StaticJsStrings;

List of commonly used strings in Javascript code.

Any strings defined here are used as a static JsString instead of allocating on the heap.

Implementations

impl StaticJsStrings

const fn find_index(candidate: &[u16]) -> usize

pub(crate) fn get_string(string: &[u16]) -> Option<JsString>

Gets the JsString corresponding to string, or None if the string doesn’t exist inside the static array.

pub(crate) fn get(index: usize) -> Option<&'static [u16]>

Gets the &[u16] slice corresponding to the provided index, or None if the index provided exceeds the size of the static array.

pub(crate) const fn empty_string() -> JsString

Gets the empty string ("") JsString.

pub(crate) const fn symbol_async_iterator() -> JsString

Gets the static JsString for "Symbol.asyncIterator".

pub(crate) const fn symbol_has_instance() -> JsString

Gets the static JsString for "Symbol.hasInstance".

pub(crate) const fn symbol_is_concat_spreadable() -> JsString

Gets the static JsString for "Symbol.isConcatSpreadable".

pub(crate) const fn symbol_iterator() -> JsString

Gets the static JsString for "Symbol.iterator".

pub(crate) const fn symbol_match() -> JsString

Gets the static JsString for "Symbol.match".

pub(crate) const fn symbol_match_all() -> JsString

Gets the static JsString for "Symbol.matchAll".

pub(crate) const fn symbol_replace() -> JsString

Gets the static JsString for "Symbol.replace".

pub(crate) const fn symbol_search() -> JsString

Gets the static JsString for "Symbol.search".

pub(crate) const fn symbol_species() -> JsString

Gets the static JsString for "Symbol.species".

pub(crate) const fn symbol_split() -> JsString

Gets the static JsString for "Symbol.split".

pub(crate) const fn symbol_to_primitive() -> JsString

Gets the static JsString for "Symbol.toPrimitive".

pub(crate) const fn symbol_to_string_tag() -> JsString

Gets the static JsString for "Symbol.toStringTag".

pub(crate) const fn symbol_unscopables() -> JsString

Gets the static JsString for "Symbol.unscopables".