Module boa_engine::syntax::ast

The Javascript Abstract Syntax Tree.

This module contains representations of Parse Nodes as defined by the ECMAScript spec. Some Parse Nodes are not represented by Boa’s AST, because a lot of grammar productions are only used to throw Early Errors, and don’t influence the evaluation of the AST itself.

Boa’s AST is mainly split in three main components: Declarations, Expressions and Statements, with StatementList being the primordial Parse Node that combines all of them to create a proper AST.

Re-exports

pub use self::declaration::Declaration;

pub use self::expression::Expression;

pub use self::keyword::Keyword;

pub use self::statement::Statement;

Modules

declaration

The Declaration Parse Node, as defined by the spec.

expression

The Expression Parse Node, as defined by the spec.

function

Functions and classes nodes, as defined by the spec.

keyword

The Keyword AST node, which represents reserved words of the JavaScript language.

pattern

A pattern binding or assignment node.

position 🔒

property

Property definition related types, used in object literals and class definitions.

punctuator 🔒

The Punctuator enum, which contains all punctuators used in JavaScript.

statement

The Statement Parse Node, as defined by the spec.

statement_list 🔒

Statement list node.

Structs

Position

A position in the JavaScript source code.

Span

A span in the JavaScript source code.

StatementList

List of statements.

Enums

ContainsSymbol 🔒

Represents all the possible symbols searched for by the Contains operation.

Punctuator

All of the punctuators used in JavaScript.

StatementListItem

An item inside a StatementList Parse Node, as defined by the spec.

Functions

block_to_string 🔒

Displays the body of a block or statement list.

join_nodes 🔒

Utility to join multiple Nodes into a single string.