Module boa_engine::syntax::ast

The Javascript Abstract Syntax Tree.

Re-exports

pub use self::constant::Const;

pub use self::keyword::Keyword;

pub use self::node::Node;

pub use self::position::Position;

pub use self::position::Span;

pub use self::punctuator::Punctuator;

Modules

constant

This module implements the Const structure, which represents the primitive values in JavaScript.

keyword

This module implements the Keyword structure, which represents reserved words of the JavaScript language.

node

This module implements the Node structure, which composes the AST.

op

This module implements various structure for logic handling.

position

This module implements the Pos structure, which represents a position in the source code.

punctuator

This module implements the Punctuator, which represents all punctuators used in JavaScript