Module boa_engine::syntax::ast::node

This module implements the Node structure, which composes the AST.

Re-exports

pub use self::array::ArrayDecl;

pub use self::await_expr::AwaitExpr;

pub use self::block::Block;

pub use self::call::Call;

pub use self::conditional::ConditionalOp;

pub use self::conditional::If;

pub use self::declaration::async_generator_decl::AsyncGeneratorDecl;

pub use self::declaration::async_generator_expr::AsyncGeneratorExpr;

pub use self::declaration::generator_decl::GeneratorDecl;

pub use self::declaration::generator_expr::GeneratorExpr;

pub use self::declaration::ArrowFunctionDecl;

pub use self::declaration::AsyncFunctionDecl;

pub use self::declaration::AsyncFunctionExpr;

pub use self::declaration::Declaration;

pub use self::declaration::DeclarationList;

pub use self::declaration::DeclarationPattern;

pub use self::declaration::FunctionDecl;

pub use self::declaration::FunctionExpr;

pub use self::field::GetConstField;

pub use self::field::GetField;

pub use self::identifier::Identifier;

pub use self::iteration::Break;

pub use self::iteration::Continue;

pub use self::iteration::DoWhileLoop;

pub use self::iteration::ForInLoop;

pub use self::iteration::ForLoop;

pub use self::iteration::ForOfLoop;

pub use self::iteration::WhileLoop;

pub use self::new::New;

pub use self::object::Object;

pub use self::operator::Assign;

pub use self::operator::BinOp;

pub use self::operator::UnaryOp;

pub use self::yield::Yield;

pub use self::return_smt::Return;

pub use self::spread::Spread;

pub use self::statement_list::RcStatementList;

pub use self::statement_list::StatementList;

pub use self::switch::Case;

pub use self::switch::Switch;

pub use self::template::TaggedTemplate;

pub use self::template::TemplateLit;

pub use self::throw::Throw;

pub use self::try_node::Catch;

pub use self::try_node::Finally;

pub use self::try_node::Try;

Modules

array

Array declaration node.

await_expr

Await expression node.

block

Block AST node.

call

conditional

Conditional nodes

declaration

Declaration nodes

field

Field nodes

identifier

Local identifier node.

iteration

Iteration nodes

new

object

Object node.

operator

Operator nodes

parameters

return_smt

spread

statement_list

Statement list node.

switch

Switch node.

template

Template literal node.

throw

try_node

yield

Structs

FormalParameter

“Formal parameter” is a fancy way of saying “function parameter”.

FormalParameterList

FormalParameterList is a list of FormalParameters that describes the parameters of a function.

Enums

Node

Functions

join_nodes

Utility to join multiple Nodes into a single string.