Enum boa_engine::syntax::ast::node::Node

pub enum Node {

    ArrayDecl(ArrayDecl),
    ArrowFunctionDecl(ArrowFunctionDecl),
    Assign(Assign),
    AsyncFunctionDecl(AsyncFunctionDecl),
    AsyncFunctionExpr(AsyncFunctionExpr),
    AsyncGeneratorExpr(AsyncGeneratorExpr),
    AsyncGeneratorDecl(AsyncGeneratorDecl),
    AwaitExpr(AwaitExpr),
    BinOp(BinOp),
    Block(Block),
    Break(Break),
    Call(Call),
    ConditionalOp(ConditionalOp),
    Const(Const),
    ConstDeclList(DeclarationList),
    Continue(Continue),
    DoWhileLoop(DoWhileLoop),
    FunctionDecl(FunctionDecl),
    FunctionExpr(FunctionExpr),
    GetConstField(GetConstField),
    GetField(GetField),
    ForLoop(ForLoop),
    ForInLoop(ForInLoop),
    ForOfLoop(ForOfLoop),
    If(If),
    LetDeclList(DeclarationList),
    Identifier(Identifier),
    New(New),
    Object(Object),
    Return(Return),
    Switch(Switch),
    Spread(Spread),
    TaggedTemplate(Box<TaggedTemplate>),
    TemplateLit(TemplateLit),
    Throw(Throw),
    Try(Box<Try>),
    This,
    UnaryOp(UnaryOp),
    VarDeclList(DeclarationList),
    WhileLoop(WhileLoop),
    Empty,
    Yield(Yield),
    GeneratorDecl(GeneratorDecl),
    GeneratorExpr(GeneratorExpr),

}

Variants

ArrayDecl(ArrayDecl)

Tuple Fields

0: ArrayDecl

Array declaration node. More information.

ArrowFunctionDecl(ArrowFunctionDecl)

Tuple Fields

0: ArrowFunctionDecl

An arrow function expression node. More information.

Assign(Assign)

Tuple Fields

0: Assign

An assignment operator node. More information.

AsyncFunctionDecl(AsyncFunctionDecl)

Tuple Fields

0: AsyncFunctionDecl

An async function declaration node. More information.

AsyncFunctionExpr(AsyncFunctionExpr)

Tuple Fields

0: AsyncFunctionExpr

An async function expression node. More information.

AsyncGeneratorExpr(AsyncGeneratorExpr)

Tuple Fields

0: AsyncGeneratorExpr

An async generator expression node.

AsyncGeneratorDecl(AsyncGeneratorDecl)

Tuple Fields

0: AsyncGeneratorDecl

An async generator declaration node.

AwaitExpr(AwaitExpr)

Tuple Fields

0: AwaitExpr

An await expression node. More information.

BinOp(BinOp)

Tuple Fields

0: BinOp

A binary operator node. More information.

Block(Block)

Tuple Fields

0: Block

A Block node. More information.

Break(Break)

Tuple Fields

0: Break

A break node. More information.

Call(Call)

Tuple Fields

0: Call

A function call. More information.

ConditionalOp(ConditionalOp)

Tuple Fields

0: ConditionalOp

A javascript conditional operand ( x ? y : z ). More information.

Const(Const)

Tuple Fields

0: Const

Literals represent values in JavaScript.

These are fixed values not variables that you literally provide in your script.

More information:

ECMAScript reference

MDN documentation

ConstDeclList(DeclarationList)

Tuple Fields

0: DeclarationList

A constant declaration list. More information.

Continue(Continue)

Tuple Fields

0: Continue

A continue statement. More information.

DoWhileLoop(DoWhileLoop)

Tuple Fields

0: DoWhileLoop

A do … while statement. More information.

FunctionDecl(FunctionDecl)

Tuple Fields

0: FunctionDecl

A function declaration node. More information.

FunctionExpr(FunctionExpr)

Tuple Fields

0: FunctionExpr

A function expression node. More information.

GetConstField(GetConstField)

Tuple Fields

0: GetConstField

Provides access to an object types’ constant properties. More information.

GetField(GetField)

Tuple Fields

0: GetField

Provides access to object fields. More information.

ForLoop(ForLoop)

Tuple Fields

0: ForLoop

A for statement. More information.

ForInLoop(ForInLoop)

Tuple Fields

0: ForInLoop

A for...of or for..in statement. More information.

ForOfLoop(ForOfLoop)

Tuple Fields

0: ForOfLoop

A for...of statement. More information.

If(If)

Tuple Fields

0: If

An ‘if’ statement. More information.

LetDeclList(DeclarationList)

Tuple Fields

0: DeclarationList

A let declaration list. More information.

Identifier(Identifier)

Tuple Fields

0: Identifier

A local identifier node. More information.

New(New)

Tuple Fields

0: New

A new expression. More information.

Object(Object)

Tuple Fields

0: Object

An object. More information.

Return(Return)

Tuple Fields

0: Return

A return statement. More information.

Switch(Switch)

Tuple Fields

0: Switch

A switch {case} statement. More information.

Spread(Spread)

Tuple Fields

0: Spread

A spread (…x) statement. More information.

TaggedTemplate(Box<TaggedTemplate>)

Tuple Fields

0: Box<TaggedTemplate>

A tagged template. More information.

TemplateLit(TemplateLit)

Tuple Fields

0: TemplateLit

A template literal. More information.

Throw(Throw)

Tuple Fields

0: Throw

A throw statement. More information.

Try(Box<Try>)

Tuple Fields

0: Box<Try>

A try...catch node. More information.

This

The JavaScript this keyword refers to the object it belongs to.

A property of an execution context (global, function or eval) that, in non–strict mode, is always a reference to an object and in strict mode can be any value.

More information:

ECMAScript reference

MDN documentation

UnaryOp(UnaryOp)

Tuple Fields

0: UnaryOp

Unary operation node. More information

VarDeclList(DeclarationList)

Tuple Fields

0: DeclarationList

Array declaration node. More information.

WhileLoop(WhileLoop)

Tuple Fields

0: WhileLoop

A ‘while {…}’ node. More information.

Empty

A empty node.

Empty statement do nothing, just return undefined.

More information:

ECMAScript reference

MDN documentation

Yield(Yield)

Tuple Fields

0: Yield

A yield node. More information.

GeneratorDecl(GeneratorDecl)

Tuple Fields

0: GeneratorDecl

A generator function declaration node. More information.

GeneratorExpr(GeneratorExpr)

Tuple Fields

0: GeneratorExpr

A generator function expression node. More information.

Implementations

impl Node

pub(crate) fn hoistable_order(a: &Self, b: &Self) -> Ordering

Returns a node ordering based on the hoistability of each node.

pub fn this() -> Self

Creates a This AST node.

fn to_indented_string(&self, interner: &Interner, indentation: usize) -> String

Creates a string of the value of the node with the given indentation. For example, an indent level of 2 would produce this:

        function hello() {
            console.log("hello");
        }
        hello();
        a = 2;

fn to_no_indent_string(&self, interner: &Interner, indentation: usize) -> String

Implements the display formatting with indentation.

This will not prefix the value with any indentation. If you want to prefix this with proper indents, use to_indented_string().