Module boa_engine::syntax::ast::statement

This module implements the Statement node.

Re-exports

pub use self::iteration::Break;

pub use self::iteration::Continue;

pub use self::iteration::DoWhileLoop;

pub use self::iteration::ForInLoop;

pub use self::iteration::ForLoop;

pub use self::iteration::ForOfLoop;

pub use self::iteration::WhileLoop;

pub use self::try::Catch;

pub use self::try::Finally;

pub use self::try::Try;

pub use self::switch::Case;

pub use self::switch::Switch;

Modules

block 🔒

Block AST node.

if 🔒

If statement

iteration

Iteration nodes

return 🔒

switch

Switch node.

throw 🔒

try

Structs

Block

A block statement (or compound statement in other languages) is used to group zero or more statements.

If

The if statement executes a statement if a specified condition is truthy. If the condition is falsy, another statement can be executed.

Return

The return statement ends function execution and specifies a value to be returned to the function caller.

Throw

The throw statement throws a user-defined exception.

Enums

Statement