Enum boa_engine::syntax::lexer::token::TokenKind

pub enum TokenKind {

    BooleanLiteral(bool),
    EOF,
    Identifier(Sym),
    PrivateIdentifier(Sym),
    Keyword((Keyword, bool)),
    NullLiteral,
    NumericLiteral(Numeric),
    Punctuator(Punctuator),
    StringLiteral(Sym),
    TemplateNoSubstitution(TemplateString),
    TemplateMiddle(TemplateString),
    RegularExpressionLiteral(Sym, Sym),
    LineTerminator,
    Comment,

}

Represents the type of Token and the data it has inside.

Variants

BooleanLiteral(bool)

A boolean literal, which is either true or false.

EOF

The end of the file.

Identifier(Sym)

An identifier.

PrivateIdentifier(Sym)

A private identifier.

Keyword((Keyword, bool))

A keyword and a flag if the keyword contains unicode escaped chars.

NullLiteral

A null literal.

NumericLiteral(Numeric)

A numeric literal.

Punctuator(Punctuator)

A piece of punctuation

StringLiteral(Sym)

A string literal.

TemplateNoSubstitution(TemplateString)

A part of a template literal without substitution.

TemplateMiddle(TemplateString)

The part of a template literal between substitutions

RegularExpressionLiteral(Sym, Sym)

A regular expression, consisting of body and flags.

LineTerminator

Indicates the end of a line (\n).

Comment

Indicates a comment, the content isn’t stored.

Implementations

impl TokenKind

pub fn boolean_literal(lit: bool) -> Self

Creates a BooleanLiteral token kind.

pub fn eof() -> Self

Creates an EOF token kind.

pub fn identifier(ident: Sym) -> Self

Creates an Identifier token type.

pub fn numeric_literal<L>(lit: L) -> Selfwhere
L: Into<Numeric>,

Creates a NumericLiteral token kind.

pub fn punctuator(punc: Punctuator) -> Self

Creates a Punctuator token type.

pub fn string_literal(lit: Sym) -> Self

Creates a StringLiteral token type.

pub fn template_middle(template_string: TemplateString) -> Self

pub fn template_no_substitution(template_string: TemplateString) -> Self

pub fn regular_expression_literal(body: Sym, flags: Sym) -> Self

Creates a RegularExpressionLiteral token kind.

pub fn line_terminator() -> Self

Creates a LineTerminator token kind.

pub fn comment() -> Self

Creates a ‘Comment’ token kind.

pub fn to_string(&self, interner: &Interner) -> String

Implements the ToString functionality for the TokenKind.