Module boa_engine::value

Boa’s ECMAScript Value implementation.

Javascript values, utility methods and conversion between Javascript values and Rust values.

Modules

conversions 🔒

display 🔒

equality 🔒

hash 🔒

integer 🔒

operations 🔒

serde_json 🔒

This module implements the conversions from and into serde_json::Value.

type 🔒

Structs

ValueDisplay

This object is used for displaying a Value.

Enums

AbstractRelation

The result of the Abstract Relational Comparison.

IntegerOrInfinity

Represents the result of the ToIntegerOrInfinity operation

JsValue

A Javascript value

Numeric

Numeric value which can be of two types Number, BigInt.

PreferredType

The preferred type to convert an object to a primitive Value.

Type

Possible types of values as defined at https://tc39.es/ecma262/#sec-typeof-operator.

Statics

TWO_E_63 🔒

TWO_E_64 🔒