Module boa_engine::vm

Boa’s ECMAScript Virtual Machine

The Virtual Machine (VM) handles generating instructions, then executing them. This module will provide an instruction set for the AST to use, various traits, plus an interpreter to execute those instructions

Modules

call_frame 🔒

CallFrame

code_block 🔒

CodeBlock

completion_record 🔒

An implementation of a CompletionRecord for Boa’s VM.

flowgraph

This module is responsible for generating the vm instruction flowgraph.

opcode 🔒

Structs

CallFrame

A CallFrame holds the state of a function call.

CodeBlock

The internal representation of a JavaScript function.

Vm

Virtual Machine.

Enums

CompletionType 🔒

Opcode

The opcodes of the vm.