Module boa_engine::vm

Boa’s ECMAScript Virtual Machine

The Virtual Machine (VM) handles generating instructions, then executing them. This module will provide an instruction set for the AST to use, various traits, plus an interpreter to execute those instructions

Modules

call_frame 🔒

CallFrame

code_block 🔒

CodeBlock

completion_record 🔒

An implementation of a CompletionRecord for Boa’s VM.

flowgraph

This module is responsible for generating the vm instruction flowgraph.

opcode 🔒

runtime_limits 🔒

Structs

CallFrame

A CallFrame holds the state of a function call.

CodeBlock

The internal representation of a JavaScript function.

RuntimeLimits

Represents the limits of different runtime operations.

Vm

Virtual Machine.

Enums

ActiveRunnable 🔒

Active runnable in the current vm context.

CompletionType 🔒

GeneratorResumeKind

Indicates how a generator function that has been called/resumed should return.