Module boa_engine::vm

Boa’s ECMAScript Virtual Machine

The Virtual Machine (VM) handles generating instructions, then executing them. This module will provide an instruction set for the AST to use, various traits, plus an interpreter to execute those instructions

Structs

CallFrame

A CallFrame holds the state of a function call.

CodeBlock

The internal representation of a JavaScript function.

RuntimeLimits

Represents the limits of different runtime operations.

Vm

Virtual Machine.

Enums

Opcode

The opcodes of the vm.