Module boa_engine::vm::call_frame

CallFrame

This module will provides everything needed to implement the CallFrame

Modules

abrupt_record 🔒

Implements an AbruptCompletionRecord struct for CallFrame tracking.

env_stack 🔒

Module for implementing a CallFrame’s environment stacks

Structs

CallFrame

A CallFrame holds the state of a function call.

Enums

GeneratorResumeKind 🔒

Indicates how a generator function that has been called/resumed should return.