Module boa_engine::vm::code_block

CodeBlock

This module is for the CodeBlock which implements a function representation in the VM

Structs

CodeBlock

The internal representation of a JavaScript function.

Traits

Readable 🔒

This represents whether a value can be read from CodeBlock code.

Functions

create_function_object 🔒

Creates a new function object.

create_generator_function_object 🔒

Creates a new generator function object.

initialize_instance_elements 🔒

InitializeInstanceElements ( O, constructor )