Struct boa_engine::vm::code_block::CodeBlock

pub struct CodeBlock {

    pub(crate) name: Sym,
    pub(crate) flags: Cell<CodeBlockFlags>,
    pub(crate) length: u32,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) bytecode: Box<[u8]>,
    pub(crate) literals: Box<[JsValue]>,
    pub(crate) names: Box<[JsString]>,
    pub(crate) private_names: Box<[PrivateName]>,
    pub(crate) bindings: Box<[BindingLocator]>,
    pub(crate) functions: Box<[Gc<Self>]>,
    pub(crate) compile_environments: Box<[Gc<GcRefCell<CompileTimeEnvironment>>]>,
    pub(crate) class_field_initializer_name: Option<Sym>,

}

The internal representation of a JavaScript function.

A CodeBlock is generated for each function compiled by the ByteCompiler. It stores the bytecode and the other attributes of the function.

Fields
name: Sym

Name of this function

flags: Cell<CodeBlockFlags>length: u32

The number of arguments expected.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

bytecode: Box<[u8]>

Bytecode

literals: Box<[JsValue]>

Literals

names: Box<[JsString]>

Property field names.

private_names: Box<[PrivateName]>

Private names.

bindings: Box<[BindingLocator]>

Locators for all bindings in the codeblock.

functions: Box<[Gc<Self>]>

Functions inside this function

compile_environments: Box<[Gc<GcRefCell<CompileTimeEnvironment>>]>

Compile time environments in this function.

class_field_initializer_name: Option<Sym>

The [[ClassFieldInitializerName]] internal slot.

Implementations

impl CodeBlock

–– CodeBlock public API ––

pub fn new(name: Sym, length: u32, strict: bool) -> Self

Creates a new CodeBlock.

pub const fn name(&self) -> Sym

Retrieves the name associated with this code block.

pub(crate) fn traceable(&self) -> bool

Check if the function is traced.

pub fn set_traceable(&self, value: bool)

Enable or disable instruction tracing to stdout.

pub(crate) fn is_class_constructor(&self) -> bool

Check if the function is a class constructor.

pub(crate) fn strict(&self) -> bool

Check if the function is in strict mode.

pub(crate) fn has_binding_identifier(&self) -> bool

Indicates if the function is an expression and has a binding identifier.

pub(crate) fn has_parameters_env_bindings(&self) -> bool

Does this function have a parameters environment.

pub(crate) fn needs_arguments_object(&self) -> bool

Does this function need a "arguments" object.

impl CodeBlock

–– CodeBlock private API ––

pub(crate) unsafe fn read_unchecked<T>(&self, offset: usize) -> Twhere T: Readable,

Read type T from code.

Safety

Does not check if read happens out-of-bounds.

pub(crate) fn read<T>(&self, offset: usize) -> Twhere T: Readable,

Read type T from code.

pub(crate) fn instruction_operands( &self, pc: &mut usize, interner: &Interner ) -> String

Get the operands after the Opcode pointed to by pc as a String. Modifies the pc to point to the next instruction.

Returns an empty String if no operands are present.

impl CodeBlock

pub fn to_graph(&self, interner: &Interner, graph: &mut SubGraph)

Output the CodeBlock VM instructions into a Graph.