Struct boa_engine::vm::code_block::CodeBlock

pub struct CodeBlock {

    pub(crate) name: Sym,
    pub(crate) has_binding_identifier: bool,
    pub(crate) length: u32,
    pub(crate) strict: bool,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) bytecode: Box<[u8]>,
    pub(crate) literals: Box<[JsValue]>,
    pub(crate) names: Box<[Identifier]>,
    pub(crate) private_names: Box<[PrivateName]>,
    pub(crate) bindings: Box<[BindingLocator]>,
    pub(crate) num_bindings: usize,
    pub(crate) functions: Box<[Gc<Self>]>,
    pub(crate) arguments_binding: Option<BindingLocator>,
    pub(crate) compile_environments: Box<[Gc<GcRefCell<CompileTimeEnvironment>>]>,
    pub(crate) is_class_constructor: bool,
    pub(crate) class_field_initializer_name: Option<Sym>,
    pub(crate) function_environment_push_location: u32,
    pub(crate) parameters_env_bindings: Option<usize>,
    pub(crate) trace: Cell<bool>,

}

The internal representation of a JavaScript function.

A CodeBlock is generated for each function compiled by the ByteCompiler. It stores the bytecode and the other attributes of the function.

Fields
name: Sym

Name of this function

has_binding_identifier: bool

Indicates if the function is an expression and has a binding identifier.

length: u32

The number of arguments expected.

strict: bool

Is this function in strict mode.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

bytecode: Box<[u8]>

Bytecode

literals: Box<[JsValue]>

Literals

names: Box<[Identifier]>

Property field names.

private_names: Box<[PrivateName]>

Private names.

bindings: Box<[BindingLocator]>

Locators for all bindings in the codeblock.

num_bindings: usize

Number of binding for the function environment.

functions: Box<[Gc<Self>]>

Functions inside this function

arguments_binding: Option<BindingLocator>

The arguments binding location of the function, if set.

compile_environments: Box<[Gc<GcRefCell<CompileTimeEnvironment>>]>

Compile time environments in this function.

is_class_constructor: bool

The [[IsClassConstructor]] internal slot.

class_field_initializer_name: Option<Sym>

The [[ClassFieldInitializerName]] internal slot.

function_environment_push_location: u32

Marks the location in the code where the function environment in pushed. This is only relevant for functions with expressions in the parameters. We execute the parameter expressions in the function code and push the function environment afterward. When the execution of the parameter expressions throws an error, we do not need to pop the function environment.

parameters_env_bindings: Option<usize>

The number of bindings in the parameters environment.

trace: Cell<bool>

Trace instruction execution to stdout.

Implementations

impl CodeBlock

–– CodeBlock public API ––

pub fn new(name: Sym, length: u32, strict: bool) -> Self

Creates a new CodeBlock.

pub const fn name(&self) -> Sym

Retrieves the name associated with this code block.

pub fn set_trace(&self, value: bool)

Enable or disable instruction tracing to stdout.

impl CodeBlock

–– CodeBlock private API ––

pub(crate) unsafe fn read_unchecked<T>(&self, offset: usize) -> Twhere T: Readable,

Read type T from code.

Safety

Does not check if read happens out-of-bounds.

pub(crate) fn read<T>(&self, offset: usize) -> Twhere T: Readable,

Read type T from code.

pub(crate) fn instruction_operands( &self, pc: &mut usize, interner: &Interner ) -> String

Get the operands after the Opcode pointed to by pc as a String. Modifies the pc to point to the next instruction.

Returns an empty String if no operands are present.

impl CodeBlock

pub fn to_graph(&self, interner: &Interner, graph: &mut SubGraph)

Output the CodeBlock VM instructions into a Graph.