Struct boa_engine::vm::code_block::CodeBlock

pub struct CodeBlock {

    pub(crate) name: Sym,
    pub(crate) has_binding_identifier: bool,
    pub(crate) length: u32,
    pub(crate) strict: bool,
    pub(crate) this_mode: ThisMode,
    pub(crate) params: FormalParameterList,
    pub(crate) bytecode: Vec<u8>,
    pub(crate) literals: Vec<JsValue>,
    pub(crate) names: Vec<Identifier>,
    pub(crate) private_names: Vec<PrivateName>,
    pub(crate) bindings: Vec<BindingLocator>,
    pub(crate) num_bindings: usize,
    pub(crate) functions: Vec<Gc<Self>>,
    pub(crate) arguments_binding: Option<BindingLocator>,
    pub(crate) compile_environments: Vec<Gc<GcCell<CompileTimeEnvironment>>>,
    pub(crate) is_class_constructor: bool,
    pub(crate) class_field_initializer_name: Option<Sym>,
    pub(crate) function_environment_push_location: u32,

}

The internal representation of a JavaScript function.

A CodeBlock is generated for each function compiled by the ByteCompiler. It stores the bytecode and the other attributes of the function.

Fields
name: Sym

Name of this function

has_binding_identifier: bool

Indicates if the function is an expression and has a binding identifier.

length: u32

The number of arguments expected.

strict: bool

Is this function in strict mode.

this_mode: ThisMode

[[ThisMode]]

params: FormalParameterList

Parameters passed to this function.

bytecode: Vec<u8>

Bytecode

literals: Vec<JsValue>

Literals

names: Vec<Identifier>

Property field names.

private_names: Vec<PrivateName>

Private names.

bindings: Vec<BindingLocator>

Locators for all bindings in the codeblock.

num_bindings: usize

Number of binding for the function environment.

functions: Vec<Gc<Self>>

Functions inside this function

arguments_binding: Option<BindingLocator>

The arguments binding location of the function, if set.

compile_environments: Vec<Gc<GcCell<CompileTimeEnvironment>>>

Compile time environments in this function.

is_class_constructor: bool

The [[IsClassConstructor]] internal slot.

class_field_initializer_name: Option<Sym>

The [[ClassFieldInitializerName]] internal slot.

function_environment_push_location: u32

Marks the location in the code where the function environment in pushed. This is only relevant for functions with expressions in the parameters. We execute the parameter expressions in the function code and push the function environment afterward. When the execution of the parameter expressions throws an error, we do not need to pop the function environment.

Implementations

impl CodeBlock

pub fn new(name: Sym, length: u32, strict: bool) -> Self

Constructs a new CodeBlock.

pub(crate) unsafe fn read_unchecked<T>(&self, offset: usize) -> Twhere
T: Readable,

Read type T from code.

Safety

Does not check if read happens out-of-bounds.

pub(crate) fn read<T>(&self, offset: usize) -> Twhere
T: Readable,

Read type T from code.

pub(crate) fn instruction_operands(
&self,
pc: &mut usize,
interner: &Interner
) -> String

Get the operands after the Opcode pointed to by pc as a String. Modifies the pc to point to the next instruction.

Returns an empty String if no operands are present.

impl CodeBlock

pub fn to_graph(&self, interner: &Interner, graph: &mut SubGraph)

Output the CodeBlock VM instructions into a Graph.