Module boa_engine::vm::opcode

Modules

await_stm ğŸ”’

binary_ops ğŸ”’

call ğŸ”’

concat ğŸ”’

control_flow ğŸ”’

copy ğŸ”’

define ğŸ”’

delete ğŸ”’

dup ğŸ”’

environment ğŸ”’

generator ğŸ”’

get ğŸ”’

iteration ğŸ”’

jump ğŸ”’

new ğŸ”’

nop ğŸ”’

pop ğŸ”’

push ğŸ”’

require ğŸ”’

rest_parameter ğŸ”’

set ğŸ”’

swap ğŸ”’

switch ğŸ”’

to ğŸ”’

unary_ops ğŸ”’

value ğŸ”’

Macros

generate_impl ğŸ”’

Enums

BindingOpcode ğŸ”’

Specific opcodes for bindings.

Opcode

The opcodes of the vm.

Traits

Operation ğŸ”’

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.