Module boa_engine::vm::opcode

Modules

await_stm ğŸ”’

binary_ops ğŸ”’

call ğŸ”’

concat ğŸ”’

control_flow ğŸ”’

copy ğŸ”’

define ğŸ”’

delete ğŸ”’

dup ğŸ”’

environment ğŸ”’

generator ğŸ”’

get ğŸ”’

iteration ğŸ”’

jump ğŸ”’

meta ğŸ”’

new ğŸ”’

nop ğŸ”’

pop ğŸ”’

push ğŸ”’

require ğŸ”’

rest_parameter ğŸ”’

set ğŸ”’

swap ğŸ”’

switch ğŸ”’

templates ğŸ”’

to ğŸ”’

unary_ops ğŸ”’

value ğŸ”’

Macros

generate_impl ğŸ”’

Enums

BindingOpcode ğŸ”’

Specific opcodes for bindings.

Opcode

The opcodes of the vm.

Constants

Add ğŸ”’

AsyncGeneratorYield ğŸ”’

Await ğŸ”’

BitAnd ğŸ”’

BitNot ğŸ”’

BitOr ğŸ”’

BitXor ğŸ”’

Break ğŸ”’

Call ğŸ”’

CallEval ğŸ”’

CallEvalSpread ğŸ”’

CallSpread ğŸ”’

Case ğŸ”’

CatchEnd ğŸ”’

CatchEnd2 ğŸ”’

CatchStart ğŸ”’

Coalesce ğŸ”’

ConcatToString ğŸ”’

Continue ğŸ”’

CopyDataProperties ğŸ”’

CreateForInIterator ğŸ”’

CreateIteratorResult ğŸ”’

Dec ğŸ”’

DecPost ğŸ”’

DefInitVar ğŸ”’

DefVar ğŸ”’

Default ğŸ”’

DefineClassGetterByName ğŸ”’

DefineClassGetterByValue ğŸ”’

DefineClassMethodByName ğŸ”’

DefineClassMethodByValue ğŸ”’

DefineClassSetterByName ğŸ”’

DefineClassSetterByValue ğŸ”’

DefineClassStaticGetterByName ğŸ”’

DefineClassStaticGetterByValue ğŸ”’

DefineClassStaticMethodByName ğŸ”’

DefineClassStaticMethodByValue ğŸ”’

DefineClassStaticSetterByName ğŸ”’

DefineClassStaticSetterByValue ğŸ”’

DefineOwnPropertyByName ğŸ”’

DefineOwnPropertyByValue ğŸ”’

DefinePrivateField ğŸ”’

DeleteName ğŸ”’

DeletePropertyByName ğŸ”’

DeletePropertyByValue ğŸ”’

DeleteSuperThrow ğŸ”’

Div ğŸ”’

Dup ğŸ”’

Eq ğŸ”’

FinallyEnd ğŸ”’

FinallyStart ğŸ”’

GeneratorDelegateNext ğŸ”’

GeneratorDelegateResume ğŸ”’

GeneratorJumpOnResumeKind ğŸ”’

GeneratorNext ğŸ”’

GeneratorResumeReturn ğŸ”’

GeneratorSetReturn ğŸ”’

GeneratorYield ğŸ”’

GetArrowFunction ğŸ”’

GetAsyncArrowFunction ğŸ”’

GetAsyncIterator ğŸ”’

GetFunction ğŸ”’

GetFunctionAsync ğŸ”’

GetGenerator ğŸ”’

GetGeneratorAsync ğŸ”’

GetIterator ğŸ”’

GetLocator ğŸ”’

GetMethod ğŸ”’

GetName ğŸ”’

GetNameAndLocator ğŸ”’

GetNameOrUndefined ğŸ”’

GetPrivateField ğŸ”’

GetPropertyByName ğŸ”’

GetPropertyByValue ğŸ”’

GetPropertyByValuePush ğŸ”’

GreaterThan ğŸ”’

GreaterThanOrEq ğŸ”’

ImportCall ğŸ”’

ImportMeta ğŸ”’

In ğŸ”’

InPrivate ğŸ”’

Inc ğŸ”’

IncPost ğŸ”’

InstanceOf ğŸ”’

IsObject ğŸ”’

IteratorDone ğŸ”’

IteratorFinishAsyncNext ğŸ”’

IteratorLoopStart ğŸ”’

IteratorNext ğŸ”’

IteratorPop ğŸ”’

IteratorResult ğŸ”’

IteratorReturn ğŸ”’

IteratorStackEmpty ğŸ”’

IteratorToArray ğŸ”’

IteratorValue ğŸ”’

Jump ğŸ”’

JumpIfFalse ğŸ”’

JumpIfNotUndefined ğŸ”’

JumpIfNullOrUndefined ğŸ”’

JumpIfTrue ğŸ”’

LabelledEnd ğŸ”’

LabelledStart ğŸ”’

LessThan ğŸ”’

LessThanOrEq ğŸ”’

LogicalAnd ğŸ”’

LogicalNot ğŸ”’

LogicalOr ğŸ”’

LoopContinue ğŸ”’

LoopEnd ğŸ”’

LoopStart ğŸ”’

LoopUpdateReturnValue ğŸ”’

Mod ğŸ”’

Mul ğŸ”’

Neg ğŸ”’

New ğŸ”’

NewSpread ğŸ”’

NewTarget ğŸ”’

Nop ğŸ”’

NotEq ğŸ”’

Pop ğŸ”’

PopEnvironment ğŸ”’

PopIfThrown ğŸ”’

PopOnReturnAdd ğŸ”’

PopOnReturnSub ğŸ”’

PopPrivateEnvironment ğŸ”’

Pos ğŸ”’

Pow ğŸ”’

PushClassField ğŸ”’

PushClassFieldPrivate ğŸ”’

PushClassPrivateGetter ğŸ”’

PushClassPrivateMethod ğŸ”’

PushClassPrivateSetter ğŸ”’

PushClassPrototype ğŸ”’

PushDeclarativeEnvironment ğŸ”’

PushElisionToArray ğŸ”’

PushEmptyObject ğŸ”’

PushFalse ğŸ”’

PushFunctionEnvironment ğŸ”’

PushInt8 ğŸ”’

PushInt16 ğŸ”’

PushInt32 ğŸ”’

PushIteratorToArray ğŸ”’

PushLiteral ğŸ”’

PushNaN ğŸ”’

PushNegativeInfinity ğŸ”’

PushNewArray ğŸ”’

PushNull ğŸ”’

PushObjectEnvironment ğŸ”’

PushOne ğŸ”’

PushPositiveInfinity ğŸ”’

PushPrivateEnvironment ğŸ”’

PushRational ğŸ”’

PushTrue ğŸ”’

PushUndefined ğŸ”’

PushValueToArray ğŸ”’

PushZero ğŸ”’

PutLexicalValue ğŸ”’

RequireObjectCoercible ğŸ”’

Reserved1 ğŸ”’

Reserved2 ğŸ”’

Reserved3 ğŸ”’

Reserved4 ğŸ”’

Reserved5 ğŸ”’

Reserved6 ğŸ”’

Reserved7 ğŸ”’

Reserved8 ğŸ”’

Reserved9 ğŸ”’

Reserved10 ğŸ”’

Reserved11 ğŸ”’

Reserved12 ğŸ”’

Reserved13 ğŸ”’

Reserved14 ğŸ”’

Reserved15 ğŸ”’

Reserved16 ğŸ”’

Reserved17 ğŸ”’

Reserved18 ğŸ”’

Reserved19 ğŸ”’

Reserved20 ğŸ”’

Reserved21 ğŸ”’

Reserved22 ğŸ”’

Reserved23 ğŸ”’

Reserved24 ğŸ”’

Reserved25 ğŸ”’

Reserved26 ğŸ”’

Reserved27 ğŸ”’

Reserved28 ğŸ”’

Reserved29 ğŸ”’

Reserved30 ğŸ”’

Reserved31 ğŸ”’

Reserved32 ğŸ”’

Reserved33 ğŸ”’

Reserved34 ğŸ”’

Reserved35 ğŸ”’

Reserved36 ğŸ”’

Reserved37 ğŸ”’

Reserved38 ğŸ”’

Reserved39 ğŸ”’

Reserved40 ğŸ”’

Reserved41 ğŸ”’

Reserved42 ğŸ”’

Reserved43 ğŸ”’

Reserved44 ğŸ”’

Reserved45 ğŸ”’

Reserved46 ğŸ”’

Reserved47 ğŸ”’

Reserved48 ğŸ”’

Reserved49 ğŸ”’

Reserved50 ğŸ”’

Reserved51 ğŸ”’

Reserved52 ğŸ”’

Reserved53 ğŸ”’

RestParameterInit ğŸ”’

RestParameterPop ğŸ”’

Return ğŸ”’

RotateLeft ğŸ”’

RotateRight ğŸ”’

SetClassPrototype ğŸ”’

SetFunctionName ğŸ”’

SetHomeObject ğŸ”’

SetName ğŸ”’

SetNameByLocator ğŸ”’

SetPrivateField ğŸ”’

SetPrivateGetter ğŸ”’

SetPrivateMethod ğŸ”’

SetPrivateSetter ğŸ”’

SetPropertyByName ğŸ”’

SetPropertyByValue ğŸ”’

SetPropertyGetterByName ğŸ”’

SetPropertyGetterByValue ğŸ”’

SetPropertySetterByName ğŸ”’

SetPropertySetterByValue ğŸ”’

SetPrototype ğŸ”’

ShiftLeft ğŸ”’

ShiftRight ğŸ”’

StrictEq ğŸ”’

StrictNotEq ğŸ”’

Sub ğŸ”’

Super ğŸ”’

SuperCall ğŸ”’

SuperCallDerived ğŸ”’

SuperCallPrepare ğŸ”’

SuperCallSpread ğŸ”’

Swap ğŸ”’

TemplateCreate ğŸ”’

TemplateLookup ğŸ”’

This ğŸ”’

Throw ğŸ”’

ThrowNewTypeError ğŸ”’

ToBoolean ğŸ”’

ToPropertyKey ğŸ”’

TryEnd ğŸ”’

TryStart ğŸ”’

TypeOf ğŸ”’

UnsignedShiftRight ğŸ”’

ValueNotNullOrUndefined ğŸ”’

Void ğŸ”’

Traits

Operation ğŸ”’

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.