Module boa_engine::vm::opcode

Modules

await ğŸ”’

binary_ops ğŸ”’

call ğŸ”’

concat ğŸ”’

control_flow ğŸ”’

copy ğŸ”’

define ğŸ”’

delete ğŸ”’

dup ğŸ”’

environment ğŸ”’

generator ğŸ”’

get ğŸ”’

iteration ğŸ”’

meta ğŸ”’

modifier ğŸ”’

new ğŸ”’

nop ğŸ”’

pop ğŸ”’

push ğŸ”’

require ğŸ”’

rest_parameter ğŸ”’

set ğŸ”’

swap ğŸ”’

switch ğŸ”’

templates ğŸ”’

to ğŸ”’

unary_ops ğŸ”’

value ğŸ”’

Macros

generate_opcodes ğŸ”’

Generate Opcodes and Instructions enums.

Structs

InstructionIterator ğŸ”’

Iterator over the instructions in the compact bytecode.

VaryingOperand ğŸ”’

Enums

BindingOpcode ğŸ”’

Specific opcodes for bindings.

Instruction ğŸ”’

This represents a VM instruction, it contains both opcode and operands.

Opcode ğŸ”’

The opcodes of the vm.

VaryingOperandKind ğŸ”’

Represents a varying operand kind.

Traits

BytecodeConversion ğŸ”’

Operation ğŸ”’

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.

Functions

read ğŸ”’

Read type T from code.

read_unchecked ğŸ”’ âš

Read type T from code.