Enum boa_engine::vm::opcode::Opcode

#[repr(u8)]
pub enum Opcode {

    Pop,
    PopIfThrown,
    Dup,
    Swap,
    PushZero,
    PushOne,
    PushInt8,
    PushInt16,
    PushInt32,
    PushRational,
    PushNaN,
    PushPositiveInfinity,
    PushNegativeInfinity,
    PushNull,
    PushTrue,
    PushFalse,
    PushUndefined,
    PushLiteral,
    PushEmptyObject,
    PushClassPrototype,
    PushNewArray,
    PushValueToArray,
    PushElisionToArray,
    PushIteratorToArray,
    Add,
    Sub,
    Div,
    Mul,
    Mod,
    Pow,
    ShiftRight,
    ShiftLeft,
    UnsignedShiftRight,
    BitOr,
    BitAnd,
    BitXor,
    BitNot,
    In,
    Eq,
    StrictEq,
    NotEq,
    StrictNotEq,
    GreaterThan,
    GreaterThanOrEq,
    LessThan,
    LessThanOrEq,
    InstanceOf,
    LogicalAnd,
    LogicalOr,
    Coalesce,
    TypeOf,
    Void,
    LogicalNot,
    Pos,
    Neg,
    Inc,
    IncPost,
    Dec,
    DecPost,
    DefInitArg,
    DefVar,
    DefInitVar,
    DefLet,
    DefInitLet,
    DefInitConst,
    GetName,
    GetNameOrUndefined,
    SetName,
    GetPropertyByName,
    GetPropertyByValue,
    SetPropertyByName,
    DefineOwnPropertyByName,
    DefineClassMethodByName,
    SetPropertyByValue,
    DefineOwnPropertyByValue,
    DefineClassMethodByValue,
    SetPropertyGetterByName,
    DefineClassGetterByName,
    SetPropertyGetterByValue,
    DefineClassGetterByValue,
    SetPropertySetterByName,
    DefineClassSetterByName,
    SetPropertySetterByValue,
    DefineClassSetterByValue,
    SetPrivateValue,
    SetPrivateSetter,
    SetPrivateGetter,
    GetPrivateField,
    PushClassComputedFieldName,
    DeletePropertyByName,
    DeletePropertyByValue,
    CopyDataProperties,
    ToPropertyKey,
    Jump,
    JumpIfFalse,
    JumpIfNotUndefined,
    Throw,
    TryStart,
    TryEnd,
    CatchStart,
    CatchEnd,
    CatchEnd2,
    FinallyStart,
    FinallyEnd,
    FinallySetJump,
    ToBoolean,
    This,
    Case,
    Default,
    GetFunction,
    GetGenerator,
    CallEval,
    CallEvalWithRest,
    Call,
    CallWithRest,
    New,
    NewWithRest,
    Return,
    PushDeclarativeEnvironment,
    PushFunctionEnvironment,
    PopEnvironment,
    LoopStart,
    LoopContinue,
    LoopEnd,
    ForInLoopInitIterator,
    InitIterator,
    IteratorNext,
    IteratorClose,
    IteratorToArray,
    ForInLoopNext,
    ConcatToString,
    RequireObjectCoercible,
    ValueNotNullOrUndefined,
    RestParameterInit,
    RestParameterPop,
    PopOnReturnAdd,
    PopOnReturnSub,
    Yield,
    GeneratorNext,
    GeneratorNextDelegate,
    Nop,

}

The opcodes of the vm.

Variants

Pop

Pop the top value from the stack.

Operands:

Stack: value =>

PopIfThrown

Pop the top value from the stack if the last try block has thrown a value.

Operands:

Stack: value =>

Dup

Push a copy of the top value on the stack.

Operands:

Stack: value => value, value

Swap

Swap the top two values on the stack.

Operands:

Stack: second, first => first, second

PushZero

Push integer 0 on the stack.

Operands:

Stack: => 0

PushOne

Push integer 1 on the stack.

Operands:

Stack: => 1

PushInt8

Push i8 value on the stack.

Operands: value: i8

Stack: => value

PushInt16

Push i16 value on the stack.

Operands: value: i16

Stack: => value

PushInt32

Push i32 value on the stack.

Operands: value: i32

Stack: => value

PushRational

Push f64 value on the stack.

Operands: value: f64

Stack: => value

PushNaN

Push NaN integer on the stack.

Operands:

Stack: => NaN

PushPositiveInfinity

Push Infinity value on the stack.

Operands:

Stack: => Infinity

PushNegativeInfinity

Push -Infinity value on the stack.

Operands:

Stack: => -Infinity

PushNull

Push null value on the stack.

Operands:

Stack: => null

PushTrue

Push true value on the stack.

Operands:

Stack: => true

PushFalse

Push false value on the stack.

Operands:

Stack: => false

PushUndefined

Push undefined value on the stack.

Operands:

Stack: => undefined

PushLiteral

Push literal value on the stack.

Like strings and bigints. The index operand is used to index into the literals array to get the value.

Operands: index: u32

Stack: => (literals[index])

PushEmptyObject

Push empty object {} value on the stack.

Operands:

Stack: => {}

PushClassPrototype

Get the prototype of a superclass and push it on the stack.

Operands:

Stack: superclass => superclass.prototype

PushNewArray

Push an empty array value on the stack.

Operands:

Stack: => []

PushValueToArray

Push a value to an array.

Operands:

Stack: array, value => array

PushElisionToArray

Push an empty element/hole to an array.

Operands:

Stack: array => array

PushIteratorToArray

Push all iterator values to an array.

Operands:

Stack: array, iterator, next_method, done => array

Add

Binary + operator.

Operands:

Stack: lhs, rhs => (lhs + rhs)

Sub

Binary - operator.

Operands:

Stack: lhs, rhs => (lhs - rhs)

Div

Binary / operator.

Operands:

Stack: lhs, rhs => (lhs / rhs)

Mul

Binary * operator.

Operands:

Stack: lhs, rhs => (lhs * rhs)

Mod

Binary % operator.

Operands:

Stack: lhs, rhs => (lhs % rhs)

Pow

Binary ** operator.

Operands:

Stack: lhs, rhs => (lhs ** rhs)

ShiftRight

Binary >> operator.

Operands:

Stack: lhs, rhs => (lhs >> rhs)

ShiftLeft

Binary << operator.

Operands:

Stack: lhs, rhs => (lhs << rhs)

UnsignedShiftRight

Binary >>> operator.

Operands:

Stack: lhs, rhs => (lhs >>> rhs)

BitOr

Binary bitwise | operator.

Operands:

Stack: lhs, rhs => (lhs | rhs)

BitAnd

Binary bitwise & operator.

Operands:

Stack: lhs, rhs => (lhs & rhs)

BitXor

Binary bitwise ^ operator.

Operands:

Stack: lhs, rhs => (lhs ^ rhs)

BitNot

Unary bitwise ~ operator.

Operands:

Stack: value => ~value

In

Binary in operator.

Operands:

Stack: lhs, rhs => (lhs in rhs)

Eq

Binary == operator.

Operands:

Stack: lhs, rhs => (lhs == rhs)

StrictEq

Binary === operator.

Operands:

Stack: lhs, rhs => (lhs === rhs)

NotEq

Binary != operator.

Operands:

Stack: lhs, rhs => (lhs != rhs)

StrictNotEq

Binary !== operator.

Operands:

Stack: lhs, rhs => (lhs !== rhs)

GreaterThan

Binary > operator.

Operands:

Stack: lhs, rhs => (lhs > rhs)

GreaterThanOrEq

Binary >= operator.

Operands:

Stack: lhs, rhs => (lhs >= rhs)

LessThan

Binary < operator.

Operands:

Stack: lhs, rhs => (lhs < rhs)

LessThanOrEq

Binary <= operator.

Operands:

Stack: lhs, rhs => (lhs <= rhs)

InstanceOf

Binary instanceof operator.

Operands:

Stack: lhs, rhs => (lhs instanceof rhs)

LogicalAnd

Binary logical && operator.

This is a short-circuit operator, if the lhs value is false, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs && rhs)

LogicalOr

Binary logical || operator.

This is a short-circuit operator, if the lhs value is true, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs || rhs)

Coalesce

Binary ?? operator.

This is a short-circuit operator, if the lhs value is not null or undefined, then it jumps to exit address.

Operands: exit: u32

Stack: lhs, rhs => (lhs && rhs)

TypeOf

Unary typeof operator.

Operands:

Stack: value => (typeof value)

Void

Unary void operator.

Operands:

Stack: value => undefined

LogicalNot

Unary logical ! operator.

Operands:

Stack: value => (!value)

Pos

Unary + operator.

Operands:

Stack: value => (+value)

Neg

Unary - operator.

Operands:

Stack: value => (-value)

Inc

Unary ++ operator.

Operands:

Stack: value => (value + 1)

IncPost

Unary postfix ++ operator.

Operands:

Stack: value => (ToNumeric(value)), (value + 1)

Dec

Unary -- operator.

Operands:

Stack: value => (value - 1)

DecPost

Unary postfix -- operator.

Operands:

Stack: value => (ToNumeric(value)), (value - 1)

DefInitArg

Declare and initialize a function argument.

Operands: name_index: u32

Stack: value =>

DefVar

Declare var type variable.

Operands: name_index: u32

Stack: =>

DefInitVar

Declare and initialize var type variable.

Operands: name_index: u32

Stack: value, has_declarative_binding =>

DefLet

Declare let type variable.

Operands: name_index: u32

Stack: =>

DefInitLet

Declare and initialize let type variable.

Operands: name_index: u32

Stack: value =>

DefInitConst

Declare and initialize const type variable.

Operands: name_index: u32

Stack: value =>

GetName

Find a binding on the environment chain and push its value.

Operands: name_index: u32

Stack: => value

GetNameOrUndefined

Find a binding on the environment chain and push its value. If the binding does not exist push undefined.

Operands: name_index: u32

Stack: => value

SetName

Find a binding on the environment chain and assign its value.

Operands: name_index: u32

Stack: value =>

GetPropertyByName

Get a property by name from an object an push it on the stack.

Like object.name

Operands: name_index: u32

Stack: object => value

GetPropertyByValue

Get a property by value from an object an push it on the stack.

Like object[key]

Operands:

Stack: key, object => value

SetPropertyByName

Sets a property by name of an object.

Like object.name = value

Operands: name_index: u32

Stack: value, object =>

DefineOwnPropertyByName

Defines a own property of an object by name.

Operands: name_index: u32

Stack: value, object =>

DefineClassMethodByName

Defines a class method by name.

Operands: name_index: u32

Stack: value, object =>

SetPropertyByValue

Sets a property by value of an object.

Like object[key] = value

Operands:

Stack: value, key, object =>

DefineOwnPropertyByValue

Defines a own property of an object by value.

Operands:

Stack: object, key, value =>

DefineClassMethodByValue

Defines a class method by value.

Operands:

Stack: object, key, value =>

SetPropertyGetterByName

Sets a getter property by name of an object.

Like get name() value

Operands: name_index: u32

Stack: value, object =>

DefineClassGetterByName

Defines a getter class method by name.

Like get name() value

Operands: name_index: u32

Stack: value, object =>

SetPropertyGetterByValue

Sets a getter property by value of an object.

Like get [key]() value

Operands:

Stack: object, key, value =>

DefineClassGetterByValue

Defines a getter class method by value.

Like get [key]() value

Operands:

Stack: object, key, value =>

SetPropertySetterByName

Sets a setter property by name of an object.

Like set name() value

Operands: name_index: u32

Stack: value, object =>

DefineClassSetterByName

Defines a setter class method by name.

Like set name() value

Operands: name_index: u32

Stack: value, object =>

SetPropertySetterByValue

Sets a setter property by value of an object.

Like set [key]() value

Operands:

Stack: object, key, value =>

DefineClassSetterByValue

Defines a setter class method by value.

Like set [key]() value

Operands:

Stack: object, key, value =>

SetPrivateValue

Set a private property by name from an object.

Like #name = value

Operands: name_index: u32

Stack: object, value =>

SetPrivateSetter

Set a private setter property by name from an object.

Like set #name() {}

Operands: name_index: u32

Stack: object, value =>

SetPrivateGetter

Set a private getter property by name from an object.

Like get #name() {}

Operands: name_index: u32

Stack: object, value =>

GetPrivateField

Get a private property by name from an object an push it on the stack.

Like object.#name

Operands: name_index: u32

Stack: object => value

PushClassComputedFieldName

Push a computed class field name to a class constructor object.

Operands:

Stack: value, object =>

DeletePropertyByName

Deletes a property by name of an object.

Like delete object.key.

Operands: name_index: u32

Stack: object =>

DeletePropertyByValue

Deletes a property by value of an object.

Like delete object[key]

Operands:

Stack: key, object =>

CopyDataProperties

Copy all properties of one object to another object.

Operands: excluded_key_count: u32

Stack: source, value, excluded_key_0 … excluded_key_n => value

ToPropertyKey

Call ToPropertyKey on the value on the stack.

Operands:

Stack: value => key

Jump

Unconditional jump to address.

Operands: address: u32

Stack: =>

JumpIfFalse

Conditional jump to address.

If the value popped is falsy then jump to address.

Operands: address: u32

Stack: cond =>

JumpIfNotUndefined

Conditional jump to address.

If the value popped is not undefined jump to address.

Operands: address: u32

Stack: value => value

Throw

Throw exception

Operands:

Stack: value =>

TryStart

Start of a try block.

Operands: next_address: u32, finally_address: u32

Stack: =>

TryEnd

End of a try block.

Operands:

Stack: =>

CatchStart

Start of a catch block.

Operands:

Stack: =>

CatchEnd

End of a catch block.

Operands:

Stack: =>

CatchEnd2

End of a catch block.

Operands:

Stack: =>

FinallyStart

Start of a finally block.

Operands:

Stack: =>

FinallyEnd

End of a finally block.

Operands:

Stack: =>

FinallySetJump

Set the address for a finally jump.

Operands:

Stack: =>

ToBoolean

Pops value converts it to boolean and pushes it back.

Operands:

Stack: value => (ToBoolean(value))

This

Pushes this value

Operands:

Stack: => this

Case

Pop the two values of the stack, strict equal compares the two values, if true jumps to address, otherwise push the second pop’ed value.

Operands: address: u32

Stack: value, cond => cond (if cond !== value).

Default

Pops the top of stack and jump to address.

Operands: address: u32

Stack: value =>

GetFunction

Get function from the pre-compiled inner functions.

Operands: address: u32

Stack: => func

GetGenerator

Get generator function from the pre-compiled inner functions.

Operands: address: u32

Stack: => func

CallEval

Call a function named “eval”.

Operands: argument_count: u32

Stack: func, this, argument_1, … argument_n => result

CallEvalWithRest

Call a function named “eval” where the last argument is a rest parameter.

Operands: argument_count: u32

Stack: func, this, argument_1, … argument_n => result

Call

Call a function.

Operands: argument_count: u32

Stack: func, this, argument_1, … argument_n => result

CallWithRest

Call a function where the last argument is a rest parameter.

Operands: argument_count: u32

Stack: func, this, argument_1, … argument_n => result

New

Call construct on a function.

Operands: argument_count: u32

Stack: func, argument_1, … argument_n => result

NewWithRest

Call construct on a function where the last argument is a rest parameter.

Operands: argument_count: u32

Stack: func, argument_1, … argument_n => result

Return

Return from a function.

Operands:

Stack: =>

PushDeclarativeEnvironment

Push a declarative environment.

Operands: num_bindings: u32, compile_environments_index: u32

Stack: =>

PushFunctionEnvironment

Push a function environment.

Operands: num_bindings: u32, compile_environments_index: u32

Stack: =>

PopEnvironment

Pop the current environment.

Operands:

Stack: =>

LoopStart

Push loop start marker.

Operands:

Stack: =>

LoopContinue

Clean up environments when a loop continues.

Operands:

Stack: =>

LoopEnd

Clean up environments at the end of a loop.

Operands:

Stack: =>

ForInLoopInitIterator

Initialize the iterator for a for..in loop or jump to after the loop if object is null or undefined.

Operands: address: u32

Stack: object => iterator, next_method, done

InitIterator

Initialize an iterator.

Operands:

Stack: object => iterator, next_method, done

IteratorNext

Advance the iterator by one and put the value on the stack.

Operands:

Stack: iterator, next_method, done => iterator, next_method, done, next_value

IteratorClose

Close an iterator.

Operands:

Stack: iterator, next_method, done =>

IteratorToArray

Consume the iterator and construct and array with all the values.

Operands:

Stack: iterator, next_method, done => iterator, next_method, done, array

ForInLoopNext

Move to the next value in a for..in loop or jump to exit of the loop if done.

Note: next_result is only pushed if the iterator is not done.

Operands: address: u32

Stack: iterator, next_method, done => iterator, next_method, done, next_result

ConcatToString

Concat multiple stack objects into a string.

Operands: value_count: u32

Stack: value_1,…value_n => string

RequireObjectCoercible

Call RequireObjectCoercible on the stack value.

Operands:

Stack: value => value

ValueNotNullOrUndefined

Require the stack value to be neither null nor undefined.

Operands:

Stack: value => value

RestParameterInit

Initialize the rest parameter value of a function from the remaining arguments.

Operands:

Stack: argument_1 .. argument_n => array

RestParameterPop

Pop the remaining arguments of a function.

Operands:

Stack: argument_1 .. argument_n =>

PopOnReturnAdd

Add one to the pop on return count.

Operands:

Stack: =>

PopOnReturnSub

Subtract one from the pop on return count.

Operands:

Stack: =>

Yield

Yield from the current execution.

Operands:

Stack: value =>

GeneratorNext

Resumes the current generator function.

Operands:

Stack: received =>

GeneratorNextDelegate

Delegates the current generator function another generator.

Operands: done_address: u32

Stack: iterator, next_method, done, received => iterator, next_method, done

Nop

No-operation instruction, does nothing.

Operands:

Stack: =>

Implementations

impl Opcode

pub unsafe fn from_raw(value: u8) -> Self

Create opcode from u8 byte.

Safety

Does not check if u8 type is a valid Opcode.

pub fn as_str(self) -> &'static str

pub fn as_instruction_str(self) -> &'static str

Name of the profiler event for this opcode