Trait boa_engine::vm::opcode::Operation

pub(crate) trait Operation {
    const NAME: &'static str;
    const INSTRUCTION: &'static str;

    // Required method
    fn execute(context: &mut Context<'_>) -> JsResult<CompletionType>;
}

The Operation trait implements the execution code along with the identifying Name and Instruction value for an Boa Opcode.

This trait should be implemented for a struct that corresponds with any arm of the OpCode enum.

Required Associated Constants

const NAME: &'static str

const INSTRUCTION: &'static str

Required Methods

fn execute(context: &mut Context<'_>) -> JsResult<CompletionType>